Check if a Concept Class Object Belongs to a Vocabulary

Check if a Concept Class Object Belongs to a Vocabulary

check_vocab_id(
  conn,
  conn_fun = "connectAthena()",
  concept_obj,
  vocab_schema = "omop_vocabulary",
  vocabulary_id,
  concept_class_id,
  cache_only = FALSE,
  skip_cache = FALSE,
  override_cache = FALSE,
  render_sql = FALSE,
  verbose = FALSE,
  sleepTime = 1
)

See also

Examples

library(chariot)
library(tidyverse)

# Checking concept id
check_concept_id(concept_id = 2160178)

#> ✓ Concept Id is a valid integer

# Doesn't matter if it is a real concept id or not. It only checks if it is a valid
# integer
check_concept_id(concept_id = 216017800)

#> ✓ Concept Id is a valid integer

# Examples of concept ids that would be invalid
check_concept_id(concept_id = "216017800A")

#> x Concept Id is not a valid integer

# Check that a concept belongs to the expected vocabulary.
# If the Concept Id does not exist in the Concept table, an alert is returned
check_vocab_id(concept_obj = 2160178,
               vocabulary_id = "HemOnc")

#> 

#> ── Is Concept Id Valid Integer? ────────────────────────────────────────────────

#> ✓ Concept Id is a valid integer

#> 

#> ── Lookup Concept Id in Concept Table ──────────────────────────────────────────

#> Connecting using PostgreSQL driver

#> [2021-03-28 18:46:12]	Postgres connection to 'athena' closed

#> x Concept Id `2160178` not found in Concept Table

check_vocab_id(concept_obj = 2160178,
               vocabulary_id = "ATC")

#> 

#> ── Is Concept Id Valid Integer? ────────────────────────────────────────────────

#> ✓ Concept Id is a valid integer

#> 

#> ── Lookup Concept Id in Concept Table ──────────────────────────────────────────

#> Connecting using PostgreSQL driver

#> [2021-03-28 18:46:12]	Postgres connection to 'athena' closed

#> x Concept Id `2160178` not found in Concept Table

# If the Concept Id exists in the Concept table and the vocabulary_id is wrong
check_vocab_id(concept_obj = 1308216,
               vocabulary_id = "ATC")

#> 

#> ── Is Concept Id Valid Integer? ────────────────────────────────────────────────

#> ✓ Concept Id is a valid integer

#> 

#> ── Lookup Concept Id in Concept Table ──────────────────────────────────────────

#> Connecting using PostgreSQL driver

#> [2021-03-28 18:46:12]	Postgres connection to 'athena' closed

#> ✓ Concept Id `1308216` found in Concept Table

#> 

#> ── Lookup Synonyms ─────────────────────────────────────────────────────────────

#> Connecting using PostgreSQL driver

#> [2021-03-28 18:46:12]	Postgres connection to 'athena' closed

#> 

#> ── Lookup 'Maps to' Concept Names ──────────────────────────────────────────────

#> Connecting using PostgreSQL driver

#> [2021-03-28 18:46:13]	Postgres connection to 'athena' closed

#> x Vocabulary Id is not ATC

# If the Concept Id exists in the Concept table and the vocabulary_id is correct
check_vocab_id(concept_obj = 1308216,
               vocabulary_id = "RxNorm")

#> 

#> ── Is Concept Id Valid Integer? ────────────────────────────────────────────────

#> ✓ Concept Id is a valid integer

#> 

#> ── Lookup Concept Id in Concept Table ──────────────────────────────────────────

#> Connecting using PostgreSQL driver

#> [2021-03-28 18:46:13]	Postgres connection to 'athena' closed

#> ✓ Concept Id `1308216` found in Concept Table

#> 

#> ── Lookup Synonyms ─────────────────────────────────────────────────────────────

#> Connecting using PostgreSQL driver

#> [2021-03-28 18:46:13]	Postgres connection to 'athena' closed

#> 

#> ── Lookup 'Maps to' Concept Names ──────────────────────────────────────────────

#> Connecting using PostgreSQL driver

#> [2021-03-28 18:46:13]	Postgres connection to 'athena' closed

#> ✓ Vocabulary Id is RxNorm