Filter Multiple Concept Strip Columns

This function performs the same style of filtering as filter_strip over multiple columns.

filter_at_all_strip(data, strip_cols, ...)

Arguments

data
Dataframe

...
Filter arguments passed to the dplyr filter function using the base Concept Table fields.

merge_cols
Character vector of column names to filter

all
Equivalent to the `all_vars()` variable predicate in the Tidyverse system. If TRUE, all `merge_cols` are filtered for. If FALSE, the equivalent to `any_vars()` is performed. Default: TRUE

Value

A tibble.

See also

Examples

library(tidyverse)
library(chariot)

test_data <-
        queryAthena("SELECT * FROM omop_vocabulary.concept ORDER BY RANDOM() LIMIT 20;",
                    cache_only = TRUE)

#> [2020-12-12 17:25:13]	Connecting using PostgreSQL driver

#> ✓ Connection is open

#> ✓ Connection is JDBC Connection

#> [2020-12-12 17:25:13]	Loading Cache...
#> [2020-12-12 17:25:13]	Cached SQL: SELECT * FROM omop_vocabulary.concept ORDER BY RANDOM() LIMIT 20;
#> [2020-12-12 17:25:13]	Cached resultset found...
#> [2020-12-12 17:25:13]	Postgres connection closed

merge_strip(data = test_data,
            into = "concept")

#> # A tibble: 20 x 2
#>    concept_id concept                                                           
#>         <dbl> <chr>                                                             
#>  1   45905682 [V] [N] 45905682 100 ML albumin human, USP 250 MG/ML Injection [A…
#>  2   44338392 [V] [C] 44338392 LEADER MUCUS DM - dextromethorphan hbr, guaifene…
#>  3   35403442 [V] [N] 35403442 memantine hydrochloride 10 MG Oral Tablet [NDC 6…
#>  4    2068957 [V] [S] 2068957 lamotrigine 100 MG Oral Tablet [LAMEPTIL] [RxNorm…
#>  5   46000567 [V] [N] 46000567 Co-amilozide 5mg/50mg/5ml oral solution (Special…
#>  6   40713545 [V] [S] 40713545 60 ACTUAT fluticasone 0.05 MG/ACTUAT Inhalant Po…
#>  7   44467708 [V] [N] 44467708 modafinil 200mg/1 ORAL TABLET [NDC 001439969] [D…
#>  8   44931227 [D] [N] 44931227 quackgrass pollen extract 40000 UNT/ML Injectabl…
#>  9   43597834 [V] [S] 43597834 1000 ML Alanine 6.95 MG/ML / Ammonium Chloride 2…
#> 10    1020048 [V] [N] 1020048 Right lacrimal gland [LOINC LP30097-7] [Observati…
#> 11   36377442 [V] [C] 36377442 nitrogen 990mL/L RESPIRATORY (INHALATION) GAS [S…
#> 12   41014174 [V] [S] 41014174 ethylene glycol 0.106 MG/MG [RxNorm Extension OM…
#> 13   45663496 [V] [C] 45663496 MEPIVACAINE HYDROCHLORIDE AND LEVONORDEFRIN - me…
#> 14   45303688 [V] [N] 45303688 Acetaminophen, Dextromethorphan Hydrobromide, Ph…
#> 15    1186708 [V] [C] 1186708 actaea spicata, aesculus hippocastanum, arnica mo…
#> 16    3035754 [V] [S] 3035754 Glucagon Ag [Presence] in Tissue by Immune stain …
#> 17   44464406 [D] [N] 44464406 Erythromycin 0.005 MG/MG Ophthalmic Ointment [ND…
#> 18   42005689 [V] [S] 42005689 北10条西20 [OSM 5541128] [Geography] [10th level]
#> 19   36375896 [V] [C] 36375896 didecyldimonium chloride and benzalkonium chlori…
#> 20   36060797 [V] [S] 36060797 febuxostat 80 MG Oral Tablet by Dr Reddy's [RxNo…

merge_strip(data = test_data,
            into = "concept",
            vocabulary_id,
            concept_class_id)

#> # A tibble: 20 x 4
#>    concept_id concept                           vocabulary_id  concept_class_id 
#>         <dbl> <chr>                             <chr>          <chr>            
#>  1   45905682 [V] [N] 45905682 100 ML albumin … NDC            11-digit NDC     
#>  2   44338392 [V] [C] 44338392 LEADER MUCUS DM… SPL            OTC Drug         
#>  3   35403442 [V] [N] 35403442 memantine hydro… NDC            11-digit NDC     
#>  4    2068957 [V] [S] 2068957 lamotrigine 100 … RxNorm Extens… Branded Drug     
#>  5   46000567 [V] [N] 46000567 Co-amilozide 5m… SNOMED         Pharma/Biol Prod…
#>  6   40713545 [V] [S] 40713545 60 ACTUAT fluti… RxNorm Extens… Marketed Product 
#>  7   44467708 [V] [N] 44467708 modafinil 200mg… NDC            9-digit NDC      
#>  8   44931227 [D] [N] 44931227 quackgrass poll… NDC            11-digit NDC     
#>  9   43597834 [V] [S] 43597834 1000 ML Alanine… RxNorm Extens… Quant Branded Dr…
#> 10    1020048 [V] [N] 1020048 Right lacrimal g… LOINC          LOINC System     
#> 11   36377442 [V] [C] 36377442 nitrogen 990mL/… SPL            Prescription Drug
#> 12   41014174 [V] [S] 41014174 ethylene glycol… RxNorm Extens… Clinical Drug Co…
#> 13   45663496 [V] [C] 45663496 MEPIVACAINE HYD… SPL            Prescription Drug
#> 14   45303688 [V] [N] 45303688 Acetaminophen, … NDC            9-digit NDC      
#> 15    1186708 [V] [C] 1186708 actaea spicata, … SPL            OTC Drug         
#> 16    3035754 [V] [S] 3035754 Glucagon Ag [Pre… LOINC          Lab Test         
#> 17   44464406 [D] [N] 44464406 Erythromycin 0.… NDC            11-digit NDC     
#> 18   42005689 [V] [S] 42005689 北10条西20 [OSM 55… OSM            10th level       
#> 19   36375896 [V] [C] 36375896 didecyldimonium… SPL            OTC Drug         
#> 20   36060797 [V] [S] 36060797 febuxostat 80 M… RxNorm Extens… Marketed Product 

test_output <- merge_strip(data = test_data,
                        into = "concept") %>%
                dplyr::select(-concept_id)

# Filter at a Strip Column
filter_strip(test_output,
             strip_col = concept,
             vocabulary_id %in% c("RxNorm"))

#> # A tibble: 2 x 1
#>   concept                                                                       
#>   <chr>                                                                         
#> 1 [V] [S] 41014174 ethylene glycol 0.106 MG/MG [RxNorm Extension OMOP2212136] […
#> 2 [V] [S] 36060797 febuxostat 80 MG Oral Tablet by Dr Reddy's [RxNorm Extension…

# Filter at more than 1 Strip Column
test_data <-
        queryAthena("SELECT * FROM omop_vocabulary.concept ORDER BY RANDOM() LIMIT 10;",
                    cache_only = TRUE) %>%
        merge_strip(into = "concept")

#> [2020-12-12 17:25:14]	Connecting using PostgreSQL driver

#> ✓ Connection is open

#> ✓ Connection is JDBC Connection

#> [2020-12-12 17:25:14]	Loading Cache...
#> [2020-12-12 17:25:14]	Cached SQL: SELECT * FROM omop_vocabulary.concept ORDER BY RANDOM() LIMIT 10;
#> [2020-12-12 17:25:14]	Cached resultset found...
#> [2020-12-12 17:25:14]	Postgres connection closed

test_data <-
        dplyr::bind_cols(test_data %>%
                                 rubix::randomize() %>%
                                 dplyr::select(concept_a = concept),
                         test_data %>%
                                 rubix::randomize() %>%
                                 dplyr::select(concept_b = concept))

filter_at_all_strip(data = test_data,
                strip_cols = c("concept_a",
                               "concept_b"),
                all = TRUE,
                vocabulary_id %in% c("RxNorm"))

#> # A tibble: 1 x 2
#>   concept_a                               concept_b                             
#>   <chr>                                   <chr>                                 
#> 1 [V] [S] 36069843 clindamycin 12 MG/ML … [V] [S] 21154268 rabeprazole 20 MG De…

filter_at_any_strip(data = test_data,
                    strip_cols = c("concept_a",
                                   "concept_b"),
                    all = TRUE,
                    vocabulary_id %in% c("RxNorm"))

#> # A tibble: 3 x 2
#>   concept_a                               concept_b                             
#>   <chr>                                   <chr>                                 
#> 1 [V] [S] 36069843 clindamycin 12 MG/ML … [V] [S] 21154268 rabeprazole 20 MG De…
#> 2 [V] [S] 21154268 rabeprazole 20 MG Del… [V] [C] 36661255 PRKAR1A gene deletio…
#> 3 [V] [S] 41089514 terbinafine 250 MG Or… [V] [S] 36069843 clindamycin 12 MG/ML…