Get a Name Cluster

A `Name Cluster` is a character vector of all the unique labels associated with a concept, including labels derived from `Maps to` relationships from the Concept Relationship table.

get_name_cluster(
  concept_obj,
  vocab_schema = "omop_vocabulary",
  conn,
  conn_fun = "connectAthena()",
  cache_only = FALSE,
  skip_cache = FALSE,
  override_cache = FALSE,
  render_sql = TRUE,
  verbose = TRUE,
  sleepTime = 1
)

Arguments

concept_obj
May be an integer or a `concept` class object.

Examples

library(chariot)
library(tidyverse)

conn <- connectAthena()

#> Connecting using PostgreSQL driver

# Drug Domain
concept_obj <- get_concept(conn = conn,
                           concept_id = 1308216)

#> 

#> ── Is Concept Id Valid Integer? ────────────────────────────────────────────────

#> ✓ Concept Id is a valid integer

#> 

#> ── Lookup Concept Id in Concept Table ──────────────────────────────────────────

#> ✓ Concept Id `1308216` found in Concept Table

#> 

#> ── Lookup Synonyms ─────────────────────────────────────────────────────────────

#> 

#> ── Lookup 'Maps to' Concept Names ──────────────────────────────────────────────

concept_obj

#> An object of class "concept"
#> Slot "concept_id":
#> [1] 1308216
#> 
#> Slot "concept_name":
#> [1] "lisinopril"
#> 
#> Slot "concept_synonym_names":
#> [1] ""
#> 
#> Slot "maps_to_concept_names":
#> [1] "lisinopril"
#> 
#> Slot "domain_id":
#> [1] "Drug"
#> 
#> Slot "vocabulary_id":
#> [1] "RxNorm"
#> 
#> Slot "concept_class_id":
#> [1] "Ingredient"
#> 
#> Slot "standard_concept":
#> [1] "S"
#> 
#> Slot "concept_code":
#> [1] "29046"
#> 
#> Slot "valid_start_date":
#> [1] "1970-01-01"
#> 
#> Slot "valid_end_date":
#> [1] "2099-12-31"
#> 
#> Slot "invalid_reason":
#> [1] NA
#> 

get_name_cluster(conn = conn,
                 concept_obj = concept_obj)

#>         concept_name maps_to_concept_name 
#>         "lisinopril"         "lisinopril" 

# Cancer Condition Domain
concept_obj <- get_concept(conn = conn,
                           concept_id = 4187868)

#> 

#> ── Is Concept Id Valid Integer? ────────────────────────────────────────────────

#> ✓ Concept Id is a valid integer

#> 

#> ── Lookup Concept Id in Concept Table ──────────────────────────────────────────

#> ✓ Concept Id `4187868` found in Concept Table

#> 

#> ── Lookup Synonyms ─────────────────────────────────────────────────────────────

#> 

#> ── Lookup 'Maps to' Concept Names ──────────────────────────────────────────────

concept_obj

#> An object of class "concept"
#> Slot "concept_id":
#> [1] 4187868
#> 
#> Slot "concept_name":
#> [1] "pTis: Ductal carcinoma in situ"
#> 
#> Slot "concept_synonym_names":
#> [1] "Ductal carcinoma in situ (breast)|pTis: Ductal carcinoma in situ (breast)|pTis: Ductal carcinoma in situ (breast) (finding)"
#> 
#> Slot "maps_to_concept_names":
#> [1] "pTis: Ductal carcinoma in situ"
#> 
#> Slot "domain_id":
#> [1] "Condition"
#> 
#> Slot "vocabulary_id":
#> [1] "SNOMED"
#> 
#> Slot "concept_class_id":
#> [1] "Clinical Finding"
#> 
#> Slot "standard_concept":
#> [1] "S"
#> 
#> Slot "concept_code":
#> [1] "373176000"
#> 
#> Slot "valid_start_date":
#> [1] "1970-01-01"
#> 
#> Slot "valid_end_date":
#> [1] "2099-12-31"
#> 
#> Slot "invalid_reason":
#> [1] NA
#> 

get_name_cluster(conn = conn,
                 concept_obj = concept_obj)

#>                                        concept_name 
#>                    "pTis: Ductal carcinoma in situ" 
#>                               concept_synonym_name1 
#>                 "Ductal carcinoma in situ (breast)" 
#>                               concept_synonym_name2 
#>           "pTis: Ductal carcinoma in situ (breast)" 
#>                               concept_synonym_name3 
#> "pTis: Ductal carcinoma in situ (breast) (finding)" 
#>                                maps_to_concept_name 
#>                    "pTis: Ductal carcinoma in situ" 

# Chronic Disease Domain
concept_obj <- get_concept(conn = conn,
                           concept_id = 319835)

#> 

#> ── Is Concept Id Valid Integer? ────────────────────────────────────────────────

#> ✓ Concept Id is a valid integer

#> 

#> ── Lookup Concept Id in Concept Table ──────────────────────────────────────────

#> ✓ Concept Id `319835` found in Concept Table

#> 

#> ── Lookup Synonyms ─────────────────────────────────────────────────────────────

#> 

#> ── Lookup 'Maps to' Concept Names ──────────────────────────────────────────────

concept_obj

#> An object of class "concept"
#> Slot "concept_id":
#> [1] 319835
#> 
#> Slot "concept_name":
#> [1] "Congestive heart failure"
#> 
#> Slot "concept_synonym_names":
#> [1] "CCF - Congestive cardiac failure|CHF - Congestive heart failure|Congestive cardiac failure|Congestive heart disease|Congestive heart failure (disorder)"
#> 
#> Slot "maps_to_concept_names":
#> [1] "Congestive heart failure"
#> 
#> Slot "domain_id":
#> [1] "Condition"
#> 
#> Slot "vocabulary_id":
#> [1] "SNOMED"
#> 
#> Slot "concept_class_id":
#> [1] "Clinical Finding"
#> 
#> Slot "standard_concept":
#> [1] "S"
#> 
#> Slot "concept_code":
#> [1] "42343007"
#> 
#> Slot "valid_start_date":
#> [1] "1970-01-01"
#> 
#> Slot "valid_end_date":
#> [1] "2099-12-31"
#> 
#> Slot "invalid_reason":
#> [1] NA
#> 

get_name_cluster(conn = conn,
                 concept_obj = concept_obj)

#>                          concept_name                 concept_synonym_name1 
#>            "Congestive heart failure"    "CCF - Congestive cardiac failure" 
#>                 concept_synonym_name2                 concept_synonym_name3 
#>      "CHF - Congestive heart failure"          "Congestive cardiac failure" 
#>                 concept_synonym_name4                 concept_synonym_name5 
#>            "Congestive heart disease" "Congestive heart failure (disorder)" 
#>                  maps_to_concept_name 
#>            "Congestive heart failure" 

# Measurement Domain
concept_obj <- get_concept(conn = conn, concept_id = 4298431)

#> 

#> ── Is Concept Id Valid Integer? ────────────────────────────────────────────────

#> ✓ Concept Id is a valid integer

#> 

#> ── Lookup Concept Id in Concept Table ──────────────────────────────────────────

#> ✓ Concept Id `4298431` found in Concept Table

#> 

#> ── Lookup Synonyms ─────────────────────────────────────────────────────────────

#> 

#> ── Lookup 'Maps to' Concept Names ──────────────────────────────────────────────

concept_obj

#> An object of class "concept"
#> Slot "concept_id":
#> [1] 4298431
#> 
#> Slot "concept_name":
#> [1] "White blood cell count"
#> 
#> Slot "concept_synonym_names":
#> [1] "WBC - White blood cell count|WBC count|WCC - White blood cell count|White blood cell count (procedure)|White blood cell count - observation"
#> 
#> Slot "maps_to_concept_names":
#> [1] "White blood cell count"
#> 
#> Slot "domain_id":
#> [1] "Measurement"
#> 
#> Slot "vocabulary_id":
#> [1] "SNOMED"
#> 
#> Slot "concept_class_id":
#> [1] "Procedure"
#> 
#> Slot "standard_concept":
#> [1] "S"
#> 
#> Slot "concept_code":
#> [1] "767002"
#> 
#> Slot "valid_start_date":
#> [1] "1970-01-01"
#> 
#> Slot "valid_end_date":
#> [1] "2099-12-31"
#> 
#> Slot "invalid_reason":
#> [1] NA
#> 

get_name_cluster(conn = conn,
                 concept_obj = concept_obj)

#>                           concept_name                  concept_synonym_name1 
#>               "White blood cell count"         "WBC - White blood cell count" 
#>                  concept_synonym_name2                  concept_synonym_name3 
#>                            "WBC count"         "WCC - White blood cell count" 
#>                  concept_synonym_name4                  concept_synonym_name5 
#>   "White blood cell count (procedure)" "White blood cell count - observation" 
#>                   maps_to_concept_name 
#>               "White blood cell count" 

dcAthena(conn = conn)

#> [2021-01-09 15:08:07]	Postgres connection to 'athena' closed