Get Unfiltered Test Concepts

Get random rows from the Concept table without any additional filters.

get_test_concepts(
  conn,
  conn_fun = "connectAthena()",
  schema = "omop_vocabulary",
  limit = 100,
  verbose = TRUE,
  render_sql = TRUE
)

See also

Examples

library(chariot)
library(tidyverse)

# Get Unfiltered Test Concepts (Invalid Concepts are included)
get_test_concepts()

#> Connecting using PostgreSQL driver
#> [2021-01-10 14:50:12]	Overriding cache... Querying Athena...
#> [2021-01-10 14:50:12] 

#> ✓ Open connection

#> [2021-01-10 14:50:12] 

#> ✓ JDBC connection

#> [2021-01-10 14:50:12]	SQL: SELECT * FROM omop_vocabulary.concept ORDER BY RANDOM() LIMIT 100
#> [2021-01-10 14:50:12]	Querying...
#> [2021-01-10 14:50:20]	Querying...complete
#> [2021-01-10 14:50:20] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 14:50:20]	Caching resultset...
#> [2021-01-10 14:50:20]	Postgres connection to 'athena' closed

#> # A tibble: 100 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id
#>         <dbl> <chr>        <chr>     <chr>         <chr>           
#>  1    4167869 Fluoroscopi… Procedure SNOMED        Procedure       
#>  2   44198939 2.5 ML Sodi… Drug      RxNorm Exten… Quant Clinical …
#>  3   42778157 Therapeutic… Procedure OPS           Procedure       
#>  4   35193928 Thrombin 50… Drug      JMDC          Drug Product    
#>  5   40310544 Urine cultu… Condition SNOMED        Clinical Finding
#>  6   45936994 Heptachlor … Condition CIEL          Diagnosis       
#>  7   43528465 Sibling Men… Observat… PPI           Answer          
#>  8   42490034 Impetigo [a… Condition KCD7          KCD7 code       
#>  9   43315795 Normal Immu… Drug      AMT           Med Product Unit
#> 10    2781801 Change Spli… Procedure ICD10PCS      ICD10PCS        
#> # … with 90 more rows, and 5 more variables: standard_concept <chr>,
#> #   concept_code <chr>, valid_start_date <date>, valid_end_date <date>,
#> #   invalid_reason <chr>

# The size of the resultset can be altered
get_test_concepts(limit = 25)

#> Connecting using PostgreSQL driver
#> [2021-01-10 14:50:20]	Overriding cache... Querying Athena...
#> [2021-01-10 14:50:20] 

#> ✓ Open connection

#> [2021-01-10 14:50:20] 

#> ✓ JDBC connection

#> [2021-01-10 14:50:20]	SQL: SELECT * FROM omop_vocabulary.concept ORDER BY RANDOM() LIMIT 25
#> [2021-01-10 14:50:20]	Querying...
#> [2021-01-10 14:50:27]	Querying...complete
#> [2021-01-10 14:50:27] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 14:50:27]	Caching resultset...
#> [2021-01-10 14:50:27]	Postgres connection to 'athena' closed

#> # A tibble: 25 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id
#>         <dbl> <chr>        <chr>     <chr>         <chr>           
#>  1   45111634 loratadine … Drug      NDC           11-digit NDC    
#>  2   43563914 Ambroxol He… Drug      RxNorm Exten… Brand Name      
#>  3   40235654 Carboxymeth… Drug      RxNorm        Branded Drug    
#>  4   44835006 Hypermobili… Condition ICD9CM        5-dig billing c…
#>  5   41262678 Allium sati… Drug      RxNorm Exten… Marketed Product
#>  6   36229324 Plasma-Lyte… Drug      RxNorm        Branded Dose Gr…
#>  7   45264405 Amoxapine 2… Drug      NDC           11-digit NDC    
#>  8   42171105 Incision An… Procedure EDI           Procedure       
#>  9   44901975 Isosorbide … Drug      NDC           11-digit NDC    
#> 10   35603295 cariprazine… Drug      RxNorm        Branded Drug Co…
#> # … with 15 more rows, and 5 more variables: standard_concept <chr>,
#> #   concept_code <chr>, valid_start_date <date>, valid_end_date <date>,
#> #   invalid_reason <chr>

# Concepts specific to the Drug domain can be sampled
# RxNorm/RxNorm Extension Concepts
get_test_drug_concepts(limit = 10)

#> Connecting using PostgreSQL driver
#> [2021-01-10 14:50:27]	Overriding cache... Querying Athena...
#> [2021-01-10 14:50:27] 

#> ✓ Open connection

#> [2021-01-10 14:50:27] 

#> ✓ JDBC connection

#> [2021-01-10 14:50:27]	SQL: SELECT * FROM omop_vocabulary.concept c WHERE c.invalid_reason IS NULL AND c.vocabulary_id IN ('RxNorm', 'RxNorm Extension') AND c.standard_concept <> 'C' AND c.domain_id = 'Drug' ORDER BY RANDOM() LIMIT 10
#> [2021-01-10 14:50:27]	Querying...
#> [2021-01-10 14:50:30]	Querying...complete
#> [2021-01-10 14:50:30] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 14:50:30]	Caching resultset...
#> [2021-01-10 14:50:30]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id
#>         <dbl> <chr>        <chr>     <chr>         <chr>           
#>  1   40985877 Arnica extr… Drug      RxNorm Exten… Branded Drug Co…
#>  2   35419352 49.9 MG Ter… Drug      RxNorm Exten… Quant Branded D…
#>  3   41500054 50000 MG PR… Drug      RxNorm Exten… Quant Clinical …
#>  4   43762024 Carbamazepi… Drug      RxNorm Exten… Marketed Product
#>  5     783266 racecadotri… Drug      RxNorm Exten… Branded Drug Fo…
#>  6   43213784 3,3-diphosp… Drug      RxNorm Exten… Clinical Drug B…
#>  7   43012808 menthol 0.0… Drug      RxNorm        Clinical Drug C…
#>  8   41305631 Bezafibrate… Drug      RxNorm Exten… Clinical Drug B…
#>  9   44123587 Coal Tar 0.… Drug      RxNorm Exten… Clinical Drug   
#> 10   41172421 Hexetidine … Drug      RxNorm Exten… Branded Drug Co…
#> # … with 5 more variables: standard_concept <chr>, concept_code <chr>,
#> #   valid_start_date <date>, valid_end_date <date>, invalid_reason <chr>

# HemOnc/ATC Classes
get_test_drug_classes(limit = 10)

#> Connecting using PostgreSQL driver
#> [2021-01-10 14:50:30]	Overriding cache... Querying Athena...
#> [2021-01-10 14:50:30] 

#> ✓ Open connection

#> [2021-01-10 14:50:30] 

#> ✓ JDBC connection

#> [2021-01-10 14:50:30]	SQL: SELECT * FROM omop_vocabulary.concept c WHERE c.invalid_reason IS NULL AND c.vocabulary_id IN ('ATC', 'HemOnc') AND c.standard_concept = 'C' AND c.domain_id = 'Drug' ORDER BY RANDOM() LIMIT 10
#> [2021-01-10 14:50:30]	Querying...
#> [2021-01-10 14:50:30]	Querying...complete
#> [2021-01-10 14:50:30] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 14:50:30]	Caching resultset...
#> [2021-01-10 14:50:30]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id
#>         <dbl> <chr>        <chr>     <chr>         <chr>           
#>  1   21603937 oxyphenbuta… Drug      ATC           ATC 5th         
#>  2   43534859 lomefloxaci… Drug      ATC           ATC 5th         
#>  3   21605010 Sympathomim… Drug      ATC           ATC 4th         
#>  4   21605560 iodocholest… Drug      ATC           ATC 5th         
#>  5   21602174 fluocinonid… Drug      ATC           ATC 5th         
#>  6   21601583 ergoloid me… Drug      ATC           ATC 5th         
#>  7   21602447 plastic IUD… Drug      ATC           ATC 5th         
#>  8   21604226 benzocaine   Drug      ATC           ATC 5th         
#>  9   21604011 tenidap      Drug      ATC           ATC 5th         
#> 10   35807304 Taxane       Drug      HemOnc        Component Class 
#> # … with 5 more variables: standard_concept <chr>, concept_code <chr>,
#> #   valid_start_date <date>, valid_end_date <date>, invalid_reason <chr>

# Test data can also be taken from tables other than Concept
get_test_data(table = "concept_relationship")

#> Connecting using PostgreSQL driver
#> [2021-01-10 14:50:31]	Overriding cache... Querying Athena...
#> [2021-01-10 14:50:31] 

#> ✓ Open connection

#> [2021-01-10 14:50:31] 

#> ✓ JDBC connection

#> [2021-01-10 14:50:31]	SQL: SELECT * FROM omop_vocabulary.concept_relationship ORDER BY RANDOM() LIMIT 100
#> [2021-01-10 14:50:31]	Querying...
#> [2021-01-10 14:51:04]	Querying...complete
#> [2021-01-10 14:51:04] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 14:51:04]	Caching resultset...
#> [2021-01-10 14:51:04]	Postgres connection to 'athena' closed

#> # A tibble: 100 x 6
#>    concept_id_1 concept_id_2 relationship_id valid_start_date valid_end_date
#>           <dbl>        <dbl> <chr>           <date>           <date>        
#>  1     44210879     40807985 Has brand name  2017-10-06       2099-12-31    
#>  2     35626973     45967677 VMP prescr sta… 2018-07-31       2099-12-31    
#>  3     42792156     42787476 Subsumes        1970-01-01       2099-12-31    
#>  4     21059350     19074842 Marketed form … 2017-06-24       2099-12-31    
#>  5     44991744     42708492 Maps to         1970-01-01       2099-12-31    
#>  6     43186945     19082104 RxNorm has dos… 2016-09-10       2099-12-31    
#>  7     46144211     35626992 Has disc indic… 2018-07-31       2099-12-31    
#>  8     37209199     46127564 Subsumes        2019-07-31       2099-12-31    
#>  9     46075308     37207390 Has CD category 2019-07-31       2099-12-31    
#> 10     21126862     21154184 Quantified for… 2015-08-17       2099-12-31    
#> # … with 90 more rows, and 1 more variable: invalid_reason <chr>

get_test_data(table = "concept_ancestor", limit = 10)

#> Connecting using PostgreSQL driver
#> [2021-01-10 14:51:04]	Overriding cache... Querying Athena...
#> [2021-01-10 14:51:04] 

#> ✓ Open connection

#> [2021-01-10 14:51:04] 

#> ✓ JDBC connection

#> [2021-01-10 14:51:04]	SQL: SELECT * FROM omop_vocabulary.concept_ancestor ORDER BY RANDOM() LIMIT 10
#> [2021-01-10 14:51:04]	Querying...
#> [2021-01-10 14:51:39]	Querying...complete
#> [2021-01-10 14:51:39] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 14:51:39]	Caching resultset...
#> [2021-01-10 14:51:39]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 4
#>    ancestor_concept_… descendant_concep… min_levels_of_sepa… max_levels_of_sepa…
#>                 <dbl>              <dbl>               <dbl>               <dbl>
#>  1           45621254           35848226                   1                   1
#>  2           21602907            1366402                   1                   1
#>  3           21602985           36895284                   2                   3
#>  4           21601909           43142423                   5                   7
#>  5           45791503           41000162                   1                   1
#>  6           36917303           43292825                   2                   2
#>  7            4274025           36552857                   4                  13
#>  8            1594348           40744737                   2                   3
#>  9           21602627           42967551                   5                   6
#> 10           36686121           36686121                   0                   0