Get Test Drug Classes

Get random records of ATC and HemOnc Drug Classes that are:

Valid,

Class,

Belong to the Drug domain

get_test_drug_classes(
  conn,
  conn_fun = "connectAthena()",
  schema = "omop_vocabulary",
  limit = 100,
  verbose = TRUE,
  render_sql = TRUE
)

See also

Examples

library(chariot)
library(tidyverse)

# Get Unfiltered Test Concepts (Invalid Concepts are included)
get_test_concepts()

#> [2021-05-26 15:05:49]	Connected to 'athena'
#> [2021-05-26 15:05:49]	Overriding cache... Querying Athena...
#> [2021-05-26 15:05:49]	✓ Open connection
#> [2021-05-26 15:05:49]	✓ JDBC connection
#> [2021-05-26 15:05:49]	SQL: SELECT * FROM omop_vocabulary.concept ORDER BY RANDOM() LIMIT 100
#> [2021-05-26 15:05:49]	Querying...
#> [2021-05-26 15:05:57]	Querying...complete
#> [2021-05-26 15:05:57]	✓ Returned data has more than 0 rows
#> [2021-05-26 15:05:57]	Caching resultset...
#> [2021-05-26 15:05:57]	Postgres connection to 'athena' closed

#> # A tibble: 100 x 10
#>    concept_id concept_name              domain_id vocabulary_id concept_class_id
#>         <int> <chr>                     <chr>     <chr>         <chr>           
#>  1   35193673 Theophylline [Unicon]     Drug      JMDC          Drug Product    
#>  2    3397148 Common variable immunode… Condition Nebraska Lex… Clinical Finding
#>  3   41457776 2.02 ML hypromellose 10 … Drug      RxNorm Exten… Marketed Product
#>  4   36764054 SDHD Protein Variant: Su… Measurem… OMOP Extensi… Protein Variant 
#>  5   45844270 nicotine polacrilex 4mg/… Drug      NDC           9-digit NDC     
#>  6    2314052 Pediatric home apnea mon… Procedure CPT4          CPT4            
#>  7   44510558 Unspecified excision of … Procedure OPCS4         Procedure       
#>  8   21365299 Neria subcutaneous drug … Device    dm+d          AMPP            
#>  9   43220166 180 ML Ambroxol 3 MG/ML … Drug      RxNorm Exten… Marketed Product
#> 10    3299256 Supervisor and general f… Observat… Nebraska Lex… Social Context  
#> # … with 90 more rows, and 5 more variables: standard_concept <chr>,
#> #   concept_code <chr>, valid_start_date <date>, valid_end_date <date>,
#> #   invalid_reason <chr>

# The size of the resultset can be altered
get_test_concepts(limit = 25)

#> [2021-05-26 15:05:58]	Connected to 'athena'
#> [2021-05-26 15:05:58]	Overriding cache... Querying Athena...
#> [2021-05-26 15:05:58]	✓ Open connection
#> [2021-05-26 15:05:58]	✓ JDBC connection
#> [2021-05-26 15:05:58]	SQL: SELECT * FROM omop_vocabulary.concept ORDER BY RANDOM() LIMIT 25
#> [2021-05-26 15:05:58]	Querying...
#> [2021-05-26 15:06:05]	Querying...complete
#> [2021-05-26 15:06:05]	✓ Returned data has more than 0 rows
#> [2021-05-26 15:06:05]	Caching resultset...
#> [2021-05-26 15:06:05]	Postgres connection to 'athena' closed

#> # A tibble: 25 x 10
#>    concept_id concept_name             domain_id  vocabulary_id concept_class_id
#>         <int> <chr>                    <chr>      <chr>         <chr>           
#>  1   44154117 Q-Fluoxetine             Drug       DPD           Brand Name      
#>  2   37018675 History of syphilis      Observati… SNOMED        Context-depende…
#>  3   46174044 Fulcionel female high w… Device     SNOMED        Physical Object 
#>  4   21394304 Redifit ileostomy bag m… Device     dm+d          AMPP            
#>  5   45720673 FOLLITROPIN BETA,RECOMB… Drug       GCN_SEQNO     GCN_SEQNO       
#>  6   41504188 80000 MG Thymus vulgari… Drug       RxNorm Exten… Quant Branded D…
#>  7   35975936 S65W (c.194C>G)          Measureme… CIViC         Variant         
#>  8   36051286 Duomed Soft class II be… Device     dm+d          AMPP            
#>  9    3575457 Form 3 NOS               Observati… SNOMED        Undefined       
#> 10   42390886 Left And Right Pulmonar… Procedure  EDI           Procedure       
#> # … with 15 more rows, and 5 more variables: standard_concept <chr>,
#> #   concept_code <chr>, valid_start_date <date>, valid_end_date <date>,
#> #   invalid_reason <chr>

# Concepts specific to the Drug domain can be sampled
# RxNorm/RxNorm Extension Concepts
get_test_drug_concepts(limit = 10)

#> [2021-05-26 15:06:05]	Connected to 'athena'
#> [2021-05-26 15:06:05]	Overriding cache... Querying Athena...
#> [2021-05-26 15:06:05]	✓ Open connection
#> [2021-05-26 15:06:05]	✓ JDBC connection
#> [2021-05-26 15:06:05]	SQL: SELECT * FROM omop_vocabulary.concept c WHERE c.invalid_reason IS NULL AND c.vocabulary_id IN ('RxNorm', 'RxNorm Extension') AND c.standard_concept <> 'C' AND c.domain_id = 'Drug' ORDER BY RANDOM() LIMIT 10
#> [2021-05-26 15:06:05]	Querying...
#> [2021-05-26 15:06:09]	Querying...complete
#> [2021-05-26 15:06:09]	✓ Returned data has more than 0 rows
#> [2021-05-26 15:06:09]	Caching resultset...
#> [2021-05-26 15:06:09]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 10
#>    concept_id concept_name             domain_id vocabulary_id  concept_class_id
#>         <int> <chr>                    <chr>     <chr>          <chr>           
#>  1   42971402 Epinephrine 0.0125 MG/M… Drug      RxNorm Extens… Branded Drug Co…
#>  2   44049548 Pulsatilla koreana whol… Drug      RxNorm Extens… Marketed Product
#>  3   44028779 Diclofenac 75 MG [DICLO… Drug      RxNorm Extens… Branded Drug Co…
#>  4   41235062 Fluoxetine 40 MG [Fluox… Drug      RxNorm Extens… Branded Drug Co…
#>  5   21129648 Ergocalciferol 3000 UNT… Drug      RxNorm Extens… Branded Drug Co…
#>  6   36789563 pyridoxine 50 MG/ML Inj… Drug      RxNorm Extens… Branded Drug    
#>  7   43766670 Sertraline 112 MG Oral … Drug      RxNorm Extens… Branded Drug    
#>  8   40839132 Verapamil 120 MG Oral T… Drug      RxNorm Extens… Branded Drug    
#>  9   41502870 70000 MG Betula pubesce… Drug      RxNorm Extens… Quant Clinical …
#> 10   40753445 MEASLES VIRUS VACCINE,L… Drug      RxNorm Extens… Clinical Drug F…
#> # … with 5 more variables: standard_concept <chr>, concept_code <chr>,
#> #   valid_start_date <date>, valid_end_date <date>, invalid_reason <chr>

# HemOnc/ATC Classes
get_test_drug_classes(limit = 10)

#> [2021-05-26 15:06:09]	Connected to 'athena'
#> [2021-05-26 15:06:09]	Overriding cache... Querying Athena...
#> [2021-05-26 15:06:09]	✓ Open connection
#> [2021-05-26 15:06:09]	✓ JDBC connection
#> [2021-05-26 15:06:09]	SQL: SELECT * FROM omop_vocabulary.concept c WHERE c.invalid_reason IS NULL AND c.vocabulary_id IN ('ATC', 'HemOnc') AND c.standard_concept = 'C' AND c.domain_id = 'Drug' ORDER BY RANDOM() LIMIT 10
#> [2021-05-26 15:06:09]	Querying...
#> [2021-05-26 15:06:09]	Querying...complete
#> [2021-05-26 15:06:09]	✓ Returned data has more than 0 rows
#> [2021-05-26 15:06:09]	Caching resultset...
#> [2021-05-26 15:06:09]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 10
#>    concept_id concept_name              domain_id vocabulary_id concept_class_id
#>         <int> <chr>                     <chr>     <chr>         <chr>           
#>  1   21600315 VASODILATORS USED IN CAR… Drug      ATC           ATC 3rd         
#>  2   21605009 DECONGESTANTS AND OTHER … Drug      ATC           ATC 3rd         
#>  3   21603463 pheniramine; oral         Drug      ATC           ATC 5th         
#>  4   21603579 Sulfonamides              Drug      ATC           ATC 4th         
#>  5   21605584 rhenium (186Re) etidroni… Drug      ATC           ATC 5th         
#>  6   35807441 Dendritic cell vaccine    Drug      HemOnc        Component Class 
#>  7   21605191 tetracycline; ophthalmic  Drug      ATC           ATC 5th         
#>  8    1502093 lifitegrast; ophthalmic   Drug      ATC           ATC 5th         
#>  9   21601296 meningococcus A, purifie… Drug      ATC           ATC 5th         
#> 10   21600775 Alpha glucosidase inhibi… Drug      ATC           ATC 4th         
#> # … with 5 more variables: standard_concept <chr>, concept_code <chr>,
#> #   valid_start_date <date>, valid_end_date <date>, invalid_reason <chr>

# Test data can also be taken from tables other than Concept
get_test_data(table = "concept_relationship")

#> [2021-05-26 15:06:09]	Connected to 'athena'
#> [2021-05-26 15:06:09]	Overriding cache... Querying Athena...
#> [2021-05-26 15:06:09]	✓ Open connection
#> [2021-05-26 15:06:09]	✓ JDBC connection
#> [2021-05-26 15:06:09]	SQL: SELECT * FROM omop_vocabulary.concept_relationship ORDER BY RANDOM() LIMIT 100
#> [2021-05-26 15:06:09]	Querying...
#> [2021-05-26 15:06:43]	Querying...complete
#> [2021-05-26 15:06:43]	✓ Returned data has more than 0 rows
#> [2021-05-26 15:06:43]	Caching resultset...
#> [2021-05-26 15:06:43]	Postgres connection to 'athena' closed

#> # A tibble: 100 x 6
#>    concept_id_1 concept_id_2 relationship_id     valid_start_date valid_end_date
#>           <int>        <int> <chr>               <date>           <date>        
#>  1      2775692      4179713 Is a                1970-01-01       2099-12-31    
#>  2      1574074        16147 Subsumes            2016-03-25       2099-12-31    
#>  3      4277343      4147543 Is a                2011-07-31       2099-12-31    
#>  4      3426656      3425646 Is a                2019-08-16       2099-12-31    
#>  5      3039678     45878148 Has Answer          1970-01-01       2099-12-31    
#>  6     41104859     40820510 Has supplier        2017-07-18       2099-12-31    
#>  7     41054914     19082573 RxNorm has dose fo… 2017-07-18       2099-12-31    
#>  8     41487666     41486542 Marketed form of    2017-07-18       2099-12-31    
#>  9      1007898     40768600 Component of        2019-12-13       2099-12-31    
#> 10     46180963     35626975 VMP has prescr stat 2018-07-31       2099-12-31    
#> # … with 90 more rows, and 1 more variable: invalid_reason <chr>

get_test_data(table = "concept_ancestor", limit = 10)

#> [2021-05-26 15:06:43]	Connected to 'athena'
#> [2021-05-26 15:06:43]	Overriding cache... Querying Athena...
#> [2021-05-26 15:06:43]	✓ Open connection
#> [2021-05-26 15:06:43]	✓ JDBC connection
#> [2021-05-26 15:06:43]	SQL: SELECT * FROM omop_vocabulary.concept_ancestor ORDER BY RANDOM() LIMIT 10
#> [2021-05-26 15:06:43]	Querying...
#> [2021-05-26 15:07:16]	Querying...complete
#> [2021-05-26 15:07:16]	✓ Returned data has more than 0 rows
#> [2021-05-26 15:07:16]	Caching resultset...
#> [2021-05-26 15:07:16]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 4
#>    ancestor_concept_… descendant_concep… min_levels_of_sepa… max_levels_of_sepa…
#>                 <int>              <int>               <int>               <int>
#>  1             441225           36553633                   1                   1
#>  2           45690023           41221464                   2                   2
#>  3           40019683           36810526                   2                   3
#>  4           42941521           42941521                   0                   0
#>  5           46297564           44127770                   2                   2
#>  6           19078558           35779537                   1                   3
#>  7           21605228           41168137                   5                   6
#>  8           46239009           36814477                   2                   2
#>  9           45706981           40853329                   2                   3
#> 10           42806363           41217845                   2                   2