Get Test Drug Concepts

Get random records of RxNorm and RxNorm Extension Concepts that are:

Valid,

Not a Class,

Belong to the Drug domain

get_test_drug_concepts(
  conn,
  conn_fun = "connectAthena()",
  schema = "omop_vocabulary",
  limit = 100,
  verbose = TRUE,
  render_sql = TRUE
)

See also

Examples

library(chariot)
library(tidyverse)

# Get Unfiltered Test Concepts (Invalid Concepts are included)
get_test_concepts()

#> Connecting using PostgreSQL driver
#> [2021-01-14 17:55:44]	Overriding cache... Querying Athena...
#> [2021-01-14 17:55:44] 

#> ✓ Open connection

#> [2021-01-14 17:55:44] 

#> ✓ JDBC connection

#> [2021-01-14 17:55:44]	SQL: SELECT * FROM omop_vocabulary.concept ORDER BY RANDOM() LIMIT 100
#> [2021-01-14 17:55:44]	Querying...
#> [2021-01-14 17:55:53]	Querying...complete
#> [2021-01-14 17:55:53] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-14 17:55:53]	Caching resultset...
#> [2021-01-14 17:55:53]	Postgres connection to 'athena' closed

#> # A tibble: 100 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id standard_concept concept_code
#>         <dbl> <chr>        <chr>     <chr>         <chr>            <chr>            <chr>       
#>  1    2518835 avobenzone … Drug      NDC           11-digit NDC     NA               49527007501 
#>  2   45409053 pancrelipas… Drug      Multum        Multum           NA               25408       
#>  3    3415979 Intramuscul… Observat… Nebraska Lex… Context-depende… NA               414533004   
#>  4   41474767 6000 MG cic… Drug      RxNorm Exten… Marketed Product S                OMOP2672729 
#>  5   44111720 Mirtazapine… Drug      RxNorm Exten… Branded Drug     S                OMOP1106351 
#>  6   44846522 alpha-ketog… Drug      NDC           9-digit NDC      NA               437420098   
#>  7   35304359 diphenhydra… Drug      NDC           9-digit NDC      NA               639410260   
#>  8   44033903 Riluzole 50… Drug      RxNorm Exten… Branded Drug     S                OMOP1028534 
#>  9   43831296 500 ML N-ac… Drug      RxNorm Exten… Marketed Product S                OMOP863249  
#> 10   43735806 10000 MG La… Drug      RxNorm Exten… Marketed Product S                OMOP778773  
#> # … with 90 more rows, and 3 more variables: valid_start_date <date>, valid_end_date <date>,
#> #   invalid_reason <chr>

# The size of the resultset can be altered
get_test_concepts(limit = 25)

#> Connecting using PostgreSQL driver
#> [2021-01-14 17:55:53]	Overriding cache... Querying Athena...
#> [2021-01-14 17:55:53] 

#> ✓ Open connection

#> [2021-01-14 17:55:53] 

#> ✓ JDBC connection

#> [2021-01-14 17:55:53]	SQL: SELECT * FROM omop_vocabulary.concept ORDER BY RANDOM() LIMIT 25
#> [2021-01-14 17:55:53]	Querying...
#> [2021-01-14 17:56:02]	Querying...complete
#> [2021-01-14 17:56:02] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-14 17:56:02]	Caching resultset...
#> [2021-01-14 17:56:02]	Postgres connection to 'athena' closed

#> # A tibble: 25 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id standard_concept concept_code
#>         <dbl> <chr>        <chr>     <chr>         <chr>            <chr>            <chr>       
#>  1   41319628 terbinafine… Drug      RxNorm Exten… Marketed Product S                OMOP2517590 
#>  2     775058 naproxen so… Drug      SPL           OTC Drug         C                9983d76d-bf…
#>  3    2748311 Fragmentati… Procedure ICD10PCS      ICD10PCS         S                0DFF0ZZ     
#>  4   41132487 Thioctate 6… Drug      RxNorm Exten… Marketed Product S                OMOP2330449 
#>  5   43768759 Betula plat… Drug      RxNorm Exten… Clinical Drug C… S                OMOP705856  
#>  6   42951157 Losartan Or… Drug      RxNorm Exten… Branded Drug Fo… S                OMOP4677729 
#>  7    3172098 Third crani… Observat… Nebraska Lex… Observable Enti… NA               11395008    
#>  8    3249830 Yersinia en… Observat… Nebraska Lex… Organism         NA               720796007   
#>  9   41263146 Salvia offi… Drug      RxNorm Exten… Clinical Drug C… S                OMOP2461108 
#> 10   42924404 ATORAN Oral… Drug      KDC           Drug Product     NA               650303020   
#> # … with 15 more rows, and 3 more variables: valid_start_date <date>, valid_end_date <date>,
#> #   invalid_reason <chr>

# Concepts specific to the Drug domain can be sampled
# RxNorm/RxNorm Extension Concepts
get_test_drug_concepts(limit = 10)

#> Connecting using PostgreSQL driver
#> [2021-01-14 17:56:02]	Overriding cache... Querying Athena...
#> [2021-01-14 17:56:02] 

#> ✓ Open connection

#> [2021-01-14 17:56:02] 

#> ✓ JDBC connection

#> [2021-01-14 17:56:02]	SQL: SELECT * FROM omop_vocabulary.concept c WHERE c.invalid_reason IS NULL AND c.vocabulary_id IN ('RxNorm', 'RxNorm Extension') AND c.standard_concept <> 'C' AND c.domain_id = 'Drug' ORDER BY RANDOM() LIMIT 10
#> [2021-01-14 17:56:02]	Querying...
#> [2021-01-14 17:56:06]	Querying...complete
#> [2021-01-14 17:56:06] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-14 17:56:06]	Caching resultset...
#> [2021-01-14 17:56:06]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id standard_concept concept_code
#>         <dbl> <chr>        <chr>     <chr>         <chr>            <chr>            <chr>       
#>  1   41446705 1.35 ML Met… Drug      RxNorm Exten… Quant Branded D… S                OMOP2644667 
#>  2   41005737 atorvastati… Drug      RxNorm Exten… Branded Drug Box S                OMOP2203699 
#>  3   21133499 lamotrigine… Drug      RxNorm Exten… Marketed Product S                OMOP409815  
#>  4   36779196 50 ML Nimod… Drug      RxNorm Exten… Marketed Product S                OMOP4767508 
#>  5   41066611 Cetirizine … Drug      RxNorm Exten… Branded Drug Box S                OMOP2264573 
#>  6   44053294 Cochineal 4… Drug      RxNorm Exten… Clinical Drug C… S                OMOP1047925 
#>  7   41183455 tamsulosin … Drug      RxNorm Exten… Branded Drug     S                OMOP2381417 
#>  8   43750162 tolterodine… Drug      RxNorm Exten… Branded Drug Co… S                OMOP695847  
#>  9   40899223 Cholecalcif… Drug      RxNorm Exten… Clinical Drug    S                OMOP2097185 
#> 10   43622985 Terazosin 2… Drug      RxNorm Exten… Branded Drug Co… S                OMOP674296  
#> # … with 3 more variables: valid_start_date <date>, valid_end_date <date>, invalid_reason <chr>

# HemOnc/ATC Classes
get_test_drug_classes(limit = 10)

#> Connecting using PostgreSQL driver
#> [2021-01-14 17:56:06]	Overriding cache... Querying Athena...
#> [2021-01-14 17:56:06] 

#> ✓ Open connection

#> [2021-01-14 17:56:06] 

#> ✓ JDBC connection

#> [2021-01-14 17:56:06]	SQL: SELECT * FROM omop_vocabulary.concept c WHERE c.invalid_reason IS NULL AND c.vocabulary_id IN ('ATC', 'HemOnc') AND c.standard_concept = 'C' AND c.domain_id = 'Drug' ORDER BY RANDOM() LIMIT 10
#> [2021-01-14 17:56:06]	Querying...
#> [2021-01-14 17:56:06]	Querying...complete
#> [2021-01-14 17:56:06] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-14 17:56:06]	Caching resultset...
#> [2021-01-14 17:56:06]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id standard_concept concept_code
#>         <dbl> <chr>        <chr>     <chr>         <chr>            <chr>            <chr>       
#>  1   21604454 bornaprine   Drug      ATC           ATC 5th          C                N04AA11     
#>  2   21605260 deferoxamine Drug      ATC           ATC 5th          C                V03AC01     
#>  3   21605051 mometasone;… Drug      ATC           ATC 5th          C                R01AD09     
#>  4   21604264 morphine, c… Drug      ATC           ATC 5th          C                N02AA51     
#>  5   21600783 Dipeptidyl … Drug      ATC           ATC 4th          C                A10BH       
#>  6   21602707 oxytocin; n… Drug      ATC           ATC 5th          C                H01BB02     
#>  7   21602478 norethister… Drug      ATC           ATC 5th          C                G03AA05     
#>  8   21605567 THERAPEUTIC… Drug      ATC           ATC 2nd          C                V10         
#>  9   21604766 caffeine; s… Drug      ATC           ATC 5th          C                N06BC01     
#> 10   21605089 lodoxamide;… Drug      ATC           ATC 5th          C                S01GX05     
#> # … with 3 more variables: valid_start_date <date>, valid_end_date <date>, invalid_reason <chr>

# Test data can also be taken from tables other than Concept
get_test_data(table = "concept_relationship")

#> Connecting using PostgreSQL driver
#> [2021-01-14 17:56:06]	Overriding cache... Querying Athena...
#> [2021-01-14 17:56:06] 

#> ✓ Open connection

#> [2021-01-14 17:56:06] 

#> ✓ JDBC connection

#> [2021-01-14 17:56:06]	SQL: SELECT * FROM omop_vocabulary.concept_relationship ORDER BY RANDOM() LIMIT 100
#> [2021-01-14 17:56:06]	Querying...
#> [2021-01-14 17:56:43]	Querying...complete
#> [2021-01-14 17:56:43] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-14 17:56:43]	Caching resultset...
#> [2021-01-14 17:56:43]	Postgres connection to 'athena' closed

#> # A tibble: 100 x 6
#>    concept_id_1 concept_id_2 relationship_id     valid_start_date valid_end_date invalid_reason
#>           <dbl>        <dbl> <chr>               <date>           <date>         <chr>         
#>  1       785075       785077 Has tradename       2019-04-01       2099-12-31     NA            
#>  2     36256220     36269649 Brand name of       2017-10-01       2099-12-31     NA            
#>  3     41109706     41109706 Maps to             2017-07-18       2099-12-31     NA            
#>  4     41517784     41517843 Available as box    2017-07-18       2099-12-31     NA            
#>  5     36404350     36404350 Mapped from         2017-10-15       2099-12-31     NA            
#>  6     46138028     35626973 VMP has prescr stat 2018-07-31       2099-12-31     NA            
#>  7      3331492      4085832 Maps to             1970-01-01       2099-12-31     NA            
#>  8     43256768     43357011 Mapped from         2016-09-30       2099-12-31     NA            
#>  9     19082077     41128115 RxNorm dose form of 2017-07-18       2099-12-31     NA            
#> 10     40077951     35774801 Has tradename       2017-08-02       2099-12-31     NA            
#> # … with 90 more rows

get_test_data(table = "concept_ancestor", limit = 10)

#> Connecting using PostgreSQL driver
#> [2021-01-14 17:56:43]	Overriding cache... Querying Athena...
#> [2021-01-14 17:56:43] 

#> ✓ Open connection

#> [2021-01-14 17:56:43] 

#> ✓ JDBC connection

#> [2021-01-14 17:56:43]	SQL: SELECT * FROM omop_vocabulary.concept_ancestor ORDER BY RANDOM() LIMIT 10
#> [2021-01-14 17:56:43]	Querying...
#> [2021-01-14 17:57:18]	Querying...complete
#> [2021-01-14 17:57:18] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-14 17:57:18]	Caching resultset...
#> [2021-01-14 17:57:18]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 4
#>    ancestor_concept_id descendant_concept_id min_levels_of_separation max_levels_of_separation
#>                  <dbl>                 <dbl>                    <dbl>                    <dbl>
#>  1            36217216              41197765                        3                        5
#>  2            41235777              40887848                        2                        3
#>  3            36371502                955411                        1                        1
#>  4             1759843              42951817                        0                        0
#>  5            35507590              42923532                        1                        1
#>  6            46303150              21170562                        2                        3
#>  7            36218475              42968035                        1                        2
#>  8            43174254              43670116                        1                        1
#>  9            46035371              46035371                        0                        0
#> 10            21604180              35142587                        5                        5