FUNCTION_TITLE

FUNCTION_DESCRIPTION

join_on_concept_synonym_name(
  kind = c("LEFT", "RIGHT", "INNER", "FULL"),
  data,
  column = NULL,
  select_data_columns = "*",
  distinct = FALSE,
  write_schema = "patelm9",
  vocab_schema = "omop_vocabulary",
  where_in_concept_synonym_field,
  where_in_concept_synonym_field_value,
  where_not_in_concept_synonym_field,
  where_not_in_concept_synonym_field_value,
  where_is_null_concept_synonym_field,
  where_is_not_null_concept_synonym_field,
  case_insensitive = TRUE,
  conn,
  conn_fun = "connectAthena()",
  verbose = TRUE,
  render_sql = TRUE,
  render_only = FALSE
)

Arguments

kind
PARAM_DESCRIPTION, Default: c("LEFT", "RIGHT", "INNER", "FULL")

data
PARAM_DESCRIPTION

column
PARAM_DESCRIPTION, Default: NULL

select_data_columns
PARAM_DESCRIPTION, Default: '*'

distinct
PARAM_DESCRIPTION, Default: FALSE

write_schema
PARAM_DESCRIPTION, Default: 'patelm9'

vocab_schema
PARAM_DESCRIPTION, Default: 'omop_vocabulary'

where_in_concept_synonym_field
PARAM_DESCRIPTION

where_in_concept_synonym_field_value
PARAM_DESCRIPTION

where_not_in_concept_synonym_field
PARAM_DESCRIPTION

where_not_in_concept_synonym_field_value
PARAM_DESCRIPTION

where_is_null_concept_synonym_field
PARAM_DESCRIPTION

where_is_not_null_concept_synonym_field
PARAM_DESCRIPTION

conn
PARAM_DESCRIPTION

conn_fun
PARAM_DESCRIPTION, Default: 'connectAthena()'

verbose
PARAM_DESCRIPTION, Default: TRUE

render_sql
PARAM_DESCRIPTION, Default: TRUE

render_only
PARAM_DESCRIPTION, Default: FALSE

Examples

library(chariot)
library(tidyverse)

test_data <-
        queryAthena(sql_statement =     "
                                        SELECT
                                                concept_id AS test_concept_id,
                                                concept_name AS test_concept_name,
                                                concept_code AS test_concept_code
                                        FROM omop_vocabulary.concept
                                        ORDER BY RANDOM()
                                        LIMIT 20;")

#> [2021-01-01 16:14:17]	Loading Cache...
#> [2021-01-01 16:14:17]	Cached SQL: 
#>                                         SELECT
#>                                                 concept_id AS test_concept_id,
#>                                                 concept_name AS test_concept_name,
#>                                                 concept_code AS test_concept_code
#>                                         FROM omop_vocabulary.concept
#>                                         ORDER BY RANDOM()
#>                                         LIMIT 20;
#> [2021-01-01 16:14:17]	Cached resultset found...

test_data

#> # A tibble: 20 x 3
#>    test_concept_id test_concept_name                      test_concept_code     
#>              <dbl> <chr>                                  <chr>                 
#>  1        44431831 Furosemide 40 MG Oral Tablet [Lasix]   68788636709           
#>  2        36153682 HIGHER CONTROL MALE CLIMAX CONTROL LU… 5f5fd127-34c5-403a-aa…
#>  3          938293 sennosides, USP 12 MG Oral Tablet      243374                
#>  4        37053310 Acetylene | Serum or Plasma | Drug to… LP387597-0            
#>  5         1532840 Medical and Surgical @ Bursae and Lig… 0MRH4                 
#>  6        35923652 Meets criteria for autopsy (a) stagin… epiglottis_anterior@2…
#>  7        41989223 Zillbach                               3120613               
#>  8        42963253 lamotrigine 25 MG Oral Tablet [LAMIAR… OMOP4689738           
#>  9        36633636 Metoprolol 50 Mg Tablets               301333                
#> 10        42048496 Mullaghinch                            4617960               
#> 11        21074774 Itraconazole 100 MG [Sporanox-Pulse]   OMOP408362            
#> 12        43194971 Pravastatin 20 MG Oral Tablet [PRAVAS… OMOP492811            
#> 13         1130266 Diphenhydramine 0.02 MG/MG Topical Ge… 212535                
#> 14        40716741 3 ML insulin, isophane 70 UNT/ML / Re… OMOP4740497           
#> 15        21449919 Polyurethane foam film dressing steri… 11262811000001101     
#> 16        42948537 Dexamethasone 0.75 MG [SAMNAM DEXAMET… OMOP4675126           
#> 17        42897206 Dilation of Right Posterior Tibial Ar… 047R0G6               
#> 18        21465174 Polyurethane foam film dressing steri… 7700511000001106      
#> 19        42904932 CEFANAN                                OMOP4631854           
#> 20         2853303 Radiation Therapy @ Ear, Nose, Mouth … D9YC                  

# `Join On` functions point to the true column that is being joined on
join_on_concept_id(
        data = test_data,
        column = "test_concept_id"
        )

#> [2021-01-01 16:14:17]	Target column: test_concept_id
#> Connecting using PostgreSQL driver

#> Error in rJava::.jcall(jdbcDriver, "Ljava/sql/Connection;", "connect",     as.character(url), p): org.postgresql.util.PSQLException: FATAL: sorry, too many clients already

join_on_concept_code(
        data = test_data,
        column = "test_concept_code"
        )

#> [2021-01-01 16:14:17]	Target column: test_concept_code
#> Connecting using PostgreSQL driver

#> Error in rJava::.jcall(jdbcDriver, "Ljava/sql/Connection;", "connect",     as.character(url), p): org.postgresql.util.PSQLException: FATAL: sorry, too many clients already

join_on_concept_name(
        data = test_data,
        column = "test_concept_name"
)

#> [2021-01-01 16:14:17]	Target column: test_concept_name
#> Connecting using PostgreSQL driver

#> Error in rJava::.jcall(jdbcDriver, "Ljava/sql/Connection;", "connect",     as.character(url), p): org.postgresql.util.PSQLException: FATAL: sorry, too many clients already

# Join on the `concept_synonym_name` field in the Concept Synonym table
join_on_concept_synonym_name(
        data = test_data,
        column = "test_concept_name"
)

#> [2021-01-01 16:14:17]	Target column: test_concept_name
#> Connecting using PostgreSQL driver

#> Error in rJava::.jcall(jdbcDriver, "Ljava/sql/Connection;", "connect",     as.character(url), p): org.postgresql.util.PSQLException: FATAL: sorry, too many clients already

# Reconfigure the kind of join
join_on_concept_synonym_name(
        data = test_data,
        column = "test_concept_name",
        kind = "INNER"
)

#> [2021-01-01 16:14:17]	Target column: test_concept_name
#> Connecting using PostgreSQL driver

#> Error in rJava::.jcall(jdbcDriver, "Ljava/sql/Connection;", "connect",     as.character(url), p): org.postgresql.util.PSQLException: FATAL: sorry, too many clients already

# Return only distinct rows
join_on_concept_synonym_name(
        data = test_data,
        column = "test_concept_name",
        kind = "INNER",
        distinct = TRUE
)

#> [2021-01-01 16:14:17]	Target column: test_concept_name
#> Connecting using PostgreSQL driver

#> Error in rJava::.jcall(jdbcDriver, "Ljava/sql/Connection;", "connect",     as.character(url), p): org.postgresql.util.PSQLException: FATAL: sorry, too many clients already

# Return only the `test_concept_code` column
join_on_concept_synonym_name(
        data = test_data,
        column = "test_concept_name",
        kind = "INNER",
        distinct = TRUE,
        select_data_columns = "test_concept_name"
)

#> [2021-01-01 16:14:17]	Target column: test_concept_name
#> Connecting using PostgreSQL driver

#> Error in rJava::.jcall(jdbcDriver, "Ljava/sql/Connection;", "connect",     as.character(url), p): org.postgresql.util.PSQLException: FATAL: sorry, too many clients already

# Filter can be applied to the vocabulary tables
join_on_concept_synonym_name(
        data = test_data,
        column = "test_concept_name",
        where_not_in_concept_synonym_field = "language_concept_id",
        where_not_in_concept_synonym_field_value = 4180186
)

#> [2021-01-01 16:14:17]	Target column: test_concept_name
#> Connecting using PostgreSQL driver

#> Error in rJava::.jcall(jdbcDriver, "Ljava/sql/Connection;", "connect",     as.character(url), p): org.postgresql.util.PSQLException: FATAL: sorry, too many clients already

# `Join For` functions join on a predetermined vocabulary field for the addition
# of the specified column in the source data

# Get all the synonyms for a concept by concept_id
join_for_concept_synonym_name(
        data = test_data,
        concept_id_column = "test_concept_id"
        )

#> [2021-01-01 16:14:17]	Target column: test_concept_id
#> Connecting using PostgreSQL driver

#> Error in rJava::.jcall(jdbcDriver, "Ljava/sql/Connection;", "connect",     as.character(url), p): org.postgresql.util.PSQLException: FATAL: sorry, too many clients already

# Get the descendant concept ids
join_for_descendant_ids(
        data = test_data,
        ancestor_id_column = "test_concept_id"
        )

#> [2021-01-01 16:14:17]	Target column: test_concept_id
#> Connecting using PostgreSQL driver

#> Error in rJava::.jcall(jdbcDriver, "Ljava/sql/Connection;", "connect",     as.character(url), p): org.postgresql.util.PSQLException: FATAL: sorry, too many clients already

# Get the ancestor concept ids
join_for_ancestor_ids(
        data = test_data,
        descendant_id_column = "test_concept_id"
        )

#> [2021-01-01 16:14:17]	Target column: test_concept_id
#> Connecting using PostgreSQL driver

#> Error in rJava::.jcall(jdbcDriver, "Ljava/sql/Connection;", "connect",     as.character(url), p): org.postgresql.util.PSQLException: FATAL: sorry, too many clients already

# To retrieve all the attributes for either the ancestor or descendants, use
# the following functions
join_for_descendants(
        data = test_data,
        ancestor_id_column = "test_concept_id"
)

#> [2021-01-01 16:14:17]	Target column: test_concept_id
#> Connecting using PostgreSQL driver

#> Error in rJava::.jcall(jdbcDriver, "Ljava/sql/Connection;", "connect",     as.character(url), p): org.postgresql.util.PSQLException: FATAL: sorry, too many clients already

join_for_ancestors(
        data = test_data,
        descendant_id_column = "test_concept_id"
)

#> [2021-01-01 16:14:17]	Target column: test_concept_id
#> Connecting using PostgreSQL driver

#> Error in rJava::.jcall(jdbcDriver, "Ljava/sql/Connection;", "connect",     as.character(url), p): org.postgresql.util.PSQLException: FATAL: sorry, too many clients already