Join a dataframe object with the Concept Table

This function has better performance than the WHERE IN statement for larger searches. A table in the format of "vunpunctuated timestamp" is written to the local Athena. A join is performed with a concept table. The table is subsequently dropped. If a dataframe column is not provided as a the column to join the concept table on, the 1st column will be used by default.

leftJoinConcept(
  .data,
  column = NULL,
  athena_schema = "public",
  concept_column = "concept_id",
  synonyms = FALSE,
  verbose = FALSE,
  conn = NULL,
  render_sql = FALSE,
  sleepTime = 1,
  ...
)

Arguments

.data
A data frame

column
Data frame column that the join will be performed on. If NULL, defaults to the column in position 1 of the data frame.

athena_schema
Schema of the OMOP Concept Table

concept_column
Column in the concept

verbose
If TRUE, prints whether the cache is being loaded or being actively queried in the Postgres database, Default: FALSE

conn
Connection object if another database is used. Default: NULL

render_sql
If TRUE, will print the SQL to the console before executing. Default: FALSE

sleepTime
Argument in seconds passed to the `Sys.sleep()` function at the end of query, Default: 1

...
Additional arguments passed to the `queryAthena()` function.

Value

A data frame

See also