Print a Concepts Hierarchy to the Console

If a level of the hierarchy has more than 10 concepts, the first 10 are printed with an ellipses is included to indicate only the first 10 rows are shown.

print_concept_hierarchy(
  concept_obj,
  level_of_separation_type = c("max", "min")
)

Arguments

concept_obj
Concept class object or a concept id.

See also

Examples

library(chariot)
library(tidyverse)

# Drug Domain
concept_obj <- get_concept(concept_id = 1308216)

#> 

#> ── Is Concept Id Valid Integer? ────────────────────────────────────────────────

#> ✓ Concept Id is a valid integer

#> 

#> ── Lookup Concept Id in Concept Table ──────────────────────────────────────────

#> ✓ Concept Id `1308216` found in Concept Table

#> 

#> ── Lookup Synonyms ─────────────────────────────────────────────────────────────

concept_obj

#> An object of class "concept"
#> Slot "concept_id":
#> [1] 1308216
#> 
#> Slot "concept_name":
#> [1] "lisinopril"
#> 
#> Slot "concept_synonym_names":
#> [1] ""
#> 
#> Slot "domain_id":
#> [1] "Drug"
#> 
#> Slot "vocabulary_id":
#> [1] "RxNorm"
#> 
#> Slot "concept_class_id":
#> [1] "Ingredient"
#> 
#> Slot "standard_concept":
#> [1] "S"
#> 
#> Slot "concept_code":
#> [1] "29046"
#> 
#> Slot "valid_start_date":
#> [1] "1970-01-01"
#> 
#> Slot "valid_end_date":
#> [1] "2099-12-31"
#> 
#> Slot "invalid_reason":
#> [1] NA
#> 

print_concept_hierarchy(concept_obj = concept_obj)

#> [2021-01-01 16:26:42]	Loading Cache...
#> [2021-01-01 16:26:42]	Cached SQL: SELECT *
#>                                                         FROM omop_vocabulary.concept c
#>                                                         WHERE c.concept_id IN (1308216)
#>                                                       
#> [2021-01-01 16:26:42]	Cached resultset found...
#> [2021-01-01 16:26:42]	Target column: concept_hierarchy_id
#> Connecting using PostgreSQL driver

#> Error in rJava::.jcall(jdbcDriver, "Ljava/sql/Connection;", "connect",     as.character(url), p): org.postgresql.util.PSQLException: Connection to localhost:5432 refused. Check that the hostname and port are correct and that the postmaster is accepting TCP/IP connections.

print_concept_hierarchy(concept_obj = concept_obj,
                        level_of_separation_type = "min")

#> [2021-01-01 16:26:42]	Loading Cache...
#> [2021-01-01 16:26:42]	Cached SQL: SELECT *
#>                                                         FROM omop_vocabulary.concept c
#>                                                         WHERE c.concept_id IN (1308216)
#>                                                       
#> [2021-01-01 16:26:42]	Cached resultset found...
#> [2021-01-01 16:26:42]	Target column: concept_hierarchy_id
#> Connecting using PostgreSQL driver

#> Error in rJava::.jcall(jdbcDriver, "Ljava/sql/Connection;", "connect",     as.character(url), p): org.postgresql.util.PSQLException: Connection to localhost:5432 refused. Check that the hostname and port are correct and that the postmaster is accepting TCP/IP connections.

# Cancer Condition Domain
concept_obj <- get_concept(concept_id = 4187868)

#> 

#> ── Is Concept Id Valid Integer? ────────────────────────────────────────────────

#> ✓ Concept Id is a valid integer

#> 

#> ── Lookup Concept Id in Concept Table ──────────────────────────────────────────

#> ✓ Concept Id `4187868` found in Concept Table

#> 

#> ── Lookup Synonyms ─────────────────────────────────────────────────────────────

concept_obj

#> An object of class "concept"
#> Slot "concept_id":
#> [1] 4187868
#> 
#> Slot "concept_name":
#> [1] "pTis: Ductal carcinoma in situ"
#> 
#> Slot "concept_synonym_names":
#> [1] "Ductal carcinoma in situ (breast)|pTis: Ductal carcinoma in situ (breast)|pTis: Ductal carcinoma in situ (breast) (finding)"
#> 
#> Slot "domain_id":
#> [1] "Condition"
#> 
#> Slot "vocabulary_id":
#> [1] "SNOMED"
#> 
#> Slot "concept_class_id":
#> [1] "Clinical Finding"
#> 
#> Slot "standard_concept":
#> [1] "S"
#> 
#> Slot "concept_code":
#> [1] "373176000"
#> 
#> Slot "valid_start_date":
#> [1] "1970-01-01"
#> 
#> Slot "valid_end_date":
#> [1] "2099-12-31"
#> 
#> Slot "invalid_reason":
#> [1] NA
#> 

print_concept_hierarchy(concept_obj = concept_obj)

#> [2021-01-01 16:26:42]	Loading Cache...
#> [2021-01-01 16:26:42]	Cached SQL: SELECT *
#>                                                         FROM omop_vocabulary.concept c
#>                                                         WHERE c.concept_id IN (4187868)
#>                                                       
#> [2021-01-01 16:26:42]	Cached resultset found...
#> [2021-01-01 16:26:42]	Target column: concept_hierarchy_id
#> Connecting using PostgreSQL driver

#> Error in rJava::.jcall(jdbcDriver, "Ljava/sql/Connection;", "connect",     as.character(url), p): org.postgresql.util.PSQLException: Connection to localhost:5432 refused. Check that the hostname and port are correct and that the postmaster is accepting TCP/IP connections.

print_concept_hierarchy(concept_obj = concept_obj,
                        level_of_separation_type = "min")

#> [2021-01-01 16:26:42]	Loading Cache...
#> [2021-01-01 16:26:42]	Cached SQL: SELECT *
#>                                                         FROM omop_vocabulary.concept c
#>                                                         WHERE c.concept_id IN (4187868)
#>                                                       
#> [2021-01-01 16:26:42]	Cached resultset found...
#> [2021-01-01 16:26:43]	Target column: concept_hierarchy_id
#> Connecting using PostgreSQL driver

#> Error in rJava::.jcall(jdbcDriver, "Ljava/sql/Connection;", "connect",     as.character(url), p): org.postgresql.util.PSQLException: Connection to localhost:5432 refused. Check that the hostname and port are correct and that the postmaster is accepting TCP/IP connections.

# Chronic Disease Domain
concept_obj <- get_concept(concept_id = 319835)

#> 

#> ── Is Concept Id Valid Integer? ────────────────────────────────────────────────

#> ✓ Concept Id is a valid integer

#> 

#> ── Lookup Concept Id in Concept Table ──────────────────────────────────────────

#> ✓ Concept Id `319835` found in Concept Table

#> 

#> ── Lookup Synonyms ─────────────────────────────────────────────────────────────

concept_obj

#> An object of class "concept"
#> Slot "concept_id":
#> [1] 319835
#> 
#> Slot "concept_name":
#> [1] "Congestive heart failure"
#> 
#> Slot "concept_synonym_names":
#> [1] "CCF - Congestive cardiac failure|CHF - Congestive heart failure|Congestive cardiac failure|Congestive heart disease|Congestive heart failure (disorder)"
#> 
#> Slot "domain_id":
#> [1] "Condition"
#> 
#> Slot "vocabulary_id":
#> [1] "SNOMED"
#> 
#> Slot "concept_class_id":
#> [1] "Clinical Finding"
#> 
#> Slot "standard_concept":
#> [1] "S"
#> 
#> Slot "concept_code":
#> [1] "42343007"
#> 
#> Slot "valid_start_date":
#> [1] "1970-01-01"
#> 
#> Slot "valid_end_date":
#> [1] "2099-12-31"
#> 
#> Slot "invalid_reason":
#> [1] NA
#> 

print_concept_hierarchy(concept_obj = concept_obj)

#> [2021-01-01 16:26:43]	Loading Cache...
#> [2021-01-01 16:26:43]	Cached SQL: SELECT *
#>                                                         FROM omop_vocabulary.concept c
#>                                                         WHERE c.concept_id IN (319835)
#>                                                       
#> [2021-01-01 16:26:43]	Cached resultset found...
#> [2021-01-01 16:26:43]	Target column: concept_hierarchy_id
#> Connecting using PostgreSQL driver

#> Error in rJava::.jcall(jdbcDriver, "Ljava/sql/Connection;", "connect",     as.character(url), p): org.postgresql.util.PSQLException: Connection to localhost:5432 refused. Check that the hostname and port are correct and that the postmaster is accepting TCP/IP connections.

print_concept_hierarchy(concept_obj = concept_obj,
                        level_of_separation_type = "min")

#> [2021-01-01 16:26:43]	Loading Cache...
#> [2021-01-01 16:26:43]	Cached SQL: SELECT *
#>                                                         FROM omop_vocabulary.concept c
#>                                                         WHERE c.concept_id IN (319835)
#>                                                       
#> [2021-01-01 16:26:43]	Cached resultset found...
#> [2021-01-01 16:26:43]	Target column: concept_hierarchy_id
#> Connecting using PostgreSQL driver

#> Error in rJava::.jcall(jdbcDriver, "Ljava/sql/Connection;", "connect",     as.character(url), p): org.postgresql.util.PSQLException: Connection to localhost:5432 refused. Check that the hostname and port are correct and that the postmaster is accepting TCP/IP connections.

# Measurement Domain
concept_obj <- get_concept(concept_id = 4298431)

#> 

#> ── Is Concept Id Valid Integer? ────────────────────────────────────────────────

#> ✓ Concept Id is a valid integer

#> 

#> ── Lookup Concept Id in Concept Table ──────────────────────────────────────────

#> ✓ Concept Id `4298431` found in Concept Table

#> 

#> ── Lookup Synonyms ─────────────────────────────────────────────────────────────

concept_obj

#> An object of class "concept"
#> Slot "concept_id":
#> [1] 4298431
#> 
#> Slot "concept_name":
#> [1] "White blood cell count"
#> 
#> Slot "concept_synonym_names":
#> [1] "WBC - White blood cell count|WBC count|WCC - White blood cell count|White blood cell count (procedure)|White blood cell count - observation"
#> 
#> Slot "domain_id":
#> [1] "Measurement"
#> 
#> Slot "vocabulary_id":
#> [1] "SNOMED"
#> 
#> Slot "concept_class_id":
#> [1] "Procedure"
#> 
#> Slot "standard_concept":
#> [1] "S"
#> 
#> Slot "concept_code":
#> [1] "767002"
#> 
#> Slot "valid_start_date":
#> [1] "1970-01-01"
#> 
#> Slot "valid_end_date":
#> [1] "2099-12-31"
#> 
#> Slot "invalid_reason":
#> [1] NA
#> 

print_concept_hierarchy(concept_obj = concept_obj)

#> [2021-01-01 16:26:43]	Loading Cache...
#> [2021-01-01 16:26:43]	Cached SQL: SELECT *
#>                                                         FROM omop_vocabulary.concept c
#>                                                         WHERE c.concept_id IN (4298431)
#>                                                       
#> [2021-01-01 16:26:43]	Cached resultset found...
#> [2021-01-01 16:26:43]	Target column: concept_hierarchy_id
#> Connecting using PostgreSQL driver

#> Error in rJava::.jcall(jdbcDriver, "Ljava/sql/Connection;", "connect",     as.character(url), p): org.postgresql.util.PSQLException: Connection to localhost:5432 refused. Check that the hostname and port are correct and that the postmaster is accepting TCP/IP connections.

print_concept_hierarchy(concept_obj = concept_obj,
                        level_of_separation_type = "min")

#> [2021-01-01 16:26:43]	Loading Cache...
#> [2021-01-01 16:26:43]	Cached SQL: SELECT *
#>                                                         FROM omop_vocabulary.concept c
#>                                                         WHERE c.concept_id IN (4298431)
#>                                                       
#> [2021-01-01 16:26:43]	Cached resultset found...
#> [2021-01-01 16:26:43]	Target column: concept_hierarchy_id
#> Connecting using PostgreSQL driver

#> Error in rJava::.jcall(jdbcDriver, "Ljava/sql/Connection;", "connect",     as.character(url), p): org.postgresql.util.PSQLException: Connection to localhost:5432 refused. Check that the hostname and port are correct and that the postmaster is accepting TCP/IP connections.