Search for a Code in the Concept Table

Search for a Code in the Concept Table

search_code(
  code_pattern,
  ignore.case = FALSE,
  perl = FALSE,
  fixed = FALSE,
  useBytes = FALSE,
  vocab_schema = "omop_vocabulary",
  limit = NULL,
  conn,
  conn_fun = "connectAthena()",
  cache_only = FALSE,
  skip_cache = FALSE,
  override_cache = FALSE,
  verbose = TRUE,
  render_sql = TRUE,
  sleepTime = 1
)

Arguments

code_pattern
Regex pattern of code

conn
Connection object. If provided, diverts queries to the connection instead of the local Athena instance without caching features.

cache_only
Loads from the cache and does not query the database. A NULL object is returned if a resultset was not cached.

skip_cache
Skip the caching altogether and directly query the database.

override_cache
If TRUE, the cache will not be loaded and will be overwritten by a new query. For override_cache to take effect, skip_cache should be FALSE.

verbose
If TRUE, prints loading and querying operations messages to the console. Default: FALSE

render_sql
If TRUE, the SQL will be printed back in the console prior to execution. Default: FALSE

sleepTime
Argument for `Sys.sleep()` in between queries to allow for halting function execution, especially in cases where other chariot functions are executing multiple queries in succession and require cancellation.

Value

resultset as a dataframe with all column types as character and trimmed white space

Examples

if(interactive()){
 search_code(code_pattern = "^C40")
 }

#> Connecting using PostgreSQL driver
#> [2021-01-10 16:05:59]	Loading Cache...
#> [2021-01-10 16:05:59]	Cached SQL: SELECT *
#> FROM omop_vocabulary.concept
#> WHERE LOWER(concept_code) LIKE '%%c40%'
#> [2021-01-10 16:05:59]	Cached resultset found...
#> [2021-01-10 16:05:59]	Postgres connection to 'athena' closed

#> # A tibble: 177 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id
#>         <dbl> <chr>        <chr>     <chr>         <chr>           
#>  1     705913 prucalopride Drug      MeSH          Suppl Concept   
#>  2     979962 mintlactone  Drug      MeSH          Suppl Concept   
#>  3    1405897 Malignant n… Condition ICD10CN       ICD10 Hierarchy 
#>  4    1405898 Malignant n… Condition ICD10CN       ICD10 code      
#>  5    1405899 Malignant n… Condition ICD10CN       ICD10 code      
#>  6    1405900 Scapula can… Condition ICD10CN       ICD10 code      
#>  7    1405901 Humerus can… Condition ICD10CN       ICD10 code      
#>  8    1405902 Ulna cancer… Condition ICD10CN       ICD10 code      
#>  9    1405903 Radial bone… Condition ICD10CN       ICD10 code      
#> 10    1405904 Elbow cance… Condition ICD10CN       ICD10 code      
#> # … with 167 more rows, and 5 more variables: standard_concept <chr>,
#> #   concept_code <chr>, valid_start_date <date>, valid_end_date <date>,
#> #   invalid_reason <chr>