Search Exact Matches to a Phrase

Search for an exact phrase in the Concept Synonym table.

search_exact_phrase(
  phrase,
  case_insensitive = TRUE,
  vocab_schema = "omop_vocabulary",
  conn,
  conn_fun = "connectAthena()",
  cache_only = FALSE,
  skip_cache = FALSE,
  override_cache = FALSE,
  cache_resultset = TRUE,
  render_sql = TRUE,
  render_only = FALSE,
  verbose = TRUE,
  sleepTime = 1
)

Arguments

phrase
String to search.

case_insensitive
Should the search ignore case?, Default: TRUE

vocab_schema
OMOP Vocabulary schema, Default: 'omop_vocabulary'

conn
Connection object. If provided, diverts queries to the connection instead of the local Athena instance without caching features.

cache_only
Loads from the cache and does not query the database. A NULL object is returned if a resultset was not cached.

skip_cache
Skip the caching altogether and directly query the database.

override_cache
If TRUE, the cache will not be loaded and will be overwritten by a new query. For override_cache to take effect, skip_cache should be FALSE.

cache_resultset
(deprecated) If TRUE, the resultset from the query will first be loaded from the cache. The query will be executed if a cached resultset is not retrieved for this particular query, after which the resultset will be cached. If FALSE, Athena or conn will be directly queried without any caching operations.

render_sql
If TRUE, the SQL will be printed back in the console prior to execution. Default: FALSE

verbose
If TRUE, prints loading and querying operations messages to the console. Default: FALSE

sleepTime
Argument for `Sys.sleep()` in between queries to allow for halting function execution, especially in cases where other chariot functions are executing multiple queries in succession and require cancellation.

Value

Tibble of all the matching Concept Table fields and an added `concept_synonyms` field of a pipe-separated aggregate of all the synonyms, including `concept_name`, in the Concept Synonym table.

See also

Examples

library(tidyverse)
library(chariot)

# Search the OMOP Vocabulary for an exact match to the phrase
# "Myocardial Infarction" insensitive to case.
search_exact_phrase(phrase = "Myocardial Infarction",
                    case_insensitive = TRUE)

#> Connecting using PostgreSQL driver
#> [2021-01-10 14:58:23]	Loading Cache...
#> [2021-01-10 14:58:23]	Cached SQL: 
#>                         SELECT c.*, STRING_AGG(cs.concept_synonym_name, '|') AS concept_synonyms
#>                         FROM omop_vocabulary.concept_synonym cs
#>                         INNER JOIN omop_vocabulary.concept c
#>                         ON c.concept_id = cs.concept_id
#>                         WHERE LOWER(cs.concept_synonym_name) = LOWER('Myocardial Infarction')
#>                         GROUP BY c.concept_id,
#>                                 c.concept_name,
#>                                 c.domain_id,
#>                                 c.vocabulary_id,
#>                                 c.concept_class_id,
#>                                 c.standard_concept,
#>                                 c.concept_code,
#>                                 c.valid_start_date,
#>                                 c.valid_end_date,
#>                                 c.invalid_reason;
#>                         
#> [2021-01-10 14:58:23]	Cached resultset found...
#> [2021-01-10 14:58:23]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 11
#>    concept_id concept_name domain_id vocabulary_id concept_class_id
#>         <dbl> <chr>        <chr>     <chr>         <chr>           
#>  1    1034841 Myocardial … Observat… LOINC         LOINC Component 
#>  2    3137847 Myocardial … Condition Nebraska Lex… Clinical Finding
#>  3    3248659 Myocardial … Condition Nebraska Lex… Clinical Finding
#>  4    4329847 Myocardial … Condition SNOMED        Clinical Finding
#>  5    4347016 Myocardial … Drug      NDFRT         Ind / CI        
#>  6   40436955 Myocardial … Condition SNOMED        Clinical Finding
#>  7   40623902 Coronary ar… Condition SNOMED        Clinical Finding
#>  8   40761359 Myocardial … Observat… LOINC         Clinical Observ…
#>  9   45616764 Myocardial … Condition MeSH          Main Heading    
#> 10   45881971 Myocardial … Meas Val… LOINC         Answer          
#> # … with 6 more variables: standard_concept <chr>, concept_code <chr>,
#> #   valid_start_date <date>, valid_end_date <date>, invalid_reason <chr>,
#> #   concept_synonyms <chr>

# Repeat with case sensitivity
search_exact_phrase(phrase = "Myocardial Infarction",
                    case_insensitive = FALSE)

#> Connecting using PostgreSQL driver
#> [2021-01-10 14:58:23]	Loading Cache...
#> [2021-01-10 14:58:23]	Cached SQL: 
#>                         SELECT c.*, STRING_AGG(cs.concept_synonym_name, '|') AS concept_synonyms
#>                         FROM omop_vocabulary.concept_synonym cs
#>                         INNER JOIN omop_vocabulary.concept c
#>                         ON c.concept_id = cs.concept_id
#>                         WHERE cs.concept_synonym_name = 'Myocardial Infarction'
#>                         GROUP BY c.concept_id,
#>                                 c.concept_name,
#>                                 c.domain_id,
#>                                 c.vocabulary_id,
#>                                 c.concept_class_id,
#>                                 c.standard_concept,
#>                                 c.concept_code,
#>                                 c.valid_start_date,
#>                                 c.valid_end_date,
#>                                 c.invalid_reason;
#>                         
#> [2021-01-10 14:58:23]	Cached resultset found...
#> [2021-01-10 14:58:23]	Postgres connection to 'athena' closed

#> # A tibble: 4 x 11
#>   concept_id concept_name domain_id vocabulary_id concept_class_id
#>        <dbl> <chr>        <chr>     <chr>         <chr>           
#> 1    4347016 Myocardial … Drug      NDFRT         Ind / CI        
#> 2   40436955 Myocardial … Condition SNOMED        Clinical Finding
#> 3   40623902 Coronary ar… Condition SNOMED        Clinical Finding
#> 4   45616764 Myocardial … Condition MeSH          Main Heading    
#> # … with 6 more variables: standard_concept <chr>, concept_code <chr>,
#> #   valid_start_date <date>, valid_end_date <date>, invalid_reason <chr>,
#> #   concept_synonyms <chr>

# Search concepts that contain the phrase "Myocardial Infarction"
search_like_phrase(phrase = "Myocardial infarction")

#> Connecting using PostgreSQL driver
#> [2021-01-10 14:58:23]	Loading Cache...
#> [2021-01-10 14:58:23]	Cached SQL: 
#>                         SELECT c.*, STRING_AGG(cs.concept_synonym_name, '|') AS concept_synonyms
#>                         FROM omop_vocabulary.concept_synonym cs
#>                         INNER JOIN omop_vocabulary.concept c
#>                         ON c.concept_id = cs.concept_id
#>                         WHERE LOWER(cs.concept_synonym_name) LIKE LOWER('%Myocardial infarction%')
#>                         GROUP BY c.concept_id,
#>                                 c.concept_name,
#>                                 c.domain_id,
#>                                 c.vocabulary_id,
#>                                 c.concept_class_id,
#>                                 c.standard_concept,
#>                                 c.concept_code,
#>                                 c.valid_start_date,
#>                                 c.valid_end_date,
#>                                 c.invalid_reason;
#>                         
#> [2021-01-10 14:58:23]	Cached resultset found...
#> [2021-01-10 14:58:24]	Postgres connection to 'athena' closed

#> # A tibble: 877 x 11
#>    concept_id concept_name domain_id vocabulary_id concept_class_id
#>         <dbl> <chr>        <chr>     <chr>         <chr>           
#>  1     312327 Acute myoca… Condition SNOMED        Clinical Finding
#>  2     314666 Old myocard… Condition SNOMED        Clinical Finding
#>  3     319038 Postmyocard… Condition SNOMED        Clinical Finding
#>  4     319039 Acute poste… Condition SNOMED        Clinical Finding
#>  5     434376 Acute myoca… Condition SNOMED        Clinical Finding
#>  6     435561 Certain cur… Condition SNOMED        Clinical Finding
#>  7     436706 Acute myoca… Condition SNOMED        Clinical Finding
#>  8     438170 Acute myoca… Condition SNOMED        Clinical Finding
#>  9     438172 Atrial sept… Condition SNOMED        Clinical Finding
#> 10     438438 Acute myoca… Condition SNOMED        Clinical Finding
#> # … with 867 more rows, and 6 more variables: standard_concept <chr>,
#> #   concept_code <chr>, valid_start_date <date>, valid_end_date <date>,
#> #   invalid_reason <chr>, concept_synonyms <chr>

search_like_phrase(phrase = "Myocardial infarction",
                   case_insensitive = FALSE)

#> Connecting using PostgreSQL driver
#> [2021-01-10 14:58:24]	Loading Cache...
#> [2021-01-10 14:58:24]	Cached SQL: 
#>                         SELECT c.*, STRING_AGG(cs.concept_synonym_name, '|') AS concept_synonyms
#>                         FROM omop_vocabulary.concept_synonym cs
#>                         INNER JOIN omop_vocabulary.concept c
#>                         ON c.concept_id = cs.concept_id
#>                         WHERE cs.concept_synonym_name LIKE '%Myocardial infarction%'
#>                         GROUP BY c.concept_id,
#>                                 c.concept_name,
#>                                 c.domain_id,
#>                                 c.vocabulary_id,
#>                                 c.concept_class_id,
#>                                 c.standard_concept,
#>                                 c.concept_code,
#>                                 c.valid_start_date,
#>                                 c.valid_end_date,
#>                                 c.invalid_reason;
#>                         
#> [2021-01-10 14:58:24]	Cached resultset found...
#> [2021-01-10 14:58:24]	Postgres connection to 'athena' closed

#> # A tibble: 61 x 11
#>    concept_id concept_name domain_id vocabulary_id concept_class_id
#>         <dbl> <chr>        <chr>     <chr>         <chr>           
#>  1    1014313 Myocardial … Observat… LOINC         LOINC Component 
#>  2    1034841 Myocardial … Observat… LOINC         LOINC Component 
#>  3    1034906 Myocardial … Observat… LOINC         LOINC Component 
#>  4    1326590 Myocardial … Condition ICD10CM       5-char billing …
#>  5    1413769 Myocardial … Condition ICD10CN       ICD10 code      
#>  6    1413811 Myocardial … Condition ICD10CN       ICD10 code      
#>  7    1413812 Myocardial … Condition ICD10CN       ICD10 code      
#>  8    3090976 H/O: Myocar… Observat… Nebraska Lex… Context-depende…
#>  9    3109458 H/O: Myocar… Observat… Nebraska Lex… Context-depende…
#> 10    3124335 Myocardial … Condition Nebraska Lex… Clinical Finding
#> # … with 51 more rows, and 6 more variables: standard_concept <chr>,
#> #   concept_code <chr>, valid_start_date <date>, valid_end_date <date>,
#> #   invalid_reason <chr>, concept_synonyms <chr>

# Search concepts that start with or end with the phrase "Myocardial Infarction"
search_starts_with_phrase(phrase = "Myocardial infarction")

#> Connecting using PostgreSQL driver
#> [2021-01-10 14:58:24]	Loading Cache...
#> [2021-01-10 14:58:24]	Cached SQL: 
#>                         SELECT c.*, STRING_AGG(cs.concept_synonym_name, '|') AS concept_synonyms
#>                         FROM omop_vocabulary.concept_synonym cs
#>                         INNER JOIN omop_vocabulary.concept c
#>                         ON c.concept_id = cs.concept_id
#>                         WHERE LOWER(cs.concept_synonym_name) LIKE LOWER('Myocardial infarction%')
#>                         GROUP BY c.concept_id,
#>                                 c.concept_name,
#>                                 c.domain_id,
#>                                 c.vocabulary_id,
#>                                 c.concept_class_id,
#>                                 c.standard_concept,
#>                                 c.concept_code,
#>                                 c.valid_start_date,
#>                                 c.valid_end_date,
#>                                 c.invalid_reason;
#>                         
#> [2021-01-10 14:58:24]	Cached resultset found...
#> [2021-01-10 14:58:24]	Postgres connection to 'athena' closed

#> # A tibble: 49 x 11
#>    concept_id concept_name domain_id vocabulary_id concept_class_id
#>         <dbl> <chr>        <chr>     <chr>         <chr>           
#>  1    1014313 Myocardial … Observat… LOINC         LOINC Component 
#>  2    1034841 Myocardial … Observat… LOINC         LOINC Component 
#>  3    1034906 Myocardial … Observat… LOINC         LOINC Component 
#>  4    1326590 Myocardial … Condition ICD10CM       5-char billing …
#>  5    1413769 Myocardial … Condition ICD10CN       ICD10 code      
#>  6    1413811 Myocardial … Condition ICD10CN       ICD10 code      
#>  7    1413812 Myocardial … Condition ICD10CN       ICD10 code      
#>  8    3124335 Myocardial … Condition Nebraska Lex… Clinical Finding
#>  9    3137847 Myocardial … Condition Nebraska Lex… Clinical Finding
#> 10    3162882 Myocardial … Observat… Nebraska Lex… Qualifier Value 
#> # … with 39 more rows, and 6 more variables: standard_concept <chr>,
#> #   concept_code <chr>, valid_start_date <date>, valid_end_date <date>,
#> #   invalid_reason <chr>, concept_synonyms <chr>

search_ends_with_phrase(phrase = "Myocardial infarction")

#> Connecting using PostgreSQL driver
#> [2021-01-10 14:58:24]	Loading Cache...
#> [2021-01-10 14:58:24]	Cached SQL: 
#>                         SELECT c.*, STRING_AGG(cs.concept_synonym_name, '|') AS concept_synonyms
#>                         FROM omop_vocabulary.concept_synonym cs
#>                         INNER JOIN omop_vocabulary.concept c
#>                         ON c.concept_id = cs.concept_id
#>                         WHERE LOWER(cs.concept_synonym_name) LIKE LOWER('%Myocardial infarction')
#>                         GROUP BY c.concept_id,
#>                                 c.concept_name,
#>                                 c.domain_id,
#>                                 c.vocabulary_id,
#>                                 c.concept_class_id,
#>                                 c.standard_concept,
#>                                 c.concept_code,
#>                                 c.valid_start_date,
#>                                 c.valid_end_date,
#>                                 c.invalid_reason;
#>                         
#> [2021-01-10 14:58:24]	Cached resultset found...
#> [2021-01-10 14:58:24]	Postgres connection to 'athena' closed

#> # A tibble: 328 x 11
#>    concept_id concept_name domain_id vocabulary_id concept_class_id
#>         <dbl> <chr>        <chr>     <chr>         <chr>           
#>  1     312327 Acute myoca… Condition SNOMED        Clinical Finding
#>  2     314666 Old myocard… Condition SNOMED        Clinical Finding
#>  3     319039 Acute poste… Condition SNOMED        Clinical Finding
#>  4     434376 Acute myoca… Condition SNOMED        Clinical Finding
#>  5     435561 Certain cur… Condition SNOMED        Clinical Finding
#>  6     436706 Acute myoca… Condition SNOMED        Clinical Finding
#>  7     438170 Acute myoca… Condition SNOMED        Clinical Finding
#>  8     438172 Atrial sept… Condition SNOMED        Clinical Finding
#>  9     438438 Acute myoca… Condition SNOMED        Clinical Finding
#> 10     438447 Acute myoca… Condition SNOMED        Clinical Finding
#> # … with 318 more rows, and 6 more variables: standard_concept <chr>,
#> #   concept_code <chr>, valid_start_date <date>, valid_end_date <date>,
#> #   invalid_reason <chr>, concept_synonyms <chr>

# Search a phrase by splitting into smaller words that are each included in a
# 'LIKE' clause
search_split_phrase(phrase = "Myocardial infarction",
                    split = " ")

#> Connecting using PostgreSQL driver
#> [2021-01-10 14:58:24]	Loading Cache...
#> [2021-01-10 14:58:24]	Cached SQL: 
#>                                                         SELECT c.*, STRING_AGG(cs.concept_synonym_name, '|') AS concept_synonyms
#>                                                         FROM omop_vocabulary.concept_synonym cs
#>                                                         INNER JOIN omop_vocabulary.concept c
#>                                                         ON c.concept_id  = cs.concept_id
#>                                                         WHERE LOWER(cs.concept_synonym_name) LIKE LOWER('%Myocardial%')
#> 								AND LOWER(cs.concept_synonym_name) LIKE LOWER('%infarction%')
#>                                                         GROUP BY
#>                                                                 c.concept_id,
#>                                                                 c.concept_name,
#>                                                                 c.domain_id,
#>                                                                 c.vocabulary_id,
#>                                                                 c.concept_class_id,
#>                                                                 c.standard_concept,
#>                                                                 c.concept_code,
#>                                                                 c.valid_start_date,
#>                                                                 c.valid_end_date,
#>                                                                 c.invalid_reason;
#> [2021-01-10 14:58:24]	Cached resultset found...
#> [2021-01-10 14:58:24]	Postgres connection to 'athena' closed

#> # A tibble: 880 x 11
#>    concept_id concept_name domain_id vocabulary_id concept_class_id
#>         <dbl> <chr>        <chr>     <chr>         <chr>           
#>  1     312327 Acute myoca… Condition SNOMED        Clinical Finding
#>  2     314666 Old myocard… Condition SNOMED        Clinical Finding
#>  3     319038 Postmyocard… Condition SNOMED        Clinical Finding
#>  4     319039 Acute poste… Condition SNOMED        Clinical Finding
#>  5     434376 Acute myoca… Condition SNOMED        Clinical Finding
#>  6     435561 Certain cur… Condition SNOMED        Clinical Finding
#>  7     436706 Acute myoca… Condition SNOMED        Clinical Finding
#>  8     438170 Acute myoca… Condition SNOMED        Clinical Finding
#>  9     438172 Atrial sept… Condition SNOMED        Clinical Finding
#> 10     438438 Acute myoca… Condition SNOMED        Clinical Finding
#> # … with 870 more rows, and 6 more variables: standard_concept <chr>,
#> #   concept_code <chr>, valid_start_date <date>, valid_end_date <date>,
#> #   invalid_reason <chr>, concept_synonyms <chr>