Filter Columns with Merged Concept Strips

This function filters a column that contains Concept Strips using Concept Table parameters. The target column may contain 1 or more merged concept strip, and the multiple strips must be separated by a new line \"\n\" for the filter to operate correctly. It is important to note that the the filter is applied to the entire Concept Strip cell and will not alter the data content within the cell otherwise. For example, if the filter `vocabulary_id == 'RxNorm'` is used for `ColumnA`, a `ColumnA` cell that contains at least 1 RxNorm concept will be filtered for though there are other non-RxNorm concepts in that same cell.

filter_strip(data, strip_col, ...)

Arguments

data
dataframe with the merged concept column

...
arguments that will be passed to the dplyr filter function using the base Concept Table field names

merge_col
column of merged concepts

Value

A tibble with the same number of columns as the input with the number of rows equal or less than that of the input.

See also

Examples

library(tidyverse)
library(chariot)

conn <- connectAthena()

#> Connecting using PostgreSQL driver

test_data <-
        queryAthena("SELECT * FROM omop_vocabulary.concept ORDER BY RANDOM() LIMIT 20;",
                    conn = conn)

#> [2021-01-09 18:30:53]	Loading Cache...
#> [2021-01-09 18:30:53]	Cached SQL: SELECT * FROM omop_vocabulary.concept ORDER BY RANDOM() LIMIT 20;
#> [2021-01-09 18:30:53]	Cached resultset found...

# Merge
merge_strip(data = test_data,
            into = "concept")

#> # A tibble: 20 x 2
#>    concept_id concept                                                           
#>         <dbl> <chr>                                                             
#>  1   46187243 [V] [N] 46187243 Levothyroxine sodium 100micrograms/5ml oral susp…
#>  2    3436653 [V] [N] 3436653 Swallowing defect amelioration [Nebraska Lexicon …
#>  3    4333553 [V] [S] 4333553 Radiofrequency destruction of trigger point [SNOM…
#>  4   44154298 [V] [N] 44154298 Phs Gold [DPD DPD127857] [Drug] [Brand Name]     
#>  5   45707694 [V] [C] 45707694 LAMOTRIGINE - lamotrigine tablet, chewable [SPL …
#>  6    3250587 [V] [N] 3250587 Intubation of intestine for decompression [Nebras…
#>  7    4343148 [V] [N] 4343148 Site of Metabolism [NDFRT N0000000024] [Drug] [Ph…
#>  8   45293012 [D] [N] 45293012 quinapril 20 MG Oral Tablet [Accupril] [NDC 5486…
#>  9   35419606 [V] [S] 35419606 7300 MG troxerutin 0.484 MG/MG Powder for Oral S…
#> 10   36723511 [V] [S] 36723511 APC Protein Variant: Substitution in position 70…
#> 11    2871840 [V] [S] 2871840 Nuclear Medicine @ Male Reproductive System @ Pla…
#> 12    3073980 [D] [N] 3073980 Retired procedure [Nebraska Lexicon 8177006] [Pro…
#> 13   36495696 [V] [N] 36495696 tramadol hydrochloride 50mg/1 ORAL TABLET [NDC 7…
#> 14    2816269 [V] [S] 2816269 Medical and Surgical @ Bursae and Ligaments @ Exc…
#> 15   42856061 [V] [N] 42856061 LEUCOVORIN CALCIUM 50MG/VIL INJ [VA Product N000…
#> 16    4228247 [U] [N] 4228247 Sodium bicarbonate 4.2% injection solution prefil…
#> 17    2768553 [V] [S] 2768553 Fusion of Right Sternoclavicular Joint with Synth…
#> 18   43710853 [V] [S] 43710853 prothipendyl 40 MG Delayed Release Oral Tablet […
#> 19   35173715 [V] [N] 35173715 Etodolac [JMDC 100000075486] [Drug] [Drug Produc…
#> 20   45000704 [V] [N] 45000704 dorzolamide 20 MG/ML / timolol 5 MG/ML Ophthalmi…

# Merge while sparing some original columns such as vocabulary_id and
# concept_class_id
merge_strip(data = test_data,
            into = "concept",
            vocabulary_id,
            concept_class_id)

#> # A tibble: 20 x 4
#>    concept_id concept                           vocabulary_id  concept_class_id 
#>         <dbl> <chr>                             <chr>          <chr>            
#>  1   46187243 [V] [N] 46187243 Levothyroxine s… SNOMED         Pharma/Biol Prod…
#>  2    3436653 [V] [N] 3436653 Swallowing defec… Nebraska Lexi… Procedure        
#>  3    4333553 [V] [S] 4333553 Radiofrequency d… SNOMED         Procedure        
#>  4   44154298 [V] [N] 44154298 Phs Gold [DPD D… DPD            Brand Name       
#>  5   45707694 [V] [C] 45707694 LAMOTRIGINE - l… SPL            Prescription Drug
#>  6    3250587 [V] [N] 3250587 Intubation of in… Nebraska Lexi… Procedure        
#>  7    4343148 [V] [N] 4343148 Site of Metaboli… NDFRT          Pharmacokinetics 
#>  8   45293012 [D] [N] 45293012 quinapril 20 MG… NDC            11-digit NDC     
#>  9   35419606 [V] [S] 35419606 7300 MG troxeru… RxNorm Extens… Quant Branded Box
#> 10   36723511 [V] [S] 36723511 APC Protein Var… OMOP Extension Protein Variant  
#> 11    2871840 [V] [S] 2871840 Nuclear Medicine… ICD10PCS       ICD10PCS Hierarc…
#> 12    3073980 [D] [N] 3073980 Retired procedur… Nebraska Lexi… Procedure        
#> 13   36495696 [V] [N] 36495696 tramadol hydroc… NDC            9-digit NDC      
#> 14    2816269 [V] [S] 2816269 Medical and Surg… ICD10PCS       ICD10PCS Hierarc…
#> 15   42856061 [V] [N] 42856061 LEUCOVORIN CALC… VA Product     VA Product       
#> 16    4228247 [U] [N] 4228247 Sodium bicarbona… SNOMED         Pharma/Biol Prod…
#> 17    2768553 [V] [S] 2768553 Fusion of Right … ICD10PCS       ICD10PCS         
#> 18   43710853 [V] [S] 43710853 prothipendyl 40… RxNorm Extens… Branded Drug     
#> 19   35173715 [V] [N] 35173715 Etodolac [JMDC … JMDC           Drug Product     
#> 20   45000704 [V] [N] 45000704 dorzolamide 20 … NDC            11-digit NDC     

# Test filtering of strip columns
test_output <- merge_strip(data = test_data,
                        into = "concept") %>%
                dplyr::select(-concept_id)

# Filter at a Strip Column
filter_strip(test_output,
             strip_col = concept,
             vocabulary_id %in% c("SNOMED"))

#> # A tibble: 3 x 1
#>   concept                                                                       
#>   <chr>                                                                         
#> 1 [V] [N] 46187243 Levothyroxine sodium 100micrograms/5ml oral suspension (Spec…
#> 2 [V] [S] 4333553 Radiofrequency destruction of trigger point [SNOMED 231400000…
#> 3 [U] [N] 4228247 Sodium bicarbonate 4.2% injection solution prefilled syringe …

# Filter at more than 1 Strip Column
test_data <-
        dplyr::bind_cols(test_output %>%
                                 rubix::randomize() %>%
                                 dplyr::select(concept_a = concept),
                         test_output %>%
                                 rubix::randomize() %>%
                                 dplyr::select(concept_b = concept))

filter_at_all_strip(data = test_data,
                strip_cols = c("concept_a",
                               "concept_b"),
                vocabulary_id %in% c("RxNorm"))

#> Error in matrix(if (is.null(value)) logical() else value, nrow = nr, dimnames = list(rn,     cn)): length of 'dimnames' [2] not equal to array extent

filter_at_any_strip(data = test_data,
                    strip_cols = c("concept_a",
                                   "concept_b"),
                    vocabulary_id %in% c("RxNorm"))

#> # A tibble: 0 x 2
#> # … with 2 variables: concept_a <chr>, concept_b <chr>

dcAthena(conn = conn)

#> [2021-01-09 18:30:53]	Postgres connection to 'athena' closed