Get Unfiltered Test Concepts

Get random rows from the Concept table without any additional filters.

get_test_concepts(
  conn,
  conn_fun = "connectAthena()",
  schema = "omop_vocabulary",
  limit = 100,
  verbose = TRUE,
  render_sql = TRUE
)

See also

Examples

library(chariot)
library(tidyverse)

# Get Unfiltered Test Concepts (Invalid Concepts are included)
get_test_concepts()

#> Connecting using PostgreSQL driver

#> [2021-04-23 19:22:52]	Overriding cache... Querying Athena...
#> [2021-04-23 19:22:52] 

#> ✓ Open connection

#> [2021-04-23 19:22:52] 

#> ✓ JDBC connection

#> [2021-04-23 19:22:52]	SQL: SELECT * FROM omop_vocabulary.concept ORDER BY RANDOM() LIMIT 100
#> [2021-04-23 19:22:52]	Querying...
#> [2021-04-23 19:23:00]	Querying...complete
#> [2021-04-23 19:23:00] 

#> ✓ Returned data has more than 0 rows

#> [2021-04-23 19:23:00]	Caching resultset...
#> [2021-04-23 19:23:00]	Postgres connection to 'athena' closed

#> # A tibble: 100 x 10
#>    concept_id concept_name             domain_id  vocabulary_id concept_class_id
#>         <int> <chr>                    <chr>      <chr>         <chr>           
#>  1   42937436 Fluconazole Oral Capsul… Drug       RxNorm Exten… Branded Drug Fo…
#>  2   46258232 Juzo Hostess class 2 (2… Device     SNOMED        Physical Object 
#>  3   40303997 O/E - perceptive deaf (… Condition  SNOMED        Clinical Finding
#>  4   36729262 CDH1 Transcript Variant… Measureme… OMOP Extensi… Transcript Vari…
#>  5   46271748 Fluoroscopy guided aspi… Procedure  SNOMED        Procedure       
#>  6    4318682 Choroid tubercles        Condition  SNOMED        Clinical Finding
#>  7   40710439 200 ACTUAT Fenoterol 0.… Drug       RxNorm Exten… Quant Clinical …
#>  8   41467234 3.13 ML Somatropin 4 MG… Drug       RxNorm Exten… Marketed Product
#>  9   37104399 Juzo Soft class 2 (23-3… Device     SNOMED        Physical Object 
#> 10   42061177 Bennett                  Geography  OSM           8th level       
#> # … with 90 more rows, and 5 more variables: standard_concept <chr>,
#> #   concept_code <chr>, valid_start_date <date>, valid_end_date <date>,
#> #   invalid_reason <chr>

# The size of the resultset can be altered
get_test_concepts(limit = 25)

#> Connecting using PostgreSQL driver

#> [2021-04-23 19:23:00]	Overriding cache... Querying Athena...
#> [2021-04-23 19:23:00] 

#> ✓ Open connection

#> [2021-04-23 19:23:00] 

#> ✓ JDBC connection

#> [2021-04-23 19:23:00]	SQL: SELECT * FROM omop_vocabulary.concept ORDER BY RANDOM() LIMIT 25
#> [2021-04-23 19:23:00]	Querying...
#> [2021-04-23 19:23:09]	Querying...complete
#> [2021-04-23 19:23:09] 

#> ✓ Returned data has more than 0 rows

#> [2021-04-23 19:23:09]	Caching resultset...
#> [2021-04-23 19:23:09]	Postgres connection to 'athena' closed

#> # A tibble: 25 x 10
#>    concept_id concept_name             domain_id  vocabulary_id concept_class_id
#>         <int> <chr>                    <chr>      <chr>         <chr>           
#>  1    1210698 zinc oxide 20g/100g TOP… Drug       SPL           OTC Drug        
#>  2   40781092 Pellitory (Parietaria o… Measureme… LOINC         LOINC Hierarchy 
#>  3   41923231 Donggou                  Geography  OSM           8th level       
#>  4   46011666 Griseofulvin 105mg/5ml … Drug       SNOMED        Pharma/Biol Pro…
#>  5   41961273 Saint-Sylvestre          Geography  OSM           8th level       
#>  6   42058306 Town of Anson            Geography  OSM           7th level       
#>  7   43636102 Hydrochlorothiazide 12.… Drug       RxNorm Exten… Branded Drug    
#>  8   43766642 gabapentin 100 MG Oral … Drug       RxNorm Exten… Branded Drug    
#>  9   37495255 clonazepam .5mg/1 ORAL … Drug       NDC           9-digit NDC     
#> 10    3093017 O/E - muscle power NOS   Condition  Nebraska Lex… Clinical Finding
#> # … with 15 more rows, and 5 more variables: standard_concept <chr>,
#> #   concept_code <chr>, valid_start_date <date>, valid_end_date <date>,
#> #   invalid_reason <chr>

# Concepts specific to the Drug domain can be sampled
# RxNorm/RxNorm Extension Concepts
get_test_drug_concepts(limit = 10)

#> Connecting using PostgreSQL driver

#> [2021-04-23 19:23:09]	Overriding cache... Querying Athena...
#> [2021-04-23 19:23:09] 

#> ✓ Open connection

#> [2021-04-23 19:23:09] 

#> ✓ JDBC connection

#> [2021-04-23 19:23:09]	SQL: SELECT * FROM omop_vocabulary.concept c WHERE c.invalid_reason IS NULL AND c.vocabulary_id IN ('RxNorm', 'RxNorm Extension') AND c.standard_concept <> 'C' AND c.domain_id = 'Drug' ORDER BY RANDOM() LIMIT 10
#> [2021-04-23 19:23:09]	Querying...
#> [2021-04-23 19:23:12]	Querying...complete
#> [2021-04-23 19:23:12] 

#> ✓ Returned data has more than 0 rows

#> [2021-04-23 19:23:12]	Caching resultset...
#> [2021-04-23 19:23:12]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 10
#>    concept_id concept_name             domain_id vocabulary_id  concept_class_id
#>         <int> <chr>                    <chr>     <chr>          <chr>           
#>  1    2052391 amoxicillin 40 MG / cla… Drug      RxNorm Extens… Branded Drug    
#>  2   36261678 Docusate 20.8 MG/ML / G… Drug      RxNorm Extens… Branded Drug Co…
#>  3   41171197 Zinc Gluconate 3 MG [Zi… Drug      RxNorm Extens… Branded Drug Co…
#>  4   40915632 Morphine 100 MG Oral Ta… Drug      RxNorm Extens… Marketed Product
#>  5   40750289 pyridoxine Oral Tablet … Drug      RxNorm Extens… Branded Drug Fo…
#>  6   36249514 0.5 ML influenza A viru… Drug      RxNorm         Quant Clinical …
#>  7   41351783 5 ML factor IX complex … Drug      RxNorm Extens… Quant Clinical …
#>  8   43744478 Oxytetracycline / predn… Drug      RxNorm Extens… Clinical Drug F…
#>  9   44164097 Ceftriaxone 1000 MG Inj… Drug      RxNorm Extens… Marketed Product
#> 10   21156532 3 ML Timolol 5 MG/ML / … Drug      RxNorm Extens… Marketed Product
#> # … with 5 more variables: standard_concept <chr>, concept_code <chr>,
#> #   valid_start_date <date>, valid_end_date <date>, invalid_reason <chr>

# HemOnc/ATC Classes
get_test_drug_classes(limit = 10)

#> Connecting using PostgreSQL driver

#> [2021-04-23 19:23:12]	Overriding cache... Querying Athena...
#> [2021-04-23 19:23:12] 

#> ✓ Open connection

#> [2021-04-23 19:23:12] 

#> ✓ JDBC connection

#> [2021-04-23 19:23:12]	SQL: SELECT * FROM omop_vocabulary.concept c WHERE c.invalid_reason IS NULL AND c.vocabulary_id IN ('ATC', 'HemOnc') AND c.standard_concept = 'C' AND c.domain_id = 'Drug' ORDER BY RANDOM() LIMIT 10
#> [2021-04-23 19:23:12]	Querying...
#> [2021-04-23 19:23:12]	Querying...complete
#> [2021-04-23 19:23:12] 

#> ✓ Returned data has more than 0 rows

#> [2021-04-23 19:23:12]	Caching resultset...
#> [2021-04-23 19:23:12]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 10
#>    concept_id concept_name              domain_id vocabulary_id concept_class_id
#>         <int> <chr>                     <chr>     <chr>         <chr>           
#>  1   21603961 piroxicam; systemic, rec… Drug      ATC           ATC 5th         
#>  2   21604523 droperidol; parenteral    Drug      ATC           ATC 5th         
#>  3   21601532 POTASSIUM-SPARING AGENTS  Drug      ATC           ATC 3rd         
#>  4   21600282 isoprenaline; systemic    Drug      ATC           ATC 5th         
#>  5    1123745 arterolane and piperaqui… Drug      ATC           ATC 5th         
#>  6   21600354 linsidomine               Drug      ATC           ATC 5th         
#>  7   21603642 flurbiprofen; ophthalmic  Drug      ATC           ATC 5th         
#>  8   21602109 fluperolone; topical      Drug      ATC           ATC 5th         
#>  9   21603193 enfuvirtide; parenteral   Drug      ATC           ATC 5th         
#> 10   21603315 hexoprenaline; systemic   Drug      ATC           ATC 5th         
#> # … with 5 more variables: standard_concept <chr>, concept_code <chr>,
#> #   valid_start_date <date>, valid_end_date <date>, invalid_reason <chr>

# Test data can also be taken from tables other than Concept
get_test_data(table = "concept_relationship")

#> Connecting using PostgreSQL driver

#> [2021-04-23 19:23:12]	Overriding cache... Querying Athena...
#> [2021-04-23 19:23:12] 

#> ✓ Open connection

#> [2021-04-23 19:23:12] 

#> ✓ JDBC connection

#> [2021-04-23 19:23:12]	SQL: SELECT * FROM omop_vocabulary.concept_relationship ORDER BY RANDOM() LIMIT 100
#> [2021-04-23 19:23:12]	Querying...
#> [2021-04-23 19:23:47]	Querying...complete
#> [2021-04-23 19:23:47] 

#> ✓ Returned data has more than 0 rows

#> [2021-04-23 19:23:47]	Caching resultset...
#> [2021-04-23 19:23:47]	Postgres connection to 'athena' closed

#> # A tibble: 100 x 6
#>    concept_id_1 concept_id_2 relationship_id     valid_start_date valid_end_date
#>           <int>        <int> <chr>               <date>           <date>        
#>  1     19001949     35763204 RxNorm dose form of 2017-08-02       2099-12-31    
#>  2     35607932     37004438 Non-avail ind of    2018-07-31       2099-12-31    
#>  3     19082573     21141048 RxNorm dose form of 2015-08-17       2099-12-31    
#>  4     44175044     40845354 Marketed form of    2017-10-06       2099-12-31    
#>  5      4334142     42853255 Physiol effect by   2012-10-01       2099-12-31    
#>  6      2735345     46271049 Is a                1970-01-01       2099-12-31    
#>  7     46122634     45989413 Has AMP             2015-01-31       2099-12-31    
#>  8     37020717     37019612 Contained in panel  1970-01-01       2099-12-31    
#>  9       968478     21146735 Has marketed form   2017-06-24       2099-12-31    
#> 10      4293448     46149845 Disp dose form of   2018-07-31       2099-12-31    
#> # … with 90 more rows, and 1 more variable: invalid_reason <chr>

get_test_data(table = "concept_ancestor", limit = 10)

#> Connecting using PostgreSQL driver

#> [2021-04-23 19:23:47]	Overriding cache... Querying Athena...
#> [2021-04-23 19:23:47] 

#> ✓ Open connection

#> [2021-04-23 19:23:47] 

#> ✓ JDBC connection

#> [2021-04-23 19:23:47]	SQL: SELECT * FROM omop_vocabulary.concept_ancestor ORDER BY RANDOM() LIMIT 10
#> [2021-04-23 19:23:47]	Querying...
#> [2021-04-23 19:24:21]	Querying...complete
#> [2021-04-23 19:24:21] 

#> ✓ Returned data has more than 0 rows

#> [2021-04-23 19:24:21]	Caching resultset...
#> [2021-04-23 19:24:21]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 4
#>    ancestor_concept_… descendant_concep… min_levels_of_sepa… max_levels_of_sepa…
#>                 <int>              <int>               <int>               <int>
#>  1           36221206           41364703                   2                   3
#>  2           36982372           43259332                   1                   1
#>  3           45688038           36261254                   2                   3
#>  4           37020070           36305659                   2                   2
#>  5           41368080           41367924                   1                   2
#>  6           42801550           41124307                   2                   2
#>  7           46302666           40904195                   2                   2
#>  8           42800011           42800011                   0                   0
#>  9           35807271           43516590                   6                   7
#> 10           45703055           21131338                   2                   3