Get Unfiltered Test Concepts

Get random rows from the Concept table without any additional filters.

get_test_concepts(
  conn,
  conn_fun = "connectAthena()",
  schema = "omop_vocabulary",
  limit = 100,
  verbose = TRUE,
  render_sql = TRUE
)

See also

Examples

library(chariot)
library(tidyverse)

# Get Unfiltered Test Concepts (Invalid Concepts are included)
get_test_concepts()

#> Connecting using PostgreSQL driver

#> [2021-03-28 23:38:52]	Overriding cache... Querying Athena...
#> [2021-03-28 23:38:52] 

#> ✓ Open connection

#> [2021-03-28 23:38:52] 

#> ✓ JDBC connection

#> [2021-03-28 23:38:52]	SQL: SELECT * FROM omop_vocabulary.concept ORDER BY RANDOM() LIMIT 100
#> [2021-03-28 23:38:52]	Querying...
#> [2021-03-28 23:39:05]	Querying...complete
#> [2021-03-28 23:39:05] 

#> ✓ Returned data has more than 0 rows

#> [2021-03-28 23:39:05]	Caching resultset...
#> [2021-03-28 23:39:05]	Postgres connection to 'athena' closed

#> # A tibble: 100 x 10
#>    concept_id concept_name              domain_id vocabulary_id concept_class_id
#>         <int> <chr>                     <chr>     <chr>         <chr>           
#>  1   40944937 Tramadol 150 MG Oral Tab… Drug      RxNorm Exten… Marketed Product
#>  2   43707174 Acetic Acid 0.009 MG/MG … Drug      RxNorm Exten… Clinical Drug B…
#>  3   44153346 Cervarix                  Drug      DPD           Brand Name      
#>  4   43370368 Methadone Hydrochloride … Drug      AMT           Med Product Unit
#>  5   21214562 Cerumol ear drops (Lexon… Drug      dm+d          AMPP            
#>  6   21421605 Sigvaris MTM Style Semit… Device    dm+d          AMP             
#>  7    2775171 Bypass Right Kidney Pelv… Procedure ICD10PCS      ICD10PCS        
#>  8   41238726 Ciprofloxacin Delayed Re… Drug      RxNorm Exten… Branded Drug Fo…
#>  9   35879247 Jianlangchen              Drug      NCCD          Brand Name      
#> 10   36990975 ethyl alcohol 70mL/100mL… Drug      NDC           9-digit NDC     
#> # … with 90 more rows, and 5 more variables: standard_concept <chr>,
#> #   concept_code <chr>, valid_start_date <date>, valid_end_date <date>,
#> #   invalid_reason <chr>

# The size of the resultset can be altered
get_test_concepts(limit = 25)

#> Connecting using PostgreSQL driver

#> [2021-03-28 23:39:05]	Overriding cache... Querying Athena...
#> [2021-03-28 23:39:05] 

#> ✓ Open connection

#> [2021-03-28 23:39:05] 

#> ✓ JDBC connection

#> [2021-03-28 23:39:05]	SQL: SELECT * FROM omop_vocabulary.concept ORDER BY RANDOM() LIMIT 25
#> [2021-03-28 23:39:05]	Querying...
#> [2021-03-28 23:39:12]	Querying...complete
#> [2021-03-28 23:39:12] 

#> ✓ Returned data has more than 0 rows

#> [2021-03-28 23:39:12]	Caching resultset...
#> [2021-03-28 23:39:12]	Postgres connection to 'athena' closed

#> # A tibble: 25 x 10
#>    concept_id concept_name              domain_id vocabulary_id concept_class_id
#>         <int> <chr>                     <chr>     <chr>         <chr>           
#>  1   45676619 LEVOCARNITINE - levocarn… Drug      SPL           Prescription Dr…
#>  2    4222292 Arteriectomy with graft … Procedure SNOMED        Procedure       
#>  3   43610087 10 ML Budesonide 1.28 MG… Drug      RxNorm Exten… Quant Branded D…
#>  4   44652867 Daytime Cold/Flu Medicin… Drug      DPD           Drug Product    
#>  5   41389417 30 ML Cromoglycate 20 MG… Drug      RxNorm Exten… Quant Branded D…
#>  6    3148212 Injury of elbow, forearm… Condition Nebraska Lex… Clinical Finding
#>  7   21227896 Ventolin 500micrograms/1… Drug      dm+d          AMPP            
#>  8   21074190 Cetrimide / Lidocaine / … Drug      RxNorm Exten… Branded Drug Fo…
#>  9   21388380 Juzo Soft class 1 (18-21… Device    dm+d          AMPP            
#> 10   45221780 meloxicam 7.5 MG Oral Ta… Drug      NDC           11-digit NDC    
#> # … with 15 more rows, and 5 more variables: standard_concept <chr>,
#> #   concept_code <chr>, valid_start_date <date>, valid_end_date <date>,
#> #   invalid_reason <chr>

# Concepts specific to the Drug domain can be sampled
# RxNorm/RxNorm Extension Concepts
get_test_drug_concepts(limit = 10)

#> Connecting using PostgreSQL driver

#> [2021-03-28 23:39:13]	Overriding cache... Querying Athena...
#> [2021-03-28 23:39:13] 

#> ✓ Open connection

#> [2021-03-28 23:39:13] 

#> ✓ JDBC connection

#> [2021-03-28 23:39:13]	SQL: SELECT * FROM omop_vocabulary.concept c WHERE c.invalid_reason IS NULL AND c.vocabulary_id IN ('RxNorm', 'RxNorm Extension') AND c.standard_concept <> 'C' AND c.domain_id = 'Drug' ORDER BY RANDOM() LIMIT 10
#> [2021-03-28 23:39:13]	Querying...
#> [2021-03-28 23:39:16]	Querying...complete
#> [2021-03-28 23:39:16] 

#> ✓ Returned data has more than 0 rows

#> [2021-03-28 23:39:16]	Caching resultset...
#> [2021-03-28 23:39:16]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 10
#>    concept_id concept_name              domain_id vocabulary_id concept_class_id
#>         <int> <chr>                     <chr>     <chr>         <chr>           
#>  1   40979307 Glyburide 3.5 MG Oral Ta… Drug      RxNorm Exten… Marketed Product
#>  2   21045460 iodixanol 550 MG/ML Pref… Drug      RxNorm Exten… Branded Drug Box
#>  3   35797657 101000 MG Urea 0.1 MG/MG… Drug      RxNorm Exten… Marketed Product
#>  4   36278750 4 ML Cefazolin 250 MG/ML… Drug      RxNorm Exten… Quant Clinical …
#>  5   41430219 150 ML Erythromycin 120 … Drug      RxNorm Exten… Quant Branded D…
#>  6   41338780 2 ML Diazepam 5 MG/ML In… Drug      RxNorm Exten… Marketed Product
#>  7   40240989 melatonin / vitamin B12 … Drug      RxNorm        Clinical Drug F…
#>  8   41295553 Xipamide 20 MG [Xipamid … Drug      RxNorm Exten… Branded Drug Co…
#>  9   41183155 tizanidine 2 MG Oral Tab… Drug      RxNorm Exten… Branded Drug Box
#> 10   40171588 allscale pollen extract … Drug      RxNorm        Clinical Drug   
#> # … with 5 more variables: standard_concept <chr>, concept_code <chr>,
#> #   valid_start_date <date>, valid_end_date <date>, invalid_reason <chr>

# HemOnc/ATC Classes
get_test_drug_classes(limit = 10)

#> Connecting using PostgreSQL driver

#> [2021-03-28 23:39:16]	Overriding cache... Querying Athena...
#> [2021-03-28 23:39:16] 

#> ✓ Open connection

#> [2021-03-28 23:39:16] 

#> ✓ JDBC connection

#> [2021-03-28 23:39:16]	SQL: SELECT * FROM omop_vocabulary.concept c WHERE c.invalid_reason IS NULL AND c.vocabulary_id IN ('ATC', 'HemOnc') AND c.standard_concept = 'C' AND c.domain_id = 'Drug' ORDER BY RANDOM() LIMIT 10
#> [2021-03-28 23:39:16]	Querying...
#> [2021-03-28 23:39:16]	Querying...complete
#> [2021-03-28 23:39:16] 

#> ✓ Returned data has more than 0 rows

#> [2021-03-28 23:39:16]	Caching resultset...
#> [2021-03-28 23:39:16]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 10
#>    concept_id concept_name              domain_id vocabulary_id concept_class_id
#>         <int> <chr>                     <chr>     <chr>         <chr>           
#>  1   21602191 desoximetasone, combinat… Drug      ATC           ATC 5th         
#>  2   21600094 enprostil; oral           Drug      ATC           ATC 5th         
#>  3   21601153 I.V. SOLUTIONS            Drug      ATC           ATC 3rd         
#>  4   45893191 cefotaxime and beta-lact… Drug      ATC           ATC 5th         
#>  5   21603643 ketorolac; ophthalmic     Drug      ATC           ATC 5th         
#>  6   21601332 encephalitis, Japanese, … Drug      ATC           ATC 5th         
#>  7   21604315 diflunisal; oral          Drug      ATC           ATC 5th         
#>  8   21603058 nitrofurantoin; oral      Drug      ATC           ATC 5th         
#>  9   21605498 technetium (99mTc) micro… Drug      ATC           ATC 5th         
#> 10    1123680 tiazotic acid             Drug      ATC           ATC 5th         
#> # … with 5 more variables: standard_concept <chr>, concept_code <chr>,
#> #   valid_start_date <date>, valid_end_date <date>, invalid_reason <chr>

# Test data can also be taken from tables other than Concept
get_test_data(table = "concept_relationship")

#> Connecting using PostgreSQL driver

#> [2021-03-28 23:39:17]	Overriding cache... Querying Athena...
#> [2021-03-28 23:39:17] 

#> ✓ Open connection

#> [2021-03-28 23:39:17] 

#> ✓ JDBC connection

#> [2021-03-28 23:39:17]	SQL: SELECT * FROM omop_vocabulary.concept_relationship ORDER BY RANDOM() LIMIT 100
#> [2021-03-28 23:39:17]	Querying...
#> [2021-03-28 23:39:52]	Querying...complete
#> [2021-03-28 23:39:52] 

#> ✓ Returned data has more than 0 rows

#> [2021-03-28 23:39:52]	Caching resultset...
#> [2021-03-28 23:39:52]	Postgres connection to 'athena' closed

#> # A tibble: 100 x 6
#>    concept_id_1 concept_id_2 relationship_id    valid_start_date valid_end_date
#>           <int>        <int> <chr>              <date>           <date>        
#>  1     45713201     40168575 Maps to            1970-01-01       2099-12-31    
#>  2     46288305     19068999 SPL - RxNorm       1970-01-01       2099-12-31    
#>  3     41284718     41118087 Tradename of       2017-07-18       2099-12-31    
#>  4     40812225     19005046 Brand name of      2017-07-18       2099-12-31    
#>  5     37110758     42534443 Numerator value of 2018-07-31       2099-12-31    
#>  6     46018009     37209200 Is a               2019-07-31       2099-12-31    
#>  7      2785255      4030747 Is a               1970-01-01       2099-12-31    
#>  8      4305880     45971988 Active ing of      2018-07-31       2099-12-31    
#>  9     19098386     19098386 Mapped from        2005-09-11       2099-12-31    
#> 10       444074      1432628 Mapped from        2016-01-01       2099-12-31    
#> # … with 90 more rows, and 1 more variable: invalid_reason <chr>

get_test_data(table = "concept_ancestor", limit = 10)

#> Connecting using PostgreSQL driver

#> [2021-03-28 23:39:52]	Overriding cache... Querying Athena...
#> [2021-03-28 23:39:52] 

#> ✓ Open connection

#> [2021-03-28 23:39:52] 

#> ✓ JDBC connection

#> [2021-03-28 23:39:52]	SQL: SELECT * FROM omop_vocabulary.concept_ancestor ORDER BY RANDOM() LIMIT 10
#> [2021-03-28 23:39:52]	Querying...
#> [2021-03-28 23:40:26]	Querying...complete
#> [2021-03-28 23:40:26] 

#> ✓ Returned data has more than 0 rows

#> [2021-03-28 23:40:26]	Caching resultset...
#> [2021-03-28 23:40:26]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 4
#>    ancestor_concept_… descendant_concep… min_levels_of_sepa… max_levels_of_sepa…
#>                 <int>              <int>               <int>               <int>
#>  1           36213917           41521752                   3                   3
#>  2           45688295           41040006                   2                   3
#>  3           46302176           41312856                   2                   2
#>  4           19093618           40228855                   0                   0
#>  5           21604364           40231839                   4                   5
#>  6             197917            4187364                   2                   2
#>  7            4201745           36555037                   6                  12
#>  8            1137529           41113245                   1                   2
#>  9           43724842           43724842                   0                   0
#> 10           36217216           21031330                   4                   5