Get Unfiltered Test Concepts

Get random rows from the Concept table without any additional filters.

get_test_concepts(
  conn,
  conn_fun = "connectAthena()",
  schema = "omop_vocabulary",
  limit = 100,
  verbose = TRUE,
  render_sql = TRUE
)

See also

Examples

library(chariot)
library(tidyverse)

# Get Unfiltered Test Concepts (Invalid Concepts are included)
get_test_concepts()

#> [2021-05-26 15:04:23]	Connected to 'athena'
#> [2021-05-26 15:04:23]	Overriding cache... Querying Athena...
#> [2021-05-26 15:04:23]	✓ Open connection
#> [2021-05-26 15:04:23]	✓ JDBC connection
#> [2021-05-26 15:04:23]	SQL: SELECT * FROM omop_vocabulary.concept ORDER BY RANDOM() LIMIT 100
#> [2021-05-26 15:04:23]	Querying...
#> [2021-05-26 15:04:31]	Querying...complete
#> [2021-05-26 15:04:31]	✓ Returned data has more than 0 rows
#> [2021-05-26 15:04:31]	Caching resultset...
#> [2021-05-26 15:04:31]	Postgres connection to 'athena' closed

#> # A tibble: 100 x 10
#>    concept_id concept_name             domain_id  vocabulary_id concept_class_id
#>         <int> <chr>                    <chr>      <chr>         <chr>           
#>  1   40838877 zolpidem 10 MG Delayed … Drug       RxNorm Exten… Branded Drug Box
#>  2   37515434 ethyl alcohol 21mL/30mL… Drug       SPL           OTC Drug        
#>  3   21400828 Jobst Opaque class 2 (2… Device     dm+d          AMP             
#>  4    1157365 Guaifenesin 40 MG/ML / … Drug       RxNorm        Branded Drug    
#>  5   43587412 Diclofenac 0.0116 MG/MG… Drug       RxNorm Exten… Branded Drug Box
#>  6   41362877 10 ML Oxytocin 3 UNT/ML… Drug       RxNorm Exten… Quant Clinical …
#>  7    3460724 Laboratory heating block Device     Nebraska Lex… Physical Object 
#>  8    3039871 Urinalysis type of non-… Measureme… LOINC         Lab Test        
#>  9   42606969 BARRIER,OSTOMY,NEW IMAG… Drug       VA Product    VA Product      
#> 10   21328301 Chailey adult pubic pre… Device     dm+d          AMP             
#> # … with 90 more rows, and 5 more variables: standard_concept <chr>,
#> #   concept_code <chr>, valid_start_date <date>, valid_end_date <date>,
#> #   invalid_reason <chr>

# The size of the resultset can be altered
get_test_concepts(limit = 25)

#> [2021-05-26 15:04:31]	Connected to 'athena'
#> [2021-05-26 15:04:31]	Overriding cache... Querying Athena...
#> [2021-05-26 15:04:31]	✓ Open connection
#> [2021-05-26 15:04:31]	✓ JDBC connection
#> [2021-05-26 15:04:31]	SQL: SELECT * FROM omop_vocabulary.concept ORDER BY RANDOM() LIMIT 25
#> [2021-05-26 15:04:31]	Querying...
#> [2021-05-26 15:04:38]	Querying...complete
#> [2021-05-26 15:04:38]	✓ Returned data has more than 0 rows
#> [2021-05-26 15:04:38]	Caching resultset...
#> [2021-05-26 15:04:38]	Postgres connection to 'athena' closed

#> # A tibble: 25 x 10
#>    concept_id concept_name             domain_id vocabulary_id  concept_class_id
#>         <int> <chr>                    <chr>     <chr>          <chr>           
#>  1   43790852 30 ML Midodrine 10 MG/M… Drug      RxNorm Extens… Quant Clinical …
#>  2   36292278 Amikacine B. Braun inf.… Drug      GGR            Drug Product    
#>  3   45511970 Faeces: mucous present   Condition Read           Read            
#>  4   21467308 Jobst Opaque class 2 (2… Device    dm+d           AMPP            
#>  5    2784359 Removal of Radioactive … Procedure ICD10PCS       ICD10PCS        
#>  6   44115720 Activated Charcoal 225 … Drug      RxNorm Extens… Marketed Product
#>  7   44672075 TIN ORAL DROPS [STANNUM… Drug      DPD            Branded Drug Fo…
#>  8   44631347 CALCIUM CARBONATE / FER… Drug      DPD            Branded Drug Fo…
#>  9   43773475 50 ML Nitroglycerin 1 M… Drug      RxNorm Extens… Marketed Product
#> 10   37206314 Mezlocillin (as mezloci… Drug      SNOMED         Clinical Drug   
#> # … with 15 more rows, and 5 more variables: standard_concept <chr>,
#> #   concept_code <chr>, valid_start_date <date>, valid_end_date <date>,
#> #   invalid_reason <chr>

# Concepts specific to the Drug domain can be sampled
# RxNorm/RxNorm Extension Concepts
get_test_drug_concepts(limit = 10)

#> [2021-05-26 15:04:39]	Connected to 'athena'
#> [2021-05-26 15:04:39]	Overriding cache... Querying Athena...
#> [2021-05-26 15:04:39]	✓ Open connection
#> [2021-05-26 15:04:39]	✓ JDBC connection
#> [2021-05-26 15:04:39]	SQL: SELECT * FROM omop_vocabulary.concept c WHERE c.invalid_reason IS NULL AND c.vocabulary_id IN ('RxNorm', 'RxNorm Extension') AND c.standard_concept <> 'C' AND c.domain_id = 'Drug' ORDER BY RANDOM() LIMIT 10
#> [2021-05-26 15:04:39]	Querying...
#> [2021-05-26 15:04:42]	Querying...complete
#> [2021-05-26 15:04:42]	✓ Returned data has more than 0 rows
#> [2021-05-26 15:04:42]	Caching resultset...
#> [2021-05-26 15:04:42]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 10
#>    concept_id concept_name              domain_id vocabulary_id concept_class_id
#>         <int> <chr>                     <chr>     <chr>         <chr>           
#>  1   35765081 Cimetidine 400 MG Delaye… Drug      RxNorm Exten… Marketed Product
#>  2   21129505 Secretin 0.016 MG Inject… Drug      RxNorm Exten… Clinical Drug   
#>  3   21115507 Nicotine 2 MG Chewing Gu… Drug      RxNorm Exten… Marketed Product
#>  4   36896926 500 ML Glucose 380 MG/ML… Drug      RxNorm Exten… Quant Clinical …
#>  5   41203663 latanoprost 0.05 MG/ML [… Drug      RxNorm Exten… Branded Drug Co…
#>  6   43144635 Ticarcillin 5000 MG [TIC… Drug      RxNorm Exten… Branded Drug Co…
#>  7   43691009 ammonium carbonate / Bar… Drug      RxNorm Exten… Clinical Drug F…
#>  8   42902653 hepatitis A vaccine (ina… Drug      RxNorm        Clinical Drug   
#>  9   43196036 Nicotine 23.6 MG Transde… Drug      RxNorm Exten… Clinical Drug   
#> 10   21069846 Hypochlorite Topical Sol… Drug      RxNorm Exten… Branded Drug Fo…
#> # … with 5 more variables: standard_concept <chr>, concept_code <chr>,
#> #   valid_start_date <date>, valid_end_date <date>, invalid_reason <chr>

# HemOnc/ATC Classes
get_test_drug_classes(limit = 10)

#> [2021-05-26 15:04:42]	Connected to 'athena'
#> [2021-05-26 15:04:42]	Overriding cache... Querying Athena...
#> [2021-05-26 15:04:42]	✓ Open connection
#> [2021-05-26 15:04:42]	✓ JDBC connection
#> [2021-05-26 15:04:42]	SQL: SELECT * FROM omop_vocabulary.concept c WHERE c.invalid_reason IS NULL AND c.vocabulary_id IN ('ATC', 'HemOnc') AND c.standard_concept = 'C' AND c.domain_id = 'Drug' ORDER BY RANDOM() LIMIT 10
#> [2021-05-26 15:04:42]	Querying...
#> [2021-05-26 15:04:42]	Querying...complete
#> [2021-05-26 15:04:42]	✓ Returned data has more than 0 rows
#> [2021-05-26 15:04:42]	Caching resultset...
#> [2021-05-26 15:04:42]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 10
#>    concept_id concept_name              domain_id vocabulary_id concept_class_id
#>         <int> <chr>                     <chr>     <chr>         <chr>           
#>  1   21600487 bromopride; systemic      Drug      ATC           ATC 5th         
#>  2    1588699 perindopril and bisoprol… Drug      ATC           ATC 5th         
#>  3   21603333 aminophylline; systemic,… Drug      ATC           ATC 5th         
#>  4   21602670 terazosin; oral           Drug      ATC           ATC 5th         
#>  5   21600933 OTHER ALIMENTARY TRACT A… Drug      ATC           ATC 3rd         
#>  6   21600689 clobenzorex               Drug      ATC           ATC 5th         
#>  7   21604833 betahistine; oral         Drug      ATC           ATC 5th         
#>  8   21602615 cyproterone; systemic     Drug      ATC           ATC 5th         
#>  9   35807495 Bowel management          Drug      HemOnc        Component Class 
#> 10    1588672 idarucizumab              Drug      ATC           ATC 5th         
#> # … with 5 more variables: standard_concept <chr>, concept_code <chr>,
#> #   valid_start_date <date>, valid_end_date <date>, invalid_reason <chr>

# Test data can also be taken from tables other than Concept
get_test_data(table = "concept_relationship")

#> [2021-05-26 15:04:42]	Connected to 'athena'
#> [2021-05-26 15:04:42]	Overriding cache... Querying Athena...
#> [2021-05-26 15:04:42]	✓ Open connection
#> [2021-05-26 15:04:42]	✓ JDBC connection
#> [2021-05-26 15:04:42]	SQL: SELECT * FROM omop_vocabulary.concept_relationship ORDER BY RANDOM() LIMIT 100
#> [2021-05-26 15:04:42]	Querying...
#> [2021-05-26 15:05:13]	Querying...complete
#> [2021-05-26 15:05:13]	✓ Returned data has more than 0 rows
#> [2021-05-26 15:05:13]	Caching resultset...
#> [2021-05-26 15:05:13]	Postgres connection to 'athena' closed

#> # A tibble: 100 x 6
#>    concept_id_1 concept_id_2 relationship_id     valid_start_date valid_end_date
#>           <int>        <int> <chr>               <date>           <date>        
#>  1     37207378      4204887 Form continuity of  2019-07-31       2099-12-31    
#>  2     42932579      1332525 Tradename of        2018-08-21       2099-12-31    
#>  3     43750456     43750456 Mapped from         2016-12-18       2099-12-31    
#>  4      2866110      2847180 Is a                2014-04-22       2099-12-31    
#>  5     40870868     40870868 Maps to             2017-07-18       2099-12-31    
#>  6     19082573     41283000 RxNorm dose form of 2017-07-18       2099-12-31    
#>  7     44795687      4023310 Concept was_a to    2019-01-31       2099-12-31    
#>  8     40910639     41377017 Has marketed form   2017-07-18       2099-12-31    
#>  9      4208390     45619665 Mapped from         2013-10-07       2099-12-31    
#> 10      4293448     46010397 Disp dose form of   2015-01-31       2099-12-31    
#> # … with 90 more rows, and 1 more variable: invalid_reason <chr>

get_test_data(table = "concept_ancestor", limit = 10)

#> [2021-05-26 15:05:14]	Connected to 'athena'
#> [2021-05-26 15:05:14]	Overriding cache... Querying Athena...
#> [2021-05-26 15:05:14]	✓ Open connection
#> [2021-05-26 15:05:14]	✓ JDBC connection
#> [2021-05-26 15:05:14]	SQL: SELECT * FROM omop_vocabulary.concept_ancestor ORDER BY RANDOM() LIMIT 10
#> [2021-05-26 15:05:14]	Querying...
#> [2021-05-26 15:05:47]	Querying...complete
#> [2021-05-26 15:05:47]	✓ Returned data has more than 0 rows
#> [2021-05-26 15:05:47]	Caching resultset...
#> [2021-05-26 15:05:47]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 4
#>    ancestor_concept_… descendant_concep… min_levels_of_sepa… max_levels_of_sepa…
#>                 <int>              <int>               <int>               <int>
#>  1           21600959           43722175                   7                   9
#>  2           45705581           43268489                   2                   2
#>  3           45894852           41309142                   2                   2
#>  4           46305158           41227122                   2                   3
#>  5           21601689           36274039                   3                   4
#>  6             817037           40740795                   2                   3
#>  7           36217214           43818200                   3                   4
#>  8           19007572           41443051                   2                   2
#>  9           45702462           41278624                   2                   2
#> 10           45643345           21128257                   2                   3