Get Unfiltered Test Concepts

Get random rows from the Concept table without any additional filters.

get_test_concepts(
  conn,
  conn_fun = "connectAthena()",
  schema = "omop_vocabulary",
  limit = 100,
  verbose = TRUE,
  render_sql = TRUE
)

See also

Examples

library(chariot)
library(tidyverse)

# Get Unfiltered Test Concepts (Invalid Concepts are included)
get_test_concepts()

#> Connecting using PostgreSQL driver
#> [2021-01-17 00:54:22]	Overriding cache... Querying Athena...
#> [2021-01-17 00:54:22] 

#> ✓ Open connection

#> [2021-01-17 00:54:22] 

#> ✓ JDBC connection

#> [2021-01-17 00:54:22]	SQL: SELECT * FROM omop_vocabulary.concept ORDER BY RANDOM() LIMIT 100
#> [2021-01-17 00:54:22]	Querying...
#> [2021-01-17 00:54:31]	Querying...complete
#> [2021-01-17 00:54:31] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-17 00:54:31]	Caching resultset...
#> [2021-01-17 00:54:31]	Postgres connection to 'athena' closed

#> # A tibble: 100 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id
#>         <dbl> <chr>        <chr>     <chr>         <chr>           
#>  1   41067630 Bile Salts … Drug      RxNorm Exten… Branded Drug Box
#>  2   37051879 Lymphocytes… Measurem… LOINC         LOINC Hierarchy 
#>  3   44667713 Dental Gel … Drug      DPD           Drug Product    
#>  4   46343714 LBEL COULEU… Drug      NDC           11-digit NDC    
#>  5   36244675 vardenafil … Drug      RxNorm        Clinical Dose G…
#>  6    1177207 Unilateral … Meas Val… LOINC         Answer          
#>  7   36982420 HAND SANITI… Drug      SPL           OTC Drug        
#>  8   42315264 Intestinal … Procedure EDI           Procedure       
#>  9   43657247 Albuterol I… Drug      RxNorm Exten… Branded Drug Fo…
#> 10      11339 Laceration … Condition ICD10CM       6-char nonbill …
#> # … with 90 more rows, and 5 more variables: standard_concept <chr>,
#> #   concept_code <chr>, valid_start_date <date>, valid_end_date <date>,
#> #   invalid_reason <chr>

# The size of the resultset can be altered
get_test_concepts(limit = 25)

#> Connecting using PostgreSQL driver
#> [2021-01-17 00:54:31]	Overriding cache... Querying Athena...
#> [2021-01-17 00:54:31] 

#> ✓ Open connection

#> [2021-01-17 00:54:31] 

#> ✓ JDBC connection

#> [2021-01-17 00:54:31]	SQL: SELECT * FROM omop_vocabulary.concept ORDER BY RANDOM() LIMIT 25
#> [2021-01-17 00:54:31]	Querying...
#> [2021-01-17 00:54:40]	Querying...complete
#> [2021-01-17 00:54:40] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-17 00:54:40]	Caching resultset...
#> [2021-01-17 00:54:40]	Postgres connection to 'athena' closed

#> # A tibble: 25 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id
#>         <dbl> <chr>        <chr>     <chr>         <chr>           
#>  1    4249743 Chronic abs… Condition SNOMED        Clinical Finding
#>  2   41516424 100000 MG h… Drug      RxNorm Exten… Quant Branded B…
#>  3   40653569 Platelet Ab… Measurem… LOINC         LOINC Group     
#>  4   45007503 Chlorzoxazo… Drug      NDC           11-digit NDC    
#>  5   44387616 quetiapine … Drug      SPL           Prescription Dr…
#>  6   41479305 13800 MG la… Drug      RxNorm Exten… Marketed Product
#>  7    2867616 Administrat… Procedure ICD10PCS      ICD10PCS Hierar…
#>  8   44901914 NITE-TIME S… Drug      NDC           11-digit NDC    
#>  9   45860377 metformin h… Drug      NDC           11-digit NDC    
#> 10   43164914 0.3 ML epoe… Drug      RxNorm Exten… Quant Clinical …
#> # … with 15 more rows, and 5 more variables: standard_concept <chr>,
#> #   concept_code <chr>, valid_start_date <date>, valid_end_date <date>,
#> #   invalid_reason <chr>

# Concepts specific to the Drug domain can be sampled
# RxNorm/RxNorm Extension Concepts
get_test_drug_concepts(limit = 10)

#> Connecting using PostgreSQL driver
#> [2021-01-17 00:54:40]	Overriding cache... Querying Athena...
#> [2021-01-17 00:54:40] 

#> ✓ Open connection

#> [2021-01-17 00:54:40] 

#> ✓ JDBC connection

#> [2021-01-17 00:54:40]	SQL: SELECT * FROM omop_vocabulary.concept c WHERE c.invalid_reason IS NULL AND c.vocabulary_id IN ('RxNorm', 'RxNorm Extension') AND c.standard_concept <> 'C' AND c.domain_id = 'Drug' ORDER BY RANDOM() LIMIT 10
#> [2021-01-17 00:54:40]	Querying...
#> [2021-01-17 00:54:44]	Querying...complete
#> [2021-01-17 00:54:44] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-17 00:54:44]	Caching resultset...
#> [2021-01-17 00:54:44]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id
#>         <dbl> <chr>        <chr>     <chr>         <chr>           
#>  1   40879895 Cefaclor 25… Drug      RxNorm Exten… Branded Drug    
#>  2   41487616 20000 MG Be… Drug      RxNorm Exten… Quant Branded D…
#>  3   40888421 Alprazolam … Drug      RxNorm Exten… Marketed Product
#>  4   43704362 219 MG Phen… Drug      RxNorm Exten… Quant Clinical …
#>  5   44058897 coffee frui… Drug      RxNorm Exten… Clinical Drug   
#>  6   40740249 Dactinomyci… Drug      RxNorm Exten… Branded Drug Box
#>  7   36265683 desloratadi… Drug      RxNorm Exten… Branded Drug Box
#>  8   41338135 1 ML cetror… Drug      RxNorm Exten… Marketed Product
#>  9   40866058 Frankincens… Drug      RxNorm Exten… Clinical Drug B…
#> 10   43753334 5 ML Bupiva… Drug      RxNorm Exten… Quant Branded B…
#> # … with 5 more variables: standard_concept <chr>, concept_code <chr>,
#> #   valid_start_date <date>, valid_end_date <date>, invalid_reason <chr>

# HemOnc/ATC Classes
get_test_drug_classes(limit = 10)

#> Connecting using PostgreSQL driver
#> [2021-01-17 00:54:44]	Overriding cache... Querying Athena...
#> [2021-01-17 00:54:44] 

#> ✓ Open connection

#> [2021-01-17 00:54:44] 

#> ✓ JDBC connection

#> [2021-01-17 00:54:44]	SQL: SELECT * FROM omop_vocabulary.concept c WHERE c.invalid_reason IS NULL AND c.vocabulary_id IN ('ATC', 'HemOnc') AND c.standard_concept = 'C' AND c.domain_id = 'Drug' ORDER BY RANDOM() LIMIT 10
#> [2021-01-17 00:54:44]	Querying...
#> [2021-01-17 00:54:44]	Querying...complete
#> [2021-01-17 00:54:44] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-17 00:54:44]	Caching resultset...
#> [2021-01-17 00:54:44]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id
#>         <dbl> <chr>        <chr>     <chr>         <chr>           
#>  1   21600753 glibornurid… Drug      ATC           ATC 5th         
#>  2   21602433 ergot alkal… Drug      ATC           ATC 5th         
#>  3   21602985 clindamycin… Drug      ATC           ATC 5th         
#>  4   21600430 endralazine… Drug      ATC           ATC 5th         
#>  5   21602696 pegvisomant… Drug      ATC           ATC 5th         
#>  6   21601788 perindopril… Drug      ATC           ATC 5th         
#>  7   21604760 modafinil; … Drug      ATC           ATC 5th         
#>  8   21603742 bleomycin    Drug      ATC           ATC 5th         
#>  9   21601000 tirofiban; … Drug      ATC           ATC 5th         
#> 10   21605585 OTHER THERA… Drug      ATC           ATC 3rd         
#> # … with 5 more variables: standard_concept <chr>, concept_code <chr>,
#> #   valid_start_date <date>, valid_end_date <date>, invalid_reason <chr>

# Test data can also be taken from tables other than Concept
get_test_data(table = "concept_relationship")

#> Connecting using PostgreSQL driver
#> [2021-01-17 00:54:44]	Overriding cache... Querying Athena...
#> [2021-01-17 00:54:44] 

#> ✓ Open connection

#> [2021-01-17 00:54:44] 

#> ✓ JDBC connection

#> [2021-01-17 00:54:44]	SQL: SELECT * FROM omop_vocabulary.concept_relationship ORDER BY RANDOM() LIMIT 100
#> [2021-01-17 00:54:44]	Querying...
#> [2021-01-17 00:55:22]	Querying...complete
#> [2021-01-17 00:55:22] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-17 00:55:22]	Caching resultset...
#> [2021-01-17 00:55:22]	Postgres connection to 'athena' closed

#> # A tibble: 100 x 6
#>    concept_id_1 concept_id_2 relationship_id valid_start_date valid_end_date
#>           <dbl>        <dbl> <chr>           <date>           <date>        
#>  1      4188203     44499594 Asso morph of   2015-09-18       2099-12-31    
#>  2     35909732     35940517 Schema to Value 1970-01-01       2099-12-31    
#>  3     41023681     40972069 Has tradename   2017-07-18       2099-12-31    
#>  4     44117861     44117861 Mapped from     2017-06-25       2099-12-31    
#>  5      1718916     44022360 Has brand name  2017-06-25       2099-12-31    
#>  6     45607216      1572752 Is a            2016-03-25       2099-12-31    
#>  7     43038064     43038064 Maps to         2019-10-11       2099-12-31    
#>  8     35810060     35817474 Has Value       2012-01-12       2099-12-31    
#>  9     43751252     43751254 Has tradename   2016-07-01       2099-12-31    
#> 10      1033751      3024359 Scale type of   2019-12-13       2099-12-31    
#> # … with 90 more rows, and 1 more variable: invalid_reason <chr>

get_test_data(table = "concept_ancestor", limit = 10)

#> Connecting using PostgreSQL driver
#> [2021-01-17 00:55:22]	Overriding cache... Querying Athena...
#> [2021-01-17 00:55:22] 

#> ✓ Open connection

#> [2021-01-17 00:55:22] 

#> ✓ JDBC connection

#> [2021-01-17 00:55:22]	SQL: SELECT * FROM omop_vocabulary.concept_ancestor ORDER BY RANDOM() LIMIT 10
#> [2021-01-17 00:55:22]	Querying...
#> [2021-01-17 00:56:00]	Querying...complete
#> [2021-01-17 00:56:00] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-17 00:56:00]	Caching resultset...
#> [2021-01-17 00:56:00]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 4
#>    ancestor_concept_… descendant_concep… min_levels_of_sepa… max_levels_of_sepa…
#>                 <dbl>              <dbl>               <dbl>               <dbl>
#>  1           21604564           43035030                   5                   7
#>  2           36889504           43220042                   1                   1
#>  3           45785229           21131158                   2                   2
#>  4           41078512           41446127                   2                   2
#>  5           44347302           44098532                   1                   1
#>  6           21600248           40165140                   3                   3
#>  7           45789910           44073854                   1                   1
#>  8           37616774           43201947                   2                   2
#>  9            4180941           40482932                   5                   5
#> 10           46289968           36062653                   2                   3