Get Unfiltered Test Concepts

Get random rows from the Concept table without any additional filters.

get_test_concepts(
  conn,
  conn_fun = "connectAthena()",
  schema = "omop_vocabulary",
  limit = 100,
  verbose = TRUE,
  render_sql = TRUE
)

See also

Examples

library(chariot)
library(tidyverse)

# Get Unfiltered Test Concepts (Invalid Concepts are included)
get_test_concepts()

#> Connecting using PostgreSQL driver
#> [2021-01-10 18:59:37]	Overriding cache... Querying Athena...
#> [2021-01-10 18:59:37] 

#> ✓ Open connection

#> [2021-01-10 18:59:37] 

#> ✓ JDBC connection

#> [2021-01-10 18:59:37]	SQL: SELECT * FROM omop_vocabulary.concept ORDER BY RANDOM() LIMIT 100
#> [2021-01-10 18:59:37]	Querying...
#> [2021-01-10 18:59:45]	Querying...complete
#> [2021-01-10 18:59:45] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 18:59:45]	Caching resultset...
#> [2021-01-10 18:59:45]	Postgres connection to 'athena' closed

#> # A tibble: 100 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id
#>         <dbl> <chr>        <chr>     <chr>         <chr>           
#>  1   42066353 Fenton       Geography OSM           8th level       
#>  2   44510634 Interval ap… Procedure OPCS4         Procedure       
#>  3    2873281 Imaging @ L… Procedure ICD10PCS      ICD10PCS Hierar…
#>  4   40899041 demelverin … Drug      RxNorm Exten… Clinical Drug   
#>  5   21140966 mercaptopur… Drug      RxNorm Exten… Clinical Drug B…
#>  6   44507546 Duavee       Drug      RxNorm        Brand Name      
#>  7   43719140 50000 MG Eu… Drug      RxNorm Exten… Quant Branded D…
#>  8   41953824 Ormoiche     Geography OSM           8th level       
#>  9    4118997 Lipiodol     Observat… SNOMED        Pharma/Biol Pro…
#> 10   43695397 Octreotide … Drug      RxNorm Exten… Branded Drug Box
#> # … with 90 more rows, and 5 more variables: standard_concept <chr>,
#> #   concept_code <chr>, valid_start_date <date>, valid_end_date <date>,
#> #   invalid_reason <chr>

# The size of the resultset can be altered
get_test_concepts(limit = 25)

#> Connecting using PostgreSQL driver
#> [2021-01-10 18:59:45]	Overriding cache... Querying Athena...
#> [2021-01-10 18:59:45] 

#> ✓ Open connection

#> [2021-01-10 18:59:45] 

#> ✓ JDBC connection

#> [2021-01-10 18:59:45]	SQL: SELECT * FROM omop_vocabulary.concept ORDER BY RANDOM() LIMIT 25
#> [2021-01-10 18:59:45]	Querying...
#> [2021-01-10 18:59:53]	Querying...complete
#> [2021-01-10 18:59:53] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 18:59:53]	Caching resultset...
#> [2021-01-10 18:59:53]	Postgres connection to 'athena' closed

#> # A tibble: 25 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id
#>         <dbl> <chr>        <chr>     <chr>         <chr>           
#>  1   36231925 Moduret Ora… Drug      RxNorm        Branded Dose Gr…
#>  2    2005477 Open reduct… Procedure ICD9Proc      4-dig billing c…
#>  3   46133690 Chlorothiaz… Drug      SNOMED        Pharma/Biol Pro…
#>  4   21464957 Hypodermic … Device    dm+d          VMPP            
#>  5   35968423 NM_000051.3… Measurem… ClinVar       Variant         
#>  6   36917489 ondansetron… Drug      SPL           Prescription Dr…
#>  7   44789822 Attention t… Procedure SNOMED        Procedure       
#>  8      15794 Strain of u… Condition ICD10CM       6-char nonbill …
#>  9   43699947 Achillea mi… Drug      RxNorm Exten… Branded Drug    
#> 10   44628218 HAMAMELIS V… Drug      DPD           Clinical Drug F…
#> # … with 15 more rows, and 5 more variables: standard_concept <chr>,
#> #   concept_code <chr>, valid_start_date <date>, valid_end_date <date>,
#> #   invalid_reason <chr>

# Concepts specific to the Drug domain can be sampled
# RxNorm/RxNorm Extension Concepts
get_test_drug_concepts(limit = 10)

#> Connecting using PostgreSQL driver
#> [2021-01-10 18:59:53]	Overriding cache... Querying Athena...
#> [2021-01-10 18:59:53] 

#> ✓ Open connection

#> [2021-01-10 18:59:53] 

#> ✓ JDBC connection

#> [2021-01-10 18:59:53]	SQL: SELECT * FROM omop_vocabulary.concept c WHERE c.invalid_reason IS NULL AND c.vocabulary_id IN ('RxNorm', 'RxNorm Extension') AND c.standard_concept <> 'C' AND c.domain_id = 'Drug' ORDER BY RANDOM() LIMIT 10
#> [2021-01-10 18:59:53]	Querying...
#> [2021-01-10 18:59:56]	Querying...complete
#> [2021-01-10 18:59:56] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 18:59:56]	Caching resultset...
#> [2021-01-10 18:59:56]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id
#>         <dbl> <chr>        <chr>     <chr>         <chr>           
#>  1   35781898 10 ML Fluor… Drug      RxNorm Exten… Marketed Product
#>  2   35129303 Tolperisone… Drug      RxNorm Exten… Branded Drug Fo…
#>  3   40021973 cetrimide T… Drug      RxNorm        Branded Drug Fo…
#>  4   41003535 Vitamin B 1… Drug      RxNorm Exten… Branded Drug Box
#>  5   36258402 desloratadi… Drug      RxNorm Exten… Marketed Product
#>  6   43669563 544 ML Gluc… Drug      RxNorm Exten… Quant Clinical …
#>  7   40886561 Esomeprazol… Drug      RxNorm Exten… Marketed Product
#>  8   35603468 Rumex aceto… Drug      RxNorm        Ingredient      
#>  9   21057737 1 ML Pentaz… Drug      RxNorm Exten… Marketed Product
#> 10   43027983 venlafaxine… Drug      RxNorm Exten… Marketed Product
#> # … with 5 more variables: standard_concept <chr>, concept_code <chr>,
#> #   valid_start_date <date>, valid_end_date <date>, invalid_reason <chr>

# HemOnc/ATC Classes
get_test_drug_classes(limit = 10)

#> Connecting using PostgreSQL driver
#> [2021-01-10 18:59:56]	Overriding cache... Querying Athena...
#> [2021-01-10 18:59:56] 

#> ✓ Open connection

#> [2021-01-10 18:59:56] 

#> ✓ JDBC connection

#> [2021-01-10 18:59:56]	SQL: SELECT * FROM omop_vocabulary.concept c WHERE c.invalid_reason IS NULL AND c.vocabulary_id IN ('ATC', 'HemOnc') AND c.standard_concept = 'C' AND c.domain_id = 'Drug' ORDER BY RANDOM() LIMIT 10
#> [2021-01-10 18:59:56]	Querying...
#> [2021-01-10 18:59:56]	Querying...complete
#> [2021-01-10 18:59:56] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 18:59:56]	Caching resultset...
#> [2021-01-10 18:59:56]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id
#>         <dbl> <chr>        <chr>     <chr>         <chr>           
#>  1   21600790 repaglinide… Drug      ATC           ATC 5th         
#>  2   21601257 immunoglobu… Drug      ATC           ATC 5th         
#>  3   35807286 Tumor vacci… Drug      HemOnc        Component Class 
#>  4   21605151 clioquinol   Drug      ATC           ATC 5th         
#>  5   21603520 tritoqualin… Drug      ATC           ATC 5th         
#>  6   21600136 dihexyverine Drug      ATC           ATC 5th         
#>  7   21601998 Proteolytic… Drug      ATC           ATC 4th         
#>  8   21601842 valsartan a… Drug      ATC           ATC 5th         
#>  9   35807241 Chemotherap… Drug      HemOnc        Component Class 
#> 10   43534779 eberconazol… Drug      ATC           ATC 5th         
#> # … with 5 more variables: standard_concept <chr>, concept_code <chr>,
#> #   valid_start_date <date>, valid_end_date <date>, invalid_reason <chr>

# Test data can also be taken from tables other than Concept
get_test_data(table = "concept_relationship")

#> Connecting using PostgreSQL driver
#> [2021-01-10 18:59:57]	Overriding cache... Querying Athena...
#> [2021-01-10 18:59:57] 

#> ✓ Open connection

#> [2021-01-10 18:59:57] 

#> ✓ JDBC connection

#> [2021-01-10 18:59:57]	SQL: SELECT * FROM omop_vocabulary.concept_relationship ORDER BY RANDOM() LIMIT 100
#> [2021-01-10 18:59:57]	Querying...
#> [2021-01-10 19:00:30]	Querying...complete
#> [2021-01-10 19:00:30] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 19:00:30]	Caching resultset...
#> [2021-01-10 19:00:30]	Postgres connection to 'athena' closed

#> # A tibble: 100 x 6
#>    concept_id_1 concept_id_2 relationship_id valid_start_date valid_end_date
#>           <dbl>        <dbl> <chr>           <date>           <date>        
#>  1      1405603       435190 Maps to         2016-01-01       2099-12-31    
#>  2     37207390     46091112 CD category of  2019-07-31       2099-12-31    
#>  3     41354959     19082103 RxNorm has dos… 2017-07-18       2099-12-31    
#>  4      3136631      3479245 Concept was_a … 2019-08-16       2099-12-31    
#>  5     41430353     41256539 Marketed form … 2017-07-18       2099-12-31    
#>  6      2844865      2840340 Subsumes        2014-04-22       2099-12-31    
#>  7     40974717     44168036 Has marketed f… 2017-10-06       2099-12-31    
#>  8     42654963      1107882 Maps to         2020-02-09       2099-12-31    
#>  9     45972176     35626975 VMP has prescr… 2018-07-31       2099-12-31    
#> 10      4040395      4041268 Subsumes        2011-07-31       2099-12-31    
#> # … with 90 more rows, and 1 more variable: invalid_reason <chr>

get_test_data(table = "concept_ancestor", limit = 10)

#> Connecting using PostgreSQL driver
#> [2021-01-10 19:00:30]	Overriding cache... Querying Athena...
#> [2021-01-10 19:00:30] 

#> ✓ Open connection

#> [2021-01-10 19:00:30] 

#> ✓ JDBC connection

#> [2021-01-10 19:00:30]	SQL: SELECT * FROM omop_vocabulary.concept_ancestor ORDER BY RANDOM() LIMIT 10
#> [2021-01-10 19:00:30]	Querying...
#> [2021-01-10 19:01:04]	Querying...complete
#> [2021-01-10 19:01:04] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 19:01:04]	Caching resultset...
#> [2021-01-10 19:01:04]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 4
#>    ancestor_concept_… descendant_concep… min_levels_of_sepa… max_levels_of_sepa…
#>                 <dbl>              <dbl>               <dbl>               <dbl>
#>  1           35804384           42716648                   4                   5
#>  2           35807271           44077383                   5                   5
#>  3           45682164             970250                   0                   0
#>  4           45676555           35141304                   2                   2
#>  5           45659173           44050220                   2                   2
#>  6            4301351            2774409                   2                   2
#>  7           42084633           42065219                   3                   3
#>  8           41063941           41482208                   0                   0
#>  9           45705065           41262205                   2                   3
#> 10            4041270            2774083                   5                   5