Get Unfiltered Test Concepts

Get random rows from the Concept table without any additional filters.

get_test_concepts(
  conn,
  conn_fun = "connectAthena()",
  schema = "omop_vocabulary",
  limit = 100,
  verbose = TRUE,
  render_sql = TRUE
)

See also

Examples

library(chariot)
library(tidyverse)

# Get Unfiltered Test Concepts (Invalid Concepts are included)
get_test_concepts()

#> Connecting using PostgreSQL driver
#> [2021-01-10 12:09:49]	Overriding cache... Querying Athena...
#> [2021-01-10 12:09:49] 

#> ✓ Open connection

#> [2021-01-10 12:09:49] 

#> ✓ JDBC connection

#> [2021-01-10 12:09:49]	SQL: SELECT * FROM omop_vocabulary.concept ORDER BY RANDOM() LIMIT 100
#> [2021-01-10 12:09:49]	Querying...
#> [2021-01-10 12:09:57]	Querying...complete
#> [2021-01-10 12:09:57] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 12:09:57]	Caching resultset...
#> [2021-01-10 12:09:57]	Postgres connection to 'athena' closed

#> # A tibble: 100 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id
#>         <dbl> <chr>        <chr>     <chr>         <chr>           
#>  1   40761593 Coagulation… Measurem… LOINC         Lab Test        
#>  2   35929582 Upper and l… Meas Val… NAACCR        NAACCR Value    
#>  3   45434920 Transport d… Observat… Read          Read            
#>  4   35178615 Isotonic So… Drug      JMDC          Drug Product    
#>  5    4308446 Ureteroneoc… Procedure SNOMED        Procedure       
#>  6   21084072 Aspirin 7.5… Drug      RxNorm Exten… Clinical Drug   
#>  7   36878772 Prucalopride Drug      RxNorm Exten… Ingredient      
#>  8   36233311 Amoxicot Or… Drug      RxNorm        Branded Dose Gr…
#>  9    1410005 Other disor… Condition ICD10CN       ICD10 code      
#> 10   43778678 5000 ML Sod… Drug      RxNorm Exten… Marketed Product
#> # … with 90 more rows, and 5 more variables: standard_concept <chr>,
#> #   concept_code <chr>, valid_start_date <date>, valid_end_date <date>,
#> #   invalid_reason <chr>

# The size of the resultset can be altered
get_test_concepts(limit = 25)

#> Connecting using PostgreSQL driver
#> [2021-01-10 12:09:57]	Overriding cache... Querying Athena...
#> [2021-01-10 12:09:57] 

#> ✓ Open connection

#> [2021-01-10 12:09:57] 

#> ✓ JDBC connection

#> [2021-01-10 12:09:57]	SQL: SELECT * FROM omop_vocabulary.concept ORDER BY RANDOM() LIMIT 25
#> [2021-01-10 12:09:57]	Querying...
#> [2021-01-10 12:10:06]	Querying...complete
#> [2021-01-10 12:10:06] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 12:10:06]	Caching resultset...
#> [2021-01-10 12:10:06]	Postgres connection to 'athena' closed

#> # A tibble: 25 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id
#>         <dbl> <chr>        <chr>     <chr>         <chr>           
#>  1   36049364 Limb protec… Device    dm+d          VMP             
#>  2    4172137 Removal of … Procedure SNOMED        Procedure       
#>  3   42964932 ezetimibe 1… Drug      RxNorm Exten… Branded Drug Co…
#>  4   46352448 ARTEMISIA V… Drug      NDC           11-digit NDC    
#>  5   21280935 Chlorothiaz… Drug      dm+d          VMPP            
#>  6   43373016 Osteoapatit… Device    AMT           Trade Product U…
#>  7   44631596 Asmanex Twi… Drug      DPD           Drug Product    
#>  8   35812285 Mean L3 in … Measurem… UK Biobank    Variable        
#>  9   45307399 candida tro… Drug      NDC           9-digit NDC     
#> 10   46140286 Boots NicAs… Drug      SNOMED        Pharma/Biol Pro…
#> # … with 15 more rows, and 5 more variables: standard_concept <chr>,
#> #   concept_code <chr>, valid_start_date <date>, valid_end_date <date>,
#> #   invalid_reason <chr>

# Concepts specific to the Drug domain can be sampled
# RxNorm/RxNorm Extension Concepts
get_test_drug_concepts(limit = 10)

#> Connecting using PostgreSQL driver
#> [2021-01-10 12:10:06]	Overriding cache... Querying Athena...
#> [2021-01-10 12:10:06] 

#> ✓ Open connection

#> [2021-01-10 12:10:06] 

#> ✓ JDBC connection

#> [2021-01-10 12:10:06]	SQL: SELECT * FROM omop_vocabulary.concept c WHERE c.invalid_reason IS NULL AND c.vocabulary_id IN ('RxNorm', 'RxNorm Extension') AND c.standard_concept <> 'C' AND c.domain_id = 'Drug' ORDER BY RANDOM() LIMIT 10
#> [2021-01-10 12:10:06]	Querying...
#> [2021-01-10 12:10:09]	Querying...complete
#> [2021-01-10 12:10:09] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 12:10:09]	Caching resultset...
#> [2021-01-10 12:10:09]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id
#>         <dbl> <chr>        <chr>     <chr>         <chr>           
#>  1   43633705 750 ML HAWT… Drug      RxNorm Exten… Marketed Product
#>  2   21051835 Azathioprin… Drug      RxNorm Exten… Clinical Drug C…
#>  3   40955732 Aesculus hi… Drug      RxNorm Exten… Clinical Drug F…
#>  4   35745549 Glucose 50 … Drug      RxNorm Exten… Clinical Drug B…
#>  5   43216874 candesartan… Drug      RxNorm Exten… Branded Drug    
#>  6   41416275 100 ML Abie… Drug      RxNorm Exten… Quant Branded D…
#>  7   41006129 Amlodipine … Drug      RxNorm Exten… Branded Drug Box
#>  8   43522880 {4 (0.5 ML … Drug      RxNorm Exten… Clinical Pack   
#>  9   40958580 Theophyllin… Drug      RxNorm Exten… Clinical Drug B…
#> 10   41198458 Guanethidin… Drug      RxNorm Exten… Marketed Product
#> # … with 5 more variables: standard_concept <chr>, concept_code <chr>,
#> #   valid_start_date <date>, valid_end_date <date>, invalid_reason <chr>

# HemOnc/ATC Classes
get_test_drug_classes(limit = 10)

#> Connecting using PostgreSQL driver
#> [2021-01-10 12:10:09]	Overriding cache... Querying Athena...
#> [2021-01-10 12:10:09] 

#> ✓ Open connection

#> [2021-01-10 12:10:09] 

#> ✓ JDBC connection

#> [2021-01-10 12:10:09]	SQL: SELECT * FROM omop_vocabulary.concept c WHERE c.invalid_reason IS NULL AND c.vocabulary_id IN ('ATC', 'HemOnc') AND c.standard_concept = 'C' AND c.domain_id = 'Drug' ORDER BY RANDOM() LIMIT 10
#> [2021-01-10 12:10:09]	Querying...
#> [2021-01-10 12:10:09]	Querying...complete
#> [2021-01-10 12:10:09] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 12:10:09]	Caching resultset...
#> [2021-01-10 12:10:09]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id
#>         <dbl> <chr>        <chr>     <chr>         <chr>           
#>  1   21602105 clobetasone… Drug      ATC           ATC 5th         
#>  2   21600262 tocainide; … Drug      ATC           ATC 5th         
#>  3   21602156 diflucortol… Drug      ATC           ATC 5th         
#>  4   21603897 sirolimus; … Drug      ATC           ATC 5th         
#>  5   35807440 Anti-IL-6 a… Drug      HemOnc        Component Class 
#>  6     715774 valbenazine  Drug      ATC           ATC 5th         
#>  7   21603503 Other antih… Drug      ATC           ATC 4th         
#>  8   21604564 ANXIOLYTICS  Drug      ATC           ATC 3rd         
#>  9   35804579 Hematopoiet… Drug      HemOnc        Component Class 
#> 10   21601320 tuberculosi… Drug      ATC           ATC 5th         
#> # … with 5 more variables: standard_concept <chr>, concept_code <chr>,
#> #   valid_start_date <date>, valid_end_date <date>, invalid_reason <chr>

# Test data can also be taken from tables other than Concept
get_test_data(table = "concept_relationship")

#> Connecting using PostgreSQL driver
#> [2021-01-10 12:10:09]	Overriding cache... Querying Athena...
#> [2021-01-10 12:10:09] 

#> ✓ Open connection

#> [2021-01-10 12:10:09] 

#> ✓ JDBC connection

#> [2021-01-10 12:10:09]	SQL: SELECT * FROM omop_vocabulary.concept_relationship ORDER BY RANDOM() LIMIT 100
#> [2021-01-10 12:10:09]	Querying...
#> [2021-01-10 12:10:47]	Querying...complete
#> [2021-01-10 12:10:47] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 12:10:47]	Caching resultset...
#> [2021-01-10 12:10:47]	Postgres connection to 'athena' closed

#> # A tibble: 100 x 6
#>    concept_id_1 concept_id_2 relationship_id valid_start_date valid_end_date
#>           <dbl>        <dbl> <chr>           <date>           <date>        
#>  1     21091531     21091531 Maps to         2016-08-07       2099-12-31    
#>  2      4153999      4077311 Has finding si… 2011-07-31       2099-12-31    
#>  3     35606120     45858011 Mapped from     2017-10-23       2099-12-31    
#>  4     37207390     46213060 CD category of  2019-07-31       2099-12-31    
#>  5     19082227     41307241 RxNorm dose fo… 2017-07-18       2099-12-31    
#>  6      3500893      3486454 Subsumes        2019-08-16       2099-12-31    
#>  7     21125411     21604256 Drug has drug … 2015-08-17       2099-12-31    
#>  8     42011745     42007445 Is a            1970-01-01       2099-12-31    
#>  9     21101523     21101523 Mapped from     2016-08-07       2099-12-31    
#> 10     40838469     41503717 Has marketed f… 2017-07-18       2099-12-31    
#> # … with 90 more rows, and 1 more variable: invalid_reason <chr>

get_test_data(table = "concept_ancestor", limit = 10)

#> Connecting using PostgreSQL driver
#> [2021-01-10 12:10:47]	Overriding cache... Querying Athena...
#> [2021-01-10 12:10:47] 

#> ✓ Open connection

#> [2021-01-10 12:10:47] 

#> ✓ JDBC connection

#> [2021-01-10 12:10:47]	SQL: SELECT * FROM omop_vocabulary.concept_ancestor ORDER BY RANDOM() LIMIT 10
#> [2021-01-10 12:10:47]	Querying...
#> [2021-01-10 12:11:23]	Querying...complete
#> [2021-01-10 12:11:23] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 12:11:23]	Caching resultset...
#> [2021-01-10 12:11:23]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 4
#>    ancestor_concept_… descendant_concep… min_levels_of_sepa… max_levels_of_sepa…
#>                 <dbl>              <dbl>               <dbl>               <dbl>
#>  1           45653761           43517213                   2                   2
#>  2             435666             760246                   3                   5
#>  3           36242520           42684063                   2                   4
#>  4           35502866           41120477                   1                   1
#>  5           45784198           21103050                   2                   3
#>  6           42705227           43213923                   2                   3
#>  7           44340943           41010746                   2                   3
#>  8           44339542           41291117                   2                   3
#>  9           45795164           19085107                   1                   1
#> 10           42807330           44062840                   2                   2