Get Unfiltered Test Concepts

Get random rows from the Concept table without any additional filters.

get_test_concepts(
  conn,
  conn_fun = "connectAthena()",
  schema = "omop_vocabulary",
  limit = 100,
  verbose = TRUE,
  render_sql = TRUE
)

See also

Examples

library(chariot)
library(tidyverse)

# Get Unfiltered Test Concepts (Invalid Concepts are included)
get_test_concepts()

#> Connecting using PostgreSQL driver
#> [2021-01-10 14:32:52]	Overriding cache... Querying Athena...
#> [2021-01-10 14:32:52] 

#> ✓ Open connection

#> [2021-01-10 14:32:53] 

#> ✓ JDBC connection

#> [2021-01-10 14:32:53]	SQL: SELECT * FROM omop_vocabulary.concept ORDER BY RANDOM() LIMIT 100
#> [2021-01-10 14:32:53]	Querying...
#> [2021-01-10 14:33:00]	Querying...complete
#> [2021-01-10 14:33:00] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 14:33:00]	Caching resultset...
#> [2021-01-10 14:33:00]	Postgres connection to 'athena' closed

#> # A tibble: 100 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id
#>         <dbl> <chr>        <chr>     <chr>         <chr>           
#>  1   36047292 Suportx shi… Device    dm+d          AMP             
#>  2   42714952 malic acid … Drug      RxNorm Exten… Branded Drug Co…
#>  3   46246804 glycerin 4g… Drug      NDC           9-digit NDC     
#>  4   40692088 Pregabalin … Drug      dm+d          AMP             
#>  5   35779038 2 ML Norepi… Drug      RxNorm Exten… Quant Clinical …
#>  6   35159444 nalfurafine… Drug      RxNorm Exten… Clinical Drug C…
#>  7    3128342 Arthritis a… Condition Nebraska Lex… Clinical Finding
#>  8   42380853 telmisartan… Drug      EDI           Drug Product    
#>  9   36632736 Alprazolam … Drug      NCCD          Drug Product    
#> 10   44356912 PRESSURE PL… Drug      SPL           OTC Drug        
#> # … with 90 more rows, and 5 more variables: standard_concept <chr>,
#> #   concept_code <chr>, valid_start_date <date>, valid_end_date <date>,
#> #   invalid_reason <chr>

# The size of the resultset can be altered
get_test_concepts(limit = 25)

#> Connecting using PostgreSQL driver
#> [2021-01-10 14:33:00]	Overriding cache... Querying Athena...
#> [2021-01-10 14:33:00] 

#> ✓ Open connection

#> [2021-01-10 14:33:00] 

#> ✓ JDBC connection

#> [2021-01-10 14:33:00]	SQL: SELECT * FROM omop_vocabulary.concept ORDER BY RANDOM() LIMIT 25
#> [2021-01-10 14:33:00]	Querying...
#> [2021-01-10 14:33:08]	Querying...complete
#> [2021-01-10 14:33:08] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 14:33:08]	Caching resultset...
#> [2021-01-10 14:33:08]	Postgres connection to 'athena' closed

#> # A tibble: 25 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id
#>         <dbl> <chr>        <chr>     <chr>         <chr>           
#>  1   36287011 Sildenafil … Drug      GGR           Drug Product    
#>  2   42835463 lansoprazol… Drug      NDC           9-digit NDC     
#>  3   46344043 ARNICA E RA… Drug      NDC           11-digit NDC    
#>  4   45263039 Naproxen so… Drug      NDC           11-digit NDC    
#>  5   43561613 METHSCOPOLA… Drug      NDFRT         Drug Interaction
#>  6   45532217 NEUROTIC EX… Condition OXMIS         OXMIS           
#>  7   36205979 Flowsheet |… Observat… LOINC         LOINC Hierarchy 
#>  8    3381632 Excisional … Procedure Nebraska Lex… Procedure       
#>  9   44600667 CAPSICUM 6 X Drug      DPD           Clinical Drug C…
#> 10    3408280 Intracrania… Condition Nebraska Lex… Clinical Finding
#> # … with 15 more rows, and 5 more variables: standard_concept <chr>,
#> #   concept_code <chr>, valid_start_date <date>, valid_end_date <date>,
#> #   invalid_reason <chr>

# Concepts specific to the Drug domain can be sampled
# RxNorm/RxNorm Extension Concepts
get_test_drug_concepts(limit = 10)

#> Connecting using PostgreSQL driver
#> [2021-01-10 14:33:09]	Overriding cache... Querying Athena...
#> [2021-01-10 14:33:09] 

#> ✓ Open connection

#> [2021-01-10 14:33:09] 

#> ✓ JDBC connection

#> [2021-01-10 14:33:09]	SQL: SELECT * FROM omop_vocabulary.concept c WHERE c.invalid_reason IS NULL AND c.vocabulary_id IN ('RxNorm', 'RxNorm Extension') AND c.standard_concept <> 'C' AND c.domain_id = 'Drug' ORDER BY RANDOM() LIMIT 10
#> [2021-01-10 14:33:09]	Querying...
#> [2021-01-10 14:33:12]	Querying...complete
#> [2021-01-10 14:33:12] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 14:33:12]	Caching resultset...
#> [2021-01-10 14:33:12]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id
#>         <dbl> <chr>        <chr>     <chr>         <chr>           
#>  1   35144708 1000 MG Thr… Drug      RxNorm Exten… Quant Clinical …
#>  2   44193365 15 ML Clobe… Drug      RxNorm Exten… Quant Branded D…
#>  3   44188444 Magnesium 0… Drug      RxNorm Exten… Branded Drug Box
#>  4   36785385 Risedronate… Drug      RxNorm Exten… Branded Drug Box
#>  5   43779120 dihydrocode… Drug      RxNorm Exten… Clinical Drug B…
#>  6   41516201 100000 MG N… Drug      RxNorm Exten… Marketed Product
#>  7   44035778 Ascorbic Ac… Drug      RxNorm Exten… Marketed Product
#>  8   40970281 gemcitabine… Drug      RxNorm Exten… Branded Drug    
#>  9   44133214 1.1 ML Saba… Drug      RxNorm Exten… Quant Clinical …
#> 10   19034354 nandrolone … Drug      RxNorm        Branded Drug    
#> # … with 5 more variables: standard_concept <chr>, concept_code <chr>,
#> #   valid_start_date <date>, valid_end_date <date>, invalid_reason <chr>

# HemOnc/ATC Classes
get_test_drug_classes(limit = 10)

#> Connecting using PostgreSQL driver
#> [2021-01-10 14:33:12]	Overriding cache... Querying Athena...
#> [2021-01-10 14:33:12] 

#> ✓ Open connection

#> [2021-01-10 14:33:12] 

#> ✓ JDBC connection

#> [2021-01-10 14:33:12]	SQL: SELECT * FROM omop_vocabulary.concept c WHERE c.invalid_reason IS NULL AND c.vocabulary_id IN ('ATC', 'HemOnc') AND c.standard_concept = 'C' AND c.domain_id = 'Drug' ORDER BY RANDOM() LIMIT 10
#> [2021-01-10 14:33:12]	Querying...
#> [2021-01-10 14:33:12]	Querying...complete
#> [2021-01-10 14:33:12] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 14:33:12]	Caching resultset...
#> [2021-01-10 14:33:12]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id
#>         <dbl> <chr>        <chr>     <chr>         <chr>           
#>  1   21602645 Drugs used … Drug      ATC           ATC 4th         
#>  2   21605044 prednisolon… Drug      ATC           ATC 5th         
#>  3    1123824 phenylephri… Drug      ATC           ATC 5th         
#>  4   21601811 fosinopril … Drug      ATC           ATC 5th         
#>  5   21601643 monoethanol… Drug      ATC           ATC 5th         
#>  6   21601019 Direct thro… Drug      ATC           ATC 4th         
#>  7   21603003 ribostamyci… Drug      ATC           ATC 5th         
#>  8   21602169 Corticoster… Drug      ATC           ATC 4th         
#>  9   21601638 heparin; to… Drug      ATC           ATC 5th         
#> 10   21600059 dihydroxial… Drug      ATC           ATC 5th         
#> # … with 5 more variables: standard_concept <chr>, concept_code <chr>,
#> #   valid_start_date <date>, valid_end_date <date>, invalid_reason <chr>

# Test data can also be taken from tables other than Concept
get_test_data(table = "concept_relationship")

#> Connecting using PostgreSQL driver
#> [2021-01-10 14:33:12]	Overriding cache... Querying Athena...
#> [2021-01-10 14:33:12] 

#> ✓ Open connection

#> [2021-01-10 14:33:12] 

#> ✓ JDBC connection

#> [2021-01-10 14:33:12]	SQL: SELECT * FROM omop_vocabulary.concept_relationship ORDER BY RANDOM() LIMIT 100
#> [2021-01-10 14:33:12]	Querying...
#> [2021-01-10 14:33:48]	Querying...complete
#> [2021-01-10 14:33:48] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 14:33:48]	Caching resultset...
#> [2021-01-10 14:33:48]	Postgres connection to 'athena' closed

#> # A tibble: 100 x 6
#>    concept_id_1 concept_id_2 relationship_id valid_start_date valid_end_date
#>           <dbl>        <dbl> <chr>           <date>           <date>        
#>  1     44207550     44207654 Has tradename   2017-10-06       2099-12-31    
#>  2      1029816      3028668 Time aspect of  2019-12-13       2099-12-31    
#>  3      4230069      4230069 Mapped from     1970-01-01       2099-12-31    
#>  4     46103114     46149399 Subsumes        2015-01-31       2099-12-31    
#>  5     35201814     35524425 Mapped from     2018-12-16       2099-12-31    
#>  6     42018386     42018386 Mapped from     2019-02-21       2099-12-31    
#>  7     43259827     46234469 RxNorm has dos… 2016-09-30       2099-12-31    
#>  8     45514026       438575 Maps to         1970-01-01       2099-12-31    
#>  9     43624112     43624112 Maps to         2016-12-18       2099-12-31    
#> 10      3291884      3364199 Subsumes        2019-08-16       2099-12-31    
#> # … with 90 more rows, and 1 more variable: invalid_reason <chr>

get_test_data(table = "concept_ancestor", limit = 10)

#> Connecting using PostgreSQL driver
#> [2021-01-10 14:33:48]	Overriding cache... Querying Athena...
#> [2021-01-10 14:33:48] 

#> ✓ Open connection

#> [2021-01-10 14:33:48] 

#> ✓ JDBC connection

#> [2021-01-10 14:33:48]	SQL: SELECT * FROM omop_vocabulary.concept_ancestor ORDER BY RANDOM() LIMIT 10
#> [2021-01-10 14:33:48]	Querying...
#> [2021-01-10 14:34:26]	Querying...complete
#> [2021-01-10 14:34:26] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 14:34:26]	Caching resultset...
#> [2021-01-10 14:34:26]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 4
#>    ancestor_concept_… descendant_concep… min_levels_of_sepa… max_levels_of_sepa…
#>                 <dbl>              <dbl>               <dbl>               <dbl>
#>  1           37021246           36204438                   2                   2
#>  2           21605008           45776602                   4                   4
#>  3             376208           37206373                   4                   5
#>  4            1597756           42953945                   1                   1
#>  5           19044756           42922294                   0                   0
#>  6           21600876           44108308                   2                   3
#>  7            4083311           40481337                   3                   3
#>  8           45783301           21042338                   2                   3
#>  9           45655431           41040007                   2                   3
#> 10           21601532           40886630                   5                   7