Get Unfiltered Test Concepts

Get random rows from the Concept table without any additional filters.

get_test_concepts(
  conn,
  conn_fun = "connectAthena()",
  schema = "omop_vocabulary",
  limit = 100,
  verbose = TRUE,
  render_sql = TRUE
)

See also

Examples

library(chariot)
library(tidyverse)

# Get Unfiltered Test Concepts (Invalid Concepts are included)
get_test_concepts()

#> Connecting using PostgreSQL driver
#> [2021-01-14 17:52:36]	Overriding cache... Querying Athena...
#> [2021-01-14 17:52:36] 

#> ✓ Open connection

#> [2021-01-14 17:52:36] 

#> ✓ JDBC connection

#> [2021-01-14 17:52:36]	SQL: SELECT * FROM omop_vocabulary.concept ORDER BY RANDOM() LIMIT 100
#> [2021-01-14 17:52:36]	Querying...
#> [2021-01-14 17:52:50]	Querying...complete
#> [2021-01-14 17:52:50] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-14 17:52:50]	Caching resultset...
#> [2021-01-14 17:52:50]	Postgres connection to 'athena' closed

#> # A tibble: 100 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id standard_concept concept_code
#>         <dbl> <chr>        <chr>     <chr>         <chr>            <chr>            <chr>       
#>  1   36692866 Venoflex Mi… Device    SNOMED        Physical Object  S                36279311000…
#>  2   45686997 MISOPROSTOL… Drug      SPL           Prescription Dr… C                26aebc95-bd…
#>  3   45946674 Open reduct… Procedure CIEL          Procedure        NA               162003      
#>  4   42302450 Medical car… Procedure EDI           Procedure        NA               A1590100    
#>  5   36000500 salicylic a… Drug      SPL           OTC Drug         C                69665906-bf…
#>  6   44415468 sertraline … Drug      NDC           11-digit NDC     NA               68788642903 
#>  7    4315891 PETROLATUM … Drug      VA Product    VA Product       NA               4008703     
#>  8   43533205 Low dose ra… Procedure HCPCS         HCPCS            S                G0458       
#>  9   36099789 nadolol 40m… Drug      NDC           11-digit NDC     NA               00904707107 
#> 10   41453425 2.02 ML Pro… Drug      RxNorm Exten… Quant Branded B… NA               OMOP2651387 
#> # … with 90 more rows, and 3 more variables: valid_start_date <date>, valid_end_date <date>,
#> #   invalid_reason <chr>

# The size of the resultset can be altered
get_test_concepts(limit = 25)

#> Connecting using PostgreSQL driver
#> [2021-01-14 17:52:50]	Overriding cache... Querying Athena...
#> [2021-01-14 17:52:50] 

#> ✓ Open connection

#> [2021-01-14 17:52:50] 

#> ✓ JDBC connection

#> [2021-01-14 17:52:50]	SQL: SELECT * FROM omop_vocabulary.concept ORDER BY RANDOM() LIMIT 25
#> [2021-01-14 17:52:50]	Querying...
#> [2021-01-14 17:52:58]	Querying...complete
#> [2021-01-14 17:52:58] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-14 17:52:58]	Caching resultset...
#> [2021-01-14 17:52:58]	Postgres connection to 'athena' closed

#> # A tibble: 25 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id standard_concept concept_code
#>         <dbl> <chr>        <chr>     <chr>         <chr>            <chr>            <chr>       
#>  1   41361893 10 ML Zinc … Drug      RxNorm Exten… Marketed Product S                OMOP2559855 
#>  2   45822859 diphenhydra… Drug      NDC           9-digit NDC      NA               469940329   
#>  3   44680403 Naprosyn 37… Drug      DPD           Drug Product     NA               2162482     
#>  4   43328408 Farine 50 M… Drug      AMT           Containered Pack NA               85049011000…
#>  5   43590671 2 MG Alchem… Drug      RxNorm Exten… Marketed Product NA               OMOP750502  
#>  6   40860711 Codeine 20 … Drug      RxNorm Exten… Branded Drug Co… S                OMOP2058673 
#>  7   21052795 olanzapine … Drug      RxNorm Exten… Marketed Product S                OMOP368414  
#>  8   45033204 Trichophyto… Drug      NDC           11-digit NDC     NA               36987212901 
#>  9   36790337 fluticasone… Drug      RxNorm Exten… Clinical Drug B… S                OMOP4778577 
#> 10    2032524 1 ML piroxi… Drug      RxNorm Exten… Marketed Product S                OMOP4931467 
#> # … with 15 more rows, and 3 more variables: valid_start_date <date>, valid_end_date <date>,
#> #   invalid_reason <chr>

# Concepts specific to the Drug domain can be sampled
# RxNorm/RxNorm Extension Concepts
get_test_drug_concepts(limit = 10)

#> Connecting using PostgreSQL driver
#> [2021-01-14 17:52:58]	Overriding cache... Querying Athena...
#> [2021-01-14 17:52:58] 

#> ✓ Open connection

#> [2021-01-14 17:52:58] 

#> ✓ JDBC connection

#> [2021-01-14 17:52:58]	SQL: SELECT * FROM omop_vocabulary.concept c WHERE c.invalid_reason IS NULL AND c.vocabulary_id IN ('RxNorm', 'RxNorm Extension') AND c.standard_concept <> 'C' AND c.domain_id = 'Drug' ORDER BY RANDOM() LIMIT 10
#> [2021-01-14 17:52:58]	Querying...
#> [2021-01-14 17:53:01]	Querying...complete
#> [2021-01-14 17:53:01] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-14 17:53:01]	Caching resultset...
#> [2021-01-14 17:53:01]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id standard_concept concept_code
#>         <dbl> <chr>        <chr>     <chr>         <chr>            <chr>            <chr>       
#>  1   36779131 50 ML Epopr… Drug      RxNorm Exten… Quant Branded B… S                OMOP4767443 
#>  2   43775723 150000 MG S… Drug      RxNorm Exten… Marketed Product S                OMOP842945  
#>  3   41178034 Physostigmi… Drug      RxNorm Exten… Clinical Drug    S                OMOP2375996 
#>  4   43261223 Gemfibrozil… Drug      RxNorm Exten… Branded Drug Fo… S                OMOP533117  
#>  5   40172298 giant wild … Drug      RxNorm        Clinical Drug F… S                900150      
#>  6   43708590 Glycine 1.8… Drug      RxNorm Exten… Clinical Drug B… S                OMOP897270  
#>  7   40999681 Maprotiline… Drug      RxNorm Exten… Branded Drug Box S                OMOP2197643 
#>  8   41307008 valsartan 8… Drug      RxNorm Exten… Branded Drug     S                OMOP2504970 
#>  9   21125177 rosuvastati… Drug      RxNorm Exten… Marketed Product S                OMOP273690  
#> 10   44086370 Fluvoxamine… Drug      RxNorm Exten… Branded Drug     S                OMOP1081001 
#> # … with 3 more variables: valid_start_date <date>, valid_end_date <date>, invalid_reason <chr>

# HemOnc/ATC Classes
get_test_drug_classes(limit = 10)

#> Connecting using PostgreSQL driver
#> [2021-01-14 17:53:01]	Overriding cache... Querying Athena...
#> [2021-01-14 17:53:01] 

#> ✓ Open connection

#> [2021-01-14 17:53:01] 

#> ✓ JDBC connection

#> [2021-01-14 17:53:01]	SQL: SELECT * FROM omop_vocabulary.concept c WHERE c.invalid_reason IS NULL AND c.vocabulary_id IN ('ATC', 'HemOnc') AND c.standard_concept = 'C' AND c.domain_id = 'Drug' ORDER BY RANDOM() LIMIT 10
#> [2021-01-14 17:53:01]	Querying...
#> [2021-01-14 17:53:02]	Querying...complete
#> [2021-01-14 17:53:02] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-14 17:53:02]	Caching resultset...
#> [2021-01-14 17:53:02]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id standard_concept concept_code
#>         <dbl> <chr>        <chr>     <chr>         <chr>            <chr>            <chr>       
#>  1   21602544 dydrogester… Drug      ATC           ATC 5th          C                G03DB01     
#>  2   43534825 tolperisone… Drug      ATC           ATC 5th          C                M02AX06     
#>  3   21600380 Other cardi… Drug      ATC           ATC 4th          C                C01EX       
#>  4   21601370 BACTERIAL A… Drug      ATC           ATC 3rd          C                J07C        
#>  5   21602666 DRUGS USED … Drug      ATC           ATC 3rd          C                G04C        
#>  6   21603045 telavancin;… Drug      ATC           ATC 5th          C                J01XA03     
#>  7   21604181 ANESTHETICS  Drug      ATC           ATC 2nd          C                N01         
#>  8   21602435 Ergot alkal… Drug      ATC           ATC 4th          C                G02AC       
#>  9   21603890 IMMUNOSUPPR… Drug      ATC           ATC 2nd          C                L04         
#> 10   21600415 urapidil; s… Drug      ATC           ATC 5th          C                C02CA06     
#> # … with 3 more variables: valid_start_date <date>, valid_end_date <date>, invalid_reason <chr>

# Test data can also be taken from tables other than Concept
get_test_data(table = "concept_relationship")

#> Connecting using PostgreSQL driver
#> [2021-01-14 17:53:02]	Overriding cache... Querying Athena...
#> [2021-01-14 17:53:02] 

#> ✓ Open connection

#> [2021-01-14 17:53:02] 

#> ✓ JDBC connection

#> [2021-01-14 17:53:02]	SQL: SELECT * FROM omop_vocabulary.concept_relationship ORDER BY RANDOM() LIMIT 100
#> [2021-01-14 17:53:02]	Querying...
#> [2021-01-14 17:53:35]	Querying...complete
#> [2021-01-14 17:53:35] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-14 17:53:35]	Caching resultset...
#> [2021-01-14 17:53:35]	Postgres connection to 'athena' closed

#> # A tibble: 100 x 6
#>    concept_id_1 concept_id_2 relationship_id    valid_start_date valid_end_date invalid_reason
#>           <dbl>        <dbl> <chr>              <date>           <date>         <chr>         
#>  1      4049503      4011234 Is a               2011-07-31       2099-12-31     NA            
#>  2      3413707      3452732 Has dir morph      2019-08-16       2099-12-31     NA            
#>  3     43783036     43783036 Mapped from        2016-12-18       2099-12-31     NA            
#>  4     41120077     41151681 Box of             2017-07-18       2099-12-31     NA            
#>  5      3472616      3455060 Subsumes           2019-08-16       2099-12-31     NA            
#>  6      4203753     37012579 Spec active ing of 2018-07-31       2099-12-31     NA            
#>  7     41457580     41032003 Quantified form of 2017-07-18       2099-12-31     NA            
#>  8     46200996     35626986 Has legal category 2018-07-31       2099-12-31     NA            
#>  9     19048660     21093123 Brand name of      2015-08-17       2099-12-31     NA            
#> 10      4264391      4264386 Is a               2011-07-31       2099-12-31     NA            
#> # … with 90 more rows

get_test_data(table = "concept_ancestor", limit = 10)

#> Connecting using PostgreSQL driver
#> [2021-01-14 17:53:35]	Overriding cache... Querying Athena...
#> [2021-01-14 17:53:35] 

#> ✓ Open connection

#> [2021-01-14 17:53:35] 

#> ✓ JDBC connection

#> [2021-01-14 17:53:35]	SQL: SELECT * FROM omop_vocabulary.concept_ancestor ORDER BY RANDOM() LIMIT 10
#> [2021-01-14 17:53:35]	Querying...
#> [2021-01-14 17:54:10]	Querying...complete
#> [2021-01-14 17:54:10] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-14 17:54:10]	Caching resultset...
#> [2021-01-14 17:54:10]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 4
#>    ancestor_concept_id descendant_concept_id min_levels_of_separation max_levels_of_separation
#>                  <dbl>                 <dbl>                    <dbl>                    <dbl>
#>  1            46300635              21131095                        2                        2
#>  2            45643169              35780622                        2                        3
#>  3            35807476              43041272                        5                        5
#>  4            21605007              36790350                        7                        7
#>  5             4078442              42896854                        3                        4
#>  6            45783536               2069737                        2                        2
#>  7            44097289              44097289                        0                        0
#>  8            45646067              43268433                        2                        2
#>  9            45628921              43521175                        1                        1
#> 10            45702355              35139979                        2                        2