Get Unfiltered Test Concepts

Get random rows from the Concept table without any additional filters.

get_test_concepts(
  conn,
  conn_fun = "connectAthena()",
  schema = "omop_vocabulary",
  limit = 100,
  verbose = TRUE,
  render_sql = TRUE
)

See also

Examples

library(chariot)
library(tidyverse)

# Get Unfiltered Test Concepts (Invalid Concepts are included)
get_test_concepts()

#> Connecting using PostgreSQL driver

#> [2021-03-28 18:46:48]	Overriding cache... Querying Athena...
#> [2021-03-28 18:46:48] 

#> ✓ Open connection

#> [2021-03-28 18:46:48] 

#> ✓ JDBC connection

#> [2021-03-28 18:46:48]	SQL: SELECT * FROM omop_vocabulary.concept ORDER BY RANDOM() LIMIT 100
#> [2021-03-28 18:46:48]	Querying...
#> [2021-03-28 18:46:56]	Querying...complete
#> [2021-03-28 18:46:56] 

#> ✓ Returned data has more than 0 rows

#> [2021-03-28 18:46:56]	Caching resultset...
#> [2021-03-28 18:46:56]	Postgres connection to 'athena' closed

#> # A tibble: 100 x 10
#>    concept_id concept_name             domain_id vocabulary_id  concept_class_id
#>         <int> <chr>                    <chr>     <chr>          <chr>           
#>  1   21318335 Phenobarbital 9.5mg/5ml… Drug      dm+d           VMPP            
#>  2   40089909 Tramadol 150 MG 24 Hour… Drug      RxNorm         Clinical Drug   
#>  3    1169732 benzalkonium chloride .… Drug      SPL            OTC Drug        
#>  4   41966179 Villeneuve-les-Genêts    Geography OSM            8th level       
#>  5   43638871 Erythromycin 1 MG/ML In… Drug      RxNorm Extens… Branded Drug Box
#>  6   21185691 Pravastatin 20mg tablets Drug      dm+d           AMP             
#>  7   43038709 1 ML Dalteparin 10000 U… Drug      RxNorm Extens… Quant Clinical …
#>  8   46006784 Bupivacaine 25mg/10ml (… Drug      SNOMED         Pharma/Biol Pro…
#>  9    3080078 Retired procedure        Procedure Nebraska Lexi… Procedure       
#> 10   41417926 100 ML Benzoate 8 MG/ML… Drug      RxNorm Extens… Quant Branded B…
#> # … with 90 more rows, and 5 more variables: standard_concept <chr>,
#> #   concept_code <chr>, valid_start_date <date>, valid_end_date <date>,
#> #   invalid_reason <chr>

# The size of the resultset can be altered
get_test_concepts(limit = 25)

#> Connecting using PostgreSQL driver

#> [2021-03-28 18:46:56]	Overriding cache... Querying Athena...
#> [2021-03-28 18:46:56] 

#> ✓ Open connection

#> [2021-03-28 18:46:56] 

#> ✓ JDBC connection

#> [2021-03-28 18:46:56]	SQL: SELECT * FROM omop_vocabulary.concept ORDER BY RANDOM() LIMIT 25
#> [2021-03-28 18:46:56]	Querying...
#> [2021-03-28 18:47:04]	Querying...complete
#> [2021-03-28 18:47:04] 

#> ✓ Returned data has more than 0 rows

#> [2021-03-28 18:47:04]	Caching resultset...
#> [2021-03-28 18:47:04]	Postgres connection to 'athena' closed

#> # A tibble: 25 x 10
#>    concept_id concept_name             domain_id  vocabulary_id concept_class_id
#>         <int> <chr>                    <chr>      <chr>         <chr>           
#>  1   41103082 Morphine 20 MG Oral Tab… Drug       RxNorm Exten… Marketed Product
#>  2   41927443 凤岗镇                   Geography  OSM           8th level       
#>  3   21448838 Juzo Inspiration class … Device     dm+d          AMPP            
#>  4   40715544 10 ML Fentanyl 0.05 MG/… Drug       RxNorm Exten… Quant Branded D…
#>  5   46020192 Haddenham Veni class 1 … Device     SNOMED        Physical Object 
#>  6   43608206 1 ML Immunoglobulin G 1… Drug       RxNorm Exten… Quant Clinical …
#>  7   37518392 ethanol 0.72 ML/ML Topi… Drug       NDC           11-digit NDC    
#>  8    3031197 HC gene rearrangements … Measureme… LOINC         Lab Test        
#>  9   45002081 Indomethacin 50 MG Oral… Drug       NDC           11-digit NDC    
#> 10   45019624 tramadol hydrochloride … Drug       NDC           11-digit NDC    
#> # … with 15 more rows, and 5 more variables: standard_concept <chr>,
#> #   concept_code <chr>, valid_start_date <date>, valid_end_date <date>,
#> #   invalid_reason <chr>

# Concepts specific to the Drug domain can be sampled
# RxNorm/RxNorm Extension Concepts
get_test_drug_concepts(limit = 10)

#> Connecting using PostgreSQL driver

#> [2021-03-28 18:47:04]	Overriding cache... Querying Athena...
#> [2021-03-28 18:47:04] 

#> ✓ Open connection

#> [2021-03-28 18:47:04] 

#> ✓ JDBC connection

#> [2021-03-28 18:47:04]	SQL: SELECT * FROM omop_vocabulary.concept c WHERE c.invalid_reason IS NULL AND c.vocabulary_id IN ('RxNorm', 'RxNorm Extension') AND c.standard_concept <> 'C' AND c.domain_id = 'Drug' ORDER BY RANDOM() LIMIT 10
#> [2021-03-28 18:47:04]	Querying...
#> [2021-03-28 18:47:06]	Querying...complete
#> [2021-03-28 18:47:06] 

#> ✓ Returned data has more than 0 rows

#> [2021-03-28 18:47:06]	Caching resultset...
#> [2021-03-28 18:47:06]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 10
#>    concept_id concept_name             domain_id vocabulary_id  concept_class_id
#>         <int> <chr>                    <chr>     <chr>          <chr>           
#>  1   41304865 Coal Tar 0.023 MG/ML To… Drug      RxNorm Extens… Clinical Drug B…
#>  2   42706624 pramoxine hydrochloride… Drug      RxNorm         Branded Drug Co…
#>  3   43736369 15 ML fluocinolone 0.25… Drug      RxNorm Extens… Quant Clinical …
#>  4   40913560 Trimipramine 100 MG Ora… Drug      RxNorm Extens… Marketed Product
#>  5   19109928 floxacillin 20 MG/ML     Drug      RxNorm         Clinical Drug C…
#>  6   43255247 Chlorhexidine 1 MG/ML I… Drug      RxNorm Extens… Branded Drug Box
#>  7   21168347 Ergonovine 0.5 MG/ML / … Drug      RxNorm Extens… Branded Drug    
#>  8   43257260 Pravastatin 40 MG Oral … Drug      RxNorm Extens… Marketed Product
#>  9   35772043 Betaxolol 2.5 MG/ML Oph… Drug      RxNorm Extens… Marketed Product
#> 10   44118845 Naphthalene 3 X [Naphta… Drug      RxNorm Extens… Branded Drug Co…
#> # … with 5 more variables: standard_concept <chr>, concept_code <chr>,
#> #   valid_start_date <date>, valid_end_date <date>, invalid_reason <chr>

# HemOnc/ATC Classes
get_test_drug_classes(limit = 10)

#> Connecting using PostgreSQL driver

#> [2021-03-28 18:47:07]	Overriding cache... Querying Athena...
#> [2021-03-28 18:47:07] 

#> ✓ Open connection

#> [2021-03-28 18:47:07] 

#> ✓ JDBC connection

#> [2021-03-28 18:47:07]	SQL: SELECT * FROM omop_vocabulary.concept c WHERE c.invalid_reason IS NULL AND c.vocabulary_id IN ('ATC', 'HemOnc') AND c.standard_concept = 'C' AND c.domain_id = 'Drug' ORDER BY RANDOM() LIMIT 10
#> [2021-03-28 18:47:07]	Querying...
#> [2021-03-28 18:47:07]	Querying...complete
#> [2021-03-28 18:47:07] 

#> ✓ Returned data has more than 0 rows

#> [2021-03-28 18:47:07]	Caching resultset...
#> [2021-03-28 18:47:07]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 10
#>    concept_id concept_name              domain_id vocabulary_id concept_class_id
#>         <int> <chr>                     <chr>     <chr>         <chr>           
#>  1   21602510 testosterone; systemic, … Drug      ATC           ATC 5th         
#>  2   21600439 Alkaloids, excl. rauwolf… Drug      ATC           ATC 4th         
#>  3   21603235 Antibiotics               Drug      ATC           ATC 4th         
#>  4   21601690 celiprolol; oral          Drug      ATC           ATC 5th         
#>  5   21601013 reteplase; parenteral     Drug      ATC           ATC 5th         
#>  6   21604631 chloralodol; oral         Drug      ATC           ATC 5th         
#>  7   21602453 Sympathomimetics, labour… Drug      ATC           ATC 4th         
#>  8   21603725 paclitaxel poliglumex     Drug      ATC           ATC 5th         
#>  9   35807217 Radiotherapy medication   Drug      HemOnc        Component Class 
#> 10   21603044 teicoplanin; parenteral   Drug      ATC           ATC 5th         
#> # … with 5 more variables: standard_concept <chr>, concept_code <chr>,
#> #   valid_start_date <date>, valid_end_date <date>, invalid_reason <chr>

# Test data can also be taken from tables other than Concept
get_test_data(table = "concept_relationship")

#> Connecting using PostgreSQL driver

#> [2021-03-28 18:47:07]	Overriding cache... Querying Athena...
#> [2021-03-28 18:47:07] 

#> ✓ Open connection

#> [2021-03-28 18:47:07] 

#> ✓ JDBC connection

#> [2021-03-28 18:47:07]	SQL: SELECT * FROM omop_vocabulary.concept_relationship ORDER BY RANDOM() LIMIT 100
#> [2021-03-28 18:47:07]	Querying...
#> [2021-03-28 18:47:41]	Querying...complete
#> [2021-03-28 18:47:41] 

#> ✓ Returned data has more than 0 rows

#> [2021-03-28 18:47:41]	Caching resultset...
#> [2021-03-28 18:47:41]	Postgres connection to 'athena' closed

#> # A tibble: 100 x 6
#>    concept_id_1 concept_id_2 relationship_id     valid_start_date valid_end_date
#>           <int>        <int> <chr>               <date>           <date>        
#>  1      4269148      4089156 Subsumes            2011-07-31       2099-12-31    
#>  2     40739215     40739215 Maps to             2018-09-19       2099-12-31    
#>  3     36916470      4265757 Maps to             2017-01-01       2099-12-31    
#>  4     40011643     36213267 RxNorm is a         2016-08-01       2099-12-31    
#>  5     40103643     40231936 Maps to             2015-09-08       2099-12-31    
#>  6      3187945      3183018 Subsumes            2019-08-16       2099-12-31    
#>  7     42899011     36665702 Mapped from         2020-09-01       2099-12-31    
#>  8     40748628     19082224 RxNorm has dose fo… 2018-06-04       2099-12-31    
#>  9     41221996     41221996 Mapped from         2017-07-18       2099-12-31    
#> 10     19135866     44179297 RxNorm dose form of 2017-10-06       2099-12-31    
#> # … with 90 more rows, and 1 more variable: invalid_reason <chr>

get_test_data(table = "concept_ancestor", limit = 10)

#> Connecting using PostgreSQL driver

#> [2021-03-28 18:47:41]	Overriding cache... Querying Athena...
#> [2021-03-28 18:47:41] 

#> ✓ Open connection

#> [2021-03-28 18:47:41] 

#> ✓ JDBC connection

#> [2021-03-28 18:47:41]	SQL: SELECT * FROM omop_vocabulary.concept_ancestor ORDER BY RANDOM() LIMIT 10
#> [2021-03-28 18:47:41]	Querying...
#> [2021-03-28 18:48:17]	Querying...complete
#> [2021-03-28 18:48:17] 

#> ✓ Returned data has more than 0 rows

#> [2021-03-28 18:48:17]	Caching resultset...
#> [2021-03-28 18:48:17]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 4
#>    ancestor_concept_… descendant_concep… min_levels_of_sepa… max_levels_of_sepa…
#>                 <int>              <int>               <int>               <int>
#>  1           36747061           36747061                   0                   0
#>  2           45654751           41153138                   2                   2
#>  3            4042140           40493028                   7                  13
#>  4           36981273           43703492                   1                   1
#>  5           36157831           41290655                   2                   3
#>  6           45675824           42480088                   2                   3
#>  7           45657090           21047219                   2                   3
#>  8           40163792           21060202                   2                   2
#>  9           36600039           40879223                   1                   1
#> 10             970779           41280100                   1                   1