Get Unfiltered Test Concepts

Get random rows from the Concept table without any additional filters.

get_test_concepts(
  conn,
  conn_fun = "connectAthena()",
  schema = "omop_vocabulary",
  limit = 100,
  verbose = TRUE,
  render_sql = TRUE
)

See also

Examples

library(chariot)
library(tidyverse)

# Get Unfiltered Test Concepts (Invalid Concepts are included)
get_test_concepts()

#> Connecting using PostgreSQL driver
#> [2021-01-10 16:32:06]	Overriding cache... Querying Athena...
#> [2021-01-10 16:32:06] 

#> ✓ Open connection

#> [2021-01-10 16:32:06] 

#> ✓ JDBC connection

#> [2021-01-10 16:32:06]	SQL: SELECT * FROM omop_vocabulary.concept ORDER BY RANDOM() LIMIT 100
#> [2021-01-10 16:32:06]	Querying...
#> [2021-01-10 16:32:14]	Querying...complete
#> [2021-01-10 16:32:14] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 16:32:14]	Caching resultset...
#> [2021-01-10 16:32:14]	Postgres connection to 'athena' closed

#> # A tibble: 100 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id
#>         <dbl> <chr>        <chr>     <chr>         <chr>           
#>  1    1175266 HIV 1 RNA i… Measurem… LOINC         Lab Test        
#>  2   45482986 Employment … Observat… Read          Read            
#>  3    3072435 Diabetes me… Condition Nebraska Lex… Clinical Finding
#>  4   36790174 Felodipine … Drug      RxNorm Exten… Branded Drug Box
#>  5   41958427 Saint-Étien… Geography OSM           8th level       
#>  6   35524661 azelaic aci… Drug      NDC           11-digit NDC    
#>  7   45667904 Benztropine… Drug      SPL           SPL             
#>  8   44988848 Oxygen 99 %… Drug      NDC           11-digit NDC    
#>  9   43728445 Verapamil D… Drug      RxNorm Exten… Branded Drug Fo…
#> 10   41135722 Hydrochloro… Drug      RxNorm Exten… Marketed Product
#> # … with 90 more rows, and 5 more variables: standard_concept <chr>,
#> #   concept_code <chr>, valid_start_date <date>, valid_end_date <date>,
#> #   invalid_reason <chr>

# The size of the resultset can be altered
get_test_concepts(limit = 25)

#> Connecting using PostgreSQL driver
#> [2021-01-10 16:32:14]	Overriding cache... Querying Athena...
#> [2021-01-10 16:32:14] 

#> ✓ Open connection

#> [2021-01-10 16:32:14] 

#> ✓ JDBC connection

#> [2021-01-10 16:32:14]	SQL: SELECT * FROM omop_vocabulary.concept ORDER BY RANDOM() LIMIT 25
#> [2021-01-10 16:32:14]	Querying...
#> [2021-01-10 16:32:21]	Querying...complete
#> [2021-01-10 16:32:21] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 16:32:21]	Caching resultset...
#> [2021-01-10 16:32:21]	Postgres connection to 'athena' closed

#> # A tibble: 25 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id
#>         <dbl> <chr>        <chr>     <chr>         <chr>           
#>  1    4171362 Nolina texa… Observat… SNOMED        Organism        
#>  2   44497456 hydromorpho… Drug      NDC           9-digit NDC     
#>  3   35742441 Doxorubicin… Drug      RxNorm Exten… Marketed Product
#>  4   40781941 Spine Facet… Observat… LOINC         LOINC Hierarchy 
#>  5   43797067 Eucalyptus … Drug      RxNorm Exten… Clinical Drug   
#>  6   42982203 DEXIROL Inj… Drug      KDC           Drug Product    
#>  7    4193043 Region of t… Spec Ana… SNOMED        Body Structure  
#>  8    4103520 Chow Chow    Observat… SNOMED        Organism        
#>  9    1029542 Ultrasound+… Observat… LOINC         LOINC Method    
#> 10    4090202 Nasopharyng… Condition SNOMED        Clinical Finding
#> # … with 15 more rows, and 5 more variables: standard_concept <chr>,
#> #   concept_code <chr>, valid_start_date <date>, valid_end_date <date>,
#> #   invalid_reason <chr>

# Concepts specific to the Drug domain can be sampled
# RxNorm/RxNorm Extension Concepts
get_test_drug_concepts(limit = 10)

#> Connecting using PostgreSQL driver
#> [2021-01-10 16:32:22]	Overriding cache... Querying Athena...
#> [2021-01-10 16:32:22] 

#> ✓ Open connection

#> [2021-01-10 16:32:22] 

#> ✓ JDBC connection

#> [2021-01-10 16:32:22]	SQL: SELECT * FROM omop_vocabulary.concept c WHERE c.invalid_reason IS NULL AND c.vocabulary_id IN ('RxNorm', 'RxNorm Extension') AND c.standard_concept <> 'C' AND c.domain_id = 'Drug' ORDER BY RANDOM() LIMIT 10
#> [2021-01-10 16:32:22]	Querying...
#> [2021-01-10 16:32:24]	Querying...complete
#> [2021-01-10 16:32:24] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 16:32:24]	Caching resultset...
#> [2021-01-10 16:32:24]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id
#>         <dbl> <chr>        <chr>     <chr>         <chr>           
#>  1   36061571 hyaluronate… Drug      RxNorm Exten… Branded Drug Fo…
#>  2   21129657 Ergocalcife… Drug      RxNorm Exten… Clinical Drug C…
#>  3   42927062 tiropramide… Drug      RxNorm Exten… Branded Drug Fo…
#>  4   43712679 gabapentin … Drug      RxNorm Exten… Branded Drug Co…
#>  5   43034969 Methylpredn… Drug      RxNorm Exten… Marketed Product
#>  6   40911551 Bromocripti… Drug      RxNorm Exten… Branded Drug Box
#>  7   43614677 380 ML Alan… Drug      RxNorm Exten… Quant Clinical …
#>  8   41273732 Codeine 0.0… Drug      RxNorm Exten… Clinical Drug B…
#>  9   41031766 imidapril 2… Drug      RxNorm Exten… Branded Drug Box
#> 10   41269941 Berberis vu… Drug      RxNorm Exten… Branded Drug Fo…
#> # … with 5 more variables: standard_concept <chr>, concept_code <chr>,
#> #   valid_start_date <date>, valid_end_date <date>, invalid_reason <chr>

# HemOnc/ATC Classes
get_test_drug_classes(limit = 10)

#> Connecting using PostgreSQL driver
#> [2021-01-10 16:32:24]	Overriding cache... Querying Athena...
#> [2021-01-10 16:32:24] 

#> ✓ Open connection

#> [2021-01-10 16:32:24] 

#> ✓ JDBC connection

#> [2021-01-10 16:32:25]	SQL: SELECT * FROM omop_vocabulary.concept c WHERE c.invalid_reason IS NULL AND c.vocabulary_id IN ('ATC', 'HemOnc') AND c.standard_concept = 'C' AND c.domain_id = 'Drug' ORDER BY RANDOM() LIMIT 10
#> [2021-01-10 16:32:25]	Querying...
#> [2021-01-10 16:32:25]	Querying...complete
#> [2021-01-10 16:32:25] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 16:32:25]	Caching resultset...
#> [2021-01-10 16:32:25]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id
#>         <dbl> <chr>        <chr>     <chr>         <chr>           
#>  1    1501770 telaprevir;… Drug      ATC           ATC 5th         
#>  2   43534861 gatifloxaci… Drug      ATC           ATC 5th         
#>  3   21601717 labetalol a… Drug      ATC           ATC 5th         
#>  4   21604266 dihydrocode… Drug      ATC           ATC 5th         
#>  5   21603545 mepixanox    Drug      ATC           ATC 5th         
#>  6    1123740 empaglifloz… Drug      ATC           ATC 5th         
#>  7   35807214 Anthracycli… Drug      HemOnc        Component Class 
#>  8   21601880 Nicotinic a… Drug      ATC           ATC 4th         
#>  9    1501886 erenumab     Drug      ATC           ATC 5th         
#> 10   21604464 levodopa; o… Drug      ATC           ATC 5th         
#> # … with 5 more variables: standard_concept <chr>, concept_code <chr>,
#> #   valid_start_date <date>, valid_end_date <date>, invalid_reason <chr>

# Test data can also be taken from tables other than Concept
get_test_data(table = "concept_relationship")

#> Connecting using PostgreSQL driver
#> [2021-01-10 16:32:25]	Overriding cache... Querying Athena...
#> [2021-01-10 16:32:25] 

#> ✓ Open connection

#> [2021-01-10 16:32:25] 

#> ✓ JDBC connection

#> [2021-01-10 16:32:25]	SQL: SELECT * FROM omop_vocabulary.concept_relationship ORDER BY RANDOM() LIMIT 100
#> [2021-01-10 16:32:25]	Querying...
#> [2021-01-10 16:32:58]	Querying...complete
#> [2021-01-10 16:32:58] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 16:32:58]	Caching resultset...
#> [2021-01-10 16:32:58]	Postgres connection to 'athena' closed

#> # A tibble: 100 x 6
#>    concept_id_1 concept_id_2 relationship_id valid_start_date valid_end_date
#>           <dbl>        <dbl> <chr>           <date>           <date>        
#>  1     35776592     19082103 RxNorm has dos… 2017-08-02       2099-12-31    
#>  2     42856236     19074907 VAProd - RxNor… 2014-12-01       2099-12-31    
#>  3     36720230     36760272 Maps to         2020-09-13       2099-12-31    
#>  4     35862460     35862460 Maps to         2020-10-02       2099-12-31    
#>  5     42081593     42081593 Maps to         2019-02-21       2099-12-31    
#>  6     46063402     45968758 Is a            2015-01-31       2099-12-31    
#>  7     19082103     36280928 RxNorm dose fo… 2017-10-01       2099-12-31    
#>  8     36896221     43630264 Has tradename   2017-06-24       2099-12-31    
#>  9     35155156     35167321 Mapped from     2019-03-11       2099-12-31    
#> 10     44075966     44084265 Marketed form … 2017-06-25       2099-12-31    
#> # … with 90 more rows, and 1 more variable: invalid_reason <chr>

get_test_data(table = "concept_ancestor", limit = 10)

#> Connecting using PostgreSQL driver
#> [2021-01-10 16:32:58]	Overriding cache... Querying Athena...
#> [2021-01-10 16:32:58] 

#> ✓ Open connection

#> [2021-01-10 16:32:58] 

#> ✓ JDBC connection

#> [2021-01-10 16:32:58]	SQL: SELECT * FROM omop_vocabulary.concept_ancestor ORDER BY RANDOM() LIMIT 10
#> [2021-01-10 16:32:58]	Querying...
#> [2021-01-10 16:33:31]	Querying...complete
#> [2021-01-10 16:33:31] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 16:33:31]	Caching resultset...
#> [2021-01-10 16:33:31]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 4
#>    ancestor_concept_… descendant_concep… min_levels_of_sepa… max_levels_of_sepa…
#>                 <dbl>              <dbl>               <dbl>               <dbl>
#>  1           21604795           42936361                   1                   2
#>  2            4017774            3021325                   2                   2
#>  3           43049429           41007315                   2                   3
#>  4           35529590           41135597                   2                   3
#>  5            4040552            4291773                   4                   5
#>  6           44344214           36810843                   2                   2
#>  7           36217214           44045025                   2                   2
#>  8           42899909           41363038                   2                   2
#>  9           21602795            2068415                   7                   7
#> 10           40491439           36549933                   6                   7