Get Unfiltered Test Concepts

Get random rows from the Concept table without any additional filters.

get_test_concepts(
  conn,
  conn_fun = "connectAthena()",
  schema = "omop_vocabulary",
  limit = 100,
  verbose = TRUE,
  render_sql = TRUE
)

See also

Examples

library(chariot)
library(tidyverse)

# Get Unfiltered Test Concepts (Invalid Concepts are included)
get_test_concepts()

#> Connecting using PostgreSQL driver
#> [2021-01-10 12:29:21]	Overriding cache... Querying Athena...
#> [2021-01-10 12:29:21] 

#> ✓ Open connection

#> [2021-01-10 12:29:21] 

#> ✓ JDBC connection

#> [2021-01-10 12:29:21]	SQL: SELECT * FROM omop_vocabulary.concept ORDER BY RANDOM() LIMIT 100
#> [2021-01-10 12:29:21]	Querying...
#> [2021-01-10 12:29:29]	Querying...complete
#> [2021-01-10 12:29:29] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 12:29:29]	Caching resultset...
#> [2021-01-10 12:29:29]	Postgres connection to 'athena' closed

#> # A tibble: 100 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id
#>         <dbl> <chr>        <chr>     <chr>         <chr>           
#>  1   40869910 1-Propanol … Drug      RxNorm Exten… Clinical Drug B…
#>  2   45534092 Osteochondr… Condition ICD10CM       6-char billing …
#>  3   45895869 ibuprofen 5… Drug      SPL           OTC Drug        
#>  4      16858 Burn of uns… Condition ICD10CM       6-char nonbill …
#>  5   43676315 potassium c… Drug      RxNorm Exten… Clinical Drug B…
#>  6   46315037 CLINDESSE -… Drug      SPL           Prescription Dr…
#>  7   42723076 Calcium Flu… Drug      RxNorm Exten… Branded Drug Co…
#>  8    4030193 Panhypopitu… Condition SNOMED        Clinical Finding
#>  9   44047063 Plantago Se… Drug      RxNorm Exten… Branded Drug    
#> 10   21153376 Flupenthixo… Drug      RxNorm Exten… Branded Drug Co…
#> # … with 90 more rows, and 5 more variables: standard_concept <chr>,
#> #   concept_code <chr>, valid_start_date <date>, valid_end_date <date>,
#> #   invalid_reason <chr>

# The size of the resultset can be altered
get_test_concepts(limit = 25)

#> Connecting using PostgreSQL driver
#> [2021-01-10 12:29:29]	Overriding cache... Querying Athena...
#> [2021-01-10 12:29:29] 

#> ✓ Open connection

#> [2021-01-10 12:29:29] 

#> ✓ JDBC connection

#> [2021-01-10 12:29:29]	SQL: SELECT * FROM omop_vocabulary.concept ORDER BY RANDOM() LIMIT 25
#> [2021-01-10 12:29:29]	Querying...
#> [2021-01-10 12:29:36]	Querying...complete
#> [2021-01-10 12:29:36] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 12:29:36]	Caching resultset...
#> [2021-01-10 12:29:36]	Postgres connection to 'athena' closed

#> # A tibble: 25 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id
#>         <dbl> <chr>        <chr>     <chr>         <chr>           
#>  1   21342918 Elasticated… Device    dm+d          VMPP            
#>  2    3139065 Child exam … Observat… Nebraska Lex… Context-depende…
#>  3   35803847 Cardegic     Drug      HemOnc        Brand Name      
#>  4   37050776 Onchocerca … Observat… LOINC         LOINC Component 
#>  5   45812783 ethacrynic … Drug      NDC           9-digit NDC     
#>  6   42513279 Neoplasm de… Condition ICDO3         ICDO Condition  
#>  7   41228896 Levetiracet… Drug      RxNorm Exten… Marketed Product
#>  8    3194501 centilitre   Unit      Nebraska Lex… Qualifier Value 
#>  9   40122715 Triclosan 2… Drug      RxNorm        Branded Drug Co…
#> 10   41046787 Mercury 20 … Drug      RxNorm Exten… Branded Drug Co…
#> # … with 15 more rows, and 5 more variables: standard_concept <chr>,
#> #   concept_code <chr>, valid_start_date <date>, valid_end_date <date>,
#> #   invalid_reason <chr>

# Concepts specific to the Drug domain can be sampled
# RxNorm/RxNorm Extension Concepts
get_test_drug_concepts(limit = 10)

#> Connecting using PostgreSQL driver
#> [2021-01-10 12:29:36]	Overriding cache... Querying Athena...
#> [2021-01-10 12:29:36] 

#> ✓ Open connection

#> [2021-01-10 12:29:36] 

#> ✓ JDBC connection

#> [2021-01-10 12:29:36]	SQL: SELECT * FROM omop_vocabulary.concept c WHERE c.invalid_reason IS NULL AND c.vocabulary_id IN ('RxNorm', 'RxNorm Extension') AND c.standard_concept <> 'C' AND c.domain_id = 'Drug' ORDER BY RANDOM() LIMIT 10
#> [2021-01-10 12:29:36]	Querying...
#> [2021-01-10 12:29:39]	Querying...complete
#> [2021-01-10 12:29:39] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 12:29:39]	Caching resultset...
#> [2021-01-10 12:29:39]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id
#>         <dbl> <chr>        <chr>     <chr>         <chr>           
#>  1   41477592 10000 MG Ta… Drug      RxNorm Exten… Quant Branded B…
#>  2   42874247 icosapent e… Drug      RxNorm        Clinical Drug C…
#>  3   40076756 pilocarpine… Drug      RxNorm        Clinical Drug F…
#>  4   43787177 Solidago vi… Drug      RxNorm Exten… Branded Drug Fo…
#>  5   43294017 Carbomer 0.… Drug      RxNorm Exten… Clinical Drug   
#>  6   41061644 Dipyrone 50… Drug      RxNorm Exten… Branded Drug Box
#>  7   43598151 1000 ML Pro… Drug      RxNorm Exten… Marketed Product
#>  8   41222664 Clomipramin… Drug      RxNorm Exten… Branded Drug    
#>  9   43041729 15 ML Carbo… Drug      RxNorm Exten… Marketed Product
#> 10   40184732 loperamide … Drug      RxNorm        Branded Drug    
#> # … with 5 more variables: standard_concept <chr>, concept_code <chr>,
#> #   valid_start_date <date>, valid_end_date <date>, invalid_reason <chr>

# HemOnc/ATC Classes
get_test_drug_classes(limit = 10)

#> Connecting using PostgreSQL driver
#> [2021-01-10 12:29:39]	Overriding cache... Querying Athena...
#> [2021-01-10 12:29:39] 

#> ✓ Open connection

#> [2021-01-10 12:29:39] 

#> ✓ JDBC connection

#> [2021-01-10 12:29:39]	SQL: SELECT * FROM omop_vocabulary.concept c WHERE c.invalid_reason IS NULL AND c.vocabulary_id IN ('ATC', 'HemOnc') AND c.standard_concept = 'C' AND c.domain_id = 'Drug' ORDER BY RANDOM() LIMIT 10
#> [2021-01-10 12:29:39]	Querying...
#> [2021-01-10 12:29:39]	Querying...complete
#> [2021-01-10 12:29:39] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 12:29:39]	Caching resultset...
#> [2021-01-10 12:29:39]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id
#>         <dbl> <chr>        <chr>     <chr>         <chr>           
#>  1   21605357 Other combi… Drug      ATC           ATC 4th         
#>  2   21603533 colfosceril… Drug      ATC           ATC 5th         
#>  3   21602158 CORTICOSTER… Drug      ATC           ATC 3rd         
#>  4   21605414 X-RAY CONTR… Drug      ATC           ATC 3rd         
#>  5    1501778 semaglutide… Drug      ATC           ATC 5th         
#>  6   21604178 Other drugs… Drug      ATC           ATC 4th         
#>  7   35807431 Alpha emitt… Drug      HemOnc        Component Class 
#>  8   21603349 proxyphylli… Drug      ATC           ATC 5th         
#>  9   21602520 estrone; or… Drug      ATC           ATC 5th         
#> 10   21605448 technetium … Drug      ATC           ATC 5th         
#> # … with 5 more variables: standard_concept <chr>, concept_code <chr>,
#> #   valid_start_date <date>, valid_end_date <date>, invalid_reason <chr>

# Test data can also be taken from tables other than Concept
get_test_data(table = "concept_relationship")

#> Connecting using PostgreSQL driver
#> [2021-01-10 12:29:40]	Overriding cache... Querying Athena...
#> [2021-01-10 12:29:40] 

#> ✓ Open connection

#> [2021-01-10 12:29:40] 

#> ✓ JDBC connection

#> [2021-01-10 12:29:40]	SQL: SELECT * FROM omop_vocabulary.concept_relationship ORDER BY RANDOM() LIMIT 100
#> [2021-01-10 12:29:40]	Querying...
#> [2021-01-10 12:30:13]	Querying...complete
#> [2021-01-10 12:30:13] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 12:30:13]	Caching resultset...
#> [2021-01-10 12:30:13]	Postgres connection to 'athena' closed

#> # A tibble: 100 x 6
#>    concept_id_1 concept_id_2 relationship_id valid_start_date valid_end_date
#>           <dbl>        <dbl> <chr>           <date>           <date>        
#>  1     46125302     35607932 Has non-avail … 2018-07-31       2099-12-31    
#>  2        32961     35811030 Unit mapped fr… 2020-12-21       2099-12-31    
#>  3     35887195     21066982 Marketed form … 2020-12-07       2099-12-31    
#>  4     45814074       781063 Maps to         2015-01-29       2099-12-31    
#>  5      4104882      2763302 Dir subst of    1970-01-01       2099-12-31    
#>  6     19016123      1124300 Brand name of   2006-01-20       2099-12-31    
#>  7     21149467     21119927 RxNorm is a     2015-08-17       2099-12-31    
#>  8      4204895     37205595 Concept replac… 2019-07-31       2099-12-31    
#>  9     36541185       139750 Is a            2015-09-18       2099-12-31    
#> 10     40949456     44110094 Marketed form … 2017-07-18       2099-12-31    
#> # … with 90 more rows, and 1 more variable: invalid_reason <chr>

get_test_data(table = "concept_ancestor", limit = 10)

#> Connecting using PostgreSQL driver
#> [2021-01-10 12:30:13]	Overriding cache... Querying Athena...
#> [2021-01-10 12:30:13] 

#> ✓ Open connection

#> [2021-01-10 12:30:13] 

#> ✓ JDBC connection

#> [2021-01-10 12:30:13]	SQL: SELECT * FROM omop_vocabulary.concept_ancestor ORDER BY RANDOM() LIMIT 10
#> [2021-01-10 12:30:13]	Querying...
#> [2021-01-10 12:30:47]	Querying...complete
#> [2021-01-10 12:30:47] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 12:30:47]	Caching resultset...
#> [2021-01-10 12:30:47]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 4
#>    ancestor_concept_… descendant_concep… min_levels_of_sepa… max_levels_of_sepa…
#>                 <dbl>              <dbl>               <dbl>               <dbl>
#>  1           45770357           44811938                   1                   1
#>  2             995125           42876447                   1                   1
#>  3           45629006           41043366                   2                   3
#>  4           37044165            3022807                   4                   4
#>  5           45802216           21163025                   2                   3
#>  6             443783            4125966                   4                   4
#>  7           45700145           21122772                   2                   3
#>  8           38004547           38004547                   0                   0
#>  9           35874297           35874296                   1                   1
#> 10           36892062           35796746                   1                   1