Get Unfiltered Test Concepts

Get random rows from the Concept table without any additional filters.

get_test_concepts(
  conn,
  conn_fun = "connectAthena()",
  schema = "omop_vocabulary",
  limit = 100,
  verbose = TRUE,
  render_sql = TRUE
)

See also

Examples

library(chariot)
library(tidyverse)

# Get Unfiltered Test Concepts (Invalid Concepts are included)
get_test_concepts()

#> [2021-05-26 15:45:19]	Connected to 'athena'
#> [2021-05-26 15:45:19]	Overriding cache... Querying Athena...
#> [2021-05-26 15:45:19]	✓ Open connection
#> [2021-05-26 15:45:19]	✓ JDBC connection
#> [2021-05-26 15:45:19]	SQL: SELECT * FROM omop_vocabulary.concept ORDER BY RANDOM() LIMIT 100
#> [2021-05-26 15:45:19]	Querying...
#> [2021-05-26 15:45:29]	Querying...complete
#> [2021-05-26 15:45:29]	✓ Returned data has more than 0 rows
#> [2021-05-26 15:45:29]	Caching resultset...
#> [2021-05-26 15:45:29]	Postgres connection to 'athena' closed

#> # A tibble: 100 x 10
#>    concept_id concept_name             domain_id  vocabulary_id concept_class_id
#>         <int> <chr>                    <chr>      <chr>         <chr>           
#>  1   35304289 rosuvastatin 10mg/1 ORA… Drug       NDC           9-digit NDC     
#>  2    1171425 benzalkonium chloride .… Drug       NDC           9-digit NDC     
#>  3   42536289 Garlic                   Observati… SNOMED        Substance       
#>  4   36738182 Chromosome 17 Variant: … Measureme… OMOP Extensi… Genomic Variant 
#>  5   42260353 Thrombectomy(Artery),Ab… Procedure  EDI           Procedure       
#>  6   35849398 sodium chloride 9 MG/ML… Drug       RxNorm Exten… Branded Drug    
#>  7   45994776 Pelican Select DuoVent … Device     SNOMED        Physical Object 
#>  8   35765035 Dutasteride 0.5 MG Exte… Drug       RxNorm Exten… Marketed Product
#>  9   41307205 Trimipramine 25 MG Oral… Drug       RxNorm Exten… Branded Drug Box
#> 10   41081993 Framycetin / Hydrocorti… Drug       RxNorm Exten… Branded Drug Fo…
#> # … with 90 more rows, and 5 more variables: standard_concept <chr>,
#> #   concept_code <chr>, valid_start_date <date>, valid_end_date <date>,
#> #   invalid_reason <chr>

# The size of the resultset can be altered
get_test_concepts(limit = 25)

#> [2021-05-26 15:45:29]	Connected to 'athena'
#> [2021-05-26 15:45:29]	Overriding cache... Querying Athena...
#> [2021-05-26 15:45:29]	✓ Open connection
#> [2021-05-26 15:45:29]	✓ JDBC connection
#> [2021-05-26 15:45:29]	SQL: SELECT * FROM omop_vocabulary.concept ORDER BY RANDOM() LIMIT 25
#> [2021-05-26 15:45:29]	Querying...
#> [2021-05-26 15:45:37]	Querying...complete
#> [2021-05-26 15:45:37]	✓ Returned data has more than 0 rows
#> [2021-05-26 15:45:37]	Caching resultset...
#> [2021-05-26 15:45:37]	Postgres connection to 'athena' closed

#> # A tibble: 25 x 10
#>    concept_id concept_name             domain_id  vocabulary_id concept_class_id
#>         <int> <chr>                    <chr>      <chr>         <chr>           
#>  1   36766621 TSC2 Transcript Variant… Measureme… OMOP Extensi… Transcript Vari…
#>  2   43557909 ethyl alcohol .75mL/mL … Drug       NDC           9-digit NDC     
#>  3    4187692 Rhodococcus marinonasce… Observati… SNOMED        Organism        
#>  4    2103412 Biopsy, soft tissue of … Procedure  CPT4          CPT4            
#>  5   45330902 gabapentin 300 MG Oral … Drug       NDC           11-digit NDC    
#>  6   45351983 alanine 9.93 MG/ML / ar… Drug       NDC           11-digit NDC    
#>  7   21126191 1 ML Pyridostigmine 4 M… Drug       RxNorm Exten… Quant Clinical …
#>  8   36171063 losatan potassium 25mg/… Drug       SPL           Prescription Dr…
#>  9   19052919 Livial                   Drug       RxNorm        Brand Name      
#> 10   45250748 atomoxetine 60 MG Oral … Drug       NDC           11-digit NDC    
#> # … with 15 more rows, and 5 more variables: standard_concept <chr>,
#> #   concept_code <chr>, valid_start_date <date>, valid_end_date <date>,
#> #   invalid_reason <chr>

# Concepts specific to the Drug domain can be sampled
# RxNorm/RxNorm Extension Concepts
get_test_drug_concepts(limit = 10)

#> [2021-05-26 15:45:37]	Connected to 'athena'
#> [2021-05-26 15:45:37]	Overriding cache... Querying Athena...
#> [2021-05-26 15:45:37]	✓ Open connection
#> [2021-05-26 15:45:37]	✓ JDBC connection
#> [2021-05-26 15:45:37]	SQL: SELECT * FROM omop_vocabulary.concept c WHERE c.invalid_reason IS NULL AND c.vocabulary_id IN ('RxNorm', 'RxNorm Extension') AND c.standard_concept <> 'C' AND c.domain_id = 'Drug' ORDER BY RANDOM() LIMIT 10
#> [2021-05-26 15:45:37]	Querying...
#> [2021-05-26 15:45:40]	Querying...complete
#> [2021-05-26 15:45:40]	✓ Returned data has more than 0 rows
#> [2021-05-26 15:45:40]	Caching resultset...
#> [2021-05-26 15:45:40]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 10
#>    concept_id concept_name              domain_id vocabulary_id concept_class_id
#>         <int> <chr>                     <chr>     <chr>         <chr>           
#>  1    2040436 kininogenase 10 IU Exten… Drug      RxNorm Exten… Clinical Drug   
#>  2   44171122 Magnesium 375 MG Oral Ca… Drug      RxNorm Exten… Marketed Product
#>  3   43623067 Ondansetron 0.64 MG/ML I… Drug      RxNorm Exten… Clinical Drug B…
#>  4   21172928 lamotrigine 1 MG/ML Oral… Drug      RxNorm Exten… Clinical Drug   
#>  5   41337755 1 ML prednisolone 50 MG/… Drug      RxNorm Exten… Quant Branded B…
#>  6   43163124 Metoprolol 190 MG Oral T… Drug      RxNorm Exten… Branded Drug Box
#>  7   35789604 0.02 ML exenatide 0.25 M… Drug      RxNorm Exten… Quant Branded B…
#>  8   44053592 ALUMINUM ZIRCONIUM TETRA… Drug      RxNorm Exten… Clinical Drug C…
#>  9   44084788 Calcium Carbonate 333 MG… Drug      RxNorm Exten… Clinical Drug   
#> 10   41474685 6000 MG Dequalinium 0.00… Drug      RxNorm Exten… Quant Branded B…
#> # … with 5 more variables: standard_concept <chr>, concept_code <chr>,
#> #   valid_start_date <date>, valid_end_date <date>, invalid_reason <chr>

# HemOnc/ATC Classes
get_test_drug_classes(limit = 10)

#> [2021-05-26 15:45:40]	Connected to 'athena'
#> [2021-05-26 15:45:40]	Overriding cache... Querying Athena...
#> [2021-05-26 15:45:40]	✓ Open connection
#> [2021-05-26 15:45:40]	✓ JDBC connection
#> [2021-05-26 15:45:40]	SQL: SELECT * FROM omop_vocabulary.concept c WHERE c.invalid_reason IS NULL AND c.vocabulary_id IN ('ATC', 'HemOnc') AND c.standard_concept = 'C' AND c.domain_id = 'Drug' ORDER BY RANDOM() LIMIT 10
#> [2021-05-26 15:45:40]	Querying...
#> [2021-05-26 15:45:40]	Querying...complete
#> [2021-05-26 15:45:40]	✓ Returned data has more than 0 rows
#> [2021-05-26 15:45:40]	Caching resultset...
#> [2021-05-26 15:45:40]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 10
#>    concept_id concept_name              domain_id vocabulary_id concept_class_id
#>         <int> <chr>                     <chr>     <chr>         <chr>           
#>  1   21601912 natamycin; topical        Drug      ATC           ATC 5th         
#>  2   21601156 fat emulsions; parenteral Drug      ATC           ATC 5th         
#>  3   21603007 Fluoroquinolones          Drug      ATC           ATC 4th         
#>  4   21600204 clidinium and psycholept… Drug      ATC           ATC 5th         
#>  5   21601478 chlorothiazide and potas… Drug      ATC           ATC 5th         
#>  6   21602805 tetracycline; systemic    Drug      ATC           ATC 5th         
#>  7   21603098 rifamycin; parenteral     Drug      ATC           ATC 5th         
#>  8   21603924 ciclosporin; systemic     Drug      ATC           ATC 5th         
#>  9   21605086 levocabastine; ophthalmic Drug      ATC           ATC 5th         
#> 10   21601075 romiplostim; parenteral   Drug      ATC           ATC 5th         
#> # … with 5 more variables: standard_concept <chr>, concept_code <chr>,
#> #   valid_start_date <date>, valid_end_date <date>, invalid_reason <chr>

# Test data can also be taken from tables other than Concept
get_test_data(table = "concept_relationship")

#> [2021-05-26 15:45:40]	Connected to 'athena'
#> [2021-05-26 15:45:40]	Overriding cache... Querying Athena...
#> [2021-05-26 15:45:40]	✓ Open connection
#> [2021-05-26 15:45:40]	✓ JDBC connection
#> [2021-05-26 15:45:40]	SQL: SELECT * FROM omop_vocabulary.concept_relationship ORDER BY RANDOM() LIMIT 100
#> [2021-05-26 15:45:40]	Querying...
#> [2021-05-26 15:46:12]	Querying...complete
#> [2021-05-26 15:46:12]	✓ Returned data has more than 0 rows
#> [2021-05-26 15:46:12]	Caching resultset...
#> [2021-05-26 15:46:12]	Postgres connection to 'athena' closed

#> # A tibble: 100 x 6
#>    concept_id_1 concept_id_2 relationship_id     valid_start_date valid_end_date
#>           <int>        <int> <chr>               <date>           <date>        
#>  1     19082103       784206 RxNorm dose form of 2019-04-01       2099-12-31    
#>  2     41016475     40804151 Has brand name      2017-07-18       2099-12-31    
#>  3     40022479      1177480 RxNorm has ing      2005-04-04       2099-12-31    
#>  4     37207390     45983337 CD category of      2019-07-31       2099-12-31    
#>  5     46024687     37209200 Is a                2019-07-31       2099-12-31    
#>  6     37207390     46266267 CD category of      2019-07-31       2099-12-31    
#>  7     19006231     19006231 Maps to             1970-01-01       2099-12-31    
#>  8     19106872     21031752 Has tradename       2015-08-17       2099-12-31    
#>  9     40809865     40875199 Brand name of       2017-07-18       2099-12-31    
#> 10     43727862     21018690 Has brand name      2017-08-24       2099-12-31    
#> # … with 90 more rows, and 1 more variable: invalid_reason <chr>

get_test_data(table = "concept_ancestor", limit = 10)

#> [2021-05-26 15:46:12]	Connected to 'athena'
#> [2021-05-26 15:46:12]	Overriding cache... Querying Athena...
#> [2021-05-26 15:46:12]	✓ Open connection
#> [2021-05-26 15:46:12]	✓ JDBC connection
#> [2021-05-26 15:46:12]	SQL: SELECT * FROM omop_vocabulary.concept_ancestor ORDER BY RANDOM() LIMIT 10
#> [2021-05-26 15:46:12]	Querying...
#> [2021-05-26 15:46:46]	Querying...complete
#> [2021-05-26 15:46:46]	✓ Returned data has more than 0 rows
#> [2021-05-26 15:46:46]	Caching resultset...
#> [2021-05-26 15:46:46]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 4
#>    ancestor_concept_… descendant_concep… min_levels_of_sepa… max_levels_of_sepa…
#>                 <int>              <int>               <int>               <int>
#>  1           45640896           40948906                   2                   3
#>  2            4199402            4346169                   4                   5
#>  3           45681866           43603360                   2                   2
#>  4           44360369           40914335                   2                   3
#>  5            4180627           37109108                   6                  10
#>  6           45659255           43851617                   2                   2
#>  7            2856643            2856643                   0                   0
#>  8           45796536           44036978                   2                   2
#>  9            4102111           36522481                   3                   7
#> 10           36790031           36790028                   1                   1