Get Unfiltered Test Concepts

Get random rows from the Concept table without any additional filters.

get_test_concepts(
  conn,
  conn_fun = "connectAthena()",
  schema = "omop_vocabulary",
  limit = 100,
  verbose = TRUE,
  render_sql = TRUE
)

See also

Examples

library(chariot)
library(tidyverse)

# Get Unfiltered Test Concepts (Invalid Concepts are included)
get_test_concepts()

#> Connecting using PostgreSQL driver
#> [2021-01-10 15:57:51]	Overriding cache... Querying Athena...
#> [2021-01-10 15:57:51] 

#> ✓ Open connection

#> [2021-01-10 15:57:51] 

#> ✓ JDBC connection

#> [2021-01-10 15:57:51]	SQL: SELECT * FROM omop_vocabulary.concept ORDER BY RANDOM() LIMIT 100
#> [2021-01-10 15:57:51]	Querying...
#> [2021-01-10 15:57:58]	Querying...complete
#> [2021-01-10 15:57:58] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 15:57:58]	Caching resultset...
#> [2021-01-10 15:57:58]	Postgres connection to 'athena' closed

#> # A tibble: 100 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id
#>         <dbl> <chr>        <chr>     <chr>         <chr>           
#>  1    1747799 miconazole … Drug      SPL           OTC Drug        
#>  2   35912573 OBSOLETE DA… Meas Val… NAACCR        NAACCR Value    
#>  3   41994719 Rosenhügel   Geography OSM           11th level      
#>  4    1013895 Can pronoun… Observat… LOINC         LOINC Component 
#>  5    4165183 Iodine comp… Observat… SNOMED        Clinical Finding
#>  6   45348055 donepezil h… Drug      NDC           11-digit NDC    
#>  7   40353590 [X]Assault … Condition SNOMED        Clinical Finding
#>  8   45193216 Hydroxychlo… Drug      NDC           11-digit NDC    
#>  9   43811594 100 ML Medi… Drug      RxNorm Exten… Quant Branded D…
#> 10   43152141 Levetiracet… Drug      RxNorm Exten… Branded Drug Box
#> # … with 90 more rows, and 5 more variables: standard_concept <chr>,
#> #   concept_code <chr>, valid_start_date <date>, valid_end_date <date>,
#> #   invalid_reason <chr>

# The size of the resultset can be altered
get_test_concepts(limit = 25)

#> Connecting using PostgreSQL driver
#> [2021-01-10 15:57:58]	Overriding cache... Querying Athena...
#> [2021-01-10 15:57:58] 

#> ✓ Open connection

#> [2021-01-10 15:57:58] 

#> ✓ JDBC connection

#> [2021-01-10 15:57:58]	SQL: SELECT * FROM omop_vocabulary.concept ORDER BY RANDOM() LIMIT 25
#> [2021-01-10 15:57:58]	Querying...
#> [2021-01-10 15:58:06]	Querying...complete
#> [2021-01-10 15:58:06] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 15:58:06]	Caching resultset...
#> [2021-01-10 15:58:06]	Postgres connection to 'athena' closed

#> # A tibble: 25 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id
#>         <dbl> <chr>        <chr>     <chr>         <chr>           
#>  1   41074641 Chromonar 7… Drug      RxNorm Exten… Marketed Product
#>  2   43207921 10 ML Glyci… Drug      RxNorm Exten… Quant Clinical …
#>  3   44553780 Unda 202 [D… Drug      DPD           Drug Product    
#>  4   44171746 Amoxicillin… Drug      RxNorm Exten… Marketed Product
#>  5   40680416 GentilFoam … Device    dm+d          AMPP            
#>  6   35803308 Sapacitabine Drug      HemOnc        Component       
#>  7   36514534 avobenzone,… Drug      NDC           11-digit NDC    
#>  8   42753009 Other opera… Procedure OPS           Procedure       
#>  9   44606105 Nu-Loxapine… Drug      DPD           Drug Product    
#> 10   36741243 Chromosome … Measurem… OMOP Extensi… Genomic Variant 
#> # … with 15 more rows, and 5 more variables: standard_concept <chr>,
#> #   concept_code <chr>, valid_start_date <date>, valid_end_date <date>,
#> #   invalid_reason <chr>

# Concepts specific to the Drug domain can be sampled
# RxNorm/RxNorm Extension Concepts
get_test_drug_concepts(limit = 10)

#> Connecting using PostgreSQL driver
#> [2021-01-10 15:58:06]	Overriding cache... Querying Athena...
#> [2021-01-10 15:58:06] 

#> ✓ Open connection

#> [2021-01-10 15:58:06] 

#> ✓ JDBC connection

#> [2021-01-10 15:58:06]	SQL: SELECT * FROM omop_vocabulary.concept c WHERE c.invalid_reason IS NULL AND c.vocabulary_id IN ('RxNorm', 'RxNorm Extension') AND c.standard_concept <> 'C' AND c.domain_id = 'Drug' ORDER BY RANDOM() LIMIT 10
#> [2021-01-10 15:58:06]	Querying...
#> [2021-01-10 15:58:09]	Querying...complete
#> [2021-01-10 15:58:09] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 15:58:09]	Caching resultset...
#> [2021-01-10 15:58:09]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id
#>         <dbl> <chr>        <chr>     <chr>         <chr>           
#>  1   42723932 Amanita mus… Drug      RxNorm Exten… Branded Drug Box
#>  2   41151414 Vincamine 1… Drug      RxNorm Exten… Branded Drug Box
#>  3   35788539 200 ML ropi… Drug      RxNorm Exten… Quant Branded B…
#>  4   43768685 travoprost … Drug      RxNorm Exten… Branded Drug    
#>  5   40880900 bendamustin… Drug      RxNorm Exten… Branded Drug Box
#>  6   21108069 50 ML Cipro… Drug      RxNorm Exten… Marketed Product
#>  7   41219370 Insulin, Po… Drug      RxNorm Exten… Branded Drug Box
#>  8   36781033 3.13 ML Som… Drug      RxNorm Exten… Quant Clinical …
#>  9   35764345 Piroxicam 1… Drug      RxNorm Exten… Marketed Product
#> 10   21040810 Dydrogester… Drug      RxNorm Exten… Clinical Drug C…
#> # … with 5 more variables: standard_concept <chr>, concept_code <chr>,
#> #   valid_start_date <date>, valid_end_date <date>, invalid_reason <chr>

# HemOnc/ATC Classes
get_test_drug_classes(limit = 10)

#> Connecting using PostgreSQL driver
#> [2021-01-10 15:58:09]	Overriding cache... Querying Athena...
#> [2021-01-10 15:58:09] 

#> ✓ Open connection

#> [2021-01-10 15:58:09] 

#> ✓ JDBC connection

#> [2021-01-10 15:58:09]	SQL: SELECT * FROM omop_vocabulary.concept c WHERE c.invalid_reason IS NULL AND c.vocabulary_id IN ('ATC', 'HemOnc') AND c.standard_concept = 'C' AND c.domain_id = 'Drug' ORDER BY RANDOM() LIMIT 10
#> [2021-01-10 15:58:09]	Querying...
#> [2021-01-10 15:58:09]	Querying...complete
#> [2021-01-10 15:58:09] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 15:58:09]	Caching resultset...
#> [2021-01-10 15:58:09]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id
#>         <dbl> <chr>        <chr>     <chr>         <chr>           
#>  1     715921 tilorone; o… Drug      ATC           ATC 5th         
#>  2   21601585 Enzymes      Drug      ATC           ATC 4th         
#>  3   21601912 natamycin; … Drug      ATC           ATC 5th         
#>  4   21602003 ANTIPRURITI… Drug      ATC           ATC 3rd         
#>  5   21602988 pristinamyc… Drug      ATC           ATC 5th         
#>  6   21602761 methylthiou… Drug      ATC           ATC 5th         
#>  7   21602558 ANDROGENS A… Drug      ATC           ATC 3rd         
#>  8    1588694 efinaconazo… Drug      ATC           ATC 5th         
#>  9   21605531 Indium (111… Drug      ATC           ATC 4th         
#> 10   21605357 Other combi… Drug      ATC           ATC 4th         
#> # … with 5 more variables: standard_concept <chr>, concept_code <chr>,
#> #   valid_start_date <date>, valid_end_date <date>, invalid_reason <chr>

# Test data can also be taken from tables other than Concept
get_test_data(table = "concept_relationship")

#> Connecting using PostgreSQL driver
#> [2021-01-10 15:58:09]	Overriding cache... Querying Athena...
#> [2021-01-10 15:58:09] 

#> ✓ Open connection

#> [2021-01-10 15:58:09] 

#> ✓ JDBC connection

#> [2021-01-10 15:58:09]	SQL: SELECT * FROM omop_vocabulary.concept_relationship ORDER BY RANDOM() LIMIT 100
#> [2021-01-10 15:58:09]	Querying...
#> [2021-01-10 15:58:41]	Querying...complete
#> [2021-01-10 15:58:41] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 15:58:41]	Caching resultset...
#> [2021-01-10 15:58:41]	Postgres connection to 'athena' closed

#> # A tibble: 100 x 6
#>    concept_id_1 concept_id_2 relationship_id valid_start_date valid_end_date
#>           <dbl>        <dbl> <chr>           <date>           <date>        
#>  1      2314019      4235356 CPT4 - SNOMED … 1980-01-01       2099-12-31    
#>  2     19016670     44117731 RxNorm ing of   2017-06-25       2099-12-31    
#>  3     19082573     41199501 RxNorm dose fo… 2017-07-18       2099-12-31    
#>  4     43040314     44023661 Has supplier    2019-10-06       2099-12-31    
#>  5     21019890     21122645 Supplier of     2015-08-17       2099-12-31    
#>  6      4226285      3179072 Mapped from     1970-01-01       2099-12-31    
#>  7     36421940     42876542 Has marketed f… 2020-01-15       2099-12-31    
#>  8      3258089      3249551 Method of       2019-08-16       2099-12-31    
#>  9     40807615     41505755 Brand name of   2017-07-18       2099-12-31    
#> 10     35909856     35942805 Schema to Value 1970-01-01       2099-12-31    
#> # … with 90 more rows, and 1 more variable: invalid_reason <chr>

get_test_data(table = "concept_ancestor", limit = 10)

#> Connecting using PostgreSQL driver
#> [2021-01-10 15:58:41]	Overriding cache... Querying Athena...
#> [2021-01-10 15:58:41] 

#> ✓ Open connection

#> [2021-01-10 15:58:41] 

#> ✓ JDBC connection

#> [2021-01-10 15:58:41]	SQL: SELECT * FROM omop_vocabulary.concept_ancestor ORDER BY RANDOM() LIMIT 10
#> [2021-01-10 15:58:41]	Querying...
#> [2021-01-10 15:59:15]	Querying...complete
#> [2021-01-10 15:59:15] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 15:59:15]	Caching resultset...
#> [2021-01-10 15:59:15]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 4
#>    ancestor_concept_… descendant_concep… min_levels_of_sepa… max_levels_of_sepa…
#>                 <dbl>              <dbl>               <dbl>               <dbl>
#>  1           21602848           43022943                   2                   3
#>  2           19110189           43198532                   2                   2
#>  3            4307750            4250189                   5                  11
#>  4           45621021           44048612                   1                   1
#>  5           35807483           42931439                   3                   3
#>  6           35807188           21096329                   4                   4
#>  7           36217220           19124017                   2                   3
#>  8             312339           42536242                   5                   5
#>  9           45875968           40767832                   2                   3
#> 10           37311678           36532101                   6                   8