Get Test Drug Classes

Get random records of ATC and HemOnc Drug Classes that are:

Valid,

Class,

Belong to the Drug domain

get_test_drug_classes(
  conn,
  conn_fun = "connectAthena()",
  schema = "omop_vocabulary",
  limit = 100,
  verbose = TRUE,
  render_sql = TRUE
)

See also

Examples

library(chariot)
library(tidyverse)

# Get Unfiltered Test Concepts (Invalid Concepts are included)
get_test_concepts()

#> [2021-05-26 15:46:47]	Connected to 'athena'
#> [2021-05-26 15:46:47]	Overriding cache... Querying Athena...
#> [2021-05-26 15:46:47]	✓ Open connection
#> [2021-05-26 15:46:47]	✓ JDBC connection
#> [2021-05-26 15:46:47]	SQL: SELECT * FROM omop_vocabulary.concept ORDER BY RANDOM() LIMIT 100
#> [2021-05-26 15:46:47]	Querying...
#> [2021-05-26 15:46:55]	Querying...complete
#> [2021-05-26 15:46:55]	✓ Returned data has more than 0 rows
#> [2021-05-26 15:46:55]	Caching resultset...
#> [2021-05-26 15:46:55]	Postgres connection to 'athena' closed

#> # A tibble: 100 x 10
#>    concept_id concept_name             domain_id  vocabulary_id concept_class_id
#>         <int> <chr>                    <chr>      <chr>         <chr>           
#>  1   21144841 varicella-zoster immune… Drug       RxNorm Exten… Branded Drug Box
#>  2    2057640 glyburide / metformin O… Drug       RxNorm Exten… Branded Drug Fo…
#>  3    4179380 Subligamental cordectomy Procedure  SNOMED        Procedure       
#>  4   42323422 Operation of Pancreatic… Procedure  EDI           Procedure       
#>  5   42347596 Genetic Tests for Somat… Measureme… EDI           Meas Class      
#>  6   44660023 TRIMIPRAMINE MALEATE 75… Drug       DPD           Branded Drug Co…
#>  7   45060786 3 ML liraglutide 6 MG/M… Drug       NDC           11-digit NDC    
#>  8   42268643 Craniotomy for Drainage… Procedure  EDI           Procedure       
#>  9    3068099 [X]Motorcycle rider inj… Observati… Nebraska Lex… Event           
#> 10    3604525 LoFric 8Ch 990800 PVC 3… Observati… SNOMED        Context-depende…
#> # … with 90 more rows, and 5 more variables: standard_concept <chr>,
#> #   concept_code <chr>, valid_start_date <date>, valid_end_date <date>,
#> #   invalid_reason <chr>

# The size of the resultset can be altered
get_test_concepts(limit = 25)

#> [2021-05-26 15:46:55]	Connected to 'athena'
#> [2021-05-26 15:46:55]	Overriding cache... Querying Athena...
#> [2021-05-26 15:46:55]	✓ Open connection
#> [2021-05-26 15:46:55]	✓ JDBC connection
#> [2021-05-26 15:46:55]	SQL: SELECT * FROM omop_vocabulary.concept ORDER BY RANDOM() LIMIT 25
#> [2021-05-26 15:46:55]	Querying...
#> [2021-05-26 15:47:03]	Querying...complete
#> [2021-05-26 15:47:03]	✓ Returned data has more than 0 rows
#> [2021-05-26 15:47:03]	Caching resultset...
#> [2021-05-26 15:47:03]	Postgres connection to 'athena' closed

#> # A tibble: 25 x 10
#>    concept_id concept_name             domain_id vocabulary_id  concept_class_id
#>         <int> <chr>                    <chr>     <chr>          <chr>           
#>  1    2767451 Replacement of Right Ac… Procedure ICD10PCS       ICD10PCS        
#>  2   43790280 20000 MG nicoboxil 0.02… Drug      RxNorm Extens… Quant Clinical …
#>  3   44867012 Metoprolol Tartrate 50 … Drug      NDC            11-digit NDC    
#>  4   45621127 terconazole 4mg/g VAGIN… Drug      SPL            Prescription Dr…
#>  5   36627459 Netilmicin Injection [D… Drug      NCCD           Drug Product    
#>  6    1414175 Conduction disorder, un… Condition ICD10CN        ICD10 code      
#>  7   44824325 Diverticula of intestine Condition ICD9CM         3-dig nonbill c…
#>  8   41035728 Carbamazepine 300 MG Or… Drug      RxNorm Extens… Branded Drug Box
#>  9   21269513 Primidone 50mg tablets … Drug      dm+d           AMPP            
#> 10   36533409 Pleomorphic carcinoma o… Condition ICDO3          ICDO Condition  
#> # … with 15 more rows, and 5 more variables: standard_concept <chr>,
#> #   concept_code <chr>, valid_start_date <date>, valid_end_date <date>,
#> #   invalid_reason <chr>

# Concepts specific to the Drug domain can be sampled
# RxNorm/RxNorm Extension Concepts
get_test_drug_concepts(limit = 10)

#> [2021-05-26 15:47:03]	Connected to 'athena'
#> [2021-05-26 15:47:03]	Overriding cache... Querying Athena...
#> [2021-05-26 15:47:03]	✓ Open connection
#> [2021-05-26 15:47:03]	✓ JDBC connection
#> [2021-05-26 15:47:03]	SQL: SELECT * FROM omop_vocabulary.concept c WHERE c.invalid_reason IS NULL AND c.vocabulary_id IN ('RxNorm', 'RxNorm Extension') AND c.standard_concept <> 'C' AND c.domain_id = 'Drug' ORDER BY RANDOM() LIMIT 10
#> [2021-05-26 15:47:03]	Querying...
#> [2021-05-26 15:47:06]	Querying...complete
#> [2021-05-26 15:47:06]	✓ Returned data has more than 0 rows
#> [2021-05-26 15:47:06]	Caching resultset...
#> [2021-05-26 15:47:06]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 10
#>    concept_id concept_name              domain_id vocabulary_id concept_class_id
#>         <int> <chr>                     <chr>     <chr>         <chr>           
#>  1   41097667 clobutinol 2 MG/ML Oral … Drug      RxNorm Exten… Branded Drug Box
#>  2   19114026 gentamicin 3 MG/ML Ophth… Drug      RxNorm        Branded Drug    
#>  3   44183458 atorvastatin Delayed Rel… Drug      RxNorm Exten… Branded Drug Fo…
#>  4   44103090 Ascorbic Acid 500 MG / N… Drug      RxNorm Exten… Marketed Product
#>  5   41041139 lansoprazole 30 MG Delay… Drug      RxNorm Exten… Marketed Product
#>  6   40733290 Nicotine 0.875 MG/HR Tra… Drug      RxNorm Exten… Branded Drug Box
#>  7   41369009 10 ML Ampicillin 100 MG/… Drug      RxNorm Exten… Quant Clinical …
#>  8   41493784 40000 MG Ichtyolammonium… Drug      RxNorm Exten… Marketed Product
#>  9   43287804 Cetirizine 10 MG Oral Ta… Drug      RxNorm Exten… Branded Drug Box
#> 10    1728545 penicillin G procaine 30… Drug      RxNorm        Clinical Drug C…
#> # … with 5 more variables: standard_concept <chr>, concept_code <chr>,
#> #   valid_start_date <date>, valid_end_date <date>, invalid_reason <chr>

# HemOnc/ATC Classes
get_test_drug_classes(limit = 10)

#> [2021-05-26 15:47:06]	Connected to 'athena'
#> [2021-05-26 15:47:06]	Overriding cache... Querying Athena...
#> [2021-05-26 15:47:06]	✓ Open connection
#> [2021-05-26 15:47:06]	✓ JDBC connection
#> [2021-05-26 15:47:06]	SQL: SELECT * FROM omop_vocabulary.concept c WHERE c.invalid_reason IS NULL AND c.vocabulary_id IN ('ATC', 'HemOnc') AND c.standard_concept = 'C' AND c.domain_id = 'Drug' ORDER BY RANDOM() LIMIT 10
#> [2021-05-26 15:47:06]	Querying...
#> [2021-05-26 15:47:06]	Querying...complete
#> [2021-05-26 15:47:06]	✓ Returned data has more than 0 rows
#> [2021-05-26 15:47:06]	Caching resultset...
#> [2021-05-26 15:47:06]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 10
#>    concept_id concept_name              domain_id vocabulary_id concept_class_id
#>         <int> <chr>                     <chr>     <chr>         <chr>           
#>  1   45893522 cangrelor; parenteral     Drug      ATC           ATC 5th         
#>  2   21601019 Direct thrombin inhibito… Drug      ATC           ATC 4th         
#>  3   21602138 difluprednate; topical    Drug      ATC           ATC 5th         
#>  4   21603407 acetyldihydrocodeine; or… Drug      ATC           ATC 5th         
#>  5   21605204 CORTICOSTEROIDS AND ANTI… Drug      ATC           ATC 3rd         
#>  6   21602303 benzoyl peroxide, combin… Drug      ATC           ATC 5th         
#>  7   21602671 silodosin; oral           Drug      ATC           ATC 5th         
#>  8   21603915 daclizumab; parenteral    Drug      ATC           ATC 5th         
#>  9   21602954 sulfamethoxypyridazine; … Drug      ATC           ATC 5th         
#> 10   21601167 IRRIGATING SOLUTIONS      Drug      ATC           ATC 3rd         
#> # … with 5 more variables: standard_concept <chr>, concept_code <chr>,
#> #   valid_start_date <date>, valid_end_date <date>, invalid_reason <chr>

# Test data can also be taken from tables other than Concept
get_test_data(table = "concept_relationship")

#> [2021-05-26 15:47:06]	Connected to 'athena'
#> [2021-05-26 15:47:06]	Overriding cache... Querying Athena...
#> [2021-05-26 15:47:06]	✓ Open connection
#> [2021-05-26 15:47:06]	✓ JDBC connection
#> [2021-05-26 15:47:06]	SQL: SELECT * FROM omop_vocabulary.concept_relationship ORDER BY RANDOM() LIMIT 100
#> [2021-05-26 15:47:06]	Querying...
#> [2021-05-26 15:47:39]	Querying...complete
#> [2021-05-26 15:47:39]	✓ Returned data has more than 0 rows
#> [2021-05-26 15:47:39]	Caching resultset...
#> [2021-05-26 15:47:39]	Postgres connection to 'athena' closed

#> # A tibble: 100 x 6
#>    concept_id_1 concept_id_2 relationship_id     valid_start_date valid_end_date
#>           <int>        <int> <chr>               <date>           <date>        
#>  1     19001949     40948180 RxNorm dose form of 2017-07-18       2099-12-31    
#>  2     19082651     35161432 RxNorm dose form of 2019-03-11       2099-12-31    
#>  3     36233811     19023906 RxNorm inverse is a 2016-08-01       2099-12-31    
#>  4     46046149     46186987 AMP of              2015-01-31       2099-12-31    
#>  5      3018699      3018699 Maps to             1970-01-01       2099-12-31    
#>  6      4217142     40479221 Indir proc site of  2011-07-31       2099-12-31    
#>  7      4312219     36522127 Asso morph of       2015-09-18       2099-12-31    
#>  8      4153616     45984594 Subsumes            2015-01-31       2099-12-31    
#>  9      4038820     40484165 Property of         2011-07-31       2099-12-31    
#> 10     43708338     40817219 Has brand name      2017-08-24       2099-12-31    
#> # … with 90 more rows, and 1 more variable: invalid_reason <chr>

get_test_data(table = "concept_ancestor", limit = 10)

#> [2021-05-26 15:47:39]	Connected to 'athena'
#> [2021-05-26 15:47:39]	Overriding cache... Querying Athena...
#> [2021-05-26 15:47:39]	✓ Open connection
#> [2021-05-26 15:47:39]	✓ JDBC connection
#> [2021-05-26 15:47:39]	SQL: SELECT * FROM omop_vocabulary.concept_ancestor ORDER BY RANDOM() LIMIT 10
#> [2021-05-26 15:47:39]	Querying...
#> [2021-05-26 15:48:13]	Querying...complete
#> [2021-05-26 15:48:13]	✓ Returned data has more than 0 rows
#> [2021-05-26 15:48:13]	Caching resultset...
#> [2021-05-26 15:48:13]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 4
#>    ancestor_concept_… descendant_concep… min_levels_of_sepa… max_levels_of_sepa…
#>                 <int>              <int>               <int>               <int>
#>  1           36217216           40990327                   3                   4
#>  2           46300746           36263859                   2                   3
#>  3           46315316           41319091                   2                   3
#>  4            4034052           37115537                   7                   8
#>  5           45658257           21031187                   2                   2
#>  6            4169366            4169366                   0                   0
#>  7            4117930            4049968                   7                   9
#>  8            2801938            2898120                   1                   1
#>  9             964407           36234811                   1                   1
#> 10           46316385           40840334                   2                   2