Get Test Drug Classes

Get random records of ATC and HemOnc Drug Classes that are:

Valid,

Class,

Belong to the Drug domain

get_test_drug_classes(
  conn,
  conn_fun = "connectAthena()",
  schema = "omop_vocabulary",
  limit = 100,
  verbose = TRUE,
  render_sql = TRUE
)

See also

Examples

library(chariot)
library(tidyverse)

# Get Unfiltered Test Concepts (Invalid Concepts are included)
get_test_concepts()

#> Connecting using PostgreSQL driver

#> [2021-04-23 19:24:23]	Overriding cache... Querying Athena...
#> [2021-04-23 19:24:23] 

#> ✓ Open connection

#> [2021-04-23 19:24:23] 

#> ✓ JDBC connection

#> [2021-04-23 19:24:23]	SQL: SELECT * FROM omop_vocabulary.concept ORDER BY RANDOM() LIMIT 100
#> [2021-04-23 19:24:23]	Querying...
#> [2021-04-23 19:24:32]	Querying...complete
#> [2021-04-23 19:24:32] 

#> ✓ Returned data has more than 0 rows

#> [2021-04-23 19:24:32]	Caching resultset...
#> [2021-04-23 19:24:32]	Postgres connection to 'athena' closed

#> # A tibble: 100 x 10
#>    concept_id concept_name             domain_id  vocabulary_id concept_class_id
#>         <int> <chr>                    <chr>      <chr>         <chr>           
#>  1   40238927 psyllium 520 MG [Wal-Mu… Drug       RxNorm        Branded Drug Co…
#>  2    4047025 Blood group antigen Dr^… Observati… SNOMED        Substance       
#>  3   40491404 Ondansetron (as ondanse… Drug       SNOMED        Pharma/Biol Pro…
#>  4   36288998 Otrivine Duo 84/70 nas.… Drug       GGR           Drug Product    
#>  5    3077761 Retired procedure        Procedure  Nebraska Lex… Procedure       
#>  6   21497945 HLA-B52(5) | donor | Bl… Measureme… LOINC         LOINC Hierarchy 
#>  7    2855901 Medical and Surgical @ … Procedure  ICD10PCS      ICD10PCS Hierar…
#>  8   40819538 Walnuss                  Drug       RxNorm Exten… Brand Name      
#>  9   44206980 13800 MG lauromacrogols… Drug       RxNorm Exten… Quant Clinical …
#> 10   46257956 Coal tar solution 5% / … Drug       SNOMED        Pharma/Biol Pro…
#> # … with 90 more rows, and 5 more variables: standard_concept <chr>,
#> #   concept_code <chr>, valid_start_date <date>, valid_end_date <date>,
#> #   invalid_reason <chr>

# The size of the resultset can be altered
get_test_concepts(limit = 25)

#> Connecting using PostgreSQL driver

#> [2021-04-23 19:24:32]	Overriding cache... Querying Athena...
#> [2021-04-23 19:24:32] 

#> ✓ Open connection

#> [2021-04-23 19:24:32] 

#> ✓ JDBC connection

#> [2021-04-23 19:24:32]	SQL: SELECT * FROM omop_vocabulary.concept ORDER BY RANDOM() LIMIT 25
#> [2021-04-23 19:24:32]	Querying...
#> [2021-04-23 19:24:39]	Querying...complete
#> [2021-04-23 19:24:39] 

#> ✓ Returned data has more than 0 rows

#> [2021-04-23 19:24:39]	Caching resultset...
#> [2021-04-23 19:24:39]	Postgres connection to 'athena' closed

#> # A tibble: 25 x 10
#>    concept_id concept_name             domain_id  vocabulary_id concept_class_id
#>         <int> <chr>                    <chr>      <chr>         <chr>           
#>  1   42088958 BONEMEDIK-S              Device     EDI           Device          
#>  2   46319645 SINUS CONGESTION AND PA… Drug       SPL           OTC Drug        
#>  3   36723913 APC Transcript Variant:… Measureme… OMOP Extensi… Transcript Vari…
#>  4   43370583 Tobramycin               Drug       AMT           AU Substance    
#>  5   21387479 Haddenham Veni class 2 … Device     dm+d          AMPP            
#>  6   41422376 100 ML Naphazoline 0.00… Drug       RxNorm Exten… Marketed Product
#>  7   42156044 Simple Tendon Suture,(S… Procedure  EDI           Procedure       
#>  8   44800289 [X]Other manic episodes  Condition  SNOMED        Clinical Finding
#>  9   42643345 oxygen 99L/100L RESPIRA… Drug       SPL           Prescription Dr…
#> 10   43812367 250 ML Mannitol 150 MG/… Drug       RxNorm Exten… Quant Branded D…
#> # … with 15 more rows, and 5 more variables: standard_concept <chr>,
#> #   concept_code <chr>, valid_start_date <date>, valid_end_date <date>,
#> #   invalid_reason <chr>

# Concepts specific to the Drug domain can be sampled
# RxNorm/RxNorm Extension Concepts
get_test_drug_concepts(limit = 10)

#> Connecting using PostgreSQL driver

#> [2021-04-23 19:24:40]	Overriding cache... Querying Athena...
#> [2021-04-23 19:24:40] 

#> ✓ Open connection

#> [2021-04-23 19:24:40] 

#> ✓ JDBC connection

#> [2021-04-23 19:24:40]	SQL: SELECT * FROM omop_vocabulary.concept c WHERE c.invalid_reason IS NULL AND c.vocabulary_id IN ('RxNorm', 'RxNorm Extension') AND c.standard_concept <> 'C' AND c.domain_id = 'Drug' ORDER BY RANDOM() LIMIT 10
#> [2021-04-23 19:24:40]	Querying...
#> [2021-04-23 19:24:44]	Querying...complete
#> [2021-04-23 19:24:44] 

#> ✓ Returned data has more than 0 rows

#> [2021-04-23 19:24:44]	Caching resultset...
#> [2021-04-23 19:24:44]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 10
#>    concept_id concept_name              domain_id vocabulary_id concept_class_id
#>         <int> <chr>                     <chr>     <chr>         <chr>           
#>  1   19035396 reproterol                Drug      RxNorm        Ingredient      
#>  2   37593113 Calcium 500 MG Chewable … Drug      RxNorm Exten… Marketed Product
#>  3    2011383 4-aminomethylbenzoic aci… Drug      RxNorm Exten… Branded Drug Fo…
#>  4   43763928 Naloxone 6.11 MG/ML / Ti… Drug      RxNorm Exten… Branded Drug Co…
#>  5   19107561 fluocortolone 50 MG Oral… Drug      RxNorm        Clinical Drug   
#>  6   44050899 Diphenhydramine 0.02 MG/… Drug      RxNorm Exten… Marketed Product
#>  7   35413965 20 ML Mepivacaine 20 MG/… Drug      RxNorm Exten… Marketed Product
#>  8   19044285 chloramphenicol 5 MG/ML … Drug      RxNorm        Branded Drug    
#>  9   43030597 Octreotide 500 MG/ML [Sa… Drug      RxNorm Exten… Branded Drug Co…
#> 10   41359127 5 ML brimonidine 2 MG/ML… Drug      RxNorm Exten… Quant Branded B…
#> # … with 5 more variables: standard_concept <chr>, concept_code <chr>,
#> #   valid_start_date <date>, valid_end_date <date>, invalid_reason <chr>

# HemOnc/ATC Classes
get_test_drug_classes(limit = 10)

#> Connecting using PostgreSQL driver

#> [2021-04-23 19:24:44]	Overriding cache... Querying Athena...
#> [2021-04-23 19:24:44] 

#> ✓ Open connection

#> [2021-04-23 19:24:44] 

#> ✓ JDBC connection

#> [2021-04-23 19:24:44]	SQL: SELECT * FROM omop_vocabulary.concept c WHERE c.invalid_reason IS NULL AND c.vocabulary_id IN ('ATC', 'HemOnc') AND c.standard_concept = 'C' AND c.domain_id = 'Drug' ORDER BY RANDOM() LIMIT 10
#> [2021-04-23 19:24:44]	Querying...
#> [2021-04-23 19:24:44]	Querying...complete
#> [2021-04-23 19:24:44] 

#> ✓ Returned data has more than 0 rows

#> [2021-04-23 19:24:44]	Caching resultset...
#> [2021-04-23 19:24:44]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 10
#>    concept_id concept_name            domain_id vocabulary_id concept_class_id
#>         <int> <chr>                   <chr>     <chr>         <chr>           
#>  1   21603764 ofatumumab              Drug      ATC           ATC 5th         
#>  2   21603924 ciclosporin; systemic   Drug      ATC           ATC 5th         
#>  3   21601711 atenolol and thiazides  Drug      ATC           ATC 5th         
#>  4   21602380 chlorquinaldol; vaginal Drug      ATC           ATC 5th         
#>  5   21604769 meclofenoxate           Drug      ATC           ATC 5th         
#>  6   21604440 lacosamide; systemic    Drug      ATC           ATC 5th         
#>  7   21602405 oxiconazole; vaginal    Drug      ATC           ATC 5th         
#>  8   21601656 diosmin                 Drug      ATC           ATC 5th         
#>  9   44819163 ocriplasmin; ophthalmic Drug      ATC           ATC 5th         
#> 10   43534790 mirabegron; oral        Drug      ATC           ATC 5th         
#> # … with 5 more variables: standard_concept <chr>, concept_code <chr>,
#> #   valid_start_date <date>, valid_end_date <date>, invalid_reason <chr>

# Test data can also be taken from tables other than Concept
get_test_data(table = "concept_relationship")

#> Connecting using PostgreSQL driver

#> [2021-04-23 19:24:45]	Overriding cache... Querying Athena...
#> [2021-04-23 19:24:45] 

#> ✓ Open connection

#> [2021-04-23 19:24:45] 

#> ✓ JDBC connection

#> [2021-04-23 19:24:45]	SQL: SELECT * FROM omop_vocabulary.concept_relationship ORDER BY RANDOM() LIMIT 100
#> [2021-04-23 19:24:45]	Querying...
#> [2021-04-23 19:25:17]	Querying...complete
#> [2021-04-23 19:25:17] 

#> ✓ Returned data has more than 0 rows

#> [2021-04-23 19:25:17]	Caching resultset...
#> [2021-04-23 19:25:17]	Postgres connection to 'athena' closed

#> # A tibble: 100 x 6
#>    concept_id_1 concept_id_2 relationship_id   valid_start_date valid_end_date
#>           <int>        <int> <chr>             <date>           <date>        
#>  1      2781537      4041249 Is a              1970-01-01       2099-12-31    
#>  2     35909706     35941821 Schema to Value   1970-01-01       2099-12-31    
#>  3      1019594     42529432 System of         2019-12-13       2099-12-31    
#>  4     45638510     40231262 SPL - RxNorm      1970-01-01       2099-12-31    
#>  5     19023563     40945388 Has marketed form 2017-07-18       2099-12-31    
#>  6     42705480     36888892 Constitutes       2017-06-24       2099-12-31    
#>  7     43751768     43600818 Marketed form of  2017-06-24       2099-12-31    
#>  8      4116830      4313834 Occurrence of     2011-07-31       2099-12-31    
#>  9     21106528     21019960 Has supplier      2015-08-17       2099-12-31    
#> 10      2743316      4214980 Is a              1970-01-01       2099-12-31    
#> # … with 90 more rows, and 1 more variable: invalid_reason <chr>

get_test_data(table = "concept_ancestor", limit = 10)

#> Connecting using PostgreSQL driver

#> [2021-04-23 19:25:17]	Overriding cache... Querying Athena...
#> [2021-04-23 19:25:17] 

#> ✓ Open connection

#> [2021-04-23 19:25:17] 

#> ✓ JDBC connection

#> [2021-04-23 19:25:17]	SQL: SELECT * FROM omop_vocabulary.concept_ancestor ORDER BY RANDOM() LIMIT 10
#> [2021-04-23 19:25:17]	Querying...
#> [2021-04-23 19:25:50]	Querying...complete
#> [2021-04-23 19:25:50] 

#> ✓ Returned data has more than 0 rows

#> [2021-04-23 19:25:50]	Caching resultset...
#> [2021-04-23 19:25:50]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 4
#>    ancestor_concept_… descendant_concep… min_levels_of_sepa… max_levels_of_sepa…
#>                 <int>              <int>               <int>               <int>
#>  1           45692516           41185350                   2                   2
#>  2             793320           36810294                   2                   2
#>  3           35507192           41319804                   2                   3
#>  4           45699186           35849384                   1                   1
#>  5           21122377           21122377                   0                   0
#>  6           40181754           35770149                   2                   2
#>  7           46319646           43283219                   1                   1
#>  8           45620297           43857844                   2                   2
#>  9           36152562           40166433                   1                   1
#> 10           45635806           40039439                   1                   2