Get Test Drug Classes

Get random records of ATC and HemOnc Drug Classes that are:

Valid,

Class,

Belong to the Drug domain

get_test_drug_classes(
  conn,
  conn_fun = "connectAthena()",
  schema = "omop_vocabulary",
  limit = 100,
  verbose = TRUE,
  render_sql = TRUE
)

See also

Examples

library(chariot)
library(tidyverse)

# Get Unfiltered Test Concepts (Invalid Concepts are included)
get_test_concepts()

#> Connecting using PostgreSQL driver
#> [2021-01-10 16:33:33]	Overriding cache... Querying Athena...
#> [2021-01-10 16:33:33] 

#> ✓ Open connection

#> [2021-01-10 16:33:33] 

#> ✓ JDBC connection

#> [2021-01-10 16:33:33]	SQL: SELECT * FROM omop_vocabulary.concept ORDER BY RANDOM() LIMIT 100
#> [2021-01-10 16:33:33]	Querying...
#> [2021-01-10 16:33:42]	Querying...complete
#> [2021-01-10 16:33:42] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 16:33:42]	Caching resultset...
#> [2021-01-10 16:33:42]	Postgres connection to 'athena' closed

#> # A tibble: 100 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id
#>         <dbl> <chr>        <chr>     <chr>         <chr>           
#>  1   45873006 clobetasol … Drug      NDC           11-digit NDC    
#>  2   40734783 Fluoride Io… Drug      RxNorm Exten… Branded Drug Co…
#>  3    1433397 Adverse eff… Observat… ICD10CN       ICD10 Hierarchy 
#>  4   46357809 ACNEZZOL BA… Drug      NDC           11-digit NDC    
#>  5   43815058 Camphor Vag… Drug      RxNorm Exten… Branded Drug Fo…
#>  6   36163308 arnica rosa… Drug      SPL           OTC Drug        
#>  7   44802968 Preferred p… Observat… SNOMED        Clinical Finding
#>  8   41519059 200000 MG T… Drug      RxNorm Exten… Marketed Product
#>  9   44034307 Marsdenia c… Drug      RxNorm Exten… Branded Drug    
#> 10   44040700 Tobramycin … Drug      RxNorm Exten… Branded Drug Co…
#> # … with 90 more rows, and 5 more variables: standard_concept <chr>,
#> #   concept_code <chr>, valid_start_date <date>, valid_end_date <date>,
#> #   invalid_reason <chr>

# The size of the resultset can be altered
get_test_concepts(limit = 25)

#> Connecting using PostgreSQL driver
#> [2021-01-10 16:33:42]	Overriding cache... Querying Athena...
#> [2021-01-10 16:33:42] 

#> ✓ Open connection

#> [2021-01-10 16:33:42] 

#> ✓ JDBC connection

#> [2021-01-10 16:33:42]	SQL: SELECT * FROM omop_vocabulary.concept ORDER BY RANDOM() LIMIT 25
#> [2021-01-10 16:33:42]	Querying...
#> [2021-01-10 16:33:51]	Querying...complete
#> [2021-01-10 16:33:51] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 16:33:51]	Caching resultset...
#> [2021-01-10 16:33:51]	Postgres connection to 'athena' closed

#> # A tibble: 25 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id
#>         <dbl> <chr>        <chr>     <chr>         <chr>           
#>  1   40565520 "Foot injur… Condition SNOMED        Clinical Finding
#>  2   42987121 "GLOCILLIN … Drug      KDC           Brand Name      
#>  3   42050608 "Sedgeford"  Geography OSM           10th level      
#>  4   42870202 "My child h… Observat… LOINC         Survey          
#>  5   41131792 "Aspirin 50… Drug      RxNorm Exten… Branded Drug    
#>  6    1423037 "Shock foll… Condition ICD10CN       ICD10 code      
#>  7   36532086 "Carcinosar… Condition ICDO3         ICDO Condition  
#>  8     576725 "SILVER 4\"… Drug      GCN_SEQNO     GCN_SEQNO       
#>  9   44151534 "Priva-Tada… Drug      DPD           Drug Product    
#> 10   40233539 "Broncotron… Drug      RxNorm        Brand Name      
#> # … with 15 more rows, and 5 more variables: standard_concept <chr>,
#> #   concept_code <chr>, valid_start_date <date>, valid_end_date <date>,
#> #   invalid_reason <chr>

# Concepts specific to the Drug domain can be sampled
# RxNorm/RxNorm Extension Concepts
get_test_drug_concepts(limit = 10)

#> Connecting using PostgreSQL driver
#> [2021-01-10 16:33:51]	Overriding cache... Querying Athena...
#> [2021-01-10 16:33:51] 

#> ✓ Open connection

#> [2021-01-10 16:33:51] 

#> ✓ JDBC connection

#> [2021-01-10 16:33:51]	SQL: SELECT * FROM omop_vocabulary.concept c WHERE c.invalid_reason IS NULL AND c.vocabulary_id IN ('RxNorm', 'RxNorm Extension') AND c.standard_concept <> 'C' AND c.domain_id = 'Drug' ORDER BY RANDOM() LIMIT 10
#> [2021-01-10 16:33:51]	Querying...
#> [2021-01-10 16:33:54]	Querying...complete
#> [2021-01-10 16:33:54] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 16:33:54]	Caching resultset...
#> [2021-01-10 16:33:54]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id
#>         <dbl> <chr>        <chr>     <chr>         <chr>           
#>  1   43138739 Glycerin 20… Drug      RxNorm Exten… Branded Drug Box
#>  2   41160747 Cimetidine … Drug      RxNorm Exten… Branded Drug    
#>  3   40707148 1010 MG Bet… Drug      RxNorm Exten… Quant Clinical …
#>  4   41259322 Methotrexat… Drug      RxNorm Exten… Marketed Product
#>  5   21114292 pioglitazon… Drug      RxNorm Exten… Marketed Product
#>  6   44048346 Bitter oran… Drug      RxNorm Exten… Branded Drug    
#>  7   41214293 trandolapri… Drug      RxNorm Exten… Branded Drug Box
#>  8   35606044 dexmedetomi… Drug      RxNorm        Clinical Drug C…
#>  9     782710 Thiotepa 15… Drug      RxNorm Exten… Marketed Product
#> 10   43169952 Risperidone… Drug      RxNorm Exten… Branded Drug Box
#> # … with 5 more variables: standard_concept <chr>, concept_code <chr>,
#> #   valid_start_date <date>, valid_end_date <date>, invalid_reason <chr>

# HemOnc/ATC Classes
get_test_drug_classes(limit = 10)

#> Connecting using PostgreSQL driver
#> [2021-01-10 16:33:54]	Overriding cache... Querying Athena...
#> [2021-01-10 16:33:54] 

#> ✓ Open connection

#> [2021-01-10 16:33:54] 

#> ✓ JDBC connection

#> [2021-01-10 16:33:54]	SQL: SELECT * FROM omop_vocabulary.concept c WHERE c.invalid_reason IS NULL AND c.vocabulary_id IN ('ATC', 'HemOnc') AND c.standard_concept = 'C' AND c.domain_id = 'Drug' ORDER BY RANDOM() LIMIT 10
#> [2021-01-10 16:33:54]	Querying...
#> [2021-01-10 16:33:55]	Querying...complete
#> [2021-01-10 16:33:55] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 16:33:55]	Caching resultset...
#> [2021-01-10 16:33:55]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id
#>         <dbl> <chr>        <chr>     <chr>         <chr>           
#>  1   21603140 valganciclo… Drug      ATC           ATC 5th         
#>  2   21602314 azelaic aci… Drug      ATC           ATC 5th         
#>  3    1123893 bisoprolol … Drug      ATC           ATC 5th         
#>  4   21602762 propylthiou… Drug      ATC           ATC 5th         
#>  5   21602981 dirithromyc… Drug      ATC           ATC 5th         
#>  6     715851 indacaterol… Drug      ATC           ATC 5th         
#>  7   43534830 fluoxetine … Drug      ATC           ATC 5th         
#>  8   21600226 camylofin a… Drug      ATC           ATC 5th         
#>  9   21600296 ibopamine; … Drug      ATC           ATC 5th         
#> 10   21602414 nifuratel; … Drug      ATC           ATC 5th         
#> # … with 5 more variables: standard_concept <chr>, concept_code <chr>,
#> #   valid_start_date <date>, valid_end_date <date>, invalid_reason <chr>

# Test data can also be taken from tables other than Concept
get_test_data(table = "concept_relationship")

#> Connecting using PostgreSQL driver
#> [2021-01-10 16:33:55]	Overriding cache... Querying Athena...
#> [2021-01-10 16:33:55] 

#> ✓ Open connection

#> [2021-01-10 16:33:55] 

#> ✓ JDBC connection

#> [2021-01-10 16:33:55]	SQL: SELECT * FROM omop_vocabulary.concept_relationship ORDER BY RANDOM() LIMIT 100
#> [2021-01-10 16:33:55]	Querying...
#> [2021-01-10 16:34:33]	Querying...complete
#> [2021-01-10 16:34:33] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 16:34:33]	Caching resultset...
#> [2021-01-10 16:34:33]	Postgres connection to 'athena' closed

#> # A tibble: 100 x 6
#>    concept_id_1 concept_id_2 relationship_id valid_start_date valid_end_date
#>           <dbl>        <dbl> <chr>           <date>           <date>        
#>  1     41347128     36895261 Tradename of    2017-07-18       2099-12-31    
#>  2     42728886     40228386 Consists of     2017-07-30       2099-12-31    
#>  3     41067929     43126126 Has brand name  2017-08-24       2099-12-31    
#>  4      3376133      3162733 Concept poss_e… 2019-08-16       2099-12-31    
#>  5     40022062     42902233 RxNorm inverse… 2014-12-01       2099-12-31    
#>  6     40380801       434866 Concept replac… 2013-12-02       2099-12-31    
#>  7     41125374     21152565 Tradename of    2017-07-18       2099-12-31    
#>  8     43619047     43619047 Maps to         2016-12-18       2099-12-31    
#>  9     21019747     21124402 Supplier of     2015-08-17       2099-12-31    
#> 10     35626975     46090966 VMP prescr sta… 2018-07-31       2099-12-31    
#> # … with 90 more rows, and 1 more variable: invalid_reason <chr>

get_test_data(table = "concept_ancestor", limit = 10)

#> Connecting using PostgreSQL driver
#> [2021-01-10 16:34:33]	Overriding cache... Querying Athena...
#> [2021-01-10 16:34:33] 

#> ✓ Open connection

#> [2021-01-10 16:34:33] 

#> ✓ JDBC connection

#> [2021-01-10 16:34:33]	SQL: SELECT * FROM omop_vocabulary.concept_ancestor ORDER BY RANDOM() LIMIT 10
#> [2021-01-10 16:34:33]	Querying...
#> [2021-01-10 16:35:11]	Querying...complete
#> [2021-01-10 16:35:11] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 16:35:11]	Caching resultset...
#> [2021-01-10 16:35:11]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 4
#>    ancestor_concept_… descendant_concep… min_levels_of_sepa… max_levels_of_sepa…
#>                 <dbl>              <dbl>               <dbl>               <dbl>
#>  1            4211641            2789595                   3                   6
#>  2           45665249           41131262                   2                   2
#>  3            4322976           37394822                   4                  10
#>  4            4042142            4170297                   1                   1
#>  5           45673105           41059524                   2                   2
#>  6            1548195           41333434                   3                   4
#>  7           21602796           41401877                   6                   7
#>  8           44092574           44072669                   1                   1
#>  9           37209201           46057415                   2                   3
#> 10           21600080             961193                   3                   3