Get Test Drug Classes

Get random records of ATC and HemOnc Drug Classes that are:

Valid,

Class,

Belong to the Drug domain

get_test_drug_classes(
  conn,
  conn_fun = "connectAthena()",
  schema = "omop_vocabulary",
  limit = 100,
  verbose = TRUE,
  render_sql = TRUE
)

See also

Examples

library(chariot)
library(tidyverse)

# Get Unfiltered Test Concepts (Invalid Concepts are included)
get_test_concepts()

#> Connecting using PostgreSQL driver

#> [2021-03-28 18:48:19]	Overriding cache... Querying Athena...
#> [2021-03-28 18:48:19] 

#> ✓ Open connection

#> [2021-03-28 18:48:19] 

#> ✓ JDBC connection

#> [2021-03-28 18:48:19]	SQL: SELECT * FROM omop_vocabulary.concept ORDER BY RANDOM() LIMIT 100
#> [2021-03-28 18:48:19]	Querying...
#> [2021-03-28 18:48:29]	Querying...complete
#> [2021-03-28 18:48:29] 

#> ✓ Returned data has more than 0 rows

#> [2021-03-28 18:48:29]	Caching resultset...
#> [2021-03-28 18:48:29]	Postgres connection to 'athena' closed

#> # A tibble: 100 x 10
#>    concept_id concept_name              domain_id vocabulary_id concept_class_id
#>         <int> <chr>                     <chr>     <chr>         <chr>           
#>  1   42836059 menthol 1.25g/100mL TOPI… Drug      NDC           9-digit NDC     
#>  2   41961044 Schweinfurt               Geography OSM           6th level       
#>  3   45163384 Ceftriaxone 500 MG Injec… Drug      NDC           11-digit NDC    
#>  4   40027400 cromolyn Oral Capsule     Drug      RxNorm        Clinical Drug F…
#>  5     924939 bisacodyl                 Drug      RxNorm        Ingredient      
#>  6   41315416 desoxycorticosterone 10 … Drug      RxNorm Exten… Branded Drug    
#>  7   41142339 Betamethasone 0.004 MG/M… Drug      RxNorm Exten… Branded Drug Co…
#>  8   40929262 Humulus lupulus whole ex… Drug      RxNorm Exten… Clinical Drug   
#>  9   21053699 Allopurinol 100 MG Oral … Drug      RxNorm Exten… Clinical Drug B…
#> 10   21130863 Aminophylline 100 MG Ext… Drug      RxNorm Exten… Branded Drug Box
#> # … with 90 more rows, and 5 more variables: standard_concept <chr>,
#> #   concept_code <chr>, valid_start_date <date>, valid_end_date <date>,
#> #   invalid_reason <chr>

# The size of the resultset can be altered
get_test_concepts(limit = 25)

#> Connecting using PostgreSQL driver

#> [2021-03-28 18:48:29]	Overriding cache... Querying Athena...
#> [2021-03-28 18:48:29] 

#> ✓ Open connection

#> [2021-03-28 18:48:29] 

#> ✓ JDBC connection

#> [2021-03-28 18:48:29]	SQL: SELECT * FROM omop_vocabulary.concept ORDER BY RANDOM() LIMIT 25
#> [2021-03-28 18:48:29]	Querying...
#> [2021-03-28 18:48:37]	Querying...complete
#> [2021-03-28 18:48:37] 

#> ✓ Returned data has more than 0 rows

#> [2021-03-28 18:48:37]	Caching resultset...
#> [2021-03-28 18:48:37]	Postgres connection to 'athena' closed

#> # A tibble: 25 x 10
#>    concept_id concept_name             domain_id  vocabulary_id concept_class_id
#>         <int> <chr>                    <chr>      <chr>         <chr>           
#>  1   41040152 olanzapine 10 MG Oral T… Drug       RxNorm Exten… Marketed Product
#>  2    2888655 Medical and Surgical @ … Procedure  ICD10PCS      ICD10PCS Hierar…
#>  3   46339102 ISAKNOX AGELESS SERUM B… Drug       NDC           11-digit NDC    
#>  4   44983590 Propranolol Hydrochlori… Drug       NDC           11-digit NDC    
#>  5    3404623 Excision vocal cord muc… Procedure  Nebraska Lex… Procedure       
#>  6   40802045 Biovital Magen           Drug       RxNorm Exten… Brand Name      
#>  7    1033536 Stress study             Observati… LOINC         LOINC System    
#>  8   44779512 allantoin 10mg/g TOPICA… Drug       SPL           OTC Drug        
#>  9   45648389 ALLERGY AND CONGESTION … Drug       SPL           OTC Drug        
#> 10   37064530 House dust Bencard       Observati… LOINC         LOINC Component 
#> # … with 15 more rows, and 5 more variables: standard_concept <chr>,
#> #   concept_code <chr>, valid_start_date <date>, valid_end_date <date>,
#> #   invalid_reason <chr>

# Concepts specific to the Drug domain can be sampled
# RxNorm/RxNorm Extension Concepts
get_test_drug_concepts(limit = 10)

#> Connecting using PostgreSQL driver

#> [2021-03-28 18:48:37]	Overriding cache... Querying Athena...
#> [2021-03-28 18:48:37] 

#> ✓ Open connection

#> [2021-03-28 18:48:37] 

#> ✓ JDBC connection

#> [2021-03-28 18:48:37]	SQL: SELECT * FROM omop_vocabulary.concept c WHERE c.invalid_reason IS NULL AND c.vocabulary_id IN ('RxNorm', 'RxNorm Extension') AND c.standard_concept <> 'C' AND c.domain_id = 'Drug' ORDER BY RANDOM() LIMIT 10
#> [2021-03-28 18:48:37]	Querying...
#> [2021-03-28 18:48:42]	Querying...complete
#> [2021-03-28 18:48:42] 

#> ✓ Returned data has more than 0 rows

#> [2021-03-28 18:48:42]	Caching resultset...
#> [2021-03-28 18:48:42]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 10
#>    concept_id concept_name              domain_id vocabulary_id concept_class_id
#>         <int> <chr>                     <chr>     <chr>         <chr>           
#>  1   41434463 200 ML iopromide 625 MG/… Drug      RxNorm Exten… Quant Clinical …
#>  2   40956209 Potassium Oral Tablet [D… Drug      RxNorm Exten… Branded Drug Fo…
#>  3   43588141 Levofloxacin 500 MG Oral… Drug      RxNorm Exten… Marketed Product
#>  4   41166050 Medigoxin 0.1 MG Oral Ta… Drug      RxNorm Exten… Marketed Product
#>  5   43773624 50 ML Clotrimazole 10 MG… Drug      RxNorm Exten… Marketed Product
#>  6   40845290 Hydrochlorothiazide 25 M… Drug      RxNorm Exten… Branded Drug Box
#>  7   35160693 olopatadine 5 MG Oral Ta… Drug      RxNorm Exten… Marketed Product
#>  8   43181919 15000 ML Nitrous Oxide 1… Drug      RxNorm Exten… Marketed Product
#>  9   40868251 Clotrimazole 10 MG/ML To… Drug      RxNorm Exten… Clinical Drug B…
#> 10   21079412 500 ML Lactulose 667 MG/… Drug      RxNorm Exten… Marketed Product
#> # … with 5 more variables: standard_concept <chr>, concept_code <chr>,
#> #   valid_start_date <date>, valid_end_date <date>, invalid_reason <chr>

# HemOnc/ATC Classes
get_test_drug_classes(limit = 10)

#> Connecting using PostgreSQL driver

#> [2021-03-28 18:48:42]	Overriding cache... Querying Athena...
#> [2021-03-28 18:48:42] 

#> ✓ Open connection

#> [2021-03-28 18:48:42] 

#> ✓ JDBC connection

#> [2021-03-28 18:48:42]	SQL: SELECT * FROM omop_vocabulary.concept c WHERE c.invalid_reason IS NULL AND c.vocabulary_id IN ('ATC', 'HemOnc') AND c.standard_concept = 'C' AND c.domain_id = 'Drug' ORDER BY RANDOM() LIMIT 10
#> [2021-03-28 18:48:42]	Querying...
#> [2021-03-28 18:48:42]	Querying...complete
#> [2021-03-28 18:48:42] 

#> ✓ Returned data has more than 0 rows

#> [2021-03-28 18:48:42]	Caching resultset...
#> [2021-03-28 18:48:42]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 10
#>    concept_id concept_name              domain_id vocabulary_id concept_class_id
#>         <int> <chr>                     <chr>     <chr>         <chr>           
#>  1   21602575 megestrol and estrogen    Drug      ATC           ATC 5th         
#>  2   43534858 ciprofloxacin; ophthalmic Drug      ATC           ATC 5th         
#>  3   21600662 Aminosalicylic acid and … Drug      ATC           ATC 4th         
#>  4   21603794 mitoguazone               Drug      ATC           ATC 5th         
#>  5   21601549 cyclopenthiazide and pot… Drug      ATC           ATC 5th         
#>  6   21600961 ANTITHROMBOTIC AGENTS     Drug      ATC           ATC 3rd         
#>  7   21603377 Expectorants - combinati… Drug      ATC           ATC 5th         
#>  8   21604325 acetylsalicylic acid, co… Drug      ATC           ATC 5th         
#>  9   45893515 donepezil, memantine and… Drug      ATC           ATC 5th         
#> 10   21603379 Mucolytics                Drug      ATC           ATC 4th         
#> # … with 5 more variables: standard_concept <chr>, concept_code <chr>,
#> #   valid_start_date <date>, valid_end_date <date>, invalid_reason <chr>

# Test data can also be taken from tables other than Concept
get_test_data(table = "concept_relationship")

#> Connecting using PostgreSQL driver

#> [2021-03-28 18:48:42]	Overriding cache... Querying Athena...
#> [2021-03-28 18:48:42] 

#> ✓ Open connection

#> [2021-03-28 18:48:42] 

#> ✓ JDBC connection

#> [2021-03-28 18:48:42]	SQL: SELECT * FROM omop_vocabulary.concept_relationship ORDER BY RANDOM() LIMIT 100
#> [2021-03-28 18:48:42]	Querying...
#> [2021-03-28 18:49:16]	Querying...complete
#> [2021-03-28 18:49:16] 

#> ✓ Returned data has more than 0 rows

#> [2021-03-28 18:49:16]	Caching resultset...
#> [2021-03-28 18:49:16]	Postgres connection to 'athena' closed

#> # A tibble: 100 x 6
#>    concept_id_1 concept_id_2 relationship_id     valid_start_date valid_end_date
#>           <int>        <int> <chr>               <date>           <date>        
#>  1     35909698     35940765 Schema to Value     1970-01-01       2099-12-31    
#>  2     41428260     41428260 Maps to             2017-07-18       2099-12-31    
#>  3      4096633      4100074 Has asso morph      2011-07-31       2099-12-31    
#>  4     40168509     42903052 Quantified form of  2014-12-01       2099-12-31    
#>  5      3223001      3415392 Method of           2019-08-16       2099-12-31    
#>  6     44498694     36535828 Histology of ICDO   2015-09-18       2099-12-31    
#>  7     44414278     40175394 Maps to             2017-06-01       2099-12-31    
#>  8     36709194     35626975 VMP has prescr stat 2018-07-31       2099-12-31    
#>  9     37209199     46212630 Subsumes            2019-07-31       2099-12-31    
#> 10     46215497     46215497 Mapped from         2018-07-31       2099-12-31    
#> # … with 90 more rows, and 1 more variable: invalid_reason <chr>

get_test_data(table = "concept_ancestor", limit = 10)

#> Connecting using PostgreSQL driver

#> [2021-03-28 18:49:16]	Overriding cache... Querying Athena...
#> [2021-03-28 18:49:16] 

#> ✓ Open connection

#> [2021-03-28 18:49:16] 

#> ✓ JDBC connection

#> [2021-03-28 18:49:16]	SQL: SELECT * FROM omop_vocabulary.concept_ancestor ORDER BY RANDOM() LIMIT 10
#> [2021-03-28 18:49:16]	Querying...
#> [2021-03-28 18:49:50]	Querying...complete
#> [2021-03-28 18:49:50] 

#> ✓ Returned data has more than 0 rows

#> [2021-03-28 18:49:50]	Caching resultset...
#> [2021-03-28 18:49:50]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 4
#>    ancestor_concept_… descendant_concep… min_levels_of_sepa… max_levels_of_sepa…
#>                 <int>              <int>               <int>               <int>
#>  1           43050991           43050991                   0                   0
#>  2           36238530           41010700                   2                   4
#>  3           35202736           41185435                   2                   2
#>  4           35953449           35953449                   0                   0
#>  5            4038502            4128624                   3                   3
#>  6            4156941           36522730                   4                   4
#>  7            4102111           36554292                   6                  12
#>  8           21602681           41412768                   7                   9
#>  9           45628377           43740086                   2                   2
#> 10           37049326           21491198                   1                   1