Get Test Drug Classes

Get random records of ATC and HemOnc Drug Classes that are:

Valid,

Class,

Belong to the Drug domain

get_test_drug_classes(
  conn,
  conn_fun = "connectAthena()",
  schema = "omop_vocabulary",
  limit = 100,
  verbose = TRUE,
  render_sql = TRUE
)

See also

Examples

library(chariot)
library(tidyverse)

# Get Unfiltered Test Concepts (Invalid Concepts are included)
get_test_concepts()

#> Connecting using PostgreSQL driver

#> [2021-03-28 23:40:29]	Overriding cache... Querying Athena...
#> [2021-03-28 23:40:29] 

#> ✓ Open connection

#> [2021-03-28 23:40:29] 

#> ✓ JDBC connection

#> [2021-03-28 23:40:29]	SQL: SELECT * FROM omop_vocabulary.concept ORDER BY RANDOM() LIMIT 100
#> [2021-03-28 23:40:29]	Querying...
#> [2021-03-28 23:40:37]	Querying...complete
#> [2021-03-28 23:40:37] 

#> ✓ Returned data has more than 0 rows

#> [2021-03-28 23:40:37]	Caching resultset...
#> [2021-03-28 23:40:37]	Postgres connection to 'athena' closed

#> # A tibble: 100 x 10
#>    concept_id concept_name             domain_id  vocabulary_id concept_class_id
#>         <int> <chr>                    <chr>      <chr>         <chr>           
#>  1   40936074 Nystatin 0.0227 MG/MG /… Drug       RxNorm Exten… Branded Drug Box
#>  2   46013646 Venosan 5000 class 3 (5… Device     SNOMED        Physical Object 
#>  3   44549319 WARFARIN SODIUM ORAL TA… Drug       DPD           Branded Drug Fo…
#>  4   37384717 Albutrepenonacog alfa 1… Drug       SNOMED        Pharma/Biol Pro…
#>  5   40486894 Measurement of Influenz… Measureme… SNOMED        Procedure       
#>  6   44449418 temozolomide 20 MG Oral… Drug       NDC           11-digit NDC    
#>  7   42842636 aluminum zirconium tric… Drug       NDC           11-digit NDC    
#>  8    3369133 Manis gigantea           Observati… Nebraska Lex… Organism        
#>  9    3079262 Gonadotrophin-releasing… Drug       Nebraska Lex… Pharma/Biol Pro…
#> 10   21411208 Jobst UlcerCARE medical… Device     dm+d          AMP             
#> # … with 90 more rows, and 5 more variables: standard_concept <chr>,
#> #   concept_code <chr>, valid_start_date <date>, valid_end_date <date>,
#> #   invalid_reason <chr>

# The size of the resultset can be altered
get_test_concepts(limit = 25)

#> Connecting using PostgreSQL driver

#> [2021-03-28 23:40:37]	Overriding cache... Querying Athena...
#> [2021-03-28 23:40:37] 

#> ✓ Open connection

#> [2021-03-28 23:40:37] 

#> ✓ JDBC connection

#> [2021-03-28 23:40:37]	SQL: SELECT * FROM omop_vocabulary.concept ORDER BY RANDOM() LIMIT 25
#> [2021-03-28 23:40:37]	Querying...
#> [2021-03-28 23:40:45]	Querying...complete
#> [2021-03-28 23:40:45] 

#> ✓ Returned data has more than 0 rows

#> [2021-03-28 23:40:45]	Caching resultset...
#> [2021-03-28 23:40:45]	Postgres connection to 'athena' closed

#> # A tibble: 25 x 10
#>    concept_id concept_name            domain_id   vocabulary_id concept_class_id
#>         <int> <chr>                   <chr>       <chr>         <chr>           
#>  1   44939890 Pentoxifylline 400 MG … Drug        NDC           11-digit NDC    
#>  2   40001746 2-ethylhexyl salicylat… Drug        RxNorm        Clinical Drug F…
#>  3    4148650 Infectious diseases se… Procedure   SNOMED        Procedure       
#>  4   21058857 30 ML Mometasone 1 MG/… Drug        RxNorm Exten… Marketed Product
#>  5    3494738 Entire distal epiphysi… Spec Anato… Nebraska Lex… Body Structure  
#>  6   42986651 GEMCIT Injectable powd… Drug        KDC           Drug Product    
#>  7   41185525 Pancreatin 300 MG Oral… Drug        RxNorm Exten… Branded Drug Box
#>  8   45983446 Pindolol 5mg tablets (… Drug        SNOMED        Pharma/Biol Pro…
#>  9   46130950 Sorbsan Plus Carbon dr… Device      SNOMED        Physical Object 
#> 10    3361553 Spirochaeta thermophila Observation Nebraska Lex… Organism        
#> # … with 15 more rows, and 5 more variables: standard_concept <chr>,
#> #   concept_code <chr>, valid_start_date <date>, valid_end_date <date>,
#> #   invalid_reason <chr>

# Concepts specific to the Drug domain can be sampled
# RxNorm/RxNorm Extension Concepts
get_test_drug_concepts(limit = 10)

#> Connecting using PostgreSQL driver

#> [2021-03-28 23:40:45]	Overriding cache... Querying Athena...
#> [2021-03-28 23:40:45] 

#> ✓ Open connection

#> [2021-03-28 23:40:45] 

#> ✓ JDBC connection

#> [2021-03-28 23:40:45]	SQL: SELECT * FROM omop_vocabulary.concept c WHERE c.invalid_reason IS NULL AND c.vocabulary_id IN ('RxNorm', 'RxNorm Extension') AND c.standard_concept <> 'C' AND c.domain_id = 'Drug' ORDER BY RANDOM() LIMIT 10
#> [2021-03-28 23:40:45]	Querying...
#> [2021-03-28 23:40:48]	Querying...complete
#> [2021-03-28 23:40:48] 

#> ✓ Returned data has more than 0 rows

#> [2021-03-28 23:40:48]	Caching resultset...
#> [2021-03-28 23:40:48]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 10
#>    concept_id concept_name              domain_id vocabulary_id concept_class_id
#>         <int> <chr>                     <chr>     <chr>         <chr>           
#>  1   35831409 radium chloride Ra-223 I… Drug      RxNorm Exten… Clinical Drug F…
#>  2     793168 24 ML nivolumab 10 MG/ML… Drug      RxNorm        Quant Branded D…
#>  3   41500387 50000 MG Estriol 0.00002… Drug      RxNorm Exten… Quant Branded D…
#>  4   41038918 Roxithromycin 150 MG Del… Drug      RxNorm Exten… Marketed Product
#>  5    2061067 factor VIII 1000 IU Inje… Drug      RxNorm Exten… Marketed Product
#>  6   43202534 ethanolamine 36 MG / lys… Drug      RxNorm Exten… Branded Drug    
#>  7   43851439 Potassium Chloride 50 MG… Drug      RxNorm Exten… Branded Drug    
#>  8   44073843 Citric Acid 12000 MG / M… Drug      RxNorm Exten… Branded Drug Box
#>  9   44816188 benzoyl peroxide 30 MG/M… Drug      RxNorm        Clinical Drug   
#> 10   41386059 30 ML Benperidol 0.0667 … Drug      RxNorm Exten… Quant Branded D…
#> # … with 5 more variables: standard_concept <chr>, concept_code <chr>,
#> #   valid_start_date <date>, valid_end_date <date>, invalid_reason <chr>

# HemOnc/ATC Classes
get_test_drug_classes(limit = 10)

#> Connecting using PostgreSQL driver

#> [2021-03-28 23:40:48]	Overriding cache... Querying Athena...
#> [2021-03-28 23:40:48] 

#> ✓ Open connection

#> [2021-03-28 23:40:48] 

#> ✓ JDBC connection

#> [2021-03-28 23:40:48]	SQL: SELECT * FROM omop_vocabulary.concept c WHERE c.invalid_reason IS NULL AND c.vocabulary_id IN ('ATC', 'HemOnc') AND c.standard_concept = 'C' AND c.domain_id = 'Drug' ORDER BY RANDOM() LIMIT 10
#> [2021-03-28 23:40:48]	Querying...
#> [2021-03-28 23:40:48]	Querying...complete
#> [2021-03-28 23:40:48] 

#> ✓ Returned data has more than 0 rows

#> [2021-03-28 23:40:48]	Caching resultset...
#> [2021-03-28 23:40:48]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 10
#>    concept_id concept_name              domain_id vocabulary_id concept_class_id
#>         <int> <chr>                     <chr>     <chr>         <chr>           
#>  1   21600650 loperamide, combinations… Drug      ATC           ATC 5th         
#>  2   21605430 gadofosveset              Drug      ATC           ATC 5th         
#>  3   21600618 sulfaguanidine; oral      Drug      ATC           ATC 5th         
#>  4   21602395 econazole; vaginal        Drug      ATC           ATC 5th         
#>  5   21600155 (2-benzhydryloxyethyl)di… Drug      ATC           ATC 5th         
#>  6   21602892 cefprozil; oral           Drug      ATC           ATC 5th         
#>  7   21602524 Synthetic estrogens, pla… Drug      ATC           ATC 4th         
#>  8   21604056 tolmetin; topical         Drug      ATC           ATC 5th         
#>  9   21605381 acetrizoic acid; oral     Drug      ATC           ATC 5th         
#> 10   21605572 yttrium (90Y) silicate c… Drug      ATC           ATC 5th         
#> # … with 5 more variables: standard_concept <chr>, concept_code <chr>,
#> #   valid_start_date <date>, valid_end_date <date>, invalid_reason <chr>

# Test data can also be taken from tables other than Concept
get_test_data(table = "concept_relationship")

#> Connecting using PostgreSQL driver

#> [2021-03-28 23:40:48]	Overriding cache... Querying Athena...
#> [2021-03-28 23:40:48] 

#> ✓ Open connection

#> [2021-03-28 23:40:48] 

#> ✓ JDBC connection

#> [2021-03-28 23:40:48]	SQL: SELECT * FROM omop_vocabulary.concept_relationship ORDER BY RANDOM() LIMIT 100
#> [2021-03-28 23:40:48]	Querying...
#> [2021-03-28 23:41:21]	Querying...complete
#> [2021-03-28 23:41:21] 

#> ✓ Returned data has more than 0 rows

#> [2021-03-28 23:41:21]	Caching resultset...
#> [2021-03-28 23:41:21]	Postgres connection to 'athena' closed

#> # A tibble: 100 x 6
#>    concept_id_1 concept_id_2 relationship_id   valid_start_date valid_end_date
#>           <int>        <int> <chr>             <date>           <date>        
#>  1      1134770     44975818 Mapped from       1970-01-01       2099-12-31    
#>  2     43738838     43612826 Marketed form of  2017-06-24       2099-12-31    
#>  3      3276266      3400171 Is a              2019-08-16       2099-12-31    
#>  4     42008948     42008948 Maps to           2019-02-21       2099-12-31    
#>  5     43611971     43611971 Maps to           2016-12-18       2099-12-31    
#>  6     43175741     43219525 Has marketed form 2017-06-24       2099-12-31    
#>  7     21395080     46149986 Maps to           2018-09-19       2099-12-31    
#>  8     40491472     36548391 Subsumes          2015-09-18       2099-12-31    
#>  9     35909817     35940942 Schema to Value   1970-01-01       2099-12-31    
#> 10     40802863     41030699 Brand name of     2017-07-18       2099-12-31    
#> # … with 90 more rows, and 1 more variable: invalid_reason <chr>

get_test_data(table = "concept_ancestor", limit = 10)

#> Connecting using PostgreSQL driver

#> [2021-03-28 23:41:21]	Overriding cache... Querying Athena...
#> [2021-03-28 23:41:21] 

#> ✓ Open connection

#> [2021-03-28 23:41:21] 

#> ✓ JDBC connection

#> [2021-03-28 23:41:21]	SQL: SELECT * FROM omop_vocabulary.concept_ancestor ORDER BY RANDOM() LIMIT 10
#> [2021-03-28 23:41:21]	Querying...
#> [2021-03-28 23:41:54]	Querying...complete
#> [2021-03-28 23:41:54] 

#> ✓ Returned data has more than 0 rows

#> [2021-03-28 23:41:54]	Caching resultset...
#> [2021-03-28 23:41:54]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 4
#>    ancestor_concept_… descendant_concep… min_levels_of_sepa… max_levels_of_sepa…
#>                 <int>              <int>               <int>               <int>
#>  1           45689442           43267763                   2                   2
#>  2           21600841           41428144                   3                   3
#>  3           44088622           44088622                   0                   0
#>  4           21600046           41164651                   6                   7
#>  5           40108192           41082605                   0                   0
#>  6           36217207           44217756                   3                   3
#>  7           46301326           43687035                   2                   2
#>  8            4113998           36530259                   5                   9
#>  9             441840            4042074                   4                   5
#> 10           21604686           44180617                   5                   5