Get Test Drug Classes

Get random records of ATC and HemOnc Drug Classes that are:

Valid,

Class,

Belong to the Drug domain

get_test_drug_classes(
  conn,
  conn_fun = "connectAthena()",
  schema = "omop_vocabulary",
  limit = 100,
  verbose = TRUE,
  render_sql = TRUE
)

See also

Examples

library(chariot)
library(tidyverse)

# Get Unfiltered Test Concepts (Invalid Concepts are included)
get_test_concepts()

#> Connecting using PostgreSQL driver
#> [2021-01-10 15:59:16]	Overriding cache... Querying Athena...
#> [2021-01-10 15:59:16] 

#> ✓ Open connection

#> [2021-01-10 15:59:16] 

#> ✓ JDBC connection

#> [2021-01-10 15:59:16]	SQL: SELECT * FROM omop_vocabulary.concept ORDER BY RANDOM() LIMIT 100
#> [2021-01-10 15:59:16]	Querying...
#> [2021-01-10 15:59:24]	Querying...complete
#> [2021-01-10 15:59:24] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 15:59:24]	Caching resultset...
#> [2021-01-10 15:59:25]	Postgres connection to 'athena' closed

#> # A tibble: 100 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id
#>         <dbl> <chr>        <chr>     <chr>         <chr>           
#>  1   44027283 mercuric su… Drug      RxNorm Exten… Clinical Drug C…
#>  2   43039486 2 ML Ipratr… Drug      RxNorm Exten… Marketed Product
#>  3   42190353 Hysteroscop… Procedure EDI           Procedure       
#>  4   44612691 MENTHA ARVE… Drug      DPD           Clinical Drug C…
#>  5   42754175 Incision, e… Procedure OPS           Procedure       
#>  6   44477331 phentermine… Drug      NDC           9-digit NDC     
#>  7   41225949 URTICA DIOI… Drug      RxNorm Exten… Marketed Product
#>  8   41120452 torsemide 5… Drug      RxNorm Exten… Branded Drug Box
#>  9    2841815 Imaging @ U… Procedure ICD10PCS      ICD10PCS Hierar…
#> 10   43174726 2 ML Doxepi… Drug      RxNorm Exten… Marketed Product
#> # … with 90 more rows, and 5 more variables: standard_concept <chr>,
#> #   concept_code <chr>, valid_start_date <date>, valid_end_date <date>,
#> #   invalid_reason <chr>

# The size of the resultset can be altered
get_test_concepts(limit = 25)

#> Connecting using PostgreSQL driver
#> [2021-01-10 15:59:25]	Overriding cache... Querying Athena...
#> [2021-01-10 15:59:25] 

#> ✓ Open connection

#> [2021-01-10 15:59:25] 

#> ✓ JDBC connection

#> [2021-01-10 15:59:25]	SQL: SELECT * FROM omop_vocabulary.concept ORDER BY RANDOM() LIMIT 25
#> [2021-01-10 15:59:25]	Querying...
#> [2021-01-10 15:59:32]	Querying...complete
#> [2021-01-10 15:59:32] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 15:59:32]	Caching resultset...
#> [2021-01-10 15:59:32]	Postgres connection to 'athena' closed

#> # A tibble: 25 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id
#>         <dbl> <chr>        <chr>     <chr>         <chr>           
#>  1    1417954 Intestinal … Condition ICD10CN       ICD10 code      
#>  2   19067782 Guiacough DM Drug      RxNorm        Brand Name      
#>  3   40888845 silybum mar… Drug      RxNorm Exten… Clinical Drug C…
#>  4   42743870 Other opera… Procedure OPS           Procedure       
#>  5   36369896 HAND SANITI… Drug      SPL           OTC Drug        
#>  6   35405558 terbinafine… Drug      NDC           11-digit NDC    
#>  7   45156413 Metronidazo… Drug      NDC           11-digit NDC    
#>  8   36177384 Modified 24… Drug      NDC           11-digit NDC    
#>  9   44879723 GOODNITES U… Device    NDC           Device          
#> 10   43322430 Seralin 50 … Drug      AMT           Containered Pack
#> # … with 15 more rows, and 5 more variables: standard_concept <chr>,
#> #   concept_code <chr>, valid_start_date <date>, valid_end_date <date>,
#> #   invalid_reason <chr>

# Concepts specific to the Drug domain can be sampled
# RxNorm/RxNorm Extension Concepts
get_test_drug_concepts(limit = 10)

#> Connecting using PostgreSQL driver
#> [2021-01-10 15:59:32]	Overriding cache... Querying Athena...
#> [2021-01-10 15:59:32] 

#> ✓ Open connection

#> [2021-01-10 15:59:32] 

#> ✓ JDBC connection

#> [2021-01-10 15:59:32]	SQL: SELECT * FROM omop_vocabulary.concept c WHERE c.invalid_reason IS NULL AND c.vocabulary_id IN ('RxNorm', 'RxNorm Extension') AND c.standard_concept <> 'C' AND c.domain_id = 'Drug' ORDER BY RANDOM() LIMIT 10
#> [2021-01-10 15:59:32]	Querying...
#> [2021-01-10 15:59:35]	Querying...complete
#> [2021-01-10 15:59:35] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 15:59:35]	Caching resultset...
#> [2021-01-10 15:59:35]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id
#>         <dbl> <chr>        <chr>     <chr>         <chr>           
#>  1   41440267 473 ML preg… Drug      RxNorm Exten… Quant Clinical …
#>  2   21139846 mecysteine … Drug      RxNorm Exten… Branded Drug    
#>  3   35789394 500 ML Phen… Drug      RxNorm Exten… Quant Clinical …
#>  4   43272434 irbesartan … Drug      RxNorm Exten… Branded Drug Box
#>  5   41172646 Ethinyl Est… Drug      RxNorm Exten… Branded Drug Co…
#>  6   21056903 Prochlorper… Drug      RxNorm Exten… Clinical Drug B…
#>  7   44165878 Metformin 5… Drug      RxNorm Exten… Branded Drug Box
#>  8   43860493 1 ML hirudo… Drug      RxNorm Exten… Quant Clinical …
#>  9   41477721 10000 MG Mo… Drug      RxNorm Exten… Marketed Product
#> 10   41047639 Dexamethaso… Drug      RxNorm Exten… Branded Drug Co…
#> # … with 5 more variables: standard_concept <chr>, concept_code <chr>,
#> #   valid_start_date <date>, valid_end_date <date>, invalid_reason <chr>

# HemOnc/ATC Classes
get_test_drug_classes(limit = 10)

#> Connecting using PostgreSQL driver
#> [2021-01-10 15:59:35]	Overriding cache... Querying Athena...
#> [2021-01-10 15:59:35] 

#> ✓ Open connection

#> [2021-01-10 15:59:35] 

#> ✓ JDBC connection

#> [2021-01-10 15:59:35]	SQL: SELECT * FROM omop_vocabulary.concept c WHERE c.invalid_reason IS NULL AND c.vocabulary_id IN ('ATC', 'HemOnc') AND c.standard_concept = 'C' AND c.domain_id = 'Drug' ORDER BY RANDOM() LIMIT 10
#> [2021-01-10 15:59:35]	Querying...
#> [2021-01-10 15:59:36]	Querying...complete
#> [2021-01-10 15:59:36] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 15:59:36]	Caching resultset...
#> [2021-01-10 15:59:36]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id
#>         <dbl> <chr>        <chr>     <chr>         <chr>           
#>  1   21600417 betanidine;… Drug      ATC           ATC 5th         
#>  2   21602163 Corticoster… Drug      ATC           ATC 4th         
#>  3   21605176 triamcinolo… Drug      ATC           ATC 5th         
#>  4   21600439 Alkaloids, … Drug      ATC           ATC 4th         
#>  5   21600050 magnesium o… Drug      ATC           ATC 5th         
#>  6   21601784 ACE inhibit… Drug      ATC           ATC 4th         
#>  7   21602093 docosanol; … Drug      ATC           ATC 5th         
#>  8   21603623 betamethaso… Drug      ATC           ATC 5th         
#>  9   21604042 etofenamate… Drug      ATC           ATC 5th         
#> 10   21601957 tolnaftate;… Drug      ATC           ATC 5th         
#> # … with 5 more variables: standard_concept <chr>, concept_code <chr>,
#> #   valid_start_date <date>, valid_end_date <date>, invalid_reason <chr>

# Test data can also be taken from tables other than Concept
get_test_data(table = "concept_relationship")

#> Connecting using PostgreSQL driver
#> [2021-01-10 15:59:36]	Overriding cache... Querying Athena...
#> [2021-01-10 15:59:36] 

#> ✓ Open connection

#> [2021-01-10 15:59:36] 

#> ✓ JDBC connection

#> [2021-01-10 15:59:36]	SQL: SELECT * FROM omop_vocabulary.concept_relationship ORDER BY RANDOM() LIMIT 100
#> [2021-01-10 15:59:36]	Querying...
#> [2021-01-10 16:00:08]	Querying...complete
#> [2021-01-10 16:00:08] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 16:00:08]	Caching resultset...
#> [2021-01-10 16:00:08]	Postgres connection to 'athena' closed

#> # A tibble: 100 x 6
#>    concept_id_1 concept_id_2 relationship_id valid_start_date valid_end_date
#>           <dbl>        <dbl> <chr>           <date>           <date>        
#>  1     46070437     46070437 Maps to         2018-07-31       2099-12-31    
#>  2     40051196     44069886 Has tradename   2017-06-25       2099-12-31    
#>  3      3413130      3493852 Has finding si… 2019-08-16       2099-12-31    
#>  4     21112436     19082573 RxNorm has dos… 2015-08-17       2099-12-31    
#>  5      3187123      3411678 Is a            2019-08-16       2099-12-31    
#>  6      3463049      3428132 Is a            2019-08-16       2099-12-31    
#>  7     42854132      4249786 NDFRT has dose… 2012-10-01       2099-12-31    
#>  8     35909711     35921142 Schema to Value 1970-01-01       2099-12-31    
#>  9     43776283     43776283 Maps to         2016-12-18       2099-12-31    
#> 10     35614833     35620616 Flavor of       2018-07-31       2099-12-31    
#> # … with 90 more rows, and 1 more variable: invalid_reason <chr>

get_test_data(table = "concept_ancestor", limit = 10)

#> Connecting using PostgreSQL driver
#> [2021-01-10 16:00:08]	Overriding cache... Querying Athena...
#> [2021-01-10 16:00:08] 

#> ✓ Open connection

#> [2021-01-10 16:00:08] 

#> ✓ JDBC connection

#> [2021-01-10 16:00:08]	SQL: SELECT * FROM omop_vocabulary.concept_ancestor ORDER BY RANDOM() LIMIT 10
#> [2021-01-10 16:00:08]	Querying...
#> [2021-01-10 16:00:42]	Querying...complete
#> [2021-01-10 16:00:42] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 16:00:42]	Caching resultset...
#> [2021-01-10 16:00:42]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 4
#>    ancestor_concept_… descendant_concep… min_levels_of_sepa… max_levels_of_sepa…
#>                 <dbl>              <dbl>               <dbl>               <dbl>
#>  1           45629907           35887511                   2                   2
#>  2            4116057            4116057                   0                   0
#>  3           21604254           40878374                   5                   6
#>  4           43678069           41307421                   1                   1
#>  5           21602700           36063842                   3                   3
#>  6           45644888           41276951                   2                   2
#>  7           21600001           40931456                   7                   8
#>  8           46240499           21163244                   2                   2
#>  9           36107150            2069896                   1                   1
#> 10           21600483           43185309                   4                   4