Get Test Drug Classes

Get random records of ATC and HemOnc Drug Classes that are:

Valid,

Class,

Belong to the Drug domain

get_test_drug_classes(
  conn,
  conn_fun = "connectAthena()",
  schema = "omop_vocabulary",
  limit = 100,
  verbose = TRUE,
  render_sql = TRUE
)

See also

Examples

library(chariot)
library(tidyverse)

# Get Unfiltered Test Concepts (Invalid Concepts are included)
get_test_concepts()

#> Connecting using PostgreSQL driver
#> [2021-01-17 00:56:02]	Overriding cache... Querying Athena...
#> [2021-01-17 00:56:02] 

#> ✓ Open connection

#> [2021-01-17 00:56:02] 

#> ✓ JDBC connection

#> [2021-01-17 00:56:02]	SQL: SELECT * FROM omop_vocabulary.concept ORDER BY RANDOM() LIMIT 100
#> [2021-01-17 00:56:02]	Querying...
#> [2021-01-17 00:56:11]	Querying...complete
#> [2021-01-17 00:56:11] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-17 00:56:11]	Caching resultset...
#> [2021-01-17 00:56:11]	Postgres connection to 'athena' closed

#> # A tibble: 100 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id
#>         <dbl> <chr>        <chr>     <chr>         <chr>           
#>  1    2098899 UNIAL Ophth… Drug      KDC           Drug Product    
#>  2    3253405 Large bowel… Condition Nebraska Lex… Clinical Finding
#>  3   43326316 Avagard Ant… Drug      AMT           Containered Pack
#>  4   37499109 chlorphenir… Drug      RxNorm        Clinical Drug   
#>  5   45009878 Codeine Pho… Drug      NDC           11-digit NDC    
#>  6   21315852 Voltarol 50… Drug      dm+d          AMP             
#>  7   44851861 hydrochloro… Drug      NDC           11-digit NDC    
#>  8    4285610 CINACALCET/… Drug      NDFRT         Drug Interaction
#>  9   40312508 Gum epulis … Procedure SNOMED        Procedure       
#> 10   36632368 Oxycodone 1… Drug      NCCD          Drug Product    
#> # … with 90 more rows, and 5 more variables: standard_concept <chr>,
#> #   concept_code <chr>, valid_start_date <date>, valid_end_date <date>,
#> #   invalid_reason <chr>

# The size of the resultset can be altered
get_test_concepts(limit = 25)

#> Connecting using PostgreSQL driver
#> [2021-01-17 00:56:11]	Overriding cache... Querying Athena...
#> [2021-01-17 00:56:11] 

#> ✓ Open connection

#> [2021-01-17 00:56:11] 

#> ✓ JDBC connection

#> [2021-01-17 00:56:11]	SQL: SELECT * FROM omop_vocabulary.concept ORDER BY RANDOM() LIMIT 25
#> [2021-01-17 00:56:11]	Querying...
#> [2021-01-17 00:56:19]	Querying...complete
#> [2021-01-17 00:56:19] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-17 00:56:19]	Caching resultset...
#> [2021-01-17 00:56:19]	Postgres connection to 'athena' closed

#> # A tibble: 25 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id
#>         <dbl> <chr>        <chr>     <chr>         <chr>           
#>  1    2837724 Medical and… Procedure ICD10PCS      ICD10PCS Hierar…
#>  2   45287759 Potassium C… Drug      NDC           11-digit NDC    
#>  3    3212242 Product con… Drug      Nebraska Lex… Pharma/Biol Pro…
#>  4   45649000 glycerin .0… Drug      SPL           OTC Drug        
#>  5    2852322 Medical and… Procedure ICD10PCS      ICD10PCS Hierar…
#>  6   35531108 ibuprofen 4… Drug      NDC           11-digit NDC    
#>  7   40680361 Sigvaris Ac… Device    dm+d          AMP             
#>  8    1434421 Need for im… Observat… ICD10CN       ICD10 Hierarchy 
#>  9   42968760 glimepiride… Drug      RxNorm Exten… Branded Drug    
#> 10   41348352 3 ML insuli… Drug      RxNorm Exten… Marketed Product
#> # … with 15 more rows, and 5 more variables: standard_concept <chr>,
#> #   concept_code <chr>, valid_start_date <date>, valid_end_date <date>,
#> #   invalid_reason <chr>

# Concepts specific to the Drug domain can be sampled
# RxNorm/RxNorm Extension Concepts
get_test_drug_concepts(limit = 10)

#> Connecting using PostgreSQL driver
#> [2021-01-17 00:56:20]	Overriding cache... Querying Athena...
#> [2021-01-17 00:56:20] 

#> ✓ Open connection

#> [2021-01-17 00:56:20] 

#> ✓ JDBC connection

#> [2021-01-17 00:56:20]	SQL: SELECT * FROM omop_vocabulary.concept c WHERE c.invalid_reason IS NULL AND c.vocabulary_id IN ('RxNorm', 'RxNorm Extension') AND c.standard_concept <> 'C' AND c.domain_id = 'Drug' ORDER BY RANDOM() LIMIT 10
#> [2021-01-17 00:56:20]	Querying...
#> [2021-01-17 00:56:24]	Querying...complete
#> [2021-01-17 00:56:24] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-17 00:56:24]	Caching resultset...
#> [2021-01-17 00:56:24]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id
#>         <dbl> <chr>        <chr>     <chr>         <chr>           
#>  1   40886742 duloxetine … Drug      RxNorm Exten… Marketed Product
#>  2   36893739 Diatrizoate… Drug      RxNorm Exten… Clinical Drug B…
#>  3   36786160 gabapentin … Drug      RxNorm Exten… Marketed Product
#>  4   21037252 Prochlorper… Drug      RxNorm Exten… Marketed Product
#>  5   44785883 sennosides,… Drug      RxNorm        Branded Drug Co…
#>  6   21152191 Pentazocine… Drug      RxNorm Exten… Marketed Product
#>  7   43665493 47 ML Bitte… Drug      RxNorm Exten… Quant Clinical …
#>  8   36787379 Metoprolol … Drug      RxNorm Exten… Marketed Product
#>  9    2049958 olanzapine … Drug      RxNorm Exten… Marketed Product
#> 10   41324870 azilsartan … Drug      RxNorm Exten… Marketed Product
#> # … with 5 more variables: standard_concept <chr>, concept_code <chr>,
#> #   valid_start_date <date>, valid_end_date <date>, invalid_reason <chr>

# HemOnc/ATC Classes
get_test_drug_classes(limit = 10)

#> Connecting using PostgreSQL driver
#> [2021-01-17 00:56:24]	Overriding cache... Querying Athena...
#> [2021-01-17 00:56:24] 

#> ✓ Open connection

#> [2021-01-17 00:56:24] 

#> ✓ JDBC connection

#> [2021-01-17 00:56:24]	SQL: SELECT * FROM omop_vocabulary.concept c WHERE c.invalid_reason IS NULL AND c.vocabulary_id IN ('ATC', 'HemOnc') AND c.standard_concept = 'C' AND c.domain_id = 'Drug' ORDER BY RANDOM() LIMIT 10
#> [2021-01-17 00:56:24]	Querying...
#> [2021-01-17 00:56:24]	Querying...complete
#> [2021-01-17 00:56:24] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-17 00:56:24]	Caching resultset...
#> [2021-01-17 00:56:24]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id
#>         <dbl> <chr>        <chr>     <chr>         <chr>           
#>  1   21601848 OTHER AGENT… Drug      ATC           ATC 3rd         
#>  2    1501761 Antivirals … Drug      ATC           ATC 4th         
#>  3   21601334 influenza, … Drug      ATC           ATC 5th         
#>  4   21602490 lynestrenol… Drug      ATC           ATC 5th         
#>  5   21602261 MEDICATED D… Drug      ATC           ATC 2nd         
#>  6   21604635 Benzodiazep… Drug      ATC           ATC 4th         
#>  7   21600486 domperidone… Drug      ATC           ATC 5th         
#>  8   21602725 aldosterone… Drug      ATC           ATC 5th         
#>  9   21600264 Antiarrhyth… Drug      ATC           ATC 4th         
#> 10   21603827 triptorelin… Drug      ATC           ATC 5th         
#> # … with 5 more variables: standard_concept <chr>, concept_code <chr>,
#> #   valid_start_date <date>, valid_end_date <date>, invalid_reason <chr>

# Test data can also be taken from tables other than Concept
get_test_data(table = "concept_relationship")

#> Connecting using PostgreSQL driver
#> [2021-01-17 00:56:24]	Overriding cache... Querying Athena...
#> [2021-01-17 00:56:24] 

#> ✓ Open connection

#> [2021-01-17 00:56:24] 

#> ✓ JDBC connection

#> [2021-01-17 00:56:24]	SQL: SELECT * FROM omop_vocabulary.concept_relationship ORDER BY RANDOM() LIMIT 100
#> [2021-01-17 00:56:24]	Querying...
#> [2021-01-17 00:56:59]	Querying...complete
#> [2021-01-17 00:56:59] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-17 00:56:59]	Caching resultset...
#> [2021-01-17 00:56:59]	Postgres connection to 'athena' closed

#> # A tibble: 100 x 6
#>    concept_id_1 concept_id_2 relationship_id valid_start_date valid_end_date
#>           <dbl>        <dbl> <chr>           <date>           <date>        
#>  1      4194499      4194499 Maps to         1970-01-01       2099-12-31    
#>  2     35918718     35909857 Variable to Sc… 1970-01-01       2099-12-31    
#>  3     35768806     43132265 Has supplier    2017-08-02       2099-12-31    
#>  4     42943104     42943104 Mapped from     2018-08-21       2099-12-31    
#>  5     46050904     46050904 Mapped from     2018-07-31       2099-12-31    
#>  6     43579811     41290067 Supplier of     2017-07-18       2099-12-31    
#>  7      4295937     43022007 Proc context of 2013-01-31       2099-12-31    
#>  8      1567339     35205673 Subsumes        2016-03-25       2099-12-31    
#>  9      2765974      4185128 Has dir proc s… 1970-01-01       2099-12-31    
#> 10     35918383     35909871 Variable to Sc… 1970-01-01       2099-12-31    
#> # … with 90 more rows, and 1 more variable: invalid_reason <chr>

get_test_data(table = "concept_ancestor", limit = 10)

#> Connecting using PostgreSQL driver
#> [2021-01-17 00:56:59]	Overriding cache... Querying Athena...
#> [2021-01-17 00:56:59] 

#> ✓ Open connection

#> [2021-01-17 00:56:59] 

#> ✓ JDBC connection

#> [2021-01-17 00:56:59]	SQL: SELECT * FROM omop_vocabulary.concept_ancestor ORDER BY RANDOM() LIMIT 10
#> [2021-01-17 00:56:59]	Querying...
#> [2021-01-17 00:57:36]	Querying...complete
#> [2021-01-17 00:57:36] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-17 00:57:36]	Caching resultset...
#> [2021-01-17 00:57:36]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 4
#>    ancestor_concept_… descendant_concep… min_levels_of_sepa… max_levels_of_sepa…
#>                 <dbl>              <dbl>               <dbl>               <dbl>
#>  1           45662068           43614644                   1                   1
#>  2           44337532           40884733                   2                   3
#>  3            4042140           36520474                   7                  12
#>  4           36216657           21159539                   2                   4
#>  5           42807308           41039173                   2                   3
#>  6            4297473            4216521                   5                   7
#>  7            1186550           21133568                   2                   3
#>  8           41975669           41975669                   0                   0
#>  9           35807475           35755840                   5                   5
#> 10            1004079            1003720                   2                   2