Get Test Drug Classes

Get random records of ATC and HemOnc Drug Classes that are:

Valid,

Class,

Belong to the Drug domain

get_test_drug_classes(
  conn,
  conn_fun = "connectAthena()",
  schema = "omop_vocabulary",
  limit = 100,
  verbose = TRUE,
  render_sql = TRUE
)

See also

Examples

library(chariot)
library(tidyverse)

# Get Unfiltered Test Concepts (Invalid Concepts are included)
get_test_concepts()

#> Connecting using PostgreSQL driver
#> [2021-01-14 17:54:12]	Overriding cache... Querying Athena...
#> [2021-01-14 17:54:12] 

#> ✓ Open connection

#> [2021-01-14 17:54:12] 

#> ✓ JDBC connection

#> [2021-01-14 17:54:12]	SQL: SELECT * FROM omop_vocabulary.concept ORDER BY RANDOM() LIMIT 100
#> [2021-01-14 17:54:12]	Querying...
#> [2021-01-14 17:54:20]	Querying...complete
#> [2021-01-14 17:54:20] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-14 17:54:20]	Caching resultset...
#> [2021-01-14 17:54:20]	Postgres connection to 'athena' closed

#> # A tibble: 100 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id standard_concept concept_code
#>         <dbl> <chr>        <chr>     <chr>         <chr>            <chr>            <chr>       
#>  1   36772456 octinoxate … Drug      SPL           OTC Drug         C                eb59ab24-64…
#>  2   36464923 avobenzone,… Drug      SPL           OTC Drug         C                7d624b00-74…
#>  3   35623578 Trifluopera… Drug      SNOMED        Clinical Drug    NA               765989000   
#>  4   44666000 TRICLOSAN .… Drug      DPD           Branded Drug Co… NA               OMOP172121  
#>  5   40815654 Quercus Ess… Drug      RxNorm Exten… Brand Name       NA               OMOP2013616 
#>  6   40934956 Pirenzepine… Drug      RxNorm Exten… Branded Drug Box S                OMOP2132918 
#>  7   46358357 ANEW PLATIN… Drug      NDC           11-digit NDC     NA               10096022403 
#>  8   41149602 Clemizole 5… Drug      RxNorm Exten… Clinical Drug B… S                OMOP2347564 
#>  9    2841961 Imaging @ U… Procedure ICD10PCS      ICD10PCS Hierar… S                B34HZZ      
#> 10   40300052 Diagnostic … Procedure SNOMED        Procedure        NA               142553008   
#> # … with 90 more rows, and 3 more variables: valid_start_date <date>, valid_end_date <date>,
#> #   invalid_reason <chr>

# The size of the resultset can be altered
get_test_concepts(limit = 25)

#> Connecting using PostgreSQL driver
#> [2021-01-14 17:54:20]	Overriding cache... Querying Athena...
#> [2021-01-14 17:54:20] 

#> ✓ Open connection

#> [2021-01-14 17:54:20] 

#> ✓ JDBC connection

#> [2021-01-14 17:54:20]	SQL: SELECT * FROM omop_vocabulary.concept ORDER BY RANDOM() LIMIT 25
#> [2021-01-14 17:54:20]	Querying...
#> [2021-01-14 17:54:28]	Querying...complete
#> [2021-01-14 17:54:28] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-14 17:54:28]	Caching resultset...
#> [2021-01-14 17:54:28]	Postgres connection to 'athena' closed

#> # A tibble: 25 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id standard_concept concept_code
#>         <dbl> <chr>        <chr>     <chr>         <chr>            <chr>            <chr>       
#>  1   41205365 proxyphylli… Drug      RxNorm Exten… Clinical Drug F… S                OMOP2403327 
#>  2   40567570 Daughter     Observat… SNOMED        Social Context   NA               394446000   
#>  3   40769986 Because of … Observat… LOINC         Clinical Observ… S                67358-2     
#>  4   35890671 Micafungin … Drug      dm+d          AMPP             NA               39138211000…
#>  5    3154682 Urine free … Procedure Nebraska Lex… Procedure        NA               390446009   
#>  6   42789960 Landmark pl… Procedure CCAM          Procedure        NA               QELH001     
#>  7   46350727 DOVE CLINIC… Drug      NDC           11-digit NDC     NA               64942139301 
#>  8    2815880 Medical and… Procedure ICD10PCS      ICD10PCS Hierar… S                09UE        
#>  9   45171677 Dexamethaso… Drug      NDC           11-digit NDC     NA               51432099005 
#> 10   36544709 Low cumulat… Condition ICDO3         ICDO Condition   S                8743/3-C69.4
#> # … with 15 more rows, and 3 more variables: valid_start_date <date>, valid_end_date <date>,
#> #   invalid_reason <chr>

# Concepts specific to the Drug domain can be sampled
# RxNorm/RxNorm Extension Concepts
get_test_drug_concepts(limit = 10)

#> Connecting using PostgreSQL driver
#> [2021-01-14 17:54:28]	Overriding cache... Querying Athena...
#> [2021-01-14 17:54:28] 

#> ✓ Open connection

#> [2021-01-14 17:54:28] 

#> ✓ JDBC connection

#> [2021-01-14 17:54:28]	SQL: SELECT * FROM omop_vocabulary.concept c WHERE c.invalid_reason IS NULL AND c.vocabulary_id IN ('RxNorm', 'RxNorm Extension') AND c.standard_concept <> 'C' AND c.domain_id = 'Drug' ORDER BY RANDOM() LIMIT 10
#> [2021-01-14 17:54:28]	Querying...
#> [2021-01-14 17:54:32]	Querying...complete
#> [2021-01-14 17:54:32] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-14 17:54:32]	Caching resultset...
#> [2021-01-14 17:54:32]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id standard_concept concept_code
#>         <dbl> <chr>        <chr>     <chr>         <chr>            <chr>            <chr>       
#>  1   35871641 yeast manna… Drug      RxNorm Exten… Clinical Drug C… S                OMOP5028427 
#>  2   37592725 mycophenola… Drug      RxNorm Exten… Marketed Product S                OMOP4782169 
#>  3   35157314 valsartan 2… Drug      RxNorm Exten… Marketed Product S                OMOP4814320 
#>  4   41036724 Caffeine 0.… Drug      RxNorm Exten… Branded Drug     S                OMOP2234686 
#>  5   21054114 Primidone 2… Drug      RxNorm Exten… Marketed Product S                OMOP392699  
#>  6   21098093 30 ML Econa… Drug      RxNorm Exten… Quant Clinical … S                OMOP316242  
#>  7   36781953 sitagliptin… Drug      RxNorm Exten… Clinical Drug B… S                OMOP4770239 
#>  8   41435721 235 ML Ange… Drug      RxNorm Exten… Quant Clinical … S                OMOP2633683 
#>  9   21107477 5 ML Midazo… Drug      RxNorm Exten… Marketed Product S                OMOP307458  
#> 10   41148552 fluticasone… Drug      RxNorm Exten… Clinical Drug    S                OMOP2346514 
#> # … with 3 more variables: valid_start_date <date>, valid_end_date <date>, invalid_reason <chr>

# HemOnc/ATC Classes
get_test_drug_classes(limit = 10)

#> Connecting using PostgreSQL driver
#> [2021-01-14 17:54:32]	Overriding cache... Querying Athena...
#> [2021-01-14 17:54:32] 

#> ✓ Open connection

#> [2021-01-14 17:54:32] 

#> ✓ JDBC connection

#> [2021-01-14 17:54:32]	SQL: SELECT * FROM omop_vocabulary.concept c WHERE c.invalid_reason IS NULL AND c.vocabulary_id IN ('ATC', 'HemOnc') AND c.standard_concept = 'C' AND c.domain_id = 'Drug' ORDER BY RANDOM() LIMIT 10
#> [2021-01-14 17:54:32]	Querying...
#> [2021-01-14 17:54:32]	Querying...complete
#> [2021-01-14 17:54:32] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-14 17:54:32]	Caching resultset...
#> [2021-01-14 17:54:32]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id standard_concept concept_code
#>         <dbl> <chr>        <chr>     <chr>         <chr>            <chr>            <chr>       
#>  1   35101524 Anticoagula… Drug      HemOnc        Component Class  C                46301       
#>  2   21602346 magnesium s… Drug      ATC           ATC 5th          C                D11AX05     
#>  3   21604642 temazepam; … Drug      ATC           ATC 5th          C                N05CD07     
#>  4   21601162 electrolyte… Drug      ATC           ATC 5th          C                B05BB02     
#>  5   21603138 cidofovir; … Drug      ATC           ATC 5th          C                J05AB12     
#>  6   21602938 sulfapyridi… Drug      ATC           ATC 5th          C                J01EB04     
#>  7   21600851 calcium pan… Drug      ATC           ATC 5th          C                A11HA31     
#>  8   21604946 fenbendazole Drug      ATC           ATC 5th          C                P02CA06     
#>  9   35804303 GnRH agonis… Drug      HemOnc        Component Class  C                9460        
#> 10    1123817 oxycodone a… Drug      ATC           ATC 5th          C                N02AJ19     
#> # … with 3 more variables: valid_start_date <date>, valid_end_date <date>, invalid_reason <chr>

# Test data can also be taken from tables other than Concept
get_test_data(table = "concept_relationship")

#> Connecting using PostgreSQL driver
#> [2021-01-14 17:54:32]	Overriding cache... Querying Athena...
#> [2021-01-14 17:54:32] 

#> ✓ Open connection

#> [2021-01-14 17:54:32] 

#> ✓ JDBC connection

#> [2021-01-14 17:54:32]	SQL: SELECT * FROM omop_vocabulary.concept_relationship ORDER BY RANDOM() LIMIT 100
#> [2021-01-14 17:54:32]	Querying...
#> [2021-01-14 17:55:07]	Querying...complete
#> [2021-01-14 17:55:07] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-14 17:55:07]	Caching resultset...
#> [2021-01-14 17:55:07]	Postgres connection to 'athena' closed

#> # A tibble: 100 x 6
#>    concept_id_1 concept_id_2 relationship_id    valid_start_date valid_end_date invalid_reason
#>           <dbl>        <dbl> <chr>              <date>           <date>         <chr>         
#>  1      3038413      1030089 Has system         2019-12-13       2099-12-31     NA            
#>  2     46145995     45956741 Has disp dose form 2018-07-31       2099-12-31     NA            
#>  3       920458     41143319 RxNorm ing of      2017-07-18       2099-12-31     NA            
#>  4     46282198     46279001 Has AMP            2015-10-01       2099-12-31     NA            
#>  5      3378613      3250776 Subsumes           2019-08-16       2099-12-31     NA            
#>  6      4179713      2744900 Subsumes           1970-01-01       2099-12-31     NA            
#>  7     46203664     46091835 Is a               2015-01-31       2099-12-31     NA            
#>  8     21017191     21074404 Brand name of      2015-08-17       2099-12-31     NA            
#>  9     46096380     46150424 Subsumes           2015-01-31       2099-12-31     NA            
#> 10       432795     37612525 Mapped from        2019-10-21       2099-12-31     NA            
#> # … with 90 more rows

get_test_data(table = "concept_ancestor", limit = 10)

#> Connecting using PostgreSQL driver
#> [2021-01-14 17:55:07]	Overriding cache... Querying Athena...
#> [2021-01-14 17:55:07] 

#> ✓ Open connection

#> [2021-01-14 17:55:07] 

#> ✓ JDBC connection

#> [2021-01-14 17:55:07]	SQL: SELECT * FROM omop_vocabulary.concept_ancestor ORDER BY RANDOM() LIMIT 10
#> [2021-01-14 17:55:07]	Querying...
#> [2021-01-14 17:55:42]	Querying...complete
#> [2021-01-14 17:55:42] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-14 17:55:42]	Caching resultset...
#> [2021-01-14 17:55:42]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 4
#>    ancestor_concept_id descendant_concept_id min_levels_of_separation max_levels_of_separation
#>                  <dbl>                 <dbl>                    <dbl>                    <dbl>
#>  1             4147164              36537109                        1                        1
#>  2            36217216              36268419                        2                        3
#>  3            36244872              21063669                        2                        3
#>  4            45706794              44077927                        2                        2
#>  5            36153082              43201290                        2                        3
#>  6            41931105              41959642                        3                        3
#>  7            36003010              41040890                        2                        3
#>  8            36107075              35407106                        2                        3
#>  9            45640299              41293848                        2                        3
#> 10            40054138              40854461                        2                        3