Get Test Drug Classes

Get random records of ATC and HemOnc Drug Classes that are:

Valid,

Class,

Belong to the Drug domain

get_test_drug_classes(
  conn,
  conn_fun = "connectAthena()",
  schema = "omop_vocabulary",
  limit = 100,
  verbose = TRUE,
  render_sql = TRUE
)

See also

Examples

library(chariot)
library(tidyverse)

# Get Unfiltered Test Concepts (Invalid Concepts are included)
get_test_concepts()

#> Connecting using PostgreSQL driver
#> [2021-01-10 19:01:06]	Overriding cache... Querying Athena...
#> [2021-01-10 19:01:06] 

#> ✓ Open connection

#> [2021-01-10 19:01:06] 

#> ✓ JDBC connection

#> [2021-01-10 19:01:06]	SQL: SELECT * FROM omop_vocabulary.concept ORDER BY RANDOM() LIMIT 100
#> [2021-01-10 19:01:06]	Querying...
#> [2021-01-10 19:01:14]	Querying...complete
#> [2021-01-10 19:01:14] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 19:01:14]	Caching resultset...
#> [2021-01-10 19:01:14]	Postgres connection to 'athena' closed

#> # A tibble: 100 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id
#>         <dbl> <chr>        <chr>     <chr>         <chr>           
#>  1   45140028 Rosuvastati… Drug      NDC           11-digit NDC    
#>  2     591561 buprenorphi… Drug      SPL           Prescription Dr…
#>  3   36484168 LOMOTIL - l… Drug      NDC           11-digit NDC    
#>  4   41456953 2.02 ML Med… Drug      RxNorm Exten… Quant Branded B…
#>  5   43574557 Rutinion     Drug      RxNorm Exten… Brand Name      
#>  6   41436590 250 ML Povi… Drug      RxNorm Exten… Marketed Product
#>  7   45308658 Warfarin So… Drug      NDC           11-digit NDC    
#>  8   45955858 idiopathic … Condition CIEL          Diagnosis       
#>  9   40969480 Ibuprofen 8… Drug      RxNorm Exten… Branded Drug    
#> 10   46205390 Jobst Ultra… Device    SNOMED        Physical Object 
#> # … with 90 more rows, and 5 more variables: standard_concept <chr>,
#> #   concept_code <chr>, valid_start_date <date>, valid_end_date <date>,
#> #   invalid_reason <chr>

# The size of the resultset can be altered
get_test_concepts(limit = 25)

#> Connecting using PostgreSQL driver
#> [2021-01-10 19:01:15]	Overriding cache... Querying Athena...
#> [2021-01-10 19:01:15] 

#> ✓ Open connection

#> [2021-01-10 19:01:15] 

#> ✓ JDBC connection

#> [2021-01-10 19:01:15]	SQL: SELECT * FROM omop_vocabulary.concept ORDER BY RANDOM() LIMIT 25
#> [2021-01-10 19:01:15]	Querying...
#> [2021-01-10 19:01:22]	Querying...complete
#> [2021-01-10 19:01:22] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 19:01:22]	Caching resultset...
#> [2021-01-10 19:01:22]	Postgres connection to 'athena' closed

#> # A tibble: 25 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id
#>         <dbl> <chr>        <chr>     <chr>         <chr>           
#>  1   35991900 Variant of … Measurem… OMOP Extensi… Transcript Vari…
#>  2   45638336 SEROQUEL - … Drug      SPL           Prescription Dr…
#>  3   40891129 Lisinopril … Drug      RxNorm Exten… Branded Drug Co…
#>  4    4281998 Cleidocrani… Condition SNOMED        Clinical Finding
#>  5   46008360 Resource Sh… Drug      SNOMED        Pharma/Biol Pro…
#>  6   21153258 Hydrochloro… Drug      RxNorm Exten… Marketed Product
#>  7   40240051 levoleucovo… Drug      RxNorm        Branded Drug    
#>  8   45179942 Lindane 10 … Drug      NDC           11-digit NDC    
#>  9   42954569 olopatadine… Drug      RxNorm Exten… Branded Drug    
#> 10   44125308 Finasteride… Drug      RxNorm Exten… Branded Drug    
#> # … with 15 more rows, and 5 more variables: standard_concept <chr>,
#> #   concept_code <chr>, valid_start_date <date>, valid_end_date <date>,
#> #   invalid_reason <chr>

# Concepts specific to the Drug domain can be sampled
# RxNorm/RxNorm Extension Concepts
get_test_drug_concepts(limit = 10)

#> Connecting using PostgreSQL driver
#> [2021-01-10 19:01:22]	Overriding cache... Querying Athena...
#> [2021-01-10 19:01:22] 

#> ✓ Open connection

#> [2021-01-10 19:01:22] 

#> ✓ JDBC connection

#> [2021-01-10 19:01:22]	SQL: SELECT * FROM omop_vocabulary.concept c WHERE c.invalid_reason IS NULL AND c.vocabulary_id IN ('RxNorm', 'RxNorm Extension') AND c.standard_concept <> 'C' AND c.domain_id = 'Drug' ORDER BY RANDOM() LIMIT 10
#> [2021-01-10 19:01:22]	Querying...
#> [2021-01-10 19:01:27]	Querying...complete
#> [2021-01-10 19:01:27] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 19:01:27]	Caching resultset...
#> [2021-01-10 19:01:27]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id
#>         <dbl> <chr>        <chr>     <chr>         <chr>           
#>  1   41335500 1 ML Dexame… Drug      RxNorm Exten… Marketed Product
#>  2    2057678 atorvastati… Drug      RxNorm Exten… Branded Drug    
#>  3   43258438 Ethanol 0.5… Drug      RxNorm Exten… Clinical Drug   
#>  4   41014449 Calendula o… Drug      RxNorm Exten… Clinical Drug C…
#>  5   35407342 Gliclazide … Drug      RxNorm Exten… Branded Drug Fo…
#>  6   40964281 Thymus vulg… Drug      RxNorm Exten… Branded Drug Box
#>  7   35785570 0.5 ML Ilop… Drug      RxNorm Exten… Marketed Product
#>  8   41370065 10 ML Cromo… Drug      RxNorm Exten… Quant Branded D…
#>  9   36272878 Omeprazole … Drug      RxNorm Exten… Branded Drug Box
#> 10   43132949 Simvastatin… Drug      RxNorm Exten… Branded Drug    
#> # … with 5 more variables: standard_concept <chr>, concept_code <chr>,
#> #   valid_start_date <date>, valid_end_date <date>, invalid_reason <chr>

# HemOnc/ATC Classes
get_test_drug_classes(limit = 10)

#> Connecting using PostgreSQL driver
#> [2021-01-10 19:01:27]	Overriding cache... Querying Athena...
#> [2021-01-10 19:01:27] 

#> ✓ Open connection

#> [2021-01-10 19:01:27] 

#> ✓ JDBC connection

#> [2021-01-10 19:01:27]	SQL: SELECT * FROM omop_vocabulary.concept c WHERE c.invalid_reason IS NULL AND c.vocabulary_id IN ('ATC', 'HemOnc') AND c.standard_concept = 'C' AND c.domain_id = 'Drug' ORDER BY RANDOM() LIMIT 10
#> [2021-01-10 19:01:27]	Querying...
#> [2021-01-10 19:01:27]	Querying...complete
#> [2021-01-10 19:01:27] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 19:01:27]	Caching resultset...
#> [2021-01-10 19:01:27]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id
#>         <dbl> <chr>        <chr>     <chr>         <chr>           
#>  1   21600314 levosimenda… Drug      ATC           ATC 5th         
#>  2   21604548 sulpiride; … Drug      ATC           ATC 5th         
#>  3   21604244 etidocaine,… Drug      ATC           ATC 5th         
#>  4    1123743 osimertinib  Drug      ATC           ATC 5th         
#>  5   21600140 benzilone; … Drug      ATC           ATC 5th         
#>  6   21602938 sulfapyridi… Drug      ATC           ATC 5th         
#>  7   40253798 voclosporin  Drug      ATC           ATC 5th         
#>  8   35806262 Bisphosphon… Drug      HemOnc        Component Class 
#>  9   21600407 mecamylamine Drug      ATC           ATC 5th         
#> 10   21603001 sisomicin; … Drug      ATC           ATC 5th         
#> # … with 5 more variables: standard_concept <chr>, concept_code <chr>,
#> #   valid_start_date <date>, valid_end_date <date>, invalid_reason <chr>

# Test data can also be taken from tables other than Concept
get_test_data(table = "concept_relationship")

#> Connecting using PostgreSQL driver
#> [2021-01-10 19:01:27]	Overriding cache... Querying Athena...
#> [2021-01-10 19:01:27] 

#> ✓ Open connection

#> [2021-01-10 19:01:27] 

#> ✓ JDBC connection

#> [2021-01-10 19:01:27]	SQL: SELECT * FROM omop_vocabulary.concept_relationship ORDER BY RANDOM() LIMIT 100
#> [2021-01-10 19:01:27]	Querying...
#> [2021-01-10 19:02:01]	Querying...complete
#> [2021-01-10 19:02:01] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 19:02:01]	Caching resultset...
#> [2021-01-10 19:02:01]	Postgres connection to 'athena' closed

#> # A tibble: 100 x 6
#>    concept_id_1 concept_id_2 relationship_id valid_start_date valid_end_date
#>           <dbl>        <dbl> <chr>           <date>           <date>        
#>  1     21078612     21117797 Has marketed f… 2015-08-17       2099-12-31    
#>  2     42479111     19026845 Has brand name  2015-08-17       2099-12-31    
#>  3     45587560      1572345 Is a            2016-03-25       2099-12-31    
#>  4     36412037     36412037 Mapped from     2017-10-15       2099-12-31    
#>  5      4335621      4279323 MoA of          2010-06-07       2099-12-31    
#>  6     19011877     19116062 Brand name of   2016-10-07       2099-12-31    
#>  7     40882275     40820327 Has supplier    2017-07-18       2099-12-31    
#>  8     41497582     19082227 RxNorm has dos… 2017-07-18       2099-12-31    
#>  9      4065239       132706 Maps to         1970-01-01       2099-12-31    
#> 10     46232233      4249123 Maps to         2018-08-29       2099-12-31    
#> # … with 90 more rows, and 1 more variable: invalid_reason <chr>

get_test_data(table = "concept_ancestor", limit = 10)

#> Connecting using PostgreSQL driver
#> [2021-01-10 19:02:01]	Overriding cache... Querying Athena...
#> [2021-01-10 19:02:01] 

#> ✓ Open connection

#> [2021-01-10 19:02:01] 

#> ✓ JDBC connection

#> [2021-01-10 19:02:01]	SQL: SELECT * FROM omop_vocabulary.concept_ancestor ORDER BY RANDOM() LIMIT 10
#> [2021-01-10 19:02:01]	Querying...
#> [2021-01-10 19:02:36]	Querying...complete
#> [2021-01-10 19:02:36] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 19:02:36]	Caching resultset...
#> [2021-01-10 19:02:36]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 4
#>    ancestor_concept_… descendant_concep… min_levels_of_sepa… max_levels_of_sepa…
#>                 <dbl>              <dbl>               <dbl>               <dbl>
#>  1           36152376           43838946                   2                   2
#>  2            1728442           35762372                   1                   1
#>  3           45706307           40835236                   2                   2
#>  4           21600835           35759241                   5                   7
#>  5           46239948           21094251                   2                   2
#>  6           41075867           41445519                   1                   1
#>  7           44351110           44351110                   0                   0
#>  8           35807212           41176890                   2                   2
#>  9           46308835           21022210                   2                   2
#> 10           41102028           41102028                   0                   0