Get Test Drug Concepts

Get random records of RxNorm and RxNorm Extension Concepts that are:

Valid,

Not a Class,

Belong to the Drug domain

Get random records of ATC and HemOnc Drug Classes that are:

Valid,

Class,

Belong to the Drug domain

get_test_drug_concepts(
  conn,
  conn_fun = "connectAthena()",
  schema = "omop_vocabulary",
  limit = 100,
  verbose = TRUE,
  render_sql = TRUE
)

get_test_drug_classes(
  conn,
  conn_fun = "connectAthena()",
  schema = "omop_vocabulary",
  limit = 100,
  verbose = TRUE,
  render_sql = TRUE
)

See also

Examples

library(chariot)
library(tidyverse)

# Get Unfiltered Test Concepts (Invalid Concepts are included)
get_test_concepts()

#> Connecting using PostgreSQL driver
#> [2021-01-10 12:11:25]	Overriding cache... Querying Athena...
#> [2021-01-10 12:11:25] 

#> ✓ Open connection

#> [2021-01-10 12:11:25] 

#> ✓ JDBC connection

#> [2021-01-10 12:11:25]	SQL: SELECT * FROM omop_vocabulary.concept ORDER BY RANDOM() LIMIT 100
#> [2021-01-10 12:11:25]	Querying...
#> [2021-01-10 12:11:34]	Querying...complete
#> [2021-01-10 12:11:34] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 12:11:34]	Caching resultset...
#> [2021-01-10 12:11:34]	Postgres connection to 'athena' closed

#> # A tibble: 100 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id
#>         <dbl> <chr>        <chr>     <chr>         <chr>           
#>  1   36077717 Lyrica 75mg… Drug      dm+d          AMP             
#>  2   45033876 chlorhexidi… Drug      NDC           11-digit NDC    
#>  3   45232173 sodium chlo… Drug      NDC           9-digit NDC     
#>  4   40754013 Risperidone… Drug      RxNorm Exten… Marketed Product
#>  5   21051318 alpha Tocop… Drug      RxNorm Exten… Branded Drug Box
#>  6   42305826 Pyloroplast… Procedure EDI           Procedure       
#>  7   41916263 Arnprior     Geography OSM           8th level       
#>  8   42548410 ALPROSTADIL… Drug      VA Product    VA Product      
#>  9   43244098 LEVACT 2,5 … Drug      BDPM          Drug Product    
#> 10   21253333 Clobazam     Drug      dm+d          Ingredient      
#> # … with 90 more rows, and 5 more variables: standard_concept <chr>,
#> #   concept_code <chr>, valid_start_date <date>, valid_end_date <date>,
#> #   invalid_reason <chr>

# The size of the resultset can be altered
get_test_concepts(limit = 25)

#> Connecting using PostgreSQL driver
#> [2021-01-10 12:11:34]	Overriding cache... Querying Athena...
#> [2021-01-10 12:11:34] 

#> ✓ Open connection

#> [2021-01-10 12:11:34] 

#> ✓ JDBC connection

#> [2021-01-10 12:11:34]	SQL: SELECT * FROM omop_vocabulary.concept ORDER BY RANDOM() LIMIT 25
#> [2021-01-10 12:11:34]	Querying...
#> [2021-01-10 12:11:42]	Querying...complete
#> [2021-01-10 12:11:42] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 12:11:42]	Caching resultset...
#> [2021-01-10 12:11:42]	Postgres connection to 'athena' closed

#> # A tibble: 25 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id
#>         <dbl> <chr>        <chr>     <chr>         <chr>           
#>  1    2837512 Medical and… Procedure ICD10PCS      ICD10PCS Hierar…
#>  2    4313837 Excision of… Procedure SNOMED        Procedure       
#>  3   43849771 0.75 MG Sen… Drug      RxNorm Exten… Marketed Product
#>  4   45917138 accidental … Condition CIEL          Diagnosis       
#>  5   38003184 Laboratory … Revenue … Revenue Code  Revenue Code    
#>  6   36818265 Generic Lep… Device    dm+d          VMP             
#>  7   36525167 Anaplastic … Condition ICDO3         ICDO Condition  
#>  8   41889370 Upper Dartb… Geography OSM           10th level      
#>  9   43676791 Lysimachia … Drug      RxNorm Exten… Branded Drug Fo…
#> 10    4320215 Hyperlipopr… Observat… SNOMED        Clinical Finding
#> # … with 15 more rows, and 5 more variables: standard_concept <chr>,
#> #   concept_code <chr>, valid_start_date <date>, valid_end_date <date>,
#> #   invalid_reason <chr>

# Concepts specific to the Drug domain can be sampled
# RxNorm/RxNorm Extension Concepts
get_test_drug_concepts(limit = 10)

#> Connecting using PostgreSQL driver
#> [2021-01-10 12:11:42]	Overriding cache... Querying Athena...
#> [2021-01-10 12:11:42] 

#> ✓ Open connection

#> [2021-01-10 12:11:42] 

#> ✓ JDBC connection

#> [2021-01-10 12:11:42]	SQL: SELECT * FROM omop_vocabulary.concept c WHERE c.invalid_reason IS NULL AND c.vocabulary_id IN ('RxNorm', 'RxNorm Extension') AND c.standard_concept <> 'C' AND c.domain_id = 'Drug' ORDER BY RANDOM() LIMIT 10
#> [2021-01-10 12:11:42]	Querying...
#> [2021-01-10 12:11:46]	Querying...complete
#> [2021-01-10 12:11:46] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 12:11:46]	Caching resultset...
#> [2021-01-10 12:11:46]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id
#>         <dbl> <chr>        <chr>     <chr>         <chr>           
#>  1   43650218 2500 MG Sal… Drug      RxNorm Exten… Marketed Product
#>  2   41502955 70000 MG me… Drug      RxNorm Exten… Quant Clinical …
#>  3   35854683 gliclazide … Drug      RxNorm Exten… Branded Drug Fo…
#>  4   43607563 1000 MG Bet… Drug      RxNorm Exten… Marketed Product
#>  5   43210092 Ramipril 10… Drug      RxNorm Exten… Marketed Product
#>  6   36420723 Methotrexat… Drug      RxNorm Exten… Clinical Drug B…
#>  7   40943759 Amoxicillin… Drug      RxNorm Exten… Branded Drug    
#>  8   41049790 Phenobarbit… Drug      RxNorm Exten… Branded Drug Fo…
#>  9   43620041 ASPARAGINAS… Drug      RxNorm Exten… Clinical Drug   
#> 10   41055613 Carbamazepi… Drug      RxNorm Exten… Clinical Drug B…
#> # … with 5 more variables: standard_concept <chr>, concept_code <chr>,
#> #   valid_start_date <date>, valid_end_date <date>, invalid_reason <chr>

# HemOnc/ATC Classes
get_test_drug_classes(limit = 10)

#> Connecting using PostgreSQL driver
#> [2021-01-10 12:11:46]	Overriding cache... Querying Athena...
#> [2021-01-10 12:11:46] 

#> ✓ Open connection

#> [2021-01-10 12:11:46] 

#> ✓ JDBC connection

#> [2021-01-10 12:11:46]	SQL: SELECT * FROM omop_vocabulary.concept c WHERE c.invalid_reason IS NULL AND c.vocabulary_id IN ('ATC', 'HemOnc') AND c.standard_concept = 'C' AND c.domain_id = 'Drug' ORDER BY RANDOM() LIMIT 10
#> [2021-01-10 12:11:46]	Querying...
#> [2021-01-10 12:11:46]	Querying...complete
#> [2021-01-10 12:11:46] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 12:11:46]	Caching resultset...
#> [2021-01-10 12:11:46]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id
#>         <dbl> <chr>        <chr>     <chr>         <chr>           
#>  1   21602339 pimecrolimu… Drug      ATC           ATC 5th         
#>  2   21602131 halometason… Drug      ATC           ATC 5th         
#>  3   21605202 prednisolon… Drug      ATC           ATC 5th         
#>  4   21600160 oxyphenoniu… Drug      ATC           ATC 5th         
#>  5   21605252 4-dimethyla… Drug      ATC           ATC 5th         
#>  6   21601577 pentoxifyll… Drug      ATC           ATC 5th         
#>  7   44819158 vortioxetin… Drug      ATC           ATC 5th         
#>  8   21602348 tiratricol;… Drug      ATC           ATC 5th         
#>  9   21603995 parecoxib; … Drug      ATC           ATC 5th         
#> 10     715884 entinostat   Drug      ATC           ATC 5th         
#> # … with 5 more variables: standard_concept <chr>, concept_code <chr>,
#> #   valid_start_date <date>, valid_end_date <date>, invalid_reason <chr>

# Test data can also be taken from tables other than Concept
get_test_data(table = "concept_relationship")

#> Connecting using PostgreSQL driver
#> [2021-01-10 12:11:46]	Overriding cache... Querying Athena...
#> [2021-01-10 12:11:46] 

#> ✓ Open connection

#> [2021-01-10 12:11:46] 

#> ✓ JDBC connection

#> [2021-01-10 12:11:46]	SQL: SELECT * FROM omop_vocabulary.concept_relationship ORDER BY RANDOM() LIMIT 100
#> [2021-01-10 12:11:46]	Querying...
#> [2021-01-10 12:12:18]	Querying...complete
#> [2021-01-10 12:12:18] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 12:12:18]	Caching resultset...
#> [2021-01-10 12:12:18]	Postgres connection to 'athena' closed

#> # A tibble: 100 x 6
#>    concept_id_1 concept_id_2 relationship_id valid_start_date valid_end_date
#>           <dbl>        <dbl> <chr>           <date>           <date>        
#>  1     41018739     35757950 Mapped from     2017-08-25       2099-12-31    
#>  2     40718969     40718969 Mapped from     2018-09-19       2099-12-31    
#>  3     46070034     35607932 Has non-avail … 2018-07-31       2099-12-31    
#>  4     44163899      1711273 Marketed form … 2017-10-06       2099-12-31    
#>  5     43677825     43677825 Maps to         2016-12-18       2099-12-31    
#>  6     37209200     46052135 Subsumes        2019-07-31       2099-12-31    
#>  7      4143069      3268912 Mapped from     1970-01-01       2099-12-31    
#>  8     21129594     19082228 RxNorm has dos… 2015-08-17       2099-12-31    
#>  9     36414201     36414232 Maps to         2018-08-02       2099-12-31    
#> 10     43726190     19082168 RxNorm has dos… 2016-07-01       2099-12-31    
#> # … with 90 more rows, and 1 more variable: invalid_reason <chr>

get_test_data(table = "concept_ancestor", limit = 10)

#> Connecting using PostgreSQL driver
#> [2021-01-10 12:12:19]	Overriding cache... Querying Athena...
#> [2021-01-10 12:12:19] 

#> ✓ Open connection

#> [2021-01-10 12:12:19] 

#> ✓ JDBC connection

#> [2021-01-10 12:12:19]	SQL: SELECT * FROM omop_vocabulary.concept_ancestor ORDER BY RANDOM() LIMIT 10
#> [2021-01-10 12:12:19]	Querying...
#> [2021-01-10 12:12:54]	Querying...complete
#> [2021-01-10 12:12:54] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 12:12:54]	Caching resultset...
#> [2021-01-10 12:12:54]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 4
#>    ancestor_concept_… descendant_concep… min_levels_of_sepa… max_levels_of_sepa…
#>                 <dbl>              <dbl>               <dbl>               <dbl>
#>  1            4022759           36714747                   6                   6
#>  2            4126705           36711456                   4                   5
#>  3           40097261           41489542                   2                   2
#>  4           45703519           41009810                   2                   3
#>  5           46239620           36265315                   2                   2
#>  6           41201689           41488776                   1                   1
#>  7           21604220           21087967                   5                   6
#>  8            4027883           36684705                   4                   4
#>  9           45670100           43754805                   1                   1
#> 10           36248739           40745180                   3                   4

library(chariot)
library(tidyverse)

# Get Unfiltered Test Concepts (Invalid Concepts are included)
get_test_concepts()

#> Connecting using PostgreSQL driver
#> [2021-01-10 12:12:54]	Overriding cache... Querying Athena...
#> [2021-01-10 12:12:54] 

#> ✓ Open connection

#> [2021-01-10 12:12:54] 

#> ✓ JDBC connection

#> [2021-01-10 12:12:54]	SQL: SELECT * FROM omop_vocabulary.concept ORDER BY RANDOM() LIMIT 100
#> [2021-01-10 12:12:54]	Querying...
#> [2021-01-10 12:13:04]	Querying...complete
#> [2021-01-10 12:13:04] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 12:13:04]	Caching resultset...
#> [2021-01-10 12:13:04]	Postgres connection to 'athena' closed

#> # A tibble: 100 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id
#>         <dbl> <chr>        <chr>     <chr>         <chr>           
#>  1   36532311 Burkitt lym… Condition ICDO3         ICDO Condition  
#>  2   35976597 V832M        Measurem… JAX           Variant         
#>  3   40752417 Calcium 100… Drug      RxNorm Exten… Marketed Product
#>  4   43621547 Ifosfamide … Drug      RxNorm Exten… Clinical Drug C…
#>  5   42482415 candesartan… Drug      RxNorm Exten… Marketed Product
#>  6   44462924 Fexofenadin… Drug      NDC           11-digit NDC    
#>  7   45746068 hydrocortis… Drug      SPL           OTC Drug        
#>  8   36281254 0.6 ML Epoe… Drug      RxNorm Exten… Quant Branded B…
#>  9   19097276 ALLERGENIC … Drug      RxNorm        Ingredient      
#> 10   35611452 Virtual rad… Drug      SNOMED        Pharma/Biol Pro…
#> # … with 90 more rows, and 5 more variables: standard_concept <chr>,
#> #   concept_code <chr>, valid_start_date <date>, valid_end_date <date>,
#> #   invalid_reason <chr>

# The size of the resultset can be altered
get_test_concepts(limit = 25)

#> Connecting using PostgreSQL driver
#> [2021-01-10 12:13:04]	Overriding cache... Querying Athena...
#> [2021-01-10 12:13:04] 

#> ✓ Open connection

#> [2021-01-10 12:13:04] 

#> ✓ JDBC connection

#> [2021-01-10 12:13:04]	SQL: SELECT * FROM omop_vocabulary.concept ORDER BY RANDOM() LIMIT 25
#> [2021-01-10 12:13:04]	Querying...
#> [2021-01-10 12:13:12]	Querying...complete
#> [2021-01-10 12:13:12] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 12:13:12]	Caching resultset...
#> [2021-01-10 12:13:12]	Postgres connection to 'athena' closed

#> # A tibble: 25 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id
#>         <dbl> <chr>        <chr>     <chr>         <chr>           
#>  1    4293637 ACETAMINOPH… Drug      VA Product    VA Product      
#>  2   35921613 Stated as T… Meas Val… NAACCR        NAACCR Value    
#>  3   21016276 Natravene L… Drug      RxNorm Exten… Brand Name      
#>  4   43631374 100 ML vale… Drug      RxNorm Exten… Marketed Product
#>  5    4329809 HYDROCHLORO… Drug      VA Product    VA Product      
#>  6   44540765 RENO-DIP     Drug      DPD           Brand Name      
#>  7   44055964 Cupric oxid… Drug      RxNorm Exten… Clinical Drug F…
#>  8   45176730 Flurazepam … Drug      NDC           11-digit NDC    
#>  9   42958071 Clarithromy… Drug      RxNorm Exten… Branded Drug Co…
#> 10   42089401 CANCELLOUS … Device    EDI           Device          
#> # … with 15 more rows, and 5 more variables: standard_concept <chr>,
#> #   concept_code <chr>, valid_start_date <date>, valid_end_date <date>,
#> #   invalid_reason <chr>

# Concepts specific to the Drug domain can be sampled
# RxNorm/RxNorm Extension Concepts
get_test_drug_concepts(limit = 10)

#> Connecting using PostgreSQL driver
#> [2021-01-10 12:13:12]	Overriding cache... Querying Athena...
#> [2021-01-10 12:13:12] 

#> ✓ Open connection

#> [2021-01-10 12:13:12] 

#> ✓ JDBC connection

#> [2021-01-10 12:13:12]	SQL: SELECT * FROM omop_vocabulary.concept c WHERE c.invalid_reason IS NULL AND c.vocabulary_id IN ('RxNorm', 'RxNorm Extension') AND c.standard_concept <> 'C' AND c.domain_id = 'Drug' ORDER BY RANDOM() LIMIT 10
#> [2021-01-10 12:13:12]	Querying...
#> [2021-01-10 12:13:16]	Querying...complete
#> [2021-01-10 12:13:16] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 12:13:16]	Caching resultset...
#> [2021-01-10 12:13:16]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id standard_concept concept_code valid_start_date valid_end_date
#>         <dbl> <chr>        <chr>     <chr>         <chr>            <chr>            <chr>        <date>           <date>        
#>  1   41032048 Hydromorpho… Drug      RxNorm Exten… Branded Drug     S                OMOP2230010  2017-08-24       2099-12-31    
#>  2   43802072 Arnica mont… Drug      RxNorm Exten… Clinical Drug    S                OMOP659236   2017-08-24       2099-12-31    
#>  3   43515870 acellular p… Drug      RxNorm Exten… Branded Drug Co… S                OMOP2802985  2017-08-29       2099-12-31    
#>  4   41270238 Achillea mi… Drug      RxNorm Exten… Branded Drug Fo… S                OMOP2468200  2017-08-24       2099-12-31    
#>  5    2064345 rosuvastati… Drug      RxNorm Exten… Branded Drug     S                OMOP4965647  2020-07-31       2099-12-31    
#>  6   42902959 nicotine 0.… Drug      RxNorm        Clinical Drug    S                1360072      2012-12-03       2099-12-31    
#>  7   36258439 Dydrogester… Drug      RxNorm Exten… Marketed Product S                OMOP3109314  2017-10-01       2099-12-31    
#>  8   43293077 alpha Tocop… Drug      RxNorm Exten… Clinical Drug F… S                OMOP567236   2017-08-24       2099-12-31    
#>  9   42954845 Roxithromyc… Drug      RxNorm Exten… Branded Drug Co… S                OMOP4681391  2018-08-21       2099-12-31    
#> 10    2045463 calcium 500… Drug      RxNorm Exten… Marketed Product S                OMOP4945609  2020-07-31       2099-12-31    
#> # … with 1 more variable: invalid_reason <chr>

# HemOnc/ATC Classes
get_test_drug_classes(limit = 10)

#> Connecting using PostgreSQL driver
#> [2021-01-10 12:13:16]	Overriding cache... Querying Athena...
#> [2021-01-10 12:13:16] 

#> ✓ Open connection

#> [2021-01-10 12:13:16] 

#> ✓ JDBC connection

#> [2021-01-10 12:13:16]	SQL: SELECT * FROM omop_vocabulary.concept c WHERE c.invalid_reason IS NULL AND c.vocabulary_id IN ('ATC', 'HemOnc') AND c.standard_concept = 'C' AND c.domain_id = 'Drug' ORDER BY RANDOM() LIMIT 10
#> [2021-01-10 12:13:16]	Querying...
#> [2021-01-10 12:13:16]	Querying...complete
#> [2021-01-10 12:13:16] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 12:13:16]	Caching resultset...
#> [2021-01-10 12:13:16]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id standard_concept concept_code valid_start_date valid_end_date
#>         <dbl> <chr>        <chr>     <chr>         <chr>            <chr>            <chr>        <date>           <date>        
#>  1   21602531 methallenes… Drug      ATC           ATC 5th          C                G03CC03      1970-01-01       2099-12-31    
#>  2   21600761 glimepiride… Drug      ATC           ATC 5th          C                A10BB12      1970-01-01       2099-12-31    
#>  3   21603709 Prostagland… Drug      ATC           ATC 4th          C                S01EE        1970-01-01       2099-12-31    
#>  4   21601753 nisoldipine… Drug      ATC           ATC 5th          C                C08CA07      1970-01-01       2099-12-31    
#>  5   21604268 codeine, co… Drug      ATC           ATC 5th          C                N02AA79      1970-01-01       2099-12-31    
#>  6   21604129 ANTIGOUT PR… Drug      ATC           ATC 3rd          C                M04A         1970-01-01       2099-12-31    
#>  7   21604493 levomeproma… Drug      ATC           ATC 5th          C                N05AA02      1970-01-01       2099-12-31    
#>  8   40252640 ticagrelor;… Drug      ATC           ATC 5th          C                B01AC24      1970-01-01       2099-12-31    
#>  9   21605116 hyaluronic … Drug      ATC           ATC 5th          C                S01KA51      1970-01-01       2099-12-31    
#> 10   21604505 trifluopera… Drug      ATC           ATC 5th          C                N05AB06      1970-01-01       2099-12-31    
#> # … with 1 more variable: invalid_reason <chr>

# Test data can also be taken from tables other than Concept
get_test_data(table = "concept_relationship")

#> Connecting using PostgreSQL driver
#> [2021-01-10 12:13:16]	Overriding cache... Querying Athena...
#> [2021-01-10 12:13:16] 

#> ✓ Open connection

#> [2021-01-10 12:13:16] 

#> ✓ JDBC connection

#> [2021-01-10 12:13:16]	SQL: SELECT * FROM omop_vocabulary.concept_relationship ORDER BY RANDOM() LIMIT 100
#> [2021-01-10 12:13:16]	Querying...
#> [2021-01-10 12:13:49]	Querying...complete
#> [2021-01-10 12:13:49] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 12:13:49]	Caching resultset...
#> [2021-01-10 12:13:49]	Postgres connection to 'athena' closed

#> # A tibble: 100 x 6
#>    concept_id_1 concept_id_2 relationship_id      valid_start_date valid_end_date invalid_reason
#>           <dbl>        <dbl> <chr>                <date>           <date>         <chr>         
#>  1     41320091       713825 Marketed form of     2017-07-18       2099-12-31     NA            
#>  2     45667420     19088587 SPL - RxNorm         1970-01-01       2099-12-31     NA            
#>  3     35151747     35128969 Has supplier         2019-03-11       2099-12-31     NA            
#>  4     37309763     46170621 Paral imprt of       2020-03-09       2099-12-31     NA            
#>  5     43580493     44170628 Supplier of          2017-10-06       2099-12-31     NA            
#>  6     35626993     46180497 Disc indicator of    2018-07-31       2099-12-31     NA            
#>  7     19097799     41266896 Has tradename        2017-07-18       2099-12-31     NA            
#>  8      2761319      2837763 Is a                 2014-04-22       2099-12-31     NA            
#>  9      1001563     45876022 Is a                 1970-01-01       2099-12-31     NA            
#> 10      4208884     37116278 Has denominator unit 2017-07-31       2099-12-31     NA            
#> # … with 90 more rows

get_test_data(table = "concept_ancestor", limit = 10)

#> Connecting using PostgreSQL driver
#> [2021-01-10 12:13:49]	Overriding cache... Querying Athena...
#> [2021-01-10 12:13:49] 

#> ✓ Open connection

#> [2021-01-10 12:13:49] 

#> ✓ JDBC connection

#> [2021-01-10 12:13:49]	SQL: SELECT * FROM omop_vocabulary.concept_ancestor ORDER BY RANDOM() LIMIT 10
#> [2021-01-10 12:13:49]	Querying...
#> [2021-01-10 12:14:23]	Querying...complete
#> [2021-01-10 12:14:23] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 12:14:23]	Caching resultset...
#> [2021-01-10 12:14:23]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 4
#>    ancestor_concept_id descendant_concept_id min_levels_of_separation max_levels_of_separation
#>                  <dbl>                 <dbl>                    <dbl>                    <dbl>
#>  1            45675698              41093962                        2                        2
#>  2             2895585               2874657                        2                        2
#>  3            45669706              41357340                        2                        2
#>  4            45897590              40870753                        2                        2
#>  5            21601463                974423                        3                        4
#>  6            46298907              42972115                        1                        1
#>  7            41970941              41980239                        3                        3
#>  8            45693311              35133456                        1                        1
#>  9            45693042              43660369                        2                        2
#> 10             4132416               4127835                        5                        5