Get Test Drug Concepts

Get random records of RxNorm and RxNorm Extension Concepts that are:

Valid,

Not a Class,

Belong to the Drug domain

get_test_drug_concepts(
  conn,
  conn_fun = "connectAthena()",
  schema = "omop_vocabulary",
  limit = 100,
  verbose = TRUE,
  render_sql = TRUE
)

See also

Examples

library(chariot)
library(tidyverse)

# Get Unfiltered Test Concepts (Invalid Concepts are included)
get_test_concepts()

#> Connecting using PostgreSQL driver

#> [2021-04-23 19:25:51]	Overriding cache... Querying Athena...
#> [2021-04-23 19:25:51] 

#> ✓ Open connection

#> [2021-04-23 19:25:51] 

#> ✓ JDBC connection

#> [2021-04-23 19:25:51]	SQL: SELECT * FROM omop_vocabulary.concept ORDER BY RANDOM() LIMIT 100
#> [2021-04-23 19:25:51]	Querying...
#> [2021-04-23 19:26:00]	Querying...complete
#> [2021-04-23 19:26:00] 

#> ✓ Returned data has more than 0 rows

#> [2021-04-23 19:26:00]	Caching resultset...
#> [2021-04-23 19:26:00]	Postgres connection to 'athena' closed

#> # A tibble: 100 x 10
#>    concept_id concept_name             domain_id  vocabulary_id concept_class_id
#>         <int> <chr>                    <chr>      <chr>         <chr>           
#>  1   41353968 5 ML Chloramphenicol 27… Drug       RxNorm Exten… Quant Clinical …
#>  2   45214343 mirtazapine 15mg/1 ORAL… Drug       NDC           9-digit NDC     
#>  3   42290632 Enterostomy-Tube,Neonat… Procedure  EDI           Procedure       
#>  4   44989648 Chlorpheniramine Maleat… Drug       NDC           11-digit NDC    
#>  5   41381453 20 ML Fenoterol 5 MG/ML… Drug       RxNorm Exten… Quant Branded D…
#>  6   46274209 tesamorelin acetate      Drug       RxNorm        Precise Ingredi…
#>  7   40771964 Major treatments - disc… Observati… LOINC         Survey          
#>  8    4203934 Alimemazine tartrate     Drug       SNOMED        Substance       
#>  9   40777455 West nile virus RNA | B… Measureme… LOINC         LOINC Hierarchy 
#> 10    3036633 Microscopic observation… Measureme… LOINC         Lab Test        
#> # … with 90 more rows, and 5 more variables: standard_concept <chr>,
#> #   concept_code <chr>, valid_start_date <date>, valid_end_date <date>,
#> #   invalid_reason <chr>

# The size of the resultset can be altered
get_test_concepts(limit = 25)

#> Connecting using PostgreSQL driver

#> [2021-04-23 19:26:00]	Overriding cache... Querying Athena...
#> [2021-04-23 19:26:00] 

#> ✓ Open connection

#> [2021-04-23 19:26:00] 

#> ✓ JDBC connection

#> [2021-04-23 19:26:00]	SQL: SELECT * FROM omop_vocabulary.concept ORDER BY RANDOM() LIMIT 25
#> [2021-04-23 19:26:00]	Querying...
#> [2021-04-23 19:26:08]	Querying...complete
#> [2021-04-23 19:26:08] 

#> ✓ Returned data has more than 0 rows

#> [2021-04-23 19:26:08]	Caching resultset...
#> [2021-04-23 19:26:08]	Postgres connection to 'athena' closed

#> # A tibble: 25 x 10
#>    concept_id concept_name              domain_id vocabulary_id concept_class_id
#>         <int> <chr>                     <chr>     <chr>         <chr>           
#>  1   45354939 DAVOL FOLEY CATH INSRT T… Device    NDC           Device          
#>  2   45611802 Hand Deformities, Acquir… Condition MeSH          Main Heading    
#>  3   42864241 CYCLOSPORINE MICROEMULSI… Drug      VA Product    VA Product      
#>  4    3259590 Main bronchus (less than… Condition Nebraska Lex… Clinical Finding
#>  5   36704980 Juzo Dynamic class 1 (18… Device    SNOMED        Physical Object 
#>  6   41513031 100000 MG Ibuprofen 0.05… Drug      RxNorm Exten… Quant Clinical …
#>  7   45843247 clopidogrel bisulfate 30… Drug      NDC           9-digit NDC     
#>  8   21280255 Rivastigmine 2mg/ml oral… Drug      dm+d          VMPP            
#>  9   21416798 Sigvaris Essential Therm… Device    dm+d          AMPP            
#> 10   21426829 Duomed Soft class III th… Device    dm+d          AMPP            
#> # … with 15 more rows, and 5 more variables: standard_concept <chr>,
#> #   concept_code <chr>, valid_start_date <date>, valid_end_date <date>,
#> #   invalid_reason <chr>

# Concepts specific to the Drug domain can be sampled
# RxNorm/RxNorm Extension Concepts
get_test_drug_concepts(limit = 10)

#> Connecting using PostgreSQL driver

#> [2021-04-23 19:26:08]	Overriding cache... Querying Athena...
#> [2021-04-23 19:26:08] 

#> ✓ Open connection

#> [2021-04-23 19:26:08] 

#> ✓ JDBC connection

#> [2021-04-23 19:26:08]	SQL: SELECT * FROM omop_vocabulary.concept c WHERE c.invalid_reason IS NULL AND c.vocabulary_id IN ('RxNorm', 'RxNorm Extension') AND c.standard_concept <> 'C' AND c.domain_id = 'Drug' ORDER BY RANDOM() LIMIT 10
#> [2021-04-23 19:26:08]	Querying...
#> [2021-04-23 19:26:12]	Querying...complete
#> [2021-04-23 19:26:12] 

#> ✓ Returned data has more than 0 rows

#> [2021-04-23 19:26:12]	Caching resultset...
#> [2021-04-23 19:26:12]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 10
#>    concept_id concept_name              domain_id vocabulary_id concept_class_id
#>         <int> <chr>                     <chr>     <chr>         <chr>           
#>  1   21129954 Nafronyl 100 MG Oral Cap… Drug      RxNorm Exten… Marketed Product
#>  2   43734571 1 ML travoprost 0.04 MG/… Drug      RxNorm Exten… Quant Branded D…
#>  3   41200436 Biotin 5 MG Oral Tablet … Drug      RxNorm Exten… Marketed Product
#>  4   40873598 Omeprazole 10 MG Oral Ca… Drug      RxNorm Exten… Branded Drug Box
#>  5   36054732 2000 ML Eucalyptus oil 8… Drug      RxNorm Exten… Marketed Product
#>  6   41146177 Valproate 150 MG Extende… Drug      RxNorm Exten… Clinical Drug B…
#>  7   40836099 Ginkgo biloba extract 40… Drug      RxNorm Exten… Clinical Drug B…
#>  8   35143344 1 ML lomefloxacin 3 MG/M… Drug      RxNorm Exten… Quant Branded D…
#>  9   41226324 temozolomide 5 MG Oral C… Drug      RxNorm Exten… Marketed Product
#> 10    2048110 ceftriaxone 1500 MG Inje… Drug      RxNorm Exten… Branded Drug    
#> # … with 5 more variables: standard_concept <chr>, concept_code <chr>,
#> #   valid_start_date <date>, valid_end_date <date>, invalid_reason <chr>

# HemOnc/ATC Classes
get_test_drug_classes(limit = 10)

#> Connecting using PostgreSQL driver

#> [2021-04-23 19:26:12]	Overriding cache... Querying Athena...
#> [2021-04-23 19:26:12] 

#> ✓ Open connection

#> [2021-04-23 19:26:12] 

#> ✓ JDBC connection

#> [2021-04-23 19:26:12]	SQL: SELECT * FROM omop_vocabulary.concept c WHERE c.invalid_reason IS NULL AND c.vocabulary_id IN ('ATC', 'HemOnc') AND c.standard_concept = 'C' AND c.domain_id = 'Drug' ORDER BY RANDOM() LIMIT 10
#> [2021-04-23 19:26:12]	Querying...
#> [2021-04-23 19:26:12]	Querying...complete
#> [2021-04-23 19:26:12] 

#> ✓ Returned data has more than 0 rows

#> [2021-04-23 19:26:12]	Caching resultset...
#> [2021-04-23 19:26:12]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 10
#>    concept_id concept_name              domain_id vocabulary_id concept_class_id
#>         <int> <chr>                     <chr>     <chr>         <chr>           
#>  1   21604407 trimethadione; oral       Drug      ATC           ATC 5th         
#>  2   21603939 kebuzone                  Drug      ATC           ATC 5th         
#>  3   21602085 aciclovir; topical        Drug      ATC           ATC 5th         
#>  4   21601132 Other antianemic prepara… Drug      ATC           ATC 4th         
#>  5   21605289 nitrogen                  Drug      ATC           ATC 5th         
#>  6   45893475 valethamate               Drug      ATC           ATC 5th         
#>  7   21604697 nortriptyline; systemic   Drug      ATC           ATC 5th         
#>  8     715890 avatrombopag; systemic    Drug      ATC           ATC 5th         
#>  9   21604686 ANTIDEPRESSANTS           Drug      ATC           ATC 3rd         
#> 10   21604413 ethosuximide, combinatio… Drug      ATC           ATC 5th         
#> # … with 5 more variables: standard_concept <chr>, concept_code <chr>,
#> #   valid_start_date <date>, valid_end_date <date>, invalid_reason <chr>

# Test data can also be taken from tables other than Concept
get_test_data(table = "concept_relationship")

#> Connecting using PostgreSQL driver

#> [2021-04-23 19:26:12]	Overriding cache... Querying Athena...
#> [2021-04-23 19:26:12] 

#> ✓ Open connection

#> [2021-04-23 19:26:12] 

#> ✓ JDBC connection

#> [2021-04-23 19:26:12]	SQL: SELECT * FROM omop_vocabulary.concept_relationship ORDER BY RANDOM() LIMIT 100
#> [2021-04-23 19:26:12]	Querying...
#> [2021-04-23 19:26:45]	Querying...complete
#> [2021-04-23 19:26:45] 

#> ✓ Returned data has more than 0 rows

#> [2021-04-23 19:26:45]	Caching resultset...
#> [2021-04-23 19:26:45]	Postgres connection to 'athena' closed

#> # A tibble: 100 x 6
#>    concept_id_1 concept_id_2 relationship_id     valid_start_date valid_end_date
#>           <int>        <int> <chr>               <date>           <date>        
#>  1     41013878     40984493 Has tradename       2017-07-18       2099-12-31    
#>  2     43278548     19000633 Has brand name      2016-09-30       2099-12-31    
#>  3      4209406      4209406 Mapped from         1970-01-01       2099-12-31    
#>  4     41516503     41512994 Has marketed form   2017-07-18       2099-12-31    
#>  5     43743686     43795790 Has quantified form 2016-07-01       2099-12-31    
#>  6     43777001     43777001 Mapped from         2016-12-18       2099-12-31    
#>  7     37209200     46077310 Subsumes            2019-07-31       2099-12-31    
#>  8     35407574     35407574 Mapped from         2017-07-17       2099-12-31    
#>  9     40234923     40238587 Has drug-drug inter 2011-05-02       2099-12-31    
#> 10      3041873      3041206 Concept replaced by 2018-06-15       2099-12-31    
#> # … with 90 more rows, and 1 more variable: invalid_reason <chr>

get_test_data(table = "concept_ancestor", limit = 10)

#> Connecting using PostgreSQL driver

#> [2021-04-23 19:26:45]	Overriding cache... Querying Athena...
#> [2021-04-23 19:26:45] 

#> ✓ Open connection

#> [2021-04-23 19:26:45] 

#> ✓ JDBC connection

#> [2021-04-23 19:26:45]	SQL: SELECT * FROM omop_vocabulary.concept_ancestor ORDER BY RANDOM() LIMIT 10
#> [2021-04-23 19:26:45]	Querying...
#> [2021-04-23 19:27:17]	Querying...complete
#> [2021-04-23 19:27:17] 

#> ✓ Returned data has more than 0 rows

#> [2021-04-23 19:27:17]	Caching resultset...
#> [2021-04-23 19:27:17]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 4
#>    ancestor_concept_… descendant_concep… min_levels_of_sepa… max_levels_of_sepa…
#>                 <int>              <int>               <int>               <int>
#>  1           36378674           36378674                   0                   0
#>  2           45697126           41200417                   2                   3
#>  3           43692712           44077749                   2                   2
#>  4           45704781           45704781                   0                   0
#>  5            4023995            4206453                   5                   9
#>  6           42662308           21093175                   2                   3
#>  7            4306780            2784789                   1                   1
#>  8           45782850           40904344                   2                   2
#>  9           46243846           42657724                   2                   3
#> 10           41249006           41491651                   0                   0