Get Test Drug Concepts

Get random records of RxNorm and RxNorm Extension Concepts that are:

Valid,

Not a Class,

Belong to the Drug domain

Get random records of ATC and HemOnc Drug Classes that are:

Valid,

Class,

Belong to the Drug domain

get_test_drug_concepts(
  conn,
  conn_fun = "connectAthena()",
  schema = "omop_vocabulary",
  limit = 100,
  verbose = TRUE,
  render_sql = TRUE
)

get_test_drug_classes(
  conn,
  conn_fun = "connectAthena()",
  schema = "omop_vocabulary",
  limit = 100,
  verbose = TRUE,
  render_sql = TRUE
)

See also

Examples

library(chariot)
library(tidyverse)

# Get Unfiltered Test Concepts (Invalid Concepts are included)
get_test_concepts()

#> Connecting using PostgreSQL driver
#> [2021-01-10 14:34:28]	Overriding cache... Querying Athena...
#> [2021-01-10 14:34:28] 

#> ✓ Open connection

#> [2021-01-10 14:34:28] 

#> ✓ JDBC connection

#> [2021-01-10 14:34:28]	SQL: SELECT * FROM omop_vocabulary.concept ORDER BY RANDOM() LIMIT 100
#> [2021-01-10 14:34:28]	Querying...
#> [2021-01-10 14:34:38]	Querying...complete
#> [2021-01-10 14:34:38] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 14:34:38]	Caching resultset...
#> [2021-01-10 14:34:38]	Postgres connection to 'athena' closed

#> # A tibble: 100 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id
#>         <dbl> <chr>        <chr>     <chr>         <chr>           
#>  1   36177118 glipizide 5… Drug      NDC           11-digit NDC    
#>  2    4303061 Clinical st… Condition SNOMED        Clinical Finding
#>  3   43821201 Mirtazapine… Drug      RxNorm Exten… Branded Drug Co…
#>  4    4167557 Bacillus an… Observat… SNOMED        Organism        
#>  5   21147100 24 HR Cloni… Drug      RxNorm Exten… Marketed Product
#>  6   44440219 acetaminoph… Drug      NDC           11-digit NDC    
#>  7   36104055 lidocaine h… Drug      NDC           11-digit NDC    
#>  8   40789156 I NOS        Observat… LOINC         LOINC Component 
#>  9   42808080 loperamide … Drug      SPL           OTC Drug        
#> 10   45008457 Antihemophi… Drug      NDC           11-digit NDC    
#> # … with 90 more rows, and 5 more variables: standard_concept <chr>,
#> #   concept_code <chr>, valid_start_date <date>, valid_end_date <date>,
#> #   invalid_reason <chr>

# The size of the resultset can be altered
get_test_concepts(limit = 25)

#> Connecting using PostgreSQL driver
#> [2021-01-10 14:34:39]	Overriding cache... Querying Athena...
#> [2021-01-10 14:34:39] 

#> ✓ Open connection

#> [2021-01-10 14:34:39] 

#> ✓ JDBC connection

#> [2021-01-10 14:34:39]	SQL: SELECT * FROM omop_vocabulary.concept ORDER BY RANDOM() LIMIT 25
#> [2021-01-10 14:34:39]	Querying...
#> [2021-01-10 14:34:48]	Querying...complete
#> [2021-01-10 14:34:48] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 14:34:48]	Caching resultset...
#> [2021-01-10 14:34:48]	Postgres connection to 'athena' closed

#> # A tibble: 25 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id
#>         <dbl> <chr>        <chr>     <chr>         <chr>           
#>  1   36174460 allium sati… Drug      SPL           OTC Drug        
#>  2   42254528 Operation o… Procedure EDI           Procedure       
#>  3   41019865 Camphor / E… Drug      RxNorm Exten… Branded Drug Fo…
#>  4   43681794 9000 MG Mag… Drug      RxNorm Exten… Quant Branded D…
#>  5    4348297 Scopulariop… Observat… SNOMED        Organism        
#>  6   36236332 Acupan Oral… Drug      RxNorm        Branded Dose Gr…
#>  7   43748834 Cefuroxime … Drug      RxNorm Exten… Branded Drug Co…
#>  8   45223494 Acetaminoph… Drug      NDC           11-digit NDC    
#>  9   40805088 Dolgit Mikr… Drug      RxNorm Exten… Brand Name      
#> 10   45036944 Tretinoin 0… Drug      NDC           11-digit NDC    
#> # … with 15 more rows, and 5 more variables: standard_concept <chr>,
#> #   concept_code <chr>, valid_start_date <date>, valid_end_date <date>,
#> #   invalid_reason <chr>

# Concepts specific to the Drug domain can be sampled
# RxNorm/RxNorm Extension Concepts
get_test_drug_concepts(limit = 10)

#> Connecting using PostgreSQL driver
#> [2021-01-10 14:34:48]	Overriding cache... Querying Athena...
#> [2021-01-10 14:34:48] 

#> ✓ Open connection

#> [2021-01-10 14:34:48] 

#> ✓ JDBC connection

#> [2021-01-10 14:34:48]	SQL: SELECT * FROM omop_vocabulary.concept c WHERE c.invalid_reason IS NULL AND c.vocabulary_id IN ('RxNorm', 'RxNorm Extension') AND c.standard_concept <> 'C' AND c.domain_id = 'Drug' ORDER BY RANDOM() LIMIT 10
#> [2021-01-10 14:34:48]	Querying...
#> [2021-01-10 14:34:52]	Querying...complete
#> [2021-01-10 14:34:52] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 14:34:52]	Caching resultset...
#> [2021-01-10 14:34:52]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id
#>         <dbl> <chr>        <chr>     <chr>         <chr>           
#>  1   43269052 Ethambutol … Drug      RxNorm Exten… Branded Drug Box
#>  2   41517991 150000 MG h… Drug      RxNorm Exten… Quant Clinical …
#>  3   41272996 factor IX 3… Drug      RxNorm Exten… Clinical Drug   
#>  4   41136453 Escin 20 MG… Drug      RxNorm Exten… Marketed Product
#>  5   43590980 3000 MG Inu… Drug      RxNorm Exten… Marketed Product
#>  6   43797943 Hydrochloro… Drug      RxNorm Exten… Marketed Product
#>  7   43790406 20 ML Bryon… Drug      RxNorm Exten… Quant Branded D…
#>  8   40923258 Calcium 140… Drug      RxNorm Exten… Branded Drug Co…
#>  9   41096249 Escitalopra… Drug      RxNorm Exten… Branded Drug    
#> 10   41060397 Ondansetron… Drug      RxNorm Exten… Branded Drug Box
#> # … with 5 more variables: standard_concept <chr>, concept_code <chr>,
#> #   valid_start_date <date>, valid_end_date <date>, invalid_reason <chr>

# HemOnc/ATC Classes
get_test_drug_classes(limit = 10)

#> Connecting using PostgreSQL driver
#> [2021-01-10 14:34:52]	Overriding cache... Querying Athena...
#> [2021-01-10 14:34:52] 

#> ✓ Open connection

#> [2021-01-10 14:34:52] 

#> ✓ JDBC connection

#> [2021-01-10 14:34:52]	SQL: SELECT * FROM omop_vocabulary.concept c WHERE c.invalid_reason IS NULL AND c.vocabulary_id IN ('ATC', 'HemOnc') AND c.standard_concept = 'C' AND c.domain_id = 'Drug' ORDER BY RANDOM() LIMIT 10
#> [2021-01-10 14:34:52]	Querying...
#> [2021-01-10 14:34:52]	Querying...complete
#> [2021-01-10 14:34:52] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 14:34:52]	Caching resultset...
#> [2021-01-10 14:34:52]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id
#>         <dbl> <chr>        <chr>     <chr>         <chr>           
#>  1   21604963 niclosamide… Drug      ATC           ATC 5th         
#>  2   35807252 Health Cana… Drug      HemOnc        Component Class 
#>  3   21604231 mepivacaine  Drug      ATC           ATC 5th         
#>  4   21603554 chloramphen… Drug      ATC           ATC 5th         
#>  5   21604188 chloroform   Drug      ATC           ATC 5th         
#>  6   21601487 Thiazides, … Drug      ATC           ATC 4th         
#>  7   21604004 glucosamino… Drug      ATC           ATC 5th         
#>  8   21604372 ergotamine,… Drug      ATC           ATC 5th         
#>  9   21604409 Succinimide… Drug      ATC           ATC 4th         
#> 10   21605006 etohexadiol  Drug      ATC           ATC 5th         
#> # … with 5 more variables: standard_concept <chr>, concept_code <chr>,
#> #   valid_start_date <date>, valid_end_date <date>, invalid_reason <chr>

# Test data can also be taken from tables other than Concept
get_test_data(table = "concept_relationship")

#> Connecting using PostgreSQL driver
#> [2021-01-10 14:34:53]	Overriding cache... Querying Athena...
#> [2021-01-10 14:34:53] 

#> ✓ Open connection

#> [2021-01-10 14:34:53] 

#> ✓ JDBC connection

#> [2021-01-10 14:34:53]	SQL: SELECT * FROM omop_vocabulary.concept_relationship ORDER BY RANDOM() LIMIT 100
#> [2021-01-10 14:34:53]	Querying...
#> [2021-01-10 14:35:25]	Querying...complete
#> [2021-01-10 14:35:25] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 14:35:25]	Caching resultset...
#> [2021-01-10 14:35:26]	Postgres connection to 'athena' closed

#> # A tibble: 100 x 6
#>    concept_id_1 concept_id_2 relationship_id valid_start_date valid_end_date
#>           <dbl>        <dbl> <chr>           <date>           <date>        
#>  1     35783173     35783069 Marketed form … 2017-08-02       2099-12-31    
#>  2      3316771     37396398 Maps to         1970-01-01       2099-12-31    
#>  3       735340      3174893 Mapped from     1970-01-01       2099-12-31    
#>  4      4196083      4282599 Subsumes        2018-07-31       2099-12-31    
#>  5     35626986     46110318 Legal category… 2018-07-31       2099-12-31    
#>  6      2813569      2793761 Subsumes        2014-04-22       2099-12-31    
#>  7     19014999     36242158 Brand name of   2016-10-03       2099-12-31    
#>  8     44091224     44091224 Mapped from     2017-06-25       2099-12-31    
#>  9     43758658     43578266 Has supplier    2016-07-01       2099-12-31    
#> 10     41913627     41915421 Subsumes        1970-01-01       2099-12-31    
#> # … with 90 more rows, and 1 more variable: invalid_reason <chr>

get_test_data(table = "concept_ancestor", limit = 10)

#> Connecting using PostgreSQL driver
#> [2021-01-10 14:35:26]	Overriding cache... Querying Athena...
#> [2021-01-10 14:35:26] 

#> ✓ Open connection

#> [2021-01-10 14:35:26] 

#> ✓ JDBC connection

#> [2021-01-10 14:35:26]	SQL: SELECT * FROM omop_vocabulary.concept_ancestor ORDER BY RANDOM() LIMIT 10
#> [2021-01-10 14:35:26]	Querying...
#> [2021-01-10 14:36:02]	Querying...complete
#> [2021-01-10 14:36:02] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 14:36:02]	Caching resultset...
#> [2021-01-10 14:36:02]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 4
#>    ancestor_concept_… descendant_concep… min_levels_of_sepa… max_levels_of_sepa…
#>                 <dbl>              <dbl>               <dbl>               <dbl>
#>  1            4296695            4002062                   2                   2
#>  2            1129625           40734665                   1                   1
#>  3           21604801           43525688                   5                   7
#>  4           36219664           41309582                   2                   3
#>  5           35807487           43708733                   5                   6
#>  6           46243684           43271649                   2                   3
#>  7           46242005           36780757                   2                   3
#>  8           37209199           46213460                   1                   1
#>  9           45634159           40934381                   2                   2
#> 10           45653048           41195350                   2                   3

library(chariot)
library(tidyverse)

# Get Unfiltered Test Concepts (Invalid Concepts are included)
get_test_concepts()

#> Connecting using PostgreSQL driver
#> [2021-01-10 14:36:02]	Overriding cache... Querying Athena...
#> [2021-01-10 14:36:02] 

#> ✓ Open connection

#> [2021-01-10 14:36:02] 

#> ✓ JDBC connection

#> [2021-01-10 14:36:02]	SQL: SELECT * FROM omop_vocabulary.concept ORDER BY RANDOM() LIMIT 100
#> [2021-01-10 14:36:02]	Querying...
#> [2021-01-10 14:36:12]	Querying...complete
#> [2021-01-10 14:36:12] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 14:36:12]	Caching resultset...
#> [2021-01-10 14:36:12]	Postgres connection to 'athena' closed

#> # A tibble: 100 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id
#>         <dbl> <chr>        <chr>     <chr>         <chr>           
#>  1   35106077 petrolatum … Drug      SPL           OTC Drug        
#>  2   21426233 OpSite Plus… Device    dm+d          AMPP            
#>  3    4278870 ANTHELMINTI… Drug      VA Class      VA Class        
#>  4   44152354 Sandoz Glic… Drug      DPD           Drug Product    
#>  5   37510175 ethanol 0.7… Drug      NDC           11-digit NDC    
#>  6   36185500 D-5 - silve… Drug      NDC           11-digit NDC    
#>  7   37009058 Juzo Soft c… Device    SNOMED        Physical Object 
#>  8   21328823 Sigvaris Co… Device    dm+d          AMP             
#>  9   35175926 Glimepiride… Drug      JMDC          Drug Product    
#> 10   40818377 Tramawieb    Drug      RxNorm Exten… Brand Name      
#> # … with 90 more rows, and 5 more variables: standard_concept <chr>,
#> #   concept_code <chr>, valid_start_date <date>, valid_end_date <date>,
#> #   invalid_reason <chr>

# The size of the resultset can be altered
get_test_concepts(limit = 25)

#> Connecting using PostgreSQL driver
#> [2021-01-10 14:36:13]	Overriding cache... Querying Athena...
#> [2021-01-10 14:36:13] 

#> ✓ Open connection

#> [2021-01-10 14:36:13] 

#> ✓ JDBC connection

#> [2021-01-10 14:36:13]	SQL: SELECT * FROM omop_vocabulary.concept ORDER BY RANDOM() LIMIT 25
#> [2021-01-10 14:36:13]	Querying...
#> [2021-01-10 14:36:23]	Querying...complete
#> [2021-01-10 14:36:23] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 14:36:23]	Caching resultset...
#> [2021-01-10 14:36:23]	Postgres connection to 'athena' closed

#> # A tibble: 25 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id
#>         <dbl> <chr>        <chr>     <chr>         <chr>           
#>  1   44570268 OCTINOXATE … Drug      DPD           Branded Drug Co…
#>  2   43825871 10000 MG ma… Drug      RxNorm Exten… Marketed Product
#>  3   37499015 Nexletol     Drug      RxNorm        Brand Name      
#>  4   44206597 10000 MG St… Drug      RxNorm Exten… Quant Clinical …
#>  5   37091400 Acute vagin… Condition ICD10GM       ICD10 code      
#>  6   42745649 Operations … Procedure OPS           Procedure       
#>  7   21058709 20 ML Trich… Drug      RxNorm Exten… Quant Clinical …
#>  8   41251732 Fluoxetine … Drug      RxNorm Exten… Branded Drug Box
#>  9    2720937 Sildenafil … Drug      HCPCS         HCPCS           
#> 10    3125020 Pneumoconio… Condition Nebraska Lex… Clinical Finding
#> # … with 15 more rows, and 5 more variables: standard_concept <chr>,
#> #   concept_code <chr>, valid_start_date <date>, valid_end_date <date>,
#> #   invalid_reason <chr>

# Concepts specific to the Drug domain can be sampled
# RxNorm/RxNorm Extension Concepts
get_test_drug_concepts(limit = 10)

#> Connecting using PostgreSQL driver
#> [2021-01-10 14:36:23]	Overriding cache... Querying Athena...
#> [2021-01-10 14:36:23] 

#> ✓ Open connection

#> [2021-01-10 14:36:23] 

#> ✓ JDBC connection

#> [2021-01-10 14:36:23]	SQL: SELECT * FROM omop_vocabulary.concept c WHERE c.invalid_reason IS NULL AND c.vocabulary_id IN ('RxNorm', 'RxNorm Extension') AND c.standard_concept <> 'C' AND c.domain_id = 'Drug' ORDER BY RANDOM() LIMIT 10
#> [2021-01-10 14:36:23]	Querying...
#> [2021-01-10 14:36:27]	Querying...complete
#> [2021-01-10 14:36:27] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 14:36:27]	Caching resultset...
#> [2021-01-10 14:36:27]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id
#>         <dbl> <chr>        <chr>     <chr>         <chr>           
#>  1    2057322 methylpredn… Drug      RxNorm Exten… Branded Drug    
#>  2    2047190 oseltamivir… Drug      RxNorm Exten… Marketed Product
#>  3   41450258 11.2 ML Ned… Drug      RxNorm Exten… Quant Clinical …
#>  4   21107056 1 ML Labeta… Drug      RxNorm Exten… Quant Clinical …
#>  5   43693446 Clindamycin… Drug      RxNorm Exten… Branded Drug Co…
#>  6   42957343 Cetirizine … Drug      RxNorm Exten… Branded Drug    
#>  7   44116894 Turpentine … Drug      RxNorm Exten… Clinical Drug C…
#>  8   43172859 trandolapri… Drug      RxNorm Exten… Branded Drug Box
#>  9   40968448 Lovastatin … Drug      RxNorm Exten… Branded Drug Box
#> 10   42716454 Trimipramin… Drug      RxNorm Exten… Branded Drug Co…
#> # … with 5 more variables: standard_concept <chr>, concept_code <chr>,
#> #   valid_start_date <date>, valid_end_date <date>, invalid_reason <chr>

# HemOnc/ATC Classes
get_test_drug_classes(limit = 10)

#> Connecting using PostgreSQL driver
#> [2021-01-10 14:36:27]	Overriding cache... Querying Athena...
#> [2021-01-10 14:36:27] 

#> ✓ Open connection

#> [2021-01-10 14:36:27] 

#> ✓ JDBC connection

#> [2021-01-10 14:36:27]	SQL: SELECT * FROM omop_vocabulary.concept c WHERE c.invalid_reason IS NULL AND c.vocabulary_id IN ('ATC', 'HemOnc') AND c.standard_concept = 'C' AND c.domain_id = 'Drug' ORDER BY RANDOM() LIMIT 10
#> [2021-01-10 14:36:27]	Querying...
#> [2021-01-10 14:36:27]	Querying...complete
#> [2021-01-10 14:36:27] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 14:36:27]	Caching resultset...
#> [2021-01-10 14:36:27]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id
#>         <dbl> <chr>        <chr>     <chr>         <chr>           
#>  1   21601463 Thiazides, … Drug      ATC           ATC 4th         
#>  2   21600613 rifaximin; … Drug      ATC           ATC 5th         
#>  3   21602060 tetracyclin… Drug      ATC           ATC 5th         
#>  4   21602349 oxaceprol; … Drug      ATC           ATC 5th         
#>  5   21605053 fluticasone… Drug      ATC           ATC 5th         
#>  6    1588656 selexipag; … Drug      ATC           ATC 5th         
#>  7   21603927 azathioprin… Drug      ATC           ATC 5th         
#>  8   45893463 evolocumab;… Drug      ATC           ATC 5th         
#>  9   21603068 linezolid; … Drug      ATC           ATC 5th         
#> 10   21600156 tiemonium i… Drug      ATC           ATC 5th         
#> # … with 5 more variables: standard_concept <chr>, concept_code <chr>,
#> #   valid_start_date <date>, valid_end_date <date>, invalid_reason <chr>

# Test data can also be taken from tables other than Concept
get_test_data(table = "concept_relationship")

#> Connecting using PostgreSQL driver
#> [2021-01-10 14:36:28]	Overriding cache... Querying Athena...
#> [2021-01-10 14:36:28] 

#> ✓ Open connection

#> [2021-01-10 14:36:28] 

#> ✓ JDBC connection

#> [2021-01-10 14:36:28]	SQL: SELECT * FROM omop_vocabulary.concept_relationship ORDER BY RANDOM() LIMIT 100
#> [2021-01-10 14:36:28]	Querying...
#> [2021-01-10 14:37:09]	Querying...complete
#> [2021-01-10 14:37:09] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 14:37:09]	Caching resultset...
#> [2021-01-10 14:37:09]	Postgres connection to 'athena' closed

#> # A tibble: 100 x 6
#>    concept_id_1 concept_id_2 relationship_id valid_start_date valid_end_date
#>           <dbl>        <dbl> <chr>           <date>           <date>        
#>  1     41208573     44109961 Concept replac… 2017-07-30       2099-12-31    
#>  2     21192013     36808688 Maps to         2019-01-28       2099-12-31    
#>  3     40077149     35133666 Has tradename   2019-03-11       2099-12-31    
#>  4      1436056      1419487 Subsumes        2016-01-01       2099-12-31    
#>  5     19082573     41477654 RxNorm dose fo… 2017-07-18       2099-12-31    
#>  6      3437243      3474420 Subsumes        2019-08-16       2099-12-31    
#>  7     19069088      4132544 Mapped from     1970-01-01       2099-12-31    
#>  8     41315313     19042583 Has brand name  2017-07-18       2099-12-31    
#>  9     40198820     35909872 Value to Schema 1970-01-01       2099-12-31    
#> 10      2866246      2866246 Mapped from     2014-04-22       2099-12-31    
#> # … with 90 more rows, and 1 more variable: invalid_reason <chr>

get_test_data(table = "concept_ancestor", limit = 10)

#> Connecting using PostgreSQL driver
#> [2021-01-10 14:37:09]	Overriding cache... Querying Athena...
#> [2021-01-10 14:37:09] 

#> ✓ Open connection

#> [2021-01-10 14:37:09] 

#> ✓ JDBC connection

#> [2021-01-10 14:37:09]	SQL: SELECT * FROM omop_vocabulary.concept_ancestor ORDER BY RANDOM() LIMIT 10
#> [2021-01-10 14:37:09]	Querying...
#> [2021-01-10 14:37:50]	Querying...complete
#> [2021-01-10 14:37:50] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 14:37:51]	Caching resultset...
#> [2021-01-10 14:37:51]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 4
#>    ancestor_concept_… descendant_concep… min_levels_of_sepa… max_levels_of_sepa…
#>                 <dbl>              <dbl>               <dbl>               <dbl>
#>  1           21168319           36414141                   1                   1
#>  2            4090035            4118600                   3                   3
#>  3           21605228           43034693                   5                   7
#>  4            4297453           45759766                   3                   3
#>  5           35807375           41247566                   3                   4
#>  6           36410009           36410009                   0                   0
#>  7           45702623           35415023                   2                   3
#>  8           45786768           44101583                   2                   2
#>  9           41929276           41939875                   2                   2
#> 10             785788           19112719                   1                   1