Get Test Drug Concepts

Get random records of RxNorm and RxNorm Extension Concepts that are:

Valid,

Not a Class,

Belong to the Drug domain

get_test_drug_concepts(
  conn,
  conn_fun = "connectAthena()",
  schema = "omop_vocabulary",
  limit = 100,
  verbose = TRUE,
  render_sql = TRUE
)

See also

Examples

library(chariot)
library(tidyverse)

# Get Unfiltered Test Concepts (Invalid Concepts are included)
get_test_concepts()

#> Connecting using PostgreSQL driver

#> [2021-03-28 18:49:52]	Overriding cache... Querying Athena...
#> [2021-03-28 18:49:52] 

#> ✓ Open connection

#> [2021-03-28 18:49:52] 

#> ✓ JDBC connection

#> [2021-03-28 18:49:52]	SQL: SELECT * FROM omop_vocabulary.concept ORDER BY RANDOM() LIMIT 100
#> [2021-03-28 18:49:52]	Querying...
#> [2021-03-28 18:50:00]	Querying...complete
#> [2021-03-28 18:50:00] 

#> ✓ Returned data has more than 0 rows

#> [2021-03-28 18:50:00]	Caching resultset...
#> [2021-03-28 18:50:00]	Postgres connection to 'athena' closed

#> # A tibble: 100 x 10
#>    concept_id concept_name             domain_id  vocabulary_id concept_class_id
#>         <int> <chr>                    <chr>      <chr>         <chr>           
#>  1   44792869 Suspended from kidney t… Observati… SNOMED        Clinical Finding
#>  2   43783777 aluminum chloride 200 M… Drug       RxNorm Exten… Branded Drug    
#>  3   43241476 CARBOCISTEINE TEVA CONS… Drug       BDPM          Drug Product    
#>  4   19098262 acetic acid 0.0092 MG/M… Drug       RxNorm        Branded Drug    
#>  5   40321903 (Acquired deformity NOS… Condition  SNOMED        Clinical Finding
#>  6   43712824 Cetirizine 10 MG/ML Ora… Drug       RxNorm Exten… Branded Drug Box
#>  7   41125238 Ketoprofen 200 MG Exten… Drug       RxNorm Exten… Branded Drug Box
#>  8   46357866 EROS AQUA BASED - glyce… Drug       NDC           11-digit NDC    
#>  9   44951330 atenolol 25mg/1 ORAL TA… Drug       NDC           9-digit NDC     
#> 10   45624084 rizatriptan benzoate 10… Drug       SPL           Prescription Dr…
#> # … with 90 more rows, and 5 more variables: standard_concept <chr>,
#> #   concept_code <chr>, valid_start_date <date>, valid_end_date <date>,
#> #   invalid_reason <chr>

# The size of the resultset can be altered
get_test_concepts(limit = 25)

#> Connecting using PostgreSQL driver

#> [2021-03-28 18:50:00]	Overriding cache... Querying Athena...
#> [2021-03-28 18:50:01] 

#> ✓ Open connection

#> [2021-03-28 18:50:01] 

#> ✓ JDBC connection

#> [2021-03-28 18:50:01]	SQL: SELECT * FROM omop_vocabulary.concept ORDER BY RANDOM() LIMIT 25
#> [2021-03-28 18:50:01]	Querying...
#> [2021-03-28 18:50:08]	Querying...complete
#> [2021-03-28 18:50:08] 

#> ✓ Returned data has more than 0 rows

#> [2021-03-28 18:50:08]	Caching resultset...
#> [2021-03-28 18:50:08]	Postgres connection to 'athena' closed

#> # A tibble: 25 x 10
#>    concept_id concept_name             domain_id  vocabulary_id concept_class_id
#>         <int> <chr>                    <chr>      <chr>         <chr>           
#>  1    2802174 Imaging @ Female Reprod… Procedure  ICD10PCS      ICD10PCS Hierar…
#>  2   42154781 Block of Peripheral Bra… Procedure  EDI           Procedure       
#>  3    2075741 ALLERVID Ophthalmic Sol… Drug       KDC           Drug Product    
#>  4   43195596 isoconazole 0.02 MG/MG … Drug       RxNorm Exten… Clinical Drug   
#>  5   41342929 2 ML Amikacin 62.5 MG/M… Drug       RxNorm Exten… Quant Branded B…
#>  6   36390056 ethanol 0.75 ML/ML Medi… Drug       NDC           11-digit NDC    
#>  7   40704511 99300 MG isopropyl myri… Drug       RxNorm Exten… Marketed Product
#>  8    3047584 MR Hand - right W contr… Measureme… LOINC         Clinical Observ…
#>  9   41101248 Terazosin 5 MG Oral Tab… Drug       RxNorm Exten… Marketed Product
#> 10   37410859 acetaminophen, guaifene… Drug       SPL           OTC Drug        
#> # … with 15 more rows, and 5 more variables: standard_concept <chr>,
#> #   concept_code <chr>, valid_start_date <date>, valid_end_date <date>,
#> #   invalid_reason <chr>

# Concepts specific to the Drug domain can be sampled
# RxNorm/RxNorm Extension Concepts
get_test_drug_concepts(limit = 10)

#> Connecting using PostgreSQL driver

#> [2021-03-28 18:50:08]	Overriding cache... Querying Athena...
#> [2021-03-28 18:50:08] 

#> ✓ Open connection

#> [2021-03-28 18:50:08] 

#> ✓ JDBC connection

#> [2021-03-28 18:50:08]	SQL: SELECT * FROM omop_vocabulary.concept c WHERE c.invalid_reason IS NULL AND c.vocabulary_id IN ('RxNorm', 'RxNorm Extension') AND c.standard_concept <> 'C' AND c.domain_id = 'Drug' ORDER BY RANDOM() LIMIT 10
#> [2021-03-28 18:50:08]	Querying...
#> [2021-03-28 18:50:11]	Querying...complete
#> [2021-03-28 18:50:11] 

#> ✓ Returned data has more than 0 rows

#> [2021-03-28 18:50:11]	Caching resultset...
#> [2021-03-28 18:50:11]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 10
#>    concept_id concept_name             domain_id vocabulary_id  concept_class_id
#>         <int> <chr>                    <chr>     <chr>          <chr>           
#>  1   21085912 dihydrocodeine 30 MG Or… Drug      RxNorm Extens… Marketed Product
#>  2   35873781 aminophylline 60 MG/ML … Drug      RxNorm Extens… Clinical Drug   
#>  3   41308399 rufinamide 0.087 MG/ML … Drug      RxNorm Extens… Branded Drug    
#>  4   43759089 500 ML Serine 2.1 MG/ML… Drug      RxNorm Extens… Quant Clinical …
#>  5   21151324 lansoprazole 15 MG Dela… Drug      RxNorm Extens… Marketed Product
#>  6   21093840 Aspirin 300 MG Delayed … Drug      RxNorm Extens… Marketed Product
#>  7   40939613 Flecainide 100 MG Oral … Drug      RxNorm Extens… Branded Drug Box
#>  8   40054807 malic acid Oral Capsule  Drug      RxNorm         Clinical Drug F…
#>  9   43032110 Bisoprolol 7.5 MG Delay… Drug      RxNorm Extens… Marketed Product
#> 10   21063797 Ibuprofen 0.1 MG/MG Vag… Drug      RxNorm Extens… Marketed Product
#> # … with 5 more variables: standard_concept <chr>, concept_code <chr>,
#> #   valid_start_date <date>, valid_end_date <date>, invalid_reason <chr>

# HemOnc/ATC Classes
get_test_drug_classes(limit = 10)

#> Connecting using PostgreSQL driver

#> [2021-03-28 18:50:11]	Overriding cache... Querying Athena...
#> [2021-03-28 18:50:11] 

#> ✓ Open connection

#> [2021-03-28 18:50:11] 

#> ✓ JDBC connection

#> [2021-03-28 18:50:11]	SQL: SELECT * FROM omop_vocabulary.concept c WHERE c.invalid_reason IS NULL AND c.vocabulary_id IN ('ATC', 'HemOnc') AND c.standard_concept = 'C' AND c.domain_id = 'Drug' ORDER BY RANDOM() LIMIT 10
#> [2021-03-28 18:50:11]	Querying...
#> [2021-03-28 18:50:11]	Querying...complete
#> [2021-03-28 18:50:11] 

#> ✓ Returned data has more than 0 rows

#> [2021-03-28 18:50:11]	Caching resultset...
#> [2021-03-28 18:50:11]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 10
#>    concept_id concept_name              domain_id vocabulary_id concept_class_id
#>         <int> <chr>                     <chr>     <chr>         <chr>           
#>  1   21603434 levodropropizine; oral    Drug      ATC           ATC 5th         
#>  2   21600946 laronidase; parenteral    Drug      ATC           ATC 5th         
#>  3   21602116 dexamethasone; topical    Drug      ATC           ATC 5th         
#>  4   21600698 DIGESTIVES, INCL. ENZYMES Drug      ATC           ATC 3rd         
#>  5   21603259 rimiterol; inhalant       Drug      ATC           ATC 5th         
#>  6     715924 amenamevir; oral          Drug      ATC           ATC 5th         
#>  7   21601801 ACE INHIBITORS, COMBINAT… Drug      ATC           ATC 3rd         
#>  8   21601167 IRRIGATING SOLUTIONS      Drug      ATC           ATC 3rd         
#>  9   21605531 Indium (111In) compounds  Drug      ATC           ATC 4th         
#> 10   40254011 ulipristal; oral          Drug      ATC           ATC 5th         
#> # … with 5 more variables: standard_concept <chr>, concept_code <chr>,
#> #   valid_start_date <date>, valid_end_date <date>, invalid_reason <chr>

# Test data can also be taken from tables other than Concept
get_test_data(table = "concept_relationship")

#> Connecting using PostgreSQL driver

#> [2021-03-28 18:50:12]	Overriding cache... Querying Athena...
#> [2021-03-28 18:50:12] 

#> ✓ Open connection

#> [2021-03-28 18:50:12] 

#> ✓ JDBC connection

#> [2021-03-28 18:50:12]	SQL: SELECT * FROM omop_vocabulary.concept_relationship ORDER BY RANDOM() LIMIT 100
#> [2021-03-28 18:50:12]	Querying...
#> [2021-03-28 18:50:46]	Querying...complete
#> [2021-03-28 18:50:46] 

#> ✓ Returned data has more than 0 rows

#> [2021-03-28 18:50:46]	Caching resultset...
#> [2021-03-28 18:50:46]	Postgres connection to 'athena' closed

#> # A tibble: 100 x 6
#>    concept_id_1 concept_id_2 relationship_id     valid_start_date valid_end_date
#>           <int>        <int> <chr>               <date>           <date>        
#>  1     37006026     35607932 Has non-avail ind   2018-07-31       2099-12-31    
#>  2     35941938     35909840 Value to Schema     1970-01-01       2099-12-31    
#>  3     41479133     19082170 RxNorm has dose fo… 2017-07-18       2099-12-31    
#>  4     43127246     43156644 Brand name of       2016-09-10       2099-12-31    
#>  5     46135634     37207390 Has CD category     2019-07-31       2099-12-31    
#>  6     46260265     36703915 Is a                2017-01-31       2099-12-31    
#>  7     40973341     41316787 Box of              2017-07-18       2099-12-31    
#>  8      2834837      4199158 Has dir proc site   1970-01-01       2099-12-31    
#>  9     41990658     41990658 Maps to             2019-02-21       2099-12-31    
#> 10      4004535      4007035 Is a                2011-07-31       2099-12-31    
#> # … with 90 more rows, and 1 more variable: invalid_reason <chr>

get_test_data(table = "concept_ancestor", limit = 10)

#> Connecting using PostgreSQL driver

#> [2021-03-28 18:50:46]	Overriding cache... Querying Athena...
#> [2021-03-28 18:50:46] 

#> ✓ Open connection

#> [2021-03-28 18:50:46] 

#> ✓ JDBC connection

#> [2021-03-28 18:50:46]	SQL: SELECT * FROM omop_vocabulary.concept_ancestor ORDER BY RANDOM() LIMIT 10
#> [2021-03-28 18:50:46]	Querying...
#> [2021-03-28 18:51:22]	Querying...complete
#> [2021-03-28 18:51:22] 

#> ✓ Returned data has more than 0 rows

#> [2021-03-28 18:51:22]	Caching resultset...
#> [2021-03-28 18:51:22]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 4
#>    ancestor_concept_… descendant_concep… min_levels_of_sepa… max_levels_of_sepa…
#>                 <int>              <int>               <int>               <int>
#>  1           43543891           40229498                   2                   3
#>  2           36217210           41372077                   3                   4
#>  3           41913545           41918146                   3                   3
#>  4           36153616           43284064                   2                   2
#>  5           36213168           43271197                   2                   3
#>  6           44339840           21100906                   2                   2
#>  7           45689449           44061763                   1                   1
#>  8           40070541           44049761                   2                   2
#>  9           42820405           40852753                   2                   3
#> 10            4213101            4017379                   4                   6