Get Test Drug Concepts

Get random records of RxNorm and RxNorm Extension Concepts that are:

Valid,

Not a Class,

Belong to the Drug domain

get_test_drug_concepts(
  conn,
  conn_fun = "connectAthena()",
  schema = "omop_vocabulary",
  limit = 100,
  verbose = TRUE,
  render_sql = TRUE
)

See also

Examples

library(chariot)
library(tidyverse)

# Get Unfiltered Test Concepts (Invalid Concepts are included)
get_test_concepts()

#> Connecting using PostgreSQL driver
#> [2021-01-10 16:35:13]	Overriding cache... Querying Athena...
#> [2021-01-10 16:35:13] 

#> ✓ Open connection

#> [2021-01-10 16:35:13] 

#> ✓ JDBC connection

#> [2021-01-10 16:35:13]	SQL: SELECT * FROM omop_vocabulary.concept ORDER BY RANDOM() LIMIT 100
#> [2021-01-10 16:35:13]	Querying...
#> [2021-01-10 16:35:22]	Querying...complete
#> [2021-01-10 16:35:22] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 16:35:22]	Caching resultset...
#> [2021-01-10 16:35:22]	Postgres connection to 'athena' closed

#> # A tibble: 100 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id
#>         <dbl> <chr>        <chr>     <chr>         <chr>           
#>  1   46294725 ILLUSION TI… Drug      SPL           OTC Drug        
#>  2   36000621 dichapetalu… Drug      SPL           OTC Drug        
#>  3   45656546 SENNA-S - d… Drug      SPL           OTC Drug        
#>  4   36562630 Follicular … Condition ICDO3         ICDO Condition  
#>  5   41154004 Piperacilli… Drug      RxNorm Exten… Branded Drug Box
#>  6   35518656 European ra… Drug      NDC           11-digit NDC    
#>  7    4140912 18 lead ele… Procedure SNOMED        Procedure       
#>  8   21393486 Generic PKU… Device    dm+d          VMPP            
#>  9   42383602 glimepiride… Drug      EDI           Drug Product    
#> 10   43351403 Chlortalido… Drug      AMT           AU Substance    
#> # … with 90 more rows, and 5 more variables: standard_concept <chr>,
#> #   concept_code <chr>, valid_start_date <date>, valid_end_date <date>,
#> #   invalid_reason <chr>

# The size of the resultset can be altered
get_test_concepts(limit = 25)

#> Connecting using PostgreSQL driver
#> [2021-01-10 16:35:22]	Overriding cache... Querying Athena...
#> [2021-01-10 16:35:22] 

#> ✓ Open connection

#> [2021-01-10 16:35:22] 

#> ✓ JDBC connection

#> [2021-01-10 16:35:22]	SQL: SELECT * FROM omop_vocabulary.concept ORDER BY RANDOM() LIMIT 25
#> [2021-01-10 16:35:22]	Querying...
#> [2021-01-10 16:35:29]	Querying...complete
#> [2021-01-10 16:35:29] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 16:35:29]	Caching resultset...
#> [2021-01-10 16:35:29]	Postgres connection to 'athena' closed

#> # A tibble: 25 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id
#>         <dbl> <chr>        <chr>     <chr>         <chr>           
#>  1   42907207 TEBINA       Drug      RxNorm Exten… Brand Name      
#>  2   44416229 carvedilol … Drug      NDC           11-digit NDC    
#>  3   44202321 2.02 ML Ris… Drug      RxNorm Exten… Quant Branded D…
#>  4   43850531 16.7 ML Pac… Drug      RxNorm Exten… Quant Clinical …
#>  5   43658165 CHROMIUM AS… Drug      RxNorm Exten… Branded Drug Co…
#>  6    4262823 Wehrdikmans… Observat… SNOMED        Organism        
#>  7   40491953 Suturing of… Procedure SNOMED        Procedure       
#>  8   35522079 aluminum zi… Drug      NDC           9-digit NDC     
#>  9   42040035 Cockermouth  Geography OSM           10th level      
#> 10   42546681 METAPROTERE… Drug      VA Product    VA Product      
#> # … with 15 more rows, and 5 more variables: standard_concept <chr>,
#> #   concept_code <chr>, valid_start_date <date>, valid_end_date <date>,
#> #   invalid_reason <chr>

# Concepts specific to the Drug domain can be sampled
# RxNorm/RxNorm Extension Concepts
get_test_drug_concepts(limit = 10)

#> Connecting using PostgreSQL driver
#> [2021-01-10 16:35:30]	Overriding cache... Querying Athena...
#> [2021-01-10 16:35:30] 

#> ✓ Open connection

#> [2021-01-10 16:35:30] 

#> ✓ JDBC connection

#> [2021-01-10 16:35:30]	SQL: SELECT * FROM omop_vocabulary.concept c WHERE c.invalid_reason IS NULL AND c.vocabulary_id IN ('RxNorm', 'RxNorm Extension') AND c.standard_concept <> 'C' AND c.domain_id = 'Drug' ORDER BY RANDOM() LIMIT 10
#> [2021-01-10 16:35:30]	Querying...
#> [2021-01-10 16:35:33]	Querying...complete
#> [2021-01-10 16:35:33] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 16:35:33]	Caching resultset...
#> [2021-01-10 16:35:33]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id
#>         <dbl> <chr>        <chr>     <chr>         <chr>           
#>  1   42717048 Articaine 4… Drug      RxNorm Exten… Branded Drug    
#>  2   35886889 avatrombopa… Drug      RxNorm Exten… Marketed Product
#>  3   42902376 fibrinogen … Drug      RxNorm        Branded Drug    
#>  4   43037948 0.5 ML Filg… Drug      RxNorm Exten… Marketed Product
#>  5   36814250 Nitrous Oxi… Drug      RxNorm Exten… Marketed Product
#>  6   41152184 thymostimul… Drug      RxNorm Exten… Branded Drug    
#>  7   41030066 Molsidomine… Drug      RxNorm Exten… Branded Drug Box
#>  8   43283712 Naloxone 0.… Drug      RxNorm Exten… Branded Drug    
#>  9   43704921 500 ML Calc… Drug      RxNorm Exten… Quant Branded B…
#> 10   41507044 99300 MG Ar… Drug      RxNorm Exten… Quant Clinical …
#> # … with 5 more variables: standard_concept <chr>, concept_code <chr>,
#> #   valid_start_date <date>, valid_end_date <date>, invalid_reason <chr>

# HemOnc/ATC Classes
get_test_drug_classes(limit = 10)

#> Connecting using PostgreSQL driver
#> [2021-01-10 16:35:33]	Overriding cache... Querying Athena...
#> [2021-01-10 16:35:33] 

#> ✓ Open connection

#> [2021-01-10 16:35:33] 

#> ✓ JDBC connection

#> [2021-01-10 16:35:33]	SQL: SELECT * FROM omop_vocabulary.concept c WHERE c.invalid_reason IS NULL AND c.vocabulary_id IN ('ATC', 'HemOnc') AND c.standard_concept = 'C' AND c.domain_id = 'Drug' ORDER BY RANDOM() LIMIT 10
#> [2021-01-10 16:35:33]	Querying...
#> [2021-01-10 16:35:33]	Querying...complete
#> [2021-01-10 16:35:33] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 16:35:33]	Caching resultset...
#> [2021-01-10 16:35:33]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id
#>         <dbl> <chr>        <chr>     <chr>         <chr>           
#>  1   21604094 carisoprodo… Drug      ATC           ATC 5th         
#>  2   35807216 PMDA approv… Drug      HemOnc        Component Class 
#>  3   43534797 emtricitabi… Drug      ATC           ATC 5th         
#>  4   21600671 lactic acid… Drug      ATC           ATC 5th         
#>  5   21600506 casopitant   Drug      ATC           ATC 5th         
#>  6   21600414 doxazosin; … Drug      ATC           ATC 5th         
#>  7   21603082 voriconazol… Drug      ATC           ATC 5th         
#>  8   21602878 cefradine; … Drug      ATC           ATC 5th         
#>  9   21601765 SELECTIVE C… Drug      ATC           ATC 3rd         
#> 10   21601893 omega-3-tri… Drug      ATC           ATC 5th         
#> # … with 5 more variables: standard_concept <chr>, concept_code <chr>,
#> #   valid_start_date <date>, valid_end_date <date>, invalid_reason <chr>

# Test data can also be taken from tables other than Concept
get_test_data(table = "concept_relationship")

#> Connecting using PostgreSQL driver
#> [2021-01-10 16:35:33]	Overriding cache... Querying Athena...
#> [2021-01-10 16:35:33] 

#> ✓ Open connection

#> [2021-01-10 16:35:33] 

#> ✓ JDBC connection

#> [2021-01-10 16:35:33]	SQL: SELECT * FROM omop_vocabulary.concept_relationship ORDER BY RANDOM() LIMIT 100
#> [2021-01-10 16:35:33]	Querying...
#> [2021-01-10 16:36:06]	Querying...complete
#> [2021-01-10 16:36:06] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 16:36:06]	Caching resultset...
#> [2021-01-10 16:36:06]	Postgres connection to 'athena' closed

#> # A tibble: 100 x 6
#>    concept_id_1 concept_id_2 relationship_id valid_start_date valid_end_date
#>           <dbl>        <dbl> <chr>           <date>           <date>        
#>  1     43197873     43197873 Maps to         2016-08-07       2099-12-31    
#>  2      4106470     42895813 Method of       1970-01-01       2099-12-31    
#>  3     45677074     19077058 SPL - RxNorm    1970-01-01       2099-12-31    
#>  4     19082573     36267862 RxNorm dose fo… 2017-10-01       2099-12-31    
#>  5     44058715      1396259 Box of          2017-06-25       2099-12-31    
#>  6     19082103     21049798 RxNorm dose fo… 2015-08-17       2099-12-31    
#>  7     44817974     44817974 Maps to         2014-06-02       2099-12-31    
#>  8     44039969       976309 RxNorm has ing  2017-06-25       2099-12-31    
#>  9     35861825     35861840 Has tradename   2020-10-02       2099-12-31    
#> 10     37310856     46209804 AMP restr ind … 2020-03-09       2099-12-31    
#> # … with 90 more rows, and 1 more variable: invalid_reason <chr>

get_test_data(table = "concept_ancestor", limit = 10)

#> Connecting using PostgreSQL driver
#> [2021-01-10 16:36:06]	Overriding cache... Querying Athena...
#> [2021-01-10 16:36:06] 

#> ✓ Open connection

#> [2021-01-10 16:36:06] 

#> ✓ JDBC connection

#> [2021-01-10 16:36:06]	SQL: SELECT * FROM omop_vocabulary.concept_ancestor ORDER BY RANDOM() LIMIT 10
#> [2021-01-10 16:36:06]	Querying...
#> [2021-01-10 16:36:40]	Querying...complete
#> [2021-01-10 16:36:40] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 16:36:40]	Caching resultset...
#> [2021-01-10 16:36:40]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 4
#>    ancestor_concept_… descendant_concep… min_levels_of_sepa… max_levels_of_sepa…
#>                 <dbl>              <dbl>               <dbl>               <dbl>
#>  1             929887           21121777                   2                   3
#>  2           45702567           35887269                   2                   2
#>  3           36151743           35142471                   2                   2
#>  4            4274025           36543427                   4                  10
#>  5           36703920           46059846                   1                   1
#>  6           45628735           43723792                   1                   1
#>  7           35902975           43035449                   2                   3
#>  8            3004296            3004296                   0                   0
#>  9           35803694             792804                   2                   4
#> 10            4095631           42536772                   3                   5