Get Test Drug Concepts

Get random records of RxNorm and RxNorm Extension Concepts that are:

Valid,

Not a Class,

Belong to the Drug domain

get_test_drug_concepts(
  conn,
  conn_fun = "connectAthena()",
  schema = "omop_vocabulary",
  limit = 100,
  verbose = TRUE,
  render_sql = TRUE
)

See also

Examples

library(chariot)
library(tidyverse)

# Get Unfiltered Test Concepts (Invalid Concepts are included)
get_test_concepts()

#> Connecting using PostgreSQL driver
#> [2021-01-10 19:02:38]	Overriding cache... Querying Athena...
#> [2021-01-10 19:02:38] 

#> ✓ Open connection

#> [2021-01-10 19:02:38] 

#> ✓ JDBC connection

#> [2021-01-10 19:02:38]	SQL: SELECT * FROM omop_vocabulary.concept ORDER BY RANDOM() LIMIT 100
#> [2021-01-10 19:02:38]	Querying...
#> [2021-01-10 19:02:47]	Querying...complete
#> [2021-01-10 19:02:47] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 19:02:47]	Caching resultset...
#> [2021-01-10 19:02:47]	Postgres connection to 'athena' closed

#> # A tibble: 100 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id
#>         <dbl> <chr>        <chr>     <chr>         <chr>           
#>  1   36658038 Sulfadiazin… Drug      NCCD          Drug Product    
#>  2   41450901 1350 MG Mag… Drug      RxNorm Exten… Quant Clinical …
#>  3   41141922 Clotrimazol… Drug      RxNorm Exten… Branded Drug Co…
#>  4   36064459 levothyroxi… Drug      RxNorm Exten… Branded Drug Co…
#>  5   42282901 Fistula For… Procedure EDI           Procedure       
#>  6   42054212 Beadle Coun… Geography OSM           6th level       
#>  7   42230747 Resection o… Procedure EDI           Procedure       
#>  8   42603443 Caprimulgus… Observat… SNOMED Veter… Organism        
#>  9   45030601 CITRULLUS C… Drug      NDC           9-digit NDC     
#> 10   35745750 Ethambutol … Drug      RxNorm Exten… Clinical Drug B…
#> # … with 90 more rows, and 5 more variables: standard_concept <chr>,
#> #   concept_code <chr>, valid_start_date <date>, valid_end_date <date>,
#> #   invalid_reason <chr>

# The size of the resultset can be altered
get_test_concepts(limit = 25)

#> Connecting using PostgreSQL driver
#> [2021-01-10 19:02:47]	Overriding cache... Querying Athena...
#> [2021-01-10 19:02:47] 

#> ✓ Open connection

#> [2021-01-10 19:02:47] 

#> ✓ JDBC connection

#> [2021-01-10 19:02:47]	SQL: SELECT * FROM omop_vocabulary.concept ORDER BY RANDOM() LIMIT 25
#> [2021-01-10 19:02:47]	Querying...
#> [2021-01-10 19:02:54]	Querying...complete
#> [2021-01-10 19:02:54] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 19:02:54]	Caching resultset...
#> [2021-01-10 19:02:54]	Postgres connection to 'athena' closed

#> # A tibble: 25 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id
#>         <dbl> <chr>        <chr>     <chr>         <chr>           
#>  1   41468078 3000 MG Ace… Drug      RxNorm Exten… Quant Branded B…
#>  2   36092984 SUAVE - tro… Drug      SPL           OTC Drug        
#>  3   46305351 OIL OF BEAU… Drug      SPL           OTC Drug        
#>  4    4130728 Intravesica… Observat… SNOMED        Qualifier Value 
#>  5   45223777 benazepril … Drug      NDC           11-digit NDC    
#>  6   45515456 Apgar at 1 … Condition Read          Read            
#>  7   44930798 amlodipine … Drug      NDC           11-digit NDC    
#>  8   45690186 camphor .1g… Drug      SPL           OTC Drug        
#>  9   45333411 Propranolol… Drug      NDC           11-digit NDC    
#> 10   42912874 SAMJIN PHEN… Drug      RxNorm Exten… Brand Name      
#> # … with 15 more rows, and 5 more variables: standard_concept <chr>,
#> #   concept_code <chr>, valid_start_date <date>, valid_end_date <date>,
#> #   invalid_reason <chr>

# Concepts specific to the Drug domain can be sampled
# RxNorm/RxNorm Extension Concepts
get_test_drug_concepts(limit = 10)

#> Connecting using PostgreSQL driver
#> [2021-01-10 19:02:54]	Overriding cache... Querying Athena...
#> [2021-01-10 19:02:54] 

#> ✓ Open connection

#> [2021-01-10 19:02:54] 

#> ✓ JDBC connection

#> [2021-01-10 19:02:54]	SQL: SELECT * FROM omop_vocabulary.concept c WHERE c.invalid_reason IS NULL AND c.vocabulary_id IN ('RxNorm', 'RxNorm Extension') AND c.standard_concept <> 'C' AND c.domain_id = 'Drug' ORDER BY RANDOM() LIMIT 10
#> [2021-01-10 19:02:54]	Querying...
#> [2021-01-10 19:02:58]	Querying...complete
#> [2021-01-10 19:02:58] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 19:02:58]	Caching resultset...
#> [2021-01-10 19:02:58]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id
#>         <dbl> <chr>        <chr>     <chr>         <chr>           
#>  1   19105773 albuterol 0… Drug      RxNorm        Clinical Drug   
#>  2   44077748 Aloe Extrac… Drug      RxNorm Exten… Marketed Product
#>  3   35142529 ranimustine… Drug      RxNorm Exten… Clinical Drug C…
#>  4   43788689 2 ML Ondans… Drug      RxNorm Exten… Quant Clinical …
#>  5   43041356 10 ML Ceftr… Drug      RxNorm Exten… Marketed Product
#>  6   36260754 Anethole Tr… Drug      RxNorm Exten… Branded Drug    
#>  7   42968428 Esomeprazol… Drug      RxNorm Exten… Branded Drug Fo…
#>  8   36261731 Amoxicillin… Drug      RxNorm Exten… Branded Drug Co…
#>  9   21052174 Phenylephri… Drug      RxNorm Exten… Marketed Product
#> 10   40823439 Aluminum ca… Drug      RxNorm Exten… Branded Drug Box
#> # … with 5 more variables: standard_concept <chr>, concept_code <chr>,
#> #   valid_start_date <date>, valid_end_date <date>, invalid_reason <chr>

# HemOnc/ATC Classes
get_test_drug_classes(limit = 10)

#> Connecting using PostgreSQL driver
#> [2021-01-10 19:02:58]	Overriding cache... Querying Athena...
#> [2021-01-10 19:02:58] 

#> ✓ Open connection

#> [2021-01-10 19:02:58] 

#> ✓ JDBC connection

#> [2021-01-10 19:02:58]	SQL: SELECT * FROM omop_vocabulary.concept c WHERE c.invalid_reason IS NULL AND c.vocabulary_id IN ('ATC', 'HemOnc') AND c.standard_concept = 'C' AND c.domain_id = 'Drug' ORDER BY RANDOM() LIMIT 10
#> [2021-01-10 19:02:58]	Querying...
#> [2021-01-10 19:02:58]	Querying...complete
#> [2021-01-10 19:02:58] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 19:02:58]	Caching resultset...
#> [2021-01-10 19:02:58]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id
#>         <dbl> <chr>        <chr>     <chr>         <chr>           
#>  1   45893192 indometacin… Drug      ATC           ATC 5th         
#>  2     715790 patent blue  Drug      ATC           ATC 5th         
#>  3   21600295 gepefrine; … Drug      ATC           ATC 5th         
#>  4   21600396 methyldopa … Drug      ATC           ATC 5th         
#>  5   21601204 sodium phos… Drug      ATC           ATC 5th         
#>  6   21603353 Leukotriene… Drug      ATC           ATC 4th         
#>  7   21602847 clometocill… Drug      ATC           ATC 5th         
#>  8   21601438 carmofur     Drug      ATC           ATC 5th         
#>  9   35807406 Anti-TRAIL-… Drug      HemOnc        Component Class 
#> 10   21602958 sulfamazone… Drug      ATC           ATC 5th         
#> # … with 5 more variables: standard_concept <chr>, concept_code <chr>,
#> #   valid_start_date <date>, valid_end_date <date>, invalid_reason <chr>

# Test data can also be taken from tables other than Concept
get_test_data(table = "concept_relationship")

#> Connecting using PostgreSQL driver
#> [2021-01-10 19:02:59]	Overriding cache... Querying Athena...
#> [2021-01-10 19:02:59] 

#> ✓ Open connection

#> [2021-01-10 19:02:59] 

#> ✓ JDBC connection

#> [2021-01-10 19:02:59]	SQL: SELECT * FROM omop_vocabulary.concept_relationship ORDER BY RANDOM() LIMIT 100
#> [2021-01-10 19:02:59]	Querying...
#> [2021-01-10 19:03:31]	Querying...complete
#> [2021-01-10 19:03:31] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 19:03:31]	Caching resultset...
#> [2021-01-10 19:03:31]	Postgres connection to 'athena' closed

#> # A tibble: 100 x 6
#>    concept_id_1 concept_id_2 relationship_id valid_start_date valid_end_date
#>           <dbl>        <dbl> <chr>           <date>           <date>        
#>  1     44027815     44085517 Constitutes     2017-06-25       2099-12-31    
#>  2      3252447      3402753 Has interprets  2019-08-16       2099-12-31    
#>  3     19048333     45781675 RxNorm - SPL    2015-01-29       2099-12-31    
#>  4      4044953      4036390 Is a            2011-07-31       2099-12-31    
#>  5      3500723      3501122 Is a            2019-08-16       2099-12-31    
#>  6      2005931     36913449 Mapped from     2017-01-01       2099-12-31    
#>  7      1127524     42939881 Has tradename   2018-08-21       2099-12-31    
#>  8      4302532      2786451 Subsumes        1970-01-01       2099-12-31    
#>  9      4284844     45534031 Mapped from     1970-01-01       2099-12-31    
#> 10      4125453     45767122 Is a            2015-03-01       2099-12-31    
#> # … with 90 more rows, and 1 more variable: invalid_reason <chr>

get_test_data(table = "concept_ancestor", limit = 10)

#> Connecting using PostgreSQL driver
#> [2021-01-10 19:03:32]	Overriding cache... Querying Athena...
#> [2021-01-10 19:03:32] 

#> ✓ Open connection

#> [2021-01-10 19:03:32] 

#> ✓ JDBC connection

#> [2021-01-10 19:03:32]	SQL: SELECT * FROM omop_vocabulary.concept_ancestor ORDER BY RANDOM() LIMIT 10
#> [2021-01-10 19:03:32]	Querying...
#> [2021-01-10 19:04:05]	Querying...complete
#> [2021-01-10 19:04:05] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 19:04:05]	Caching resultset...
#> [2021-01-10 19:04:05]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 4
#>    ancestor_concept_… descendant_concep… min_levels_of_sepa… max_levels_of_sepa…
#>                 <dbl>              <dbl>               <dbl>               <dbl>
#>  1           46096260           37107054                   1                   1
#>  2            2006974            4194672                   3                   3
#>  3            4169265           46194566                   3                   4
#>  4             196931           36553751                   6                   7
#>  5            4189640           36546808                   5                   6
#>  6           40897220           40977483                   1                   2
#>  7             920458           41056000                   3                   4
#>  8           21602099           41513798                   5                   6
#>  9           21604297           41092595                   3                   4
#> 10            4274025             761500                   4                   8