Get Test Drug Concepts

Get random records of RxNorm and RxNorm Extension Concepts that are:

Valid,

Not a Class,

Belong to the Drug domain

get_test_drug_concepts(
  conn,
  conn_fun = "connectAthena()",
  schema = "omop_vocabulary",
  limit = 100,
  verbose = TRUE,
  render_sql = TRUE
)

See also

Examples

library(chariot)
library(tidyverse)

# Get Unfiltered Test Concepts (Invalid Concepts are included)
get_test_concepts()

#> Connecting using PostgreSQL driver
#> [2021-01-10 16:00:44]	Overriding cache... Querying Athena...
#> [2021-01-10 16:00:44] 

#> ✓ Open connection

#> [2021-01-10 16:00:44] 

#> ✓ JDBC connection

#> [2021-01-10 16:00:44]	SQL: SELECT * FROM omop_vocabulary.concept ORDER BY RANDOM() LIMIT 100
#> [2021-01-10 16:00:44]	Querying...
#> [2021-01-10 16:00:52]	Querying...complete
#> [2021-01-10 16:00:52] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 16:00:52]	Caching resultset...
#> [2021-01-10 16:00:52]	Postgres connection to 'athena' closed

#> # A tibble: 100 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id
#>         <dbl> <chr>        <chr>     <chr>         <chr>           
#>  1   36290574 Abilify tab… Drug      GGR           Drug Product    
#>  2    4076934 Urobilinoge… Measurem… SNOMED        Clinical Finding
#>  3   45011609 Theophyllin… Drug      NDC           11-digit NDC    
#>  4   43560560 sarkosyl     Drug      RxNorm        Ingredient      
#>  5    4347467 Urofollitro… Drug      NDFRT         Chemical Struct…
#>  6   42119596 Excision of… Procedure EDI           Procedure       
#>  7   44627208 SELENIUM HV… Drug      DPD           Clinical Drug C…
#>  8   41232266 Tacalcitol … Drug      RxNorm Exten… Clinical Drug C…
#>  9   40864802 Vitex agnus… Drug      RxNorm Exten… Clinical Drug B…
#> 10   37067129 Asialogangl… Measurem… LOINC         LOINC Hierarchy 
#> # … with 90 more rows, and 5 more variables: standard_concept <chr>,
#> #   concept_code <chr>, valid_start_date <date>, valid_end_date <date>,
#> #   invalid_reason <chr>

# The size of the resultset can be altered
get_test_concepts(limit = 25)

#> Connecting using PostgreSQL driver
#> [2021-01-10 16:00:52]	Overriding cache... Querying Athena...
#> [2021-01-10 16:00:52] 

#> ✓ Open connection

#> [2021-01-10 16:00:52] 

#> ✓ JDBC connection

#> [2021-01-10 16:00:52]	SQL: SELECT * FROM omop_vocabulary.concept ORDER BY RANDOM() LIMIT 25
#> [2021-01-10 16:00:52]	Querying...
#> [2021-01-10 16:00:59]	Querying...complete
#> [2021-01-10 16:00:59] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 16:00:59]	Caching resultset...
#> [2021-01-10 16:00:59]	Postgres connection to 'athena' closed

#> # A tibble: 25 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id
#>         <dbl> <chr>        <chr>     <chr>         <chr>           
#>  1   36497165 povidone-io… Drug      NDC           11-digit NDC    
#>  2   45037586 Amoxicillin… Drug      NDC           11-digit NDC    
#>  3   41954437 Pérignat-lè… Geography OSM           8th level       
#>  4   42654442 agnus castu… Drug      NDC           9-digit NDC     
#>  5    2007109 Microscopic… Procedure ICD9Proc      3-dig nonbill c…
#>  6   37509034 isopropyl a… Drug      NDC           9-digit NDC     
#>  7   42194328 Repair of L… Procedure EDI           Procedure       
#>  8   43544929 alcohol 70m… Drug      SPL           OTC Drug        
#>  9    4144296 Penetrating… Condition SNOMED        Clinical Finding
#> 10   40933299 tapentadol … Drug      RxNorm Exten… Branded Drug Box
#> # … with 15 more rows, and 5 more variables: standard_concept <chr>,
#> #   concept_code <chr>, valid_start_date <date>, valid_end_date <date>,
#> #   invalid_reason <chr>

# Concepts specific to the Drug domain can be sampled
# RxNorm/RxNorm Extension Concepts
get_test_drug_concepts(limit = 10)

#> Connecting using PostgreSQL driver
#> [2021-01-10 16:01:00]	Overriding cache... Querying Athena...
#> [2021-01-10 16:01:00] 

#> ✓ Open connection

#> [2021-01-10 16:01:00] 

#> ✓ JDBC connection

#> [2021-01-10 16:01:00]	SQL: SELECT * FROM omop_vocabulary.concept c WHERE c.invalid_reason IS NULL AND c.vocabulary_id IN ('RxNorm', 'RxNorm Extension') AND c.standard_concept <> 'C' AND c.domain_id = 'Drug' ORDER BY RANDOM() LIMIT 10
#> [2021-01-10 16:01:00]	Querying...
#> [2021-01-10 16:01:03]	Querying...complete
#> [2021-01-10 16:01:03] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 16:01:03]	Caching resultset...
#> [2021-01-10 16:01:03]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id
#>         <dbl> <chr>        <chr>     <chr>         <chr>           
#>  1   41099306 bendamustin… Drug      RxNorm Exten… Branded Drug    
#>  2   43134787 Hydrochloro… Drug      RxNorm Exten… Branded Drug Co…
#>  3   21082241 quetiapine … Drug      RxNorm Exten… Marketed Product
#>  4   43739660 100 ML Lact… Drug      RxNorm Exten… Quant Branded D…
#>  5   44111415 Ramipril 2.… Drug      RxNorm Exten… Branded Drug    
#>  6   21023206 Risperidone… Drug      RxNorm Exten… Marketed Product
#>  7   41136244 Famotidine … Drug      RxNorm Exten… Marketed Product
#>  8   40755484 Coal Tar 0.… Drug      RxNorm Exten… Marketed Product
#>  9   43044086 500 ML Povi… Drug      RxNorm Exten… Marketed Product
#> 10   40715374 10 ML Nitro… Drug      RxNorm Exten… Marketed Product
#> # … with 5 more variables: standard_concept <chr>, concept_code <chr>,
#> #   valid_start_date <date>, valid_end_date <date>, invalid_reason <chr>

# HemOnc/ATC Classes
get_test_drug_classes(limit = 10)

#> Connecting using PostgreSQL driver
#> [2021-01-10 16:01:03]	Overriding cache... Querying Athena...
#> [2021-01-10 16:01:03] 

#> ✓ Open connection

#> [2021-01-10 16:01:03] 

#> ✓ JDBC connection

#> [2021-01-10 16:01:03]	SQL: SELECT * FROM omop_vocabulary.concept c WHERE c.invalid_reason IS NULL AND c.vocabulary_id IN ('ATC', 'HemOnc') AND c.standard_concept = 'C' AND c.domain_id = 'Drug' ORDER BY RANDOM() LIMIT 10
#> [2021-01-10 16:01:03]	Querying...
#> [2021-01-10 16:01:03]	Querying...complete
#> [2021-01-10 16:01:03] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 16:01:03]	Caching resultset...
#> [2021-01-10 16:01:03]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id
#>         <dbl> <chr>        <chr>     <chr>         <chr>           
#>  1   21602918 Monobactams  Drug      ATC           ATC 4th         
#>  2     715918 atazanavir … Drug      ATC           ATC 5th         
#>  3   21604028 aurothioglu… Drug      ATC           ATC 5th         
#>  4   35803499 Serotonin 5… Drug      HemOnc        Component Class 
#>  5   21603240 gramicidin;… Drug      ATC           ATC 5th         
#>  6   21602045 Psoralens f… Drug      ATC           ATC 4th         
#>  7   21602421 copper usna… Drug      ATC           ATC 5th         
#>  8    1123814 cinitapride  Drug      ATC           ATC 5th         
#>  9   21600430 endralazine… Drug      ATC           ATC 5th         
#> 10   21604750 agomelatine… Drug      ATC           ATC 5th         
#> # … with 5 more variables: standard_concept <chr>, concept_code <chr>,
#> #   valid_start_date <date>, valid_end_date <date>, invalid_reason <chr>

# Test data can also be taken from tables other than Concept
get_test_data(table = "concept_relationship")

#> Connecting using PostgreSQL driver
#> [2021-01-10 16:01:03]	Overriding cache... Querying Athena...
#> [2021-01-10 16:01:03] 

#> ✓ Open connection

#> [2021-01-10 16:01:03] 

#> ✓ JDBC connection

#> [2021-01-10 16:01:03]	SQL: SELECT * FROM omop_vocabulary.concept_relationship ORDER BY RANDOM() LIMIT 100
#> [2021-01-10 16:01:03]	Querying...
#> [2021-01-10 16:01:35]	Querying...complete
#> [2021-01-10 16:01:35] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 16:01:35]	Caching resultset...
#> [2021-01-10 16:01:35]	Postgres connection to 'athena' closed

#> # A tibble: 100 x 6
#>    concept_id_1 concept_id_2 relationship_id valid_start_date valid_end_date
#>           <dbl>        <dbl> <chr>           <date>           <date>        
#>  1     44073777     19082573 RxNorm has dos… 2017-06-25       2099-12-31    
#>  2      1000995     21017712 Has brand name  2015-08-17       2099-12-31    
#>  3     44912578     19070343 Maps to         2011-10-25       2099-12-31    
#>  4     40970233     40970233 Mapped from     2017-07-18       2099-12-31    
#>  5     46171894     35626985 Has legal cate… 2018-07-31       2099-12-31    
#>  6     19082573     35152771 RxNorm dose fo… 2019-03-11       2099-12-31    
#>  7     35626975     46188084 VMP prescr sta… 2018-07-31       2099-12-31    
#>  8      3421215      3496382 Has finding si… 2019-08-16       2099-12-31    
#>  9      4216534      4194454 Subsumes        2011-07-31       2099-12-31    
#> 10     37207390     46165532 CD category of  2019-07-31       2099-12-31    
#> # … with 90 more rows, and 1 more variable: invalid_reason <chr>

get_test_data(table = "concept_ancestor", limit = 10)

#> Connecting using PostgreSQL driver
#> [2021-01-10 16:01:36]	Overriding cache... Querying Athena...
#> [2021-01-10 16:01:36] 

#> ✓ Open connection

#> [2021-01-10 16:01:36] 

#> ✓ JDBC connection

#> [2021-01-10 16:01:36]	SQL: SELECT * FROM omop_vocabulary.concept_ancestor ORDER BY RANDOM() LIMIT 10
#> [2021-01-10 16:01:36]	Querying...
#> [2021-01-10 16:02:09]	Querying...complete
#> [2021-01-10 16:02:09] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 16:02:09]	Caching resultset...
#> [2021-01-10 16:02:09]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 4
#>    ancestor_concept_… descendant_concep… min_levels_of_sepa… max_levels_of_sepa…
#>                 <dbl>              <dbl>               <dbl>               <dbl>
#>  1           44359645           41484845                   2                   3
#>  2           35807272           43270276                   5                   6
#>  3           45783644           36506309                   2                   2
#>  4           35988932           35988932                   0                   0
#>  5           45781572           21024019                   2                   3
#>  6           44363379           42969478                   1                   1
#>  7            4128342            4134278                   5                   6
#>  8            4199402            4083788                   4                   4
#>  9           45670158           44130255                   2                   2
#> 10           21601032             793105                   4                   4