Get Test Drug Concepts

Get random records of RxNorm and RxNorm Extension Concepts that are:

Valid,

Not a Class,

Belong to the Drug domain

get_test_drug_concepts(
  conn,
  conn_fun = "connectAthena()",
  schema = "omop_vocabulary",
  limit = 100,
  verbose = TRUE,
  render_sql = TRUE
)

See also

Examples

library(chariot)
library(tidyverse)

# Get Unfiltered Test Concepts (Invalid Concepts are included)
get_test_concepts()

#> [2021-05-26 15:48:15]	Connected to 'athena'
#> [2021-05-26 15:48:15]	Overriding cache... Querying Athena...
#> [2021-05-26 15:48:15]	✓ Open connection
#> [2021-05-26 15:48:15]	✓ JDBC connection
#> [2021-05-26 15:48:15]	SQL: SELECT * FROM omop_vocabulary.concept ORDER BY RANDOM() LIMIT 100
#> [2021-05-26 15:48:15]	Querying...
#> [2021-05-26 15:48:23]	Querying...complete
#> [2021-05-26 15:48:23]	✓ Returned data has more than 0 rows
#> [2021-05-26 15:48:23]	Caching resultset...
#> [2021-05-26 15:48:23]	Postgres connection to 'athena' closed

#> # A tibble: 100 x 10
#>    concept_id concept_name            domain_id   vocabulary_id concept_class_id
#>         <int> <chr>                   <chr>       <chr>         <chr>           
#>  1   42315393 Serosal Repair or Prim… Procedure   EDI           Procedure       
#>  2   45162233 Aspirin 325 MG / butal… Drug        NDC           11-digit NDC    
#>  3   46274921 entecavir anhydrous     Drug        RxNorm        Precise Ingredi…
#>  4    3268240 Machinery fitter/machi… Observation Nebraska Lex… Social Context  
#>  5    3484455 Entire layer of anteri… Spec Anato… Nebraska Lex… Body Structure  
#>  6   44141385 Plectranthus Fruticosus Drug        DPD           Ingredient      
#>  7   41111597 Caffeine / Gelsemium S… Drug        RxNorm Exten… Clinical Drug F…
#>  8   44065239 Aloe Extract 22 MG / A… Drug        RxNorm Exten… Marketed Product
#>  9   44871039 ixabepilone 2 MG/ML In… Drug        NDC           11-digit NDC    
#> 10   35813773 Area of G-temporal-inf… Measurement UK Biobank    Variable        
#> # … with 90 more rows, and 5 more variables: standard_concept <chr>,
#> #   concept_code <chr>, valid_start_date <date>, valid_end_date <date>,
#> #   invalid_reason <chr>

# The size of the resultset can be altered
get_test_concepts(limit = 25)

#> [2021-05-26 15:48:23]	Connected to 'athena'
#> [2021-05-26 15:48:23]	Overriding cache... Querying Athena...
#> [2021-05-26 15:48:23]	✓ Open connection
#> [2021-05-26 15:48:23]	✓ JDBC connection
#> [2021-05-26 15:48:23]	SQL: SELECT * FROM omop_vocabulary.concept ORDER BY RANDOM() LIMIT 25
#> [2021-05-26 15:48:23]	Querying...
#> [2021-05-26 15:48:31]	Querying...complete
#> [2021-05-26 15:48:31]	✓ Returned data has more than 0 rows
#> [2021-05-26 15:48:31]	Caching resultset...
#> [2021-05-26 15:48:31]	Postgres connection to 'athena' closed

#> # A tibble: 25 x 10
#>    concept_id concept_name             domain_id  vocabulary_id concept_class_id
#>         <int> <chr>                    <chr>      <chr>         <chr>           
#>  1   40520528 [V]Screening for other … Observati… SNOMED        Context-depende…
#>  2   43604723 Bupivacaine 2.5 MG/ML I… Drug       RxNorm Exten… Branded Drug    
#>  3   42769281 Operations on other bon… Procedure  OPS           Procedure       
#>  4   45300235 Acetaminophen 650 MG / … Drug       NDC           11-digit NDC    
#>  5   45429482 Blood sample taken from… Observati… Read          Read            
#>  6   42303914 Total Coloprotectomy(Wi… Procedure  EDI           Procedure       
#>  7   21146235 1 ML Carbimazole 0.4 MG… Drug       RxNorm Exten… Quant Clinical …
#>  8   40954352 Chloramphenicol 250 MG … Drug       RxNorm Exten… Branded Drug Co…
#>  9    3606914 Pro-Viron                Observati… SNOMED        Qualifier Value 
#> 10   40971587 Diltiazem 90 MG Extende… Drug       RxNorm Exten… Branded Drug Box
#> # … with 15 more rows, and 5 more variables: standard_concept <chr>,
#> #   concept_code <chr>, valid_start_date <date>, valid_end_date <date>,
#> #   invalid_reason <chr>

# Concepts specific to the Drug domain can be sampled
# RxNorm/RxNorm Extension Concepts
get_test_drug_concepts(limit = 10)

#> [2021-05-26 15:48:31]	Connected to 'athena'
#> [2021-05-26 15:48:31]	Overriding cache... Querying Athena...
#> [2021-05-26 15:48:31]	✓ Open connection
#> [2021-05-26 15:48:31]	✓ JDBC connection
#> [2021-05-26 15:48:31]	SQL: SELECT * FROM omop_vocabulary.concept c WHERE c.invalid_reason IS NULL AND c.vocabulary_id IN ('RxNorm', 'RxNorm Extension') AND c.standard_concept <> 'C' AND c.domain_id = 'Drug' ORDER BY RANDOM() LIMIT 10
#> [2021-05-26 15:48:31]	Querying...
#> [2021-05-26 15:48:34]	Querying...complete
#> [2021-05-26 15:48:34]	✓ Returned data has more than 0 rows
#> [2021-05-26 15:48:34]	Caching resultset...
#> [2021-05-26 15:48:34]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 10
#>    concept_id concept_name             domain_id vocabulary_id  concept_class_id
#>         <int> <chr>                    <chr>     <chr>          <chr>           
#>  1   41011825 Clonidine 0.3 MG Oral T… Drug      RxNorm Extens… Marketed Product
#>  2   40841647 Phenylbutazone 100 MG /… Drug      RxNorm Extens… Branded Drug Box
#>  3   35407144 Ethanol 500 ML Topical … Drug      RxNorm Extens… Marketed Product
#>  4   21045831 Gonadorelin 0.1 MG Inje… Drug      RxNorm Extens… Marketed Product
#>  5   43210340 Carbimazole 20 MG Oral … Drug      RxNorm Extens… Clinical Drug B…
#>  6   35159247 1010 MG Cimetidine 0.4 … Drug      RxNorm Extens… Quant Clinical …
#>  7   43280917 20 ML Cholecalciferol 0… Drug      RxNorm Extens… Quant Branded D…
#>  8   43150014 Carboplatin 10 MG/ML Pr… Drug      RxNorm Extens… Branded Drug Box
#>  9   41403711 75 ML Cefaclor 6.67 MG/… Drug      RxNorm Extens… Quant Clinical …
#> 10   35130145 lecithine-bound iodine … Drug      RxNorm Extens… Clinical Drug F…
#> # … with 5 more variables: standard_concept <chr>, concept_code <chr>,
#> #   valid_start_date <date>, valid_end_date <date>, invalid_reason <chr>

# HemOnc/ATC Classes
get_test_drug_classes(limit = 10)

#> [2021-05-26 15:48:34]	Connected to 'athena'
#> [2021-05-26 15:48:34]	Overriding cache... Querying Athena...
#> [2021-05-26 15:48:34]	✓ Open connection
#> [2021-05-26 15:48:34]	✓ JDBC connection
#> [2021-05-26 15:48:34]	SQL: SELECT * FROM omop_vocabulary.concept c WHERE c.invalid_reason IS NULL AND c.vocabulary_id IN ('ATC', 'HemOnc') AND c.standard_concept = 'C' AND c.domain_id = 'Drug' ORDER BY RANDOM() LIMIT 10
#> [2021-05-26 15:48:34]	Querying...
#> [2021-05-26 15:48:34]	Querying...complete
#> [2021-05-26 15:48:34]	✓ Returned data has more than 0 rows
#> [2021-05-26 15:48:34]	Caching resultset...
#> [2021-05-26 15:48:34]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 10
#>    concept_id concept_name              domain_id vocabulary_id concept_class_id
#>         <int> <chr>                     <chr>     <chr>         <chr>           
#>  1   21601146 hemoglobin raffimer; par… Drug      ATC           ATC 5th         
#>  2   21603361 fenspiride; systemic      Drug      ATC           ATC 5th         
#>  3   21604648 cinolazepam               Drug      ATC           ATC 5th         
#>  4   21600731 insulin (pork); parenter… Drug      ATC           ATC 5th         
#>  5   21600805 multivitamins and other … Drug      ATC           ATC 5th         
#>  6   35807216 PMDA approved drug        Drug      HemOnc        Component Class 
#>  7   21602749 Anticorticosteroids       Drug      ATC           ATC 4th         
#>  8   21603692 Carbonic anhydrase inhib… Drug      ATC           ATC 4th         
#>  9   21602046 trioxysalen; oral         Drug      ATC           ATC 5th         
#> 10   21601820 trandolapril and verapam… Drug      ATC           ATC 5th         
#> # … with 5 more variables: standard_concept <chr>, concept_code <chr>,
#> #   valid_start_date <date>, valid_end_date <date>, invalid_reason <chr>

# Test data can also be taken from tables other than Concept
get_test_data(table = "concept_relationship")

#> [2021-05-26 15:48:34]	Connected to 'athena'
#> [2021-05-26 15:48:34]	Overriding cache... Querying Athena...
#> [2021-05-26 15:48:34]	✓ Open connection
#> [2021-05-26 15:48:34]	✓ JDBC connection
#> [2021-05-26 15:48:34]	SQL: SELECT * FROM omop_vocabulary.concept_relationship ORDER BY RANDOM() LIMIT 100
#> [2021-05-26 15:48:34]	Querying...
#> [2021-05-26 15:49:06]	Querying...complete
#> [2021-05-26 15:49:06]	✓ Returned data has more than 0 rows
#> [2021-05-26 15:49:06]	Caching resultset...
#> [2021-05-26 15:49:06]	Postgres connection to 'athena' closed

#> # A tibble: 100 x 6
#>    concept_id_1 concept_id_2 relationship_id     valid_start_date valid_end_date
#>           <int>        <int> <chr>               <date>           <date>        
#>  1     35607932     46050925 Non-avail ind of    2018-07-31       2099-12-31    
#>  2     40769075     45875977 Is a                1970-01-01       2099-12-31    
#>  3      3656648      3656625 Disp dose form of   2020-07-31       2099-12-31    
#>  4      3209845      3181503 Subsumes            2019-08-16       2099-12-31    
#>  5     40738328      1518261 Marketed form of    2018-06-04       2099-12-31    
#>  6     35416756     35416742 Marketed form of    2017-07-17       2099-12-31    
#>  7      3260849      3408603 Is a                2019-08-16       2099-12-31    
#>  8     43210226     43522272 Has quantified form 2017-08-29       2099-12-31    
#>  9     41027658     41027658 Mapped from         2017-07-18       2099-12-31    
#> 10     35909693     36524614 Schema to ICDO      1970-01-01       2099-12-31    
#> # … with 90 more rows, and 1 more variable: invalid_reason <chr>

get_test_data(table = "concept_ancestor", limit = 10)

#> [2021-05-26 15:49:07]	Connected to 'athena'
#> [2021-05-26 15:49:07]	Overriding cache... Querying Athena...
#> [2021-05-26 15:49:07]	✓ Open connection
#> [2021-05-26 15:49:07]	✓ JDBC connection
#> [2021-05-26 15:49:07]	SQL: SELECT * FROM omop_vocabulary.concept_ancestor ORDER BY RANDOM() LIMIT 10
#> [2021-05-26 15:49:07]	Querying...
#> [2021-05-26 15:49:43]	Querying...complete
#> [2021-05-26 15:49:43]	✓ Returned data has more than 0 rows
#> [2021-05-26 15:49:43]	Caching resultset...
#> [2021-05-26 15:49:43]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 4
#>    ancestor_concept_… descendant_concep… min_levels_of_sepa… max_levels_of_sepa…
#>                 <int>              <int>               <int>               <int>
#>  1           45876002            3007242                   1                   1
#>  2           45787136           43273227                   2                   2
#>  3            4048234            4295595                   2                   2
#>  4           46312365           41252886                   2                   2
#>  5           36107167            1164849                   1                   1
#>  6            4194163            4004070                  11                  12
#>  7           45661826           43283132                   2                   3
#>  8           36475002           36409139                   2                   2
#>  9           45803042           21082824                   2                   2
#> 10           36159317            2062257                   2                   2