Get Test Drug Concepts

Get random records of RxNorm and RxNorm Extension Concepts that are:

Valid,

Not a Class,

Belong to the Drug domain

Get random records of ATC and HemOnc Drug Classes that are:

Valid,

Class,

Belong to the Drug domain

get_test_drug_concepts(
  conn,
  conn_fun = "connectAthena()",
  schema = "omop_vocabulary",
  limit = 100,
  verbose = TRUE,
  render_sql = TRUE
)

get_test_drug_classes(
  conn,
  conn_fun = "connectAthena()",
  schema = "omop_vocabulary",
  limit = 100,
  verbose = TRUE,
  render_sql = TRUE
)

See also

Examples

library(chariot)
library(tidyverse)

# Get Unfiltered Test Concepts (Invalid Concepts are included)
get_test_concepts()

#> Connecting using PostgreSQL driver
#> [2021-01-10 12:30:50]	Overriding cache... Querying Athena...
#> [2021-01-10 12:30:50] 

#> ✓ Open connection

#> [2021-01-10 12:30:50] 

#> ✓ JDBC connection

#> [2021-01-10 12:30:50]	SQL: SELECT * FROM omop_vocabulary.concept ORDER BY RANDOM() LIMIT 100
#> [2021-01-10 12:30:50]	Querying...
#> [2021-01-10 12:30:58]	Querying...complete
#> [2021-01-10 12:30:58] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 12:30:59]	Caching resultset...
#> [2021-01-10 12:30:59]	Postgres connection to 'athena' closed

#> # A tibble: 100 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id
#>         <dbl> <chr>        <chr>     <chr>         <chr>           
#>  1   40285649 Decreased t… Condition SNOMED        Clinical Finding
#>  2   40893144 Cetylpyridi… Drug      RxNorm Exten… Clinical Drug F…
#>  3   21085544 Glucosamine… Drug      RxNorm Exten… Marketed Product
#>  4   41888299 Roto         Geography OSM           10th level      
#>  5   44558139 CALCIUM 100… Drug      DPD           Branded Drug Co…
#>  6   35864189 repaglinide… Drug      RxNorm Exten… Branded Drug    
#>  7   44510806 Subtotal ex… Procedure OPCS4         Procedure       
#>  8   42770792 Operations … Procedure OPS           Procedure       
#>  9   41921915 Shuangyang … Geography OSM           6th level       
#> 10   37074084 Ketones | V… Measurem… LOINC         LOINC Hierarchy 
#> # … with 90 more rows, and 5 more variables: standard_concept <chr>,
#> #   concept_code <chr>, valid_start_date <date>, valid_end_date <date>,
#> #   invalid_reason <chr>

# The size of the resultset can be altered
get_test_concepts(limit = 25)

#> Connecting using PostgreSQL driver
#> [2021-01-10 12:30:59]	Overriding cache... Querying Athena...
#> [2021-01-10 12:30:59] 

#> ✓ Open connection

#> [2021-01-10 12:30:59] 

#> ✓ JDBC connection

#> [2021-01-10 12:30:59]	SQL: SELECT * FROM omop_vocabulary.concept ORDER BY RANDOM() LIMIT 25
#> [2021-01-10 12:30:59]	Querying...
#> [2021-01-10 12:31:06]	Querying...complete
#> [2021-01-10 12:31:06] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 12:31:06]	Caching resultset...
#> [2021-01-10 12:31:06]	Postgres connection to 'athena' closed

#> # A tibble: 25 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id
#>         <dbl> <chr>        <chr>     <chr>         <chr>           
#>  1   35917509 Levels IV-V… Measurem… NAACCR        NAACCR Variable 
#>  2   43242547 ATORVASTATI… Drug      BDPM          Drug Product    
#>  3   45997067 Celiprolol … Drug      SNOMED        Pharma/Biol Pro…
#>  4   44812953 [X]Confined… Observat… SNOMED        Event           
#>  5   44931567 BACLOFEN PO… Drug      NDC           11-digit NDC    
#>  6   40134968 Isotretinoi… Drug      RxNorm        Branded Drug Co…
#>  7   42083844 Overlook     Geography OSM           10th level      
#>  8   43042415 26 ML benda… Drug      RxNorm Exten… Quant Clinical …
#>  9   35112522 acetaminoph… Drug      SPL           OTC Drug        
#> 10   40874328 montelukast… Drug      RxNorm Exten… Branded Drug Box
#> # … with 15 more rows, and 5 more variables: standard_concept <chr>,
#> #   concept_code <chr>, valid_start_date <date>, valid_end_date <date>,
#> #   invalid_reason <chr>

# Concepts specific to the Drug domain can be sampled
# RxNorm/RxNorm Extension Concepts
get_test_drug_concepts(limit = 10)

#> Connecting using PostgreSQL driver
#> [2021-01-10 12:31:07]	Overriding cache... Querying Athena...
#> [2021-01-10 12:31:07] 

#> ✓ Open connection

#> [2021-01-10 12:31:07] 

#> ✓ JDBC connection

#> [2021-01-10 12:31:07]	SQL: SELECT * FROM omop_vocabulary.concept c WHERE c.invalid_reason IS NULL AND c.vocabulary_id IN ('RxNorm', 'RxNorm Extension') AND c.standard_concept <> 'C' AND c.domain_id = 'Drug' ORDER BY RANDOM() LIMIT 10
#> [2021-01-10 12:31:07]	Querying...
#> [2021-01-10 12:31:10]	Querying...complete
#> [2021-01-10 12:31:10] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 12:31:10]	Caching resultset...
#> [2021-01-10 12:31:10]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id
#>         <dbl> <chr>        <chr>     <chr>         <chr>           
#>  1    1361078 influenza A… Drug      RxNorm        Branded Drug Co…
#>  2   19088856 norepinephr… Drug      RxNorm        Branded Drug Co…
#>  3   43613229 100 ML Camp… Drug      RxNorm Exten… Marketed Product
#>  4   44117969 pantothenat… Drug      RxNorm Exten… Clinical Drug C…
#>  5   41111471 Ginkgo bilo… Drug      RxNorm Exten… Clinical Drug F…
#>  6   40705548 30000 MG be… Drug      RxNorm Exten… Quant Clinical …
#>  7   41101726 rizatriptan… Drug      RxNorm Exten… Marketed Product
#>  8   41161703 Budesonide … Drug      RxNorm Exten… Branded Drug    
#>  9   43826849 25 ML Furos… Drug      RxNorm Exten… Marketed Product
#> 10   41040180 olanzapine … Drug      RxNorm Exten… Marketed Product
#> # … with 5 more variables: standard_concept <chr>, concept_code <chr>,
#> #   valid_start_date <date>, valid_end_date <date>, invalid_reason <chr>

# HemOnc/ATC Classes
get_test_drug_classes(limit = 10)

#> Connecting using PostgreSQL driver
#> [2021-01-10 12:31:10]	Overriding cache... Querying Athena...
#> [2021-01-10 12:31:10] 

#> ✓ Open connection

#> [2021-01-10 12:31:10] 

#> ✓ JDBC connection

#> [2021-01-10 12:31:10]	SQL: SELECT * FROM omop_vocabulary.concept c WHERE c.invalid_reason IS NULL AND c.vocabulary_id IN ('ATC', 'HemOnc') AND c.standard_concept = 'C' AND c.domain_id = 'Drug' ORDER BY RANDOM() LIMIT 10
#> [2021-01-10 12:31:10]	Querying...
#> [2021-01-10 12:31:10]	Querying...complete
#> [2021-01-10 12:31:10] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 12:31:10]	Caching resultset...
#> [2021-01-10 12:31:10]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id
#>         <dbl> <chr>        <chr>     <chr>         <chr>           
#>  1   40252128 rilpivirine… Drug      ATC           ATC 5th         
#>  2   21601211 cardioplegi… Drug      ATC           ATC 5th         
#>  3   21602429 OTHER GYNEC… Drug      ATC           ATC 2nd         
#>  4   21602100 Corticoster… Drug      ATC           ATC 4th         
#>  5   35807358 FGFR inhibi… Drug      HemOnc        Component Class 
#>  6   21602591 medroxyprog… Drug      ATC           ATC 5th         
#>  7   21600702 pepsin       Drug      ATC           ATC 5th         
#>  8   21602537 PROGESTOGENS Drug      ATC           ATC 3rd         
#>  9   21604214 propofol     Drug      ATC           ATC 5th         
#> 10   44819166 formoterol … Drug      ATC           ATC 5th         
#> # … with 5 more variables: standard_concept <chr>, concept_code <chr>,
#> #   valid_start_date <date>, valid_end_date <date>, invalid_reason <chr>

# Test data can also be taken from tables other than Concept
get_test_data(table = "concept_relationship")

#> Connecting using PostgreSQL driver
#> [2021-01-10 12:31:10]	Overriding cache... Querying Athena...
#> [2021-01-10 12:31:10] 

#> ✓ Open connection

#> [2021-01-10 12:31:10] 

#> ✓ JDBC connection

#> [2021-01-10 12:31:10]	SQL: SELECT * FROM omop_vocabulary.concept_relationship ORDER BY RANDOM() LIMIT 100
#> [2021-01-10 12:31:10]	Querying...
#> [2021-01-10 12:31:44]	Querying...complete
#> [2021-01-10 12:31:44] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 12:31:44]	Caching resultset...
#> [2021-01-10 12:31:44]	Postgres connection to 'athena' closed

#> # A tibble: 100 x 6
#>    concept_id_1 concept_id_2 relationship_id valid_start_date valid_end_date
#>           <dbl>        <dbl> <chr>           <date>           <date>        
#>  1      1419631      1419650 Subsumes        2016-01-01       2099-12-31    
#>  2     44159480     44161782 Marketed form … 2017-10-06       2099-12-31    
#>  3     35626950     46016292 Excipient of    2018-07-31       2099-12-31    
#>  4      4237692      4168626 Part of         2011-07-31       2099-12-31    
#>  5     43619030     43568672 Has brand name  2016-07-01       2099-12-31    
#>  6     46166828     46028631 Has AMP         2015-01-31       2099-12-31    
#>  7     19056402      1717240 Maps to         2020-05-04       2099-12-31    
#>  8     36530386      1537778 ICDO to Chapter 2020-09-23       2099-12-31    
#>  9      3033223      1026064 Has method      2019-12-13       2099-12-31    
#> 10     41508545     41508545 Maps to         2017-07-18       2099-12-31    
#> # … with 90 more rows, and 1 more variable: invalid_reason <chr>

get_test_data(table = "concept_ancestor", limit = 10)

#> Connecting using PostgreSQL driver
#> [2021-01-10 12:31:44]	Overriding cache... Querying Athena...
#> [2021-01-10 12:31:44] 

#> ✓ Open connection

#> [2021-01-10 12:31:44] 

#> ✓ JDBC connection

#> [2021-01-10 12:31:44]	SQL: SELECT * FROM omop_vocabulary.concept_ancestor ORDER BY RANDOM() LIMIT 10
#> [2021-01-10 12:31:44]	Querying...
#> [2021-01-10 12:32:21]	Querying...complete
#> [2021-01-10 12:32:21] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 12:32:21]	Caching resultset...
#> [2021-01-10 12:32:21]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 4
#>    ancestor_concept_… descendant_concep… min_levels_of_sepa… max_levels_of_sepa…
#>                 <dbl>              <dbl>               <dbl>               <dbl>
#>  1           45657601           44115793                   2                   2
#>  2           35529462           41039773                   2                   3
#>  3             438112           42513969                   3                   3
#>  4           45800968           45800968                   0                   0
#>  5           35807389           21068422                   4                   5
#>  6           45783587           44112497                   1                   1
#>  7           21600001           40231976                   6                   8
#>  8           45779164           40907788                   2                   2
#>  9           37014570           37014570                   0                   0
#> 10            1332418           41068448                   3                   4

library(chariot)
library(tidyverse)

# Get Unfiltered Test Concepts (Invalid Concepts are included)
get_test_concepts()

#> Connecting using PostgreSQL driver
#> [2021-01-10 12:32:21]	Overriding cache... Querying Athena...
#> [2021-01-10 12:32:21] 

#> ✓ Open connection

#> [2021-01-10 12:32:21] 

#> ✓ JDBC connection

#> [2021-01-10 12:32:21]	SQL: SELECT * FROM omop_vocabulary.concept ORDER BY RANDOM() LIMIT 100
#> [2021-01-10 12:32:21]	Querying...
#> [2021-01-10 12:32:31]	Querying...complete
#> [2021-01-10 12:32:31] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 12:32:31]	Caching resultset...
#> [2021-01-10 12:32:31]	Postgres connection to 'athena' closed

#> # A tibble: 100 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id
#>         <dbl> <chr>        <chr>     <chr>         <chr>           
#>  1    4245109 pT1a1 (IA1)… Condition SNOMED        Clinical Finding
#>  2   35905053 iodine 5 MG… Drug      NDC           11-digit NDC    
#>  3    1146092 artesunate … Drug      RxNorm        Clinical Drug   
#>  4   45282866 Erythromyci… Drug      NDC           11-digit NDC    
#>  5   41332151 1 ML Silico… Drug      RxNorm Exten… Quant Clinical …
#>  6   41979275 Niedertrebra Geography OSM           8th level       
#>  7   43759254 6900 MG Pot… Drug      RxNorm Exten… Quant Branded B…
#>  8   35825623 builder, bu… Observat… UK Biobank    Answer          
#>  9   40986585 Cascara sag… Drug      RxNorm Exten… Clinical Drug F…
#> 10   46069653 Haddenham V… Device    SNOMED        Physical Object 
#> # … with 90 more rows, and 5 more variables: standard_concept <chr>,
#> #   concept_code <chr>, valid_start_date <date>, valid_end_date <date>,
#> #   invalid_reason <chr>

# The size of the resultset can be altered
get_test_concepts(limit = 25)

#> Connecting using PostgreSQL driver
#> [2021-01-10 12:32:31]	Overriding cache... Querying Athena...
#> [2021-01-10 12:32:31] 

#> ✓ Open connection

#> [2021-01-10 12:32:31] 

#> ✓ JDBC connection

#> [2021-01-10 12:32:31]	SQL: SELECT * FROM omop_vocabulary.concept ORDER BY RANDOM() LIMIT 25
#> [2021-01-10 12:32:31]	Querying...
#> [2021-01-10 12:32:40]	Querying...complete
#> [2021-01-10 12:32:40] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 12:32:40]	Caching resultset...
#> [2021-01-10 12:32:40]	Postgres connection to 'athena' closed

#> # A tibble: 25 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id
#>         <dbl> <chr>        <chr>     <chr>         <chr>           
#>  1    1775805 Cefoxitin 9… Drug      RxNorm        Clinical Drug C…
#>  2   45460280 Obstetric p… Condition Read          Read            
#>  3   43133946 valerian ro… Drug      RxNorm Exten… Branded Drug Co…
#>  4   43314230 Terbix 250 … Drug      AMT           Trade Product U…
#>  5   46201629 Sigvaris Es… Device    SNOMED        Physical Object 
#>  6   43755800 75000 MG CH… Drug      RxNorm Exten… Quant Branded D…
#>  7   43558295 isopropyl a… Drug      NDC           11-digit NDC    
#>  8   21480492 Juzo Soft c… Device    dm+d          AMPP            
#>  9    3249351 Gap in tend… Condition Nebraska Lex… Clinical Finding
#> 10   37600420 lovastatin … Drug      SPL           Prescription Dr…
#> # … with 15 more rows, and 5 more variables: standard_concept <chr>,
#> #   concept_code <chr>, valid_start_date <date>, valid_end_date <date>,
#> #   invalid_reason <chr>

# Concepts specific to the Drug domain can be sampled
# RxNorm/RxNorm Extension Concepts
get_test_drug_concepts(limit = 10)

#> Connecting using PostgreSQL driver
#> [2021-01-10 12:32:40]	Overriding cache... Querying Athena...
#> [2021-01-10 12:32:40] 

#> ✓ Open connection

#> [2021-01-10 12:32:40] 

#> ✓ JDBC connection

#> [2021-01-10 12:32:40]	SQL: SELECT * FROM omop_vocabulary.concept c WHERE c.invalid_reason IS NULL AND c.vocabulary_id IN ('RxNorm', 'RxNorm Extension') AND c.standard_concept <> 'C' AND c.domain_id = 'Drug' ORDER BY RANDOM() LIMIT 10
#> [2021-01-10 12:32:40]	Querying...
#> [2021-01-10 12:32:44]	Querying...complete
#> [2021-01-10 12:32:44] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 12:32:44]	Caching resultset...
#> [2021-01-10 12:32:44]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id
#>         <dbl> <chr>        <chr>     <chr>         <chr>           
#>  1   41333482 1 ML Vitami… Drug      RxNorm Exten… Quant Clinical …
#>  2   44089965 Carbamazepi… Drug      RxNorm Exten… Marketed Product
#>  3   40961097 Escherichia… Drug      RxNorm Exten… Clinical Drug   
#>  4   21094776 diethylamin… Drug      RxNorm Exten… Branded Drug Fo…
#>  5   43293093 Calcium 100… Drug      RxNorm Exten… Branded Drug Box
#>  6   19131278 valine 1.61… Drug      RxNorm        Clinical Drug C…
#>  7   21127799 0.5 ML Suma… Drug      RxNorm Exten… Marketed Product
#>  8   41409482 0.5 ML Filg… Drug      RxNorm Exten… Marketed Product
#>  9   42730898 Sulfadiazin… Drug      RxNorm Exten… Clinical Drug C…
#> 10   42800244 pramoxine /… Drug      RxNorm        Branded Drug Fo…
#> # … with 5 more variables: standard_concept <chr>, concept_code <chr>,
#> #   valid_start_date <date>, valid_end_date <date>, invalid_reason <chr>

# HemOnc/ATC Classes
get_test_drug_classes(limit = 10)

#> Connecting using PostgreSQL driver
#> [2021-01-10 12:32:44]	Overriding cache... Querying Athena...
#> [2021-01-10 12:32:44] 

#> ✓ Open connection

#> [2021-01-10 12:32:44] 

#> ✓ JDBC connection

#> [2021-01-10 12:32:44]	SQL: SELECT * FROM omop_vocabulary.concept c WHERE c.invalid_reason IS NULL AND c.vocabulary_id IN ('ATC', 'HemOnc') AND c.standard_concept = 'C' AND c.domain_id = 'Drug' ORDER BY RANDOM() LIMIT 10
#> [2021-01-10 12:32:44]	Querying...
#> [2021-01-10 12:32:44]	Querying...complete
#> [2021-01-10 12:32:44] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 12:32:44]	Caching resultset...
#> [2021-01-10 12:32:44]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id
#>         <dbl> <chr>        <chr>     <chr>         <chr>           
#>  1   21605267 calcium ace… Drug      ATC           ATC 5th         
#>  2   21600674 albumin tan… Drug      ATC           ATC 5th         
#>  3     715852 indacaterol… Drug      ATC           ATC 5th         
#>  4   21600188 trimethyldi… Drug      ATC           ATC 5th         
#>  5   21603381 bromhexine;… Drug      ATC           ATC 5th         
#>  6   21600965 warfarin; s… Drug      ATC           ATC 5th         
#>  7   21604163 dibotermin … Drug      ATC           ATC 5th         
#>  8   21603389 Mucolytics … Drug      ATC           ATC 5th         
#>  9   21602884 cefuroxime;… Drug      ATC           ATC 5th         
#> 10   21601514 Other low-c… Drug      ATC           ATC 4th         
#> # … with 5 more variables: standard_concept <chr>, concept_code <chr>,
#> #   valid_start_date <date>, valid_end_date <date>, invalid_reason <chr>

# Test data can also be taken from tables other than Concept
get_test_data(table = "concept_relationship")

#> Connecting using PostgreSQL driver
#> [2021-01-10 12:32:44]	Overriding cache... Querying Athena...
#> [2021-01-10 12:32:44] 

#> ✓ Open connection

#> [2021-01-10 12:32:44] 

#> ✓ JDBC connection

#> [2021-01-10 12:32:44]	SQL: SELECT * FROM omop_vocabulary.concept_relationship ORDER BY RANDOM() LIMIT 100
#> [2021-01-10 12:32:44]	Querying...
#> [2021-01-10 12:33:18]	Querying...complete
#> [2021-01-10 12:33:18] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 12:33:18]	Caching resultset...
#> [2021-01-10 12:33:18]	Postgres connection to 'athena' closed

#> # A tibble: 100 x 6
#>    concept_id_1 concept_id_2 relationship_id valid_start_date valid_end_date
#>           <dbl>        <dbl> <chr>           <date>           <date>        
#>  1      1036576      1036576 Mapped from     1970-01-01       2099-12-31    
#>  2     35909782     40198787 Schema to Value 1970-01-01       2099-12-31    
#>  3     44918098     19033656 Maps to         2009-08-03       2099-12-31    
#>  4     41207685     41207685 Maps to         2017-07-18       2099-12-31    
#>  5     43566421     41326709 Brand name of   2017-07-18       2099-12-31    
#>  6     41190373     40971322 Available as b… 2017-07-18       2099-12-31    
#>  7      4061459      4331463 Mapped from     2018-07-31       2099-12-31    
#>  8      4175309      2762733 Subsumes        1970-01-01       2099-12-31    
#>  9      4030871      4039785 Interprets of   2017-01-31       2099-12-31    
#> 10     21044271     21044271 Mapped from     2016-08-07       2099-12-31    
#> # … with 90 more rows, and 1 more variable: invalid_reason <chr>

get_test_data(table = "concept_ancestor", limit = 10)

#> Connecting using PostgreSQL driver
#> [2021-01-10 12:33:18]	Overriding cache... Querying Athena...
#> [2021-01-10 12:33:18] 

#> ✓ Open connection

#> [2021-01-10 12:33:18] 

#> ✓ JDBC connection

#> [2021-01-10 12:33:18]	SQL: SELECT * FROM omop_vocabulary.concept_ancestor ORDER BY RANDOM() LIMIT 10
#> [2021-01-10 12:33:18]	Querying...
#> [2021-01-10 12:33:54]	Querying...complete
#> [2021-01-10 12:33:54] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 12:33:54]	Caching resultset...
#> [2021-01-10 12:33:54]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 4
#>    ancestor_concept_… descendant_concep… min_levels_of_sepa… max_levels_of_sepa…
#>                 <dbl>              <dbl>               <dbl>               <dbl>
#>  1           42898944           41076875                   1                   1
#>  2           41971696           41979830                   1                   1
#>  3           45786318           21035227                   2                   2
#>  4           21601461           35766182                   4                   4
#>  5             733523           43695315                   1                   1
#>  6           36217206            2057185                   2                   2
#>  7           40481827            4089108                   5                  18
#>  8           46239887           42950573                   1                   1
#>  9           42803544            2054510                   2                   2
#> 10            4072484            4224918                   2                   2