Get Test Drug Concepts

Get random records of RxNorm and RxNorm Extension Concepts that are:

Valid,

Not a Class,

Belong to the Drug domain

get_test_drug_concepts(
  conn,
  conn_fun = "connectAthena()",
  schema = "omop_vocabulary",
  limit = 100,
  verbose = TRUE,
  render_sql = TRUE
)

See also

Examples

library(chariot)
library(tidyverse)

# Get Unfiltered Test Concepts (Invalid Concepts are included)
get_test_concepts()

#> Connecting using PostgreSQL driver
#> [2021-01-17 00:57:38]	Overriding cache... Querying Athena...
#> [2021-01-17 00:57:38] 

#> ✓ Open connection

#> [2021-01-17 00:57:38] 

#> ✓ JDBC connection

#> [2021-01-17 00:57:38]	SQL: SELECT * FROM omop_vocabulary.concept ORDER BY RANDOM() LIMIT 100
#> [2021-01-17 00:57:38]	Querying...
#> [2021-01-17 00:57:48]	Querying...complete
#> [2021-01-17 00:57:48] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-17 00:57:48]	Caching resultset...
#> [2021-01-17 00:57:48]	Postgres connection to 'athena' closed

#> # A tibble: 100 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id
#>         <dbl> <chr>        <chr>     <chr>         <chr>           
#>  1   40681556 Juzo Dynami… Device    dm+d          AMP             
#>  2    3168185 SNOMED CT c… Metadata  Nebraska Lex… Model Comp      
#>  3   46305991 ALFALFA - a… Drug      SPL           OTC Drug        
#>  4   45284853 Acetazolami… Drug      NDC           11-digit NDC    
#>  5   40802601 Carvedilol … Drug      RxNorm Exten… Brand Name      
#>  6    3005610 Chills [HIV… Observat… LOINC         Survey          
#>  7   40849686 Benactyzine… Drug      RxNorm Exten… Branded Drug    
#>  8   36294661 misoprostol  Drug      GGR           Ingredient      
#>  9    2850783 Medical and… Procedure ICD10PCS      ICD10PCS Hierar…
#> 10   45987923 Gabitril 15… Drug      SNOMED        Pharma/Biol Pro…
#> # … with 90 more rows, and 5 more variables: standard_concept <chr>,
#> #   concept_code <chr>, valid_start_date <date>, valid_end_date <date>,
#> #   invalid_reason <chr>

# The size of the resultset can be altered
get_test_concepts(limit = 25)

#> Connecting using PostgreSQL driver
#> [2021-01-17 00:57:48]	Overriding cache... Querying Athena...
#> [2021-01-17 00:57:48] 

#> ✓ Open connection

#> [2021-01-17 00:57:48] 

#> ✓ JDBC connection

#> [2021-01-17 00:57:48]	SQL: SELECT * FROM omop_vocabulary.concept ORDER BY RANDOM() LIMIT 25
#> [2021-01-17 00:57:48]	Querying...
#> [2021-01-17 00:57:56]	Querying...complete
#> [2021-01-17 00:57:56] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-17 00:57:56]	Caching resultset...
#> [2021-01-17 00:57:56]	Postgres connection to 'athena' closed

#> # A tibble: 25 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id
#>         <dbl> <chr>        <chr>     <chr>         <chr>           
#>  1   44494974 metaxalone … Drug      NDC           9-digit NDC     
#>  2   21237495 Folic acid … Drug      dm+d          AMP             
#>  3   45927398 Second Degr… Condition CIEL          Diagnosis       
#>  4   43209070 125 ML Meth… Drug      RxNorm Exten… Quant Branded D…
#>  5   36223471 dexpantheno… Drug      RxNorm        Clinical Dose G…
#>  6   46361464 PLEO STROPH… Drug      NDC           11-digit NDC    
#>  7   40837229 Clemizole 5… Drug      RxNorm Exten… Clinical Drug B…
#>  8   21348727 Welland Sil… Device    dm+d          AMP             
#>  9   37016656 Specimen un… Condition SNOMED        Clinical Finding
#> 10   43253748 Paracetamol… Drug      RxNorm Exten… Brand Name      
#> # … with 15 more rows, and 5 more variables: standard_concept <chr>,
#> #   concept_code <chr>, valid_start_date <date>, valid_end_date <date>,
#> #   invalid_reason <chr>

# Concepts specific to the Drug domain can be sampled
# RxNorm/RxNorm Extension Concepts
get_test_drug_concepts(limit = 10)

#> Connecting using PostgreSQL driver
#> [2021-01-17 00:57:56]	Overriding cache... Querying Athena...
#> [2021-01-17 00:57:56] 

#> ✓ Open connection

#> [2021-01-17 00:57:56] 

#> ✓ JDBC connection

#> [2021-01-17 00:57:56]	SQL: SELECT * FROM omop_vocabulary.concept c WHERE c.invalid_reason IS NULL AND c.vocabulary_id IN ('RxNorm', 'RxNorm Extension') AND c.standard_concept <> 'C' AND c.domain_id = 'Drug' ORDER BY RANDOM() LIMIT 10
#> [2021-01-17 00:57:56]	Querying...
#> [2021-01-17 00:58:00]	Querying...complete
#> [2021-01-17 00:58:00] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-17 00:58:00]	Caching resultset...
#> [2021-01-17 00:58:00]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id
#>         <dbl> <chr>        <chr>     <chr>         <chr>           
#>  1   41122788 pantoprazol… Drug      RxNorm Exten… Branded Drug Box
#>  2   41243516 buflomedil … Drug      RxNorm Exten… Clinical Drug B…
#>  3   43759128 500 ML Argi… Drug      RxNorm Exten… Quant Branded D…
#>  4   40057381 iron sorbit… Drug      RxNorm        Clinical Drug F…
#>  5     587659 Immunoglobu… Drug      RxNorm Exten… Clinical Drug B…
#>  6   43611365 50000 MG Ka… Drug      RxNorm Exten… Marketed Product
#>  7   43140120 2000 ML Nit… Drug      RxNorm Exten… Marketed Product
#>  8   41290280 Mianserin 6… Drug      RxNorm Exten… Marketed Product
#>  9   21026901 ammonium ca… Drug      RxNorm Exten… Clinical Drug F…
#> 10   35851320 naphazoline… Drug      RxNorm Exten… Clinical Drug   
#> # … with 5 more variables: standard_concept <chr>, concept_code <chr>,
#> #   valid_start_date <date>, valid_end_date <date>, invalid_reason <chr>

# HemOnc/ATC Classes
get_test_drug_classes(limit = 10)

#> Connecting using PostgreSQL driver
#> [2021-01-17 00:58:00]	Overriding cache... Querying Athena...
#> [2021-01-17 00:58:00] 

#> ✓ Open connection

#> [2021-01-17 00:58:00] 

#> ✓ JDBC connection

#> [2021-01-17 00:58:00]	SQL: SELECT * FROM omop_vocabulary.concept c WHERE c.invalid_reason IS NULL AND c.vocabulary_id IN ('ATC', 'HemOnc') AND c.standard_concept = 'C' AND c.domain_id = 'Drug' ORDER BY RANDOM() LIMIT 10
#> [2021-01-17 00:58:00]	Querying...
#> [2021-01-17 00:58:00]	Querying...complete
#> [2021-01-17 00:58:00] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-17 00:58:00]	Caching resultset...
#> [2021-01-17 00:58:00]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id
#>         <dbl> <chr>        <chr>     <chr>         <chr>           
#>  1   21602175 fluocortolo… Drug      ATC           ATC 5th         
#>  2   21604676 niaprazine   Drug      ATC           ATC 5th         
#>  3   21601368 papillomavi… Drug      ATC           ATC 5th         
#>  4   21601398 Alkyl sulfo… Drug      ATC           ATC 4th         
#>  5    1123903 etelcalceti… Drug      ATC           ATC 5th         
#>  6   21601218 lysine; par… Drug      ATC           ATC 5th         
#>  7   21601340 combination… Drug      ATC           ATC 5th         
#>  8    1588677 emtricitabi… Drug      ATC           ATC 5th         
#>  9   21600714 Insulins an… Drug      ATC           ATC 4th         
#> 10   21600722 Insulins an… Drug      ATC           ATC 4th         
#> # … with 5 more variables: standard_concept <chr>, concept_code <chr>,
#> #   valid_start_date <date>, valid_end_date <date>, invalid_reason <chr>

# Test data can also be taken from tables other than Concept
get_test_data(table = "concept_relationship")

#> Connecting using PostgreSQL driver
#> [2021-01-17 00:58:01]	Overriding cache... Querying Athena...
#> [2021-01-17 00:58:01] 

#> ✓ Open connection

#> [2021-01-17 00:58:01] 

#> ✓ JDBC connection

#> [2021-01-17 00:58:01]	SQL: SELECT * FROM omop_vocabulary.concept_relationship ORDER BY RANDOM() LIMIT 100
#> [2021-01-17 00:58:01]	Querying...
#> [2021-01-17 00:58:38]	Querying...complete
#> [2021-01-17 00:58:38] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-17 00:58:38]	Caching resultset...
#> [2021-01-17 00:58:38]	Postgres connection to 'athena' closed

#> # A tibble: 100 x 6
#>    concept_id_1 concept_id_2 relationship_id valid_start_date valid_end_date
#>           <dbl>        <dbl> <chr>           <date>           <date>        
#>  1     19082224     43607687 RxNorm dose fo… 2016-07-01       2099-12-31    
#>  2     36811498     43515171 Marketed form … 2019-01-28       2099-12-31    
#>  3      3486679      3486682 Part of         2019-08-16       2099-12-31    
#>  4      3499640      3387309 Finding site of 2019-08-16       2099-12-31    
#>  5      2020041      2069813 Brand name of   2020-07-31       2099-12-31    
#>  6     40905934     19092339 Has brand name  2017-07-18       2099-12-31    
#>  7     40873858     41488616 Has marketed f… 2017-07-18       2099-12-31    
#>  8     35856368     36621664 Mapped from     2020-10-02       2099-12-31    
#>  9     45987963     37310856 Has AMP restr … 2020-03-09       2099-12-31    
#> 10      3436334      3300074 Interprets of   2019-08-16       2099-12-31    
#> # … with 90 more rows, and 1 more variable: invalid_reason <chr>

get_test_data(table = "concept_ancestor", limit = 10)

#> Connecting using PostgreSQL driver
#> [2021-01-17 00:58:38]	Overriding cache... Querying Athena...
#> [2021-01-17 00:58:38] 

#> ✓ Open connection

#> [2021-01-17 00:58:38] 

#> ✓ JDBC connection

#> [2021-01-17 00:58:38]	SQL: SELECT * FROM omop_vocabulary.concept_ancestor ORDER BY RANDOM() LIMIT 10
#> [2021-01-17 00:58:38]	Querying...
#> [2021-01-17 00:59:25]	Querying...complete
#> [2021-01-17 00:59:25] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-17 00:59:25]	Caching resultset...
#> [2021-01-17 00:59:25]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 4
#>    ancestor_concept_… descendant_concep… min_levels_of_sepa… max_levels_of_sepa…
#>                 <dbl>              <dbl>               <dbl>               <dbl>
#>  1           45659927           21100442                   2                   3
#>  2           44097951           44126071                   0                   0
#>  3           44350120           43293676                   2                   3
#>  4           45665452           43652118                   2                   3
#>  5           46294058           43701589                   2                   3
#>  6            4278311           42596627                   1                   1
#>  7           40112580           35769613                   2                   2
#>  8           36878884           43733898                   2                   2
#>  9           44360148            2012773                   2                   2
#> 10           45701084           43834257                   2                   2