Get Test Drug Concepts

Get random records of RxNorm and RxNorm Extension Concepts that are:

Valid,

Not a Class,

Belong to the Drug domain

get_test_drug_concepts(
  conn,
  conn_fun = "connectAthena()",
  schema = "omop_vocabulary",
  limit = 100,
  verbose = TRUE,
  render_sql = TRUE
)

See also

Examples

library(chariot)
library(tidyverse)

# Get Unfiltered Test Concepts (Invalid Concepts are included)
get_test_concepts()

#> Connecting using PostgreSQL driver

#> [2021-03-28 23:41:55]	Overriding cache... Querying Athena...
#> [2021-03-28 23:41:55] 

#> ✓ Open connection

#> [2021-03-28 23:41:55] 

#> ✓ JDBC connection

#> [2021-03-28 23:41:55]	SQL: SELECT * FROM omop_vocabulary.concept ORDER BY RANDOM() LIMIT 100
#> [2021-03-28 23:41:55]	Querying...
#> [2021-03-28 23:42:03]	Querying...complete
#> [2021-03-28 23:42:03] 

#> ✓ Returned data has more than 0 rows

#> [2021-03-28 23:42:03]	Caching resultset...
#> [2021-03-28 23:42:03]	Postgres connection to 'athena' closed

#> # A tibble: 100 x 10
#>    concept_id concept_name             domain_id vocabulary_id  concept_class_id
#>         <int> <chr>                    <chr>     <chr>          <chr>           
#>  1   45625444 terbinafine hydrochlori… Drug      SPL            OTC Drug        
#>  2    2854923 Imaging @ Non-Axial Low… Procedure ICD10PCS       ICD10PCS Hierar…
#>  3   44635789 AVOBENZONE 2.2% / ENZAC… Drug      DPD            Branded Drug Co…
#>  4   45452317 Other specified operati… Procedure Read           Read            
#>  5   21080201 Estradiol 0.025 MG Vagi… Drug      RxNorm Extens… Marketed Product
#>  6     576310 SODIUM PHOS,M-BASIC-D-B… Drug      GCN_SEQNO      GCN_SEQNO       
#>  7   42769175 Associations and remova… Procedure OPS            Procedure       
#>  8   46036706 Galpharm Athletes Foot … Drug      SNOMED         Pharma/Biol Pro…
#>  9   40886672 Enalapril 2.5 MG Oral T… Drug      RxNorm Extens… Marketed Product
#> 10   21301403 Allopurinol 100mg table… Drug      dm+d           AMPP            
#> # … with 90 more rows, and 5 more variables: standard_concept <chr>,
#> #   concept_code <chr>, valid_start_date <date>, valid_end_date <date>,
#> #   invalid_reason <chr>

# The size of the resultset can be altered
get_test_concepts(limit = 25)

#> Connecting using PostgreSQL driver

#> [2021-03-28 23:42:03]	Overriding cache... Querying Athena...
#> [2021-03-28 23:42:03] 

#> ✓ Open connection

#> [2021-03-28 23:42:03] 

#> ✓ JDBC connection

#> [2021-03-28 23:42:03]	SQL: SELECT * FROM omop_vocabulary.concept ORDER BY RANDOM() LIMIT 25
#> [2021-03-28 23:42:03]	Querying...
#> [2021-03-28 23:42:11]	Querying...complete
#> [2021-03-28 23:42:11] 

#> ✓ Returned data has more than 0 rows

#> [2021-03-28 23:42:11]	Caching resultset...
#> [2021-03-28 23:42:11]	Postgres connection to 'athena' closed

#> # A tibble: 25 x 10
#>    concept_id concept_name             domain_id  vocabulary_id concept_class_id
#>         <int> <chr>                    <chr>      <chr>         <chr>           
#>  1   40795735 Trihexyphenidyl          Observati… LOINC         LOINC Component 
#>  2   42211112 Tracheal or Bronchial R… Procedure  EDI           Procedure       
#>  3   41025747 zolmitriptan 5 MG Disin… Drug       RxNorm Exten… Branded Drug    
#>  4   45827717 avobenzone, homosalate,… Drug       NDC           9-digit NDC     
#>  5    3244906 Biological agent therap… Procedure  Nebraska Lex… Procedure       
#>  6   21268327 Mirtazapine 45mg tablet… Drug       dm+d          AMPP            
#>  7   37036053 Vein | Radio fluoroscop… Measureme… LOINC         LOINC Hierarchy 
#>  8   46136099 Scanlux 370mg/ml soluti… Drug       SNOMED        Pharma/Biol Pro…
#>  9   44125523 Diclofenac 50 MG Oral T… Drug       RxNorm Exten… Branded Drug    
#> 10   40097460 Vitamin D 0.01 MG        Drug       RxNorm        Clinical Drug C…
#> # … with 15 more rows, and 5 more variables: standard_concept <chr>,
#> #   concept_code <chr>, valid_start_date <date>, valid_end_date <date>,
#> #   invalid_reason <chr>

# Concepts specific to the Drug domain can be sampled
# RxNorm/RxNorm Extension Concepts
get_test_drug_concepts(limit = 10)

#> Connecting using PostgreSQL driver

#> [2021-03-28 23:42:11]	Overriding cache... Querying Athena...
#> [2021-03-28 23:42:11] 

#> ✓ Open connection

#> [2021-03-28 23:42:11] 

#> ✓ JDBC connection

#> [2021-03-28 23:42:11]	SQL: SELECT * FROM omop_vocabulary.concept c WHERE c.invalid_reason IS NULL AND c.vocabulary_id IN ('RxNorm', 'RxNorm Extension') AND c.standard_concept <> 'C' AND c.domain_id = 'Drug' ORDER BY RANDOM() LIMIT 10
#> [2021-03-28 23:42:11]	Querying...
#> [2021-03-28 23:42:13]	Querying...complete
#> [2021-03-28 23:42:13] 

#> ✓ Returned data has more than 0 rows

#> [2021-03-28 23:42:13]	Caching resultset...
#> [2021-03-28 23:42:13]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 10
#>    concept_id concept_name              domain_id vocabulary_id concept_class_id
#>         <int> <chr>                     <chr>     <chr>         <chr>           
#>  1   35413187 2 ML Adenosine 3 MG/ML I… Drug      RxNorm Exten… Quant Branded D…
#>  2   42931095 Alprostadil 0.005 MG/ML … Drug      RxNorm Exten… Branded Drug Co…
#>  3   43633094 500 ML dexpanthenol 1 MG… Drug      RxNorm Exten… Quant Branded B…
#>  4   41402574 60 ML dipropizine, (S)-i… Drug      RxNorm Exten… Quant Clinical …
#>  5   35147277 levothyroxine 0.075 MG O… Drug      RxNorm Exten… Branded Drug    
#>  6   40708435 1000 MG imiquimod 0.05 M… Drug      RxNorm Exten… Marketed Product
#>  7   44072446 Verapamil 120 MG Extende… Drug      RxNorm Exten… Branded Drug    
#>  8   41466595 2990 MG Acetylcysteine 0… Drug      RxNorm Exten… Quant Branded B…
#>  9   43623348 Clarithromycin 50 MG/ML … Drug      RxNorm Exten… Branded Drug Box
#> 10   19072935 alprazolam 0.1 MG/ML Ora… Drug      RxNorm        Clinical Drug   
#> # … with 5 more variables: standard_concept <chr>, concept_code <chr>,
#> #   valid_start_date <date>, valid_end_date <date>, invalid_reason <chr>

# HemOnc/ATC Classes
get_test_drug_classes(limit = 10)

#> Connecting using PostgreSQL driver

#> [2021-03-28 23:42:14]	Overriding cache... Querying Athena...
#> [2021-03-28 23:42:14] 

#> ✓ Open connection

#> [2021-03-28 23:42:14] 

#> ✓ JDBC connection

#> [2021-03-28 23:42:14]	SQL: SELECT * FROM omop_vocabulary.concept c WHERE c.invalid_reason IS NULL AND c.vocabulary_id IN ('ATC', 'HemOnc') AND c.standard_concept = 'C' AND c.domain_id = 'Drug' ORDER BY RANDOM() LIMIT 10
#> [2021-03-28 23:42:14]	Querying...
#> [2021-03-28 23:42:14]	Querying...complete
#> [2021-03-28 23:42:14] 

#> ✓ Returned data has more than 0 rows

#> [2021-03-28 23:42:14]	Caching resultset...
#> [2021-03-28 23:42:14]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 10
#>    concept_id concept_name              domain_id vocabulary_id concept_class_id
#>         <int> <chr>                     <chr>     <chr>         <chr>           
#>  1   21604570 potassium clorazepate; o… Drug      ATC           ATC 5th         
#>  2   21604352 phenacetin, combinations… Drug      ATC           ATC 5th         
#>  3    1588672 idarucizumab              Drug      ATC           ATC 5th         
#>  4   21600650 loperamide, combinations… Drug      ATC           ATC 5th         
#>  5   21605473 chromium (51Cr) edetate   Drug      ATC           ATC 5th         
#>  6   21604589 Diphenylmethane derivati… Drug      ATC           ATC 4th         
#>  7   21602903 cefodizime; parenteral    Drug      ATC           ATC 5th         
#>  8   21601217 alanyl glutamine; parent… Drug      ATC           ATC 5th         
#>  9   21600155 (2-benzhydryloxyethyl)di… Drug      ATC           ATC 5th         
#> 10     715900 medroxyprogesterone and … Drug      ATC           ATC 5th         
#> # … with 5 more variables: standard_concept <chr>, concept_code <chr>,
#> #   valid_start_date <date>, valid_end_date <date>, invalid_reason <chr>

# Test data can also be taken from tables other than Concept
get_test_data(table = "concept_relationship")

#> Connecting using PostgreSQL driver

#> [2021-03-28 23:42:14]	Overriding cache... Querying Athena...
#> [2021-03-28 23:42:14] 

#> ✓ Open connection

#> [2021-03-28 23:42:14] 

#> ✓ JDBC connection

#> [2021-03-28 23:42:14]	SQL: SELECT * FROM omop_vocabulary.concept_relationship ORDER BY RANDOM() LIMIT 100
#> [2021-03-28 23:42:14]	Querying...
#> [2021-03-28 23:42:46]	Querying...complete
#> [2021-03-28 23:42:46] 

#> ✓ Returned data has more than 0 rows

#> [2021-03-28 23:42:46]	Caching resultset...
#> [2021-03-28 23:42:46]	Postgres connection to 'athena' closed

#> # A tibble: 100 x 6
#>    concept_id_1 concept_id_2 relationship_id     valid_start_date valid_end_date
#>           <int>        <int> <chr>               <date>           <date>        
#>  1     46196489     35626975 VMP has prescr stat 2018-07-31       2099-12-31    
#>  2     43732482     43732482 Maps to             2016-12-18       2099-12-31    
#>  3      1025532     37021417 System of           2019-12-13       2099-12-31    
#>  4     43520826     46234469 RxNorm has dose fo… 2017-08-29       2099-12-31    
#>  5      4116007         9696 Mapped from         2017-10-01       2099-12-31    
#>  6     37218704     35626993 Has disc indicator  2020-03-09       2099-12-31    
#>  7     41043539     41043539 Maps to             2017-07-18       2099-12-31    
#>  8     37078842     40776014 Is a                1970-01-01       2099-12-31    
#>  9     45875980      3053130 Subsumes            1970-01-01       2099-12-31    
#> 10     45865057     42707316 Maps to             2015-01-29       2099-12-31    
#> # … with 90 more rows, and 1 more variable: invalid_reason <chr>

get_test_data(table = "concept_ancestor", limit = 10)

#> Connecting using PostgreSQL driver

#> [2021-03-28 23:42:46]	Overriding cache... Querying Athena...
#> [2021-03-28 23:42:46] 

#> ✓ Open connection

#> [2021-03-28 23:42:46] 

#> ✓ JDBC connection

#> [2021-03-28 23:42:46]	SQL: SELECT * FROM omop_vocabulary.concept_ancestor ORDER BY RANDOM() LIMIT 10
#> [2021-03-28 23:42:46]	Querying...
#> [2021-03-28 23:43:19]	Querying...complete
#> [2021-03-28 23:43:19] 

#> ✓ Returned data has more than 0 rows

#> [2021-03-28 23:43:19]	Caching resultset...
#> [2021-03-28 23:43:19]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 4
#>    ancestor_concept_… descendant_concep… min_levels_of_sepa… max_levels_of_sepa…
#>                 <int>              <int>               <int>               <int>
#>  1           41422097           41420388                   1                   1
#>  2           44347241           41321605                   2                   3
#>  3           45692618           43764282                   2                   2
#>  4             581567           21074418                   2                   2
#>  5           21605008           36264552                   4                   5
#>  6            4126705           46072365                   4                   5
#>  7            1326020            2063723                   1                   1
#>  8            4294168           42895162                   1                   1
#>  9           45778388           41215729                   1                   1
#> 10           40140735           43296571                   2                   2