Get Test Drug Concepts

Get random records of RxNorm and RxNorm Extension Concepts that are:

Valid,

Not a Class,

Belong to the Drug domain

get_test_drug_concepts(
  conn,
  conn_fun = "connectAthena()",
  schema = "omop_vocabulary",
  limit = 100,
  verbose = TRUE,
  render_sql = TRUE
)

See also

Examples

library(chariot)
library(tidyverse)

# Get Unfiltered Test Concepts (Invalid Concepts are included)
get_test_concepts()

#> [2021-05-26 15:07:17]	Connected to 'athena'
#> [2021-05-26 15:07:17]	Overriding cache... Querying Athena...
#> [2021-05-26 15:07:17]	✓ Open connection
#> [2021-05-26 15:07:17]	✓ JDBC connection
#> [2021-05-26 15:07:17]	SQL: SELECT * FROM omop_vocabulary.concept ORDER BY RANDOM() LIMIT 100
#> [2021-05-26 15:07:17]	Querying...
#> [2021-05-26 15:07:26]	Querying...complete
#> [2021-05-26 15:07:26]	✓ Returned data has more than 0 rows
#> [2021-05-26 15:07:26]	Caching resultset...
#> [2021-05-26 15:07:26]	Postgres connection to 'athena' closed

#> # A tibble: 100 x 10
#>    concept_id concept_name              domain_id vocabulary_id concept_class_id
#>         <int> <chr>                     <chr>     <chr>         <chr>           
#>  1   45930810 Massive Aspiration Syndr… Condition CIEL          Diagnosis       
#>  2   44846305 bacitracin zinc 0.5 UNT/… Drug      NDC           11-digit NDC    
#>  3   42808872 cetirizine hydrochloride… Drug      SPL           OTC Drug        
#>  4    2813500 Medical and Surgical @ F… Procedure ICD10PCS      ICD10PCS Hierar…
#>  5   46350556 CY BETTER LIPS BALM HUME… Drug      NDC           11-digit NDC    
#>  6   36056381 500 ML aluminum hydroxid… Drug      RxNorm Exten… Marketed Product
#>  7   36258324 montelukast 4 MG Chewabl… Drug      RxNorm Exten… Marketed Product
#>  8   21407569 Haddenham Veni class 1 (… Device    dm+d          AMPP            
#>  9   40960458 ST. JOHN'S WORT EXTRACT … Drug      RxNorm Exten… Clinical Drug B…
#> 10   41420028 100 ML Fosfomycin 26.4 M… Drug      RxNorm Exten… Quant Clinical …
#> # … with 90 more rows, and 5 more variables: standard_concept <chr>,
#> #   concept_code <chr>, valid_start_date <date>, valid_end_date <date>,
#> #   invalid_reason <chr>

# The size of the resultset can be altered
get_test_concepts(limit = 25)

#> [2021-05-26 15:07:26]	Connected to 'athena'
#> [2021-05-26 15:07:26]	Overriding cache... Querying Athena...
#> [2021-05-26 15:07:26]	✓ Open connection
#> [2021-05-26 15:07:26]	✓ JDBC connection
#> [2021-05-26 15:07:26]	SQL: SELECT * FROM omop_vocabulary.concept ORDER BY RANDOM() LIMIT 25
#> [2021-05-26 15:07:26]	Querying...
#> [2021-05-26 15:07:34]	Querying...complete
#> [2021-05-26 15:07:34]	✓ Returned data has more than 0 rows
#> [2021-05-26 15:07:34]	Caching resultset...
#> [2021-05-26 15:07:34]	Postgres connection to 'athena' closed

#> # A tibble: 25 x 10
#>    concept_id concept_name             domain_id vocabulary_id  concept_class_id
#>         <int> <chr>                    <chr>     <chr>          <chr>           
#>  1   21244051 Hydroxycarbamide 300mg … Drug      dm+d           VMPP            
#>  2   45999698 Flecainide 50mg/5ml ora… Drug      SNOMED         Pharma/Biol Pro…
#>  3   46351586 NUTRICEL ADDITIVE SYSTE… Drug      NDC            11-digit NDC    
#>  4    2048081 ceftriaxone 2000 MG Inj… Drug      RxNorm Extens… Marketed Product
#>  5   21207677 Ipratropium bromide 250… Drug      dm+d           VMPP            
#>  6   42830904 famotidine 40 MG Oral T… Drug      NDC            11-digit NDC    
#>  7   42100703 SPEEDYFLAP               Device    EDI            Device          
#>  8   19059294 Congestant D             Drug      RxNorm         Brand Name      
#>  9   42944784 Carboxymethylcellulose … Drug      RxNorm Extens… Branded Drug Co…
#> 10   35904497 fluconazole 150mg/1 ORA… Drug      NDC            9-digit NDC     
#> # … with 15 more rows, and 5 more variables: standard_concept <chr>,
#> #   concept_code <chr>, valid_start_date <date>, valid_end_date <date>,
#> #   invalid_reason <chr>

# Concepts specific to the Drug domain can be sampled
# RxNorm/RxNorm Extension Concepts
get_test_drug_concepts(limit = 10)

#> [2021-05-26 15:07:34]	Connected to 'athena'
#> [2021-05-26 15:07:34]	Overriding cache... Querying Athena...
#> [2021-05-26 15:07:34]	✓ Open connection
#> [2021-05-26 15:07:34]	✓ JDBC connection
#> [2021-05-26 15:07:34]	SQL: SELECT * FROM omop_vocabulary.concept c WHERE c.invalid_reason IS NULL AND c.vocabulary_id IN ('RxNorm', 'RxNorm Extension') AND c.standard_concept <> 'C' AND c.domain_id = 'Drug' ORDER BY RANDOM() LIMIT 10
#> [2021-05-26 15:07:34]	Querying...
#> [2021-05-26 15:07:38]	Querying...complete
#> [2021-05-26 15:07:38]	✓ Returned data has more than 0 rows
#> [2021-05-26 15:07:38]	Caching resultset...
#> [2021-05-26 15:07:38]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 10
#>    concept_id concept_name              domain_id vocabulary_id concept_class_id
#>         <int> <chr>                     <chr>     <chr>         <chr>           
#>  1   44097041 Palladium 6 X Oral Granu… Drug      RxNorm Exten… Clinical Drug   
#>  2    2034342 0.2 ML bemiparin 175 MG/… Drug      RxNorm Exten… Marketed Product
#>  3   43204612 Amoxicillin / Clavulanat… Drug      RxNorm Exten… Branded Drug Fo…
#>  4   43632573 250 ML Leucine 8.8 MG/ML… Drug      RxNorm Exten… Marketed Product
#>  5   41040052 olanzapine 2.5 MG Oral T… Drug      RxNorm Exten… Marketed Product
#>  6   41271076 Vitamin A 3000 UNT Oral … Drug      RxNorm Exten… Clinical Drug B…
#>  7   40979310 Glyburide 1.75 MG Oral T… Drug      RxNorm Exten… Marketed Product
#>  8   35772868 Glipizide 5 MG Extended … Drug      RxNorm Exten… Marketed Product
#>  9   40864534 Antithrombin III Injecta… Drug      RxNorm Exten… Branded Drug Fo…
#> 10   43834517 Dimenhydrinate 50 MG Ora… Drug      RxNorm Exten… Marketed Product
#> # … with 5 more variables: standard_concept <chr>, concept_code <chr>,
#> #   valid_start_date <date>, valid_end_date <date>, invalid_reason <chr>

# HemOnc/ATC Classes
get_test_drug_classes(limit = 10)

#> [2021-05-26 15:07:38]	Connected to 'athena'
#> [2021-05-26 15:07:38]	Overriding cache... Querying Athena...
#> [2021-05-26 15:07:38]	✓ Open connection
#> [2021-05-26 15:07:38]	✓ JDBC connection
#> [2021-05-26 15:07:38]	SQL: SELECT * FROM omop_vocabulary.concept c WHERE c.invalid_reason IS NULL AND c.vocabulary_id IN ('ATC', 'HemOnc') AND c.standard_concept = 'C' AND c.domain_id = 'Drug' ORDER BY RANDOM() LIMIT 10
#> [2021-05-26 15:07:38]	Querying...
#> [2021-05-26 15:07:38]	Querying...complete
#> [2021-05-26 15:07:38]	✓ Returned data has more than 0 rows
#> [2021-05-26 15:07:38]	Caching resultset...
#> [2021-05-26 15:07:38]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 10
#>    concept_id concept_name              domain_id vocabulary_id concept_class_id
#>         <int> <chr>                     <chr>     <chr>         <chr>           
#>  1   21601022 argatroban; parenteral    Drug      ATC           ATC 5th         
#>  2   21600045 various; oral             Drug      ATC           ATC 5th         
#>  3   21600873 calcium laevulate; paren… Drug      ATC           ATC 5th         
#>  4   21604521 bromperidol; systemic     Drug      ATC           ATC 5th         
#>  5   21600654 hydrocortisone; rectal    Drug      ATC           ATC 5th         
#>  6   21604089 Other muscle relaxants, … Drug      ATC           ATC 4th         
#>  7   21601758 barnidipine; oral         Drug      ATC           ATC 5th         
#>  8   21602070 rifaximin; topical        Drug      ATC           ATC 5th         
#>  9   21601837 irbesartan and diuretics  Drug      ATC           ATC 5th         
#> 10   21603144 tromantadine; systemic    Drug      ATC           ATC 5th         
#> # … with 5 more variables: standard_concept <chr>, concept_code <chr>,
#> #   valid_start_date <date>, valid_end_date <date>, invalid_reason <chr>

# Test data can also be taken from tables other than Concept
get_test_data(table = "concept_relationship")

#> [2021-05-26 15:07:38]	Connected to 'athena'
#> [2021-05-26 15:07:38]	Overriding cache... Querying Athena...
#> [2021-05-26 15:07:38]	✓ Open connection
#> [2021-05-26 15:07:38]	✓ JDBC connection
#> [2021-05-26 15:07:38]	SQL: SELECT * FROM omop_vocabulary.concept_relationship ORDER BY RANDOM() LIMIT 100
#> [2021-05-26 15:07:38]	Querying...
#> [2021-05-26 15:08:11]	Querying...complete
#> [2021-05-26 15:08:11]	✓ Returned data has more than 0 rows
#> [2021-05-26 15:08:11]	Caching resultset...
#> [2021-05-26 15:08:11]	Postgres connection to 'athena' closed

#> # A tibble: 100 x 6
#>    concept_id_1 concept_id_2 relationship_id     valid_start_date valid_end_date
#>           <int>        <int> <chr>               <date>           <date>        
#>  1      1146647      1146646 RxNorm is a         2020-08-03       2099-12-31    
#>  2     37418515       955432 Maps to             2020-06-14       2099-12-31    
#>  3     45766528      1553319 Asso morph of       1970-01-01       2099-12-31    
#>  4     42480110     21099187 Has marketed form   2017-06-24       2099-12-31    
#>  5     43586502     40890508 Maps to             2017-08-25       2099-12-31    
#>  6     40876082     41385097 Has quantified form 2017-07-18       2099-12-31    
#>  7     41425137     41424092 Has marketed form   2017-07-18       2099-12-31    
#>  8     19044580     40940845 Brand name of       2017-07-18       2099-12-31    
#>  9     46282895     36708233 Is a                2017-01-31       2099-12-31    
#> 10     40971712     40971712 Maps to             2017-07-18       2099-12-31    
#> # … with 90 more rows, and 1 more variable: invalid_reason <chr>

get_test_data(table = "concept_ancestor", limit = 10)

#> [2021-05-26 15:08:11]	Connected to 'athena'
#> [2021-05-26 15:08:11]	Overriding cache... Querying Athena...
#> [2021-05-26 15:08:11]	✓ Open connection
#> [2021-05-26 15:08:11]	✓ JDBC connection
#> [2021-05-26 15:08:11]	SQL: SELECT * FROM omop_vocabulary.concept_ancestor ORDER BY RANDOM() LIMIT 10
#> [2021-05-26 15:08:11]	Querying...
#> [2021-05-26 15:08:45]	Querying...complete
#> [2021-05-26 15:08:45]	✓ Returned data has more than 0 rows
#> [2021-05-26 15:08:45]	Caching resultset...
#> [2021-05-26 15:08:45]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 4
#>    ancestor_concept_… descendant_concep… min_levels_of_sepa… max_levels_of_sepa…
#>                 <int>              <int>               <int>               <int>
#>  1           41928939           41944399                   2                   2
#>  2            4040721            2741789                   6                   6
#>  3            4024528           46272181                   3                   3
#>  4            4006473           37397000                   1                   1
#>  5           35509017           43024545                   2                   3
#>  6            4322976           44791167                   7                  11
#>  7            4042140           37108634                   4                   8
#>  8           35807480           41033266                   4                   5
#>  9           42806731           41030852                   2                   2
#> 10           45647261           43139419                   2                   2