Get Test Drug Concepts

Get random records of RxNorm and RxNorm Extension Concepts that are:

Valid,

Not a Class,

Belong to the Drug domain

Get random records of ATC and HemOnc Drug Classes that are:

Valid,

Class,

Belong to the Drug domain

get_test_drug_concepts(
  conn,
  conn_fun = "connectAthena()",
  schema = "omop_vocabulary",
  limit = 100,
  verbose = TRUE,
  render_sql = TRUE
)

get_test_drug_classes(
  conn,
  conn_fun = "connectAthena()",
  schema = "omop_vocabulary",
  limit = 100,
  verbose = TRUE,
  render_sql = TRUE
)

See also

Examples

library(chariot)
library(tidyverse)

# Get Unfiltered Test Concepts (Invalid Concepts are included)
get_test_concepts()

#> Connecting using PostgreSQL driver
#> [2021-01-10 14:51:41]	Overriding cache... Querying Athena...
#> [2021-01-10 14:51:41] 

#> ✓ Open connection

#> [2021-01-10 14:51:41] 

#> ✓ JDBC connection

#> [2021-01-10 14:51:41]	SQL: SELECT * FROM omop_vocabulary.concept ORDER BY RANDOM() LIMIT 100
#> [2021-01-10 14:51:41]	Querying...
#> [2021-01-10 14:51:49]	Querying...complete
#> [2021-01-10 14:51:49] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 14:51:49]	Caching resultset...
#> [2021-01-10 14:51:49]	Postgres connection to 'athena' closed

#> # A tibble: 100 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id
#>         <dbl> <chr>        <chr>     <chr>         <chr>           
#>  1   44587692 BETA-CAROTE… Drug      DPD           Clinical Drug   
#>  2   36478647 levetiracet… Drug      NDC           11-digit NDC    
#>  3   43192354 Epoprosteno… Drug      RxNorm Exten… Branded Drug    
#>  4   45417484 chlorphenir… Drug      Multum        Multum          
#>  5   42553322 EPLERENONE … Drug      VA Product    VA Product      
#>  6    4174347 Structure o… Spec Ana… SNOMED        Body Structure  
#>  7   41080457 Benzocaine … Drug      RxNorm Exten… Clinical Drug F…
#>  8    4312044 Entire groo… Spec Ana… SNOMED        Body Structure  
#>  9   35771364 atorvastati… Drug      RxNorm Exten… Branded Drug Box
#> 10    1566692 aspirin 81 … Drug      NDC           11-digit NDC    
#> # … with 90 more rows, and 5 more variables: standard_concept <chr>,
#> #   concept_code <chr>, valid_start_date <date>, valid_end_date <date>,
#> #   invalid_reason <chr>

# The size of the resultset can be altered
get_test_concepts(limit = 25)

#> Connecting using PostgreSQL driver
#> [2021-01-10 14:51:49]	Overriding cache... Querying Athena...
#> [2021-01-10 14:51:49] 

#> ✓ Open connection

#> [2021-01-10 14:51:49] 

#> ✓ JDBC connection

#> [2021-01-10 14:51:49]	SQL: SELECT * FROM omop_vocabulary.concept ORDER BY RANDOM() LIMIT 25
#> [2021-01-10 14:51:49]	Querying...
#> [2021-01-10 14:51:57]	Querying...complete
#> [2021-01-10 14:51:57] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 14:51:57]	Caching resultset...
#> [2021-01-10 14:51:57]	Postgres connection to 'athena' closed

#> # A tibble: 25 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id
#>         <dbl> <chr>        <chr>     <chr>         <chr>           
#>  1   45931828 Congenital … Condition CIEL          Diagnosis       
#>  2   36985814 safety hypo… Drug      SPL           OTC Drug        
#>  3    3303934 Adherent sc… Condition Nebraska Lex… Clinical Finding
#>  4   21044052 Mannitol 10… Drug      RxNorm Exten… Clinical Drug B…
#>  5   46080380 Aprotinin 1… Drug      SNOMED        Pharma/Biol Pro…
#>  6   40745105 insulin, is… Drug      RxNorm Exten… Branded Drug    
#>  7   43366730 Panadeine C… Drug      AMT           Containered Pack
#>  8   46015109 Prednisolon… Drug      SNOMED        Pharma/Biol Pro…
#>  9   37054898 Leptospira … Observat… LOINC         LOINC Component 
#> 10   36627893 E 6.4 Mg Vi… Drug      NCCD          Drug Product    
#> # … with 15 more rows, and 5 more variables: standard_concept <chr>,
#> #   concept_code <chr>, valid_start_date <date>, valid_end_date <date>,
#> #   invalid_reason <chr>

# Concepts specific to the Drug domain can be sampled
# RxNorm/RxNorm Extension Concepts
get_test_drug_concepts(limit = 10)

#> Connecting using PostgreSQL driver
#> [2021-01-10 14:51:57]	Overriding cache... Querying Athena...
#> [2021-01-10 14:51:57] 

#> ✓ Open connection

#> [2021-01-10 14:51:57] 

#> ✓ JDBC connection

#> [2021-01-10 14:51:57]	SQL: SELECT * FROM omop_vocabulary.concept c WHERE c.invalid_reason IS NULL AND c.vocabulary_id IN ('RxNorm', 'RxNorm Extension') AND c.standard_concept <> 'C' AND c.domain_id = 'Drug' ORDER BY RANDOM() LIMIT 10
#> [2021-01-10 14:51:57]	Querying...
#> [2021-01-10 14:52:00]	Querying...complete
#> [2021-01-10 14:52:00] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 14:52:00]	Caching resultset...
#> [2021-01-10 14:52:00]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id
#>         <dbl> <chr>        <chr>     <chr>         <chr>           
#>  1   19094125 insulin, re… Drug      RxNorm        Clinical Drug C…
#>  2   41061732 metylperon … Drug      RxNorm Exten… Branded Drug    
#>  3   43776444 250 ML Prol… Drug      RxNorm Exten… Marketed Product
#>  4   44099584 Diltiazem 3… Drug      RxNorm Exten… Branded Drug    
#>  5   43208883 1.5 ML Soma… Drug      RxNorm Exten… Marketed Product
#>  6   21163116 Azithromyci… Drug      RxNorm Exten… Marketed Product
#>  7   42717599 Thuja occid… Drug      RxNorm Exten… Clinical Drug C…
#>  8   44073591 Enalapril 2… Drug      RxNorm Exten… Branded Drug    
#>  9   41217542 Nifedipine … Drug      RxNorm Exten… Branded Drug Box
#> 10   21160023 Citalopram … Drug      RxNorm Exten… Marketed Product
#> # … with 5 more variables: standard_concept <chr>, concept_code <chr>,
#> #   valid_start_date <date>, valid_end_date <date>, invalid_reason <chr>

# HemOnc/ATC Classes
get_test_drug_classes(limit = 10)

#> Connecting using PostgreSQL driver
#> [2021-01-10 14:52:00]	Overriding cache... Querying Athena...
#> [2021-01-10 14:52:00] 

#> ✓ Open connection

#> [2021-01-10 14:52:00] 

#> ✓ JDBC connection

#> [2021-01-10 14:52:00]	SQL: SELECT * FROM omop_vocabulary.concept c WHERE c.invalid_reason IS NULL AND c.vocabulary_id IN ('ATC', 'HemOnc') AND c.standard_concept = 'C' AND c.domain_id = 'Drug' ORDER BY RANDOM() LIMIT 10
#> [2021-01-10 14:52:00]	Querying...
#> [2021-01-10 14:52:00]	Querying...complete
#> [2021-01-10 14:52:00] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 14:52:00]	Caching resultset...
#> [2021-01-10 14:52:00]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id
#>         <dbl> <chr>        <chr>     <chr>         <chr>           
#>  1   21603937 oxyphenbuta… Drug      ATC           ATC 5th         
#>  2   21602087 inosine; to… Drug      ATC           ATC 5th         
#>  3   21601392 melphalan    Drug      ATC           ATC 5th         
#>  4   21601964 ANTIFUNGALS… Drug      ATC           ATC 3rd         
#>  5   21602408 Triazole de… Drug      ATC           ATC 4th         
#>  6   21600909 sodium sele… Drug      ATC           ATC 5th         
#>  7   21604446 trihexyphen… Drug      ATC           ATC 5th         
#>  8   21604713 paroxetine;… Drug      ATC           ATC 5th         
#>  9   21602758 thyroid gla… Drug      ATC           ATC 5th         
#> 10   21601424 raltitrexed  Drug      ATC           ATC 5th         
#> # … with 5 more variables: standard_concept <chr>, concept_code <chr>,
#> #   valid_start_date <date>, valid_end_date <date>, invalid_reason <chr>

# Test data can also be taken from tables other than Concept
get_test_data(table = "concept_relationship")

#> Connecting using PostgreSQL driver
#> [2021-01-10 14:52:00]	Overriding cache... Querying Athena...
#> [2021-01-10 14:52:00] 

#> ✓ Open connection

#> [2021-01-10 14:52:00] 

#> ✓ JDBC connection

#> [2021-01-10 14:52:00]	SQL: SELECT * FROM omop_vocabulary.concept_relationship ORDER BY RANDOM() LIMIT 100
#> [2021-01-10 14:52:00]	Querying...
#> [2021-01-10 14:52:32]	Querying...complete
#> [2021-01-10 14:52:32] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 14:52:32]	Caching resultset...
#> [2021-01-10 14:52:32]	Postgres connection to 'athena' closed

#> # A tibble: 100 x 6
#>    concept_id_1 concept_id_2 relationship_id valid_start_date valid_end_date
#>           <dbl>        <dbl> <chr>           <date>           <date>        
#>  1     45765787     36683184 Active ing of   2019-01-31       2099-12-31    
#>  2     19082103     41095395 RxNorm dose fo… 2017-07-18       2099-12-31    
#>  3     43795301     21020276 Has supplier    2017-07-30       2099-12-31    
#>  4      4336541     36907616 Mapped from     2017-01-01       2099-12-31    
#>  5      2801210      2801210 Mapped from     2014-04-22       2099-12-31    
#>  6     35922950     35922950 Maps to         2018-03-02       2099-12-31    
#>  7     43862998     46234469 RxNorm has dos… 2016-07-01       2099-12-31    
#>  8     41928948     41930085 Subsumes        1970-01-01       2099-12-31    
#>  9     42021897     42021897 Maps to         2019-02-21       2099-12-31    
#> 10     45574039      4050690 Maps to         1970-01-01       2099-12-31    
#> # … with 90 more rows, and 1 more variable: invalid_reason <chr>

get_test_data(table = "concept_ancestor", limit = 10)

#> Connecting using PostgreSQL driver
#> [2021-01-10 14:52:32]	Overriding cache... Querying Athena...
#> [2021-01-10 14:52:32] 

#> ✓ Open connection

#> [2021-01-10 14:52:32] 

#> ✓ JDBC connection

#> [2021-01-10 14:52:32]	SQL: SELECT * FROM omop_vocabulary.concept_ancestor ORDER BY RANDOM() LIMIT 10
#> [2021-01-10 14:52:32]	Querying...
#> [2021-01-10 14:53:06]	Querying...complete
#> [2021-01-10 14:53:06] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 14:53:06]	Caching resultset...
#> [2021-01-10 14:53:06]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 4
#>    ancestor_concept_… descendant_concep… min_levels_of_sepa… max_levels_of_sepa…
#>                 <dbl>              <dbl>               <dbl>               <dbl>
#>  1           21602410           19112469                   3                   4
#>  2           45700921           40853773                   2                   3
#>  3            4082264            4056941                   3                   3
#>  4           36217216           40988789                   2                   2
#>  5           40087909           43163917                   2                   2
#>  6           35807212           44164530                   2                   2
#>  7           40154717           43852307                   1                   1
#>  8           45659412           45659412                   0                   0
#>  9           36217220           42900541                   2                   3
#> 10           45781884           37593199                   2                   2

library(chariot)
library(tidyverse)

# Get Unfiltered Test Concepts (Invalid Concepts are included)
get_test_concepts()

#> Connecting using PostgreSQL driver
#> [2021-01-10 14:53:06]	Overriding cache... Querying Athena...
#> [2021-01-10 14:53:06] 

#> ✓ Open connection

#> [2021-01-10 14:53:06] 

#> ✓ JDBC connection

#> [2021-01-10 14:53:06]	SQL: SELECT * FROM omop_vocabulary.concept ORDER BY RANDOM() LIMIT 100
#> [2021-01-10 14:53:06]	Querying...
#> [2021-01-10 14:53:14]	Querying...complete
#> [2021-01-10 14:53:14] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 14:53:14]	Caching resultset...
#> [2021-01-10 14:53:14]	Postgres connection to 'athena' closed

#> # A tibble: 100 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id
#>         <dbl> <chr>        <chr>     <chr>         <chr>           
#>  1   19053134 Roccal       Drug      RxNorm        Brand Name      
#>  2   42895790 Dilation of… Procedure ICD10PCS      ICD10PCS        
#>  3   45907768 Antimongoli… Condition CIEL          Diagnosis       
#>  4   36486859 APIS MELLIF… Drug      NDC           11-digit NDC    
#>  5   46232799 Open repair… Procedure OPCS4         Procedure       
#>  6   21345749 Juzo Soft c… Device    dm+d          AMPP            
#>  7   36725923 APC Transcr… Measurem… OMOP Extensi… Transcript Vari…
#>  8   43769093 Hamamelis v… Drug      RxNorm Exten… Branded Drug Fo…
#>  9    2082065 DOXOTEL Inj… Drug      KDC           Drug Product    
#> 10   36693911 Venoflex Mi… Device    SNOMED        Physical Object 
#> # … with 90 more rows, and 5 more variables: standard_concept <chr>,
#> #   concept_code <chr>, valid_start_date <date>, valid_end_date <date>,
#> #   invalid_reason <chr>

# The size of the resultset can be altered
get_test_concepts(limit = 25)

#> Connecting using PostgreSQL driver
#> [2021-01-10 14:53:14]	Overriding cache... Querying Athena...
#> [2021-01-10 14:53:14] 

#> ✓ Open connection

#> [2021-01-10 14:53:14] 

#> ✓ JDBC connection

#> [2021-01-10 14:53:14]	SQL: SELECT * FROM omop_vocabulary.concept ORDER BY RANDOM() LIMIT 25
#> [2021-01-10 14:53:14]	Querying...
#> [2021-01-10 14:53:21]	Querying...complete
#> [2021-01-10 14:53:21] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 14:53:21]	Caching resultset...
#> [2021-01-10 14:53:21]	Postgres connection to 'athena' closed

#> # A tibble: 25 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id
#>         <dbl> <chr>        <chr>     <chr>         <chr>           
#>  1   42530378 Ribs - unil… Observat… LOINC         LOINC System    
#>  2   45818148 hydrocortis… Drug      NDC           9-digit NDC     
#>  3   42652418 alcohol 665… Drug      SPL           OTC Drug        
#>  4   21355187 Calogen Ext… Device    dm+d          AMPP            
#>  5   45222908 carvedilol … Drug      NDC           11-digit NDC    
#>  6   35767329 Clomipramin… Drug      RxNorm Exten… Marketed Product
#>  7   21444477 Juzo Soft c… Device    dm+d          AMP             
#>  8   41932462 Assencières  Geography OSM           8th level       
#>  9   45908119 Sprains and… Condition CIEL          Diagnosis       
#> 10   21040516 fluticasone… Drug      RxNorm Exten… Marketed Product
#> # … with 15 more rows, and 5 more variables: standard_concept <chr>,
#> #   concept_code <chr>, valid_start_date <date>, valid_end_date <date>,
#> #   invalid_reason <chr>

# Concepts specific to the Drug domain can be sampled
# RxNorm/RxNorm Extension Concepts
get_test_drug_concepts(limit = 10)

#> Connecting using PostgreSQL driver
#> [2021-01-10 14:53:21]	Overriding cache... Querying Athena...
#> [2021-01-10 14:53:21] 

#> ✓ Open connection

#> [2021-01-10 14:53:21] 

#> ✓ JDBC connection

#> [2021-01-10 14:53:21]	SQL: SELECT * FROM omop_vocabulary.concept c WHERE c.invalid_reason IS NULL AND c.vocabulary_id IN ('RxNorm', 'RxNorm Extension') AND c.standard_concept <> 'C' AND c.domain_id = 'Drug' ORDER BY RANDOM() LIMIT 10
#> [2021-01-10 14:53:21]	Querying...
#> [2021-01-10 14:53:24]	Querying...complete
#> [2021-01-10 14:53:24] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 14:53:24]	Caching resultset...
#> [2021-01-10 14:53:24]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id
#>         <dbl> <chr>        <chr>     <chr>         <chr>           
#>  1   41300317 Escin / hep… Drug      RxNorm Exten… Branded Drug Fo…
#>  2   43644125 1 ML Spongi… Drug      RxNorm Exten… Quant Clinical …
#>  3   43027637 levothyroxi… Drug      RxNorm Exten… Branded Drug Fo…
#>  4   40231139 hyoscyamine… Drug      RxNorm        Clinical Drug C…
#>  5   40866998 hyaluronate… Drug      RxNorm Exten… Clinical Drug   
#>  6   43718815 25 ML Pacli… Drug      RxNorm Exten… Marketed Product
#>  7   44040088 Eugenol 1 M… Drug      RxNorm Exten… Clinical Drug C…
#>  8    2029450 2 ML hyalur… Drug      RxNorm Exten… Quant Branded D…
#>  9   44078136 Lilium lanc… Drug      RxNorm Exten… Clinical Drug C…
#> 10   43624329 olmesartan … Drug      RxNorm Exten… Branded Drug Co…
#> # … with 5 more variables: standard_concept <chr>, concept_code <chr>,
#> #   valid_start_date <date>, valid_end_date <date>, invalid_reason <chr>

# HemOnc/ATC Classes
get_test_drug_classes(limit = 10)

#> Connecting using PostgreSQL driver
#> [2021-01-10 14:53:24]	Overriding cache... Querying Athena...
#> [2021-01-10 14:53:24] 

#> ✓ Open connection

#> [2021-01-10 14:53:24] 

#> ✓ JDBC connection

#> [2021-01-10 14:53:24]	SQL: SELECT * FROM omop_vocabulary.concept c WHERE c.invalid_reason IS NULL AND c.vocabulary_id IN ('ATC', 'HemOnc') AND c.standard_concept = 'C' AND c.domain_id = 'Drug' ORDER BY RANDOM() LIMIT 10
#> [2021-01-10 14:53:24]	Querying...
#> [2021-01-10 14:53:25]	Querying...complete
#> [2021-01-10 14:53:25] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 14:53:25]	Caching resultset...
#> [2021-01-10 14:53:25]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 10
#>    concept_id concept_name domain_id vocabulary_id concept_class_id
#>         <dbl> <chr>        <chr>     <chr>         <chr>           
#>  1   21603596 silver comp… Drug      ATC           ATC 5th         
#>  2   43534844 salbutamol … Drug      ATC           ATC 5th         
#>  3   21604507 thiopropera… Drug      ATC           ATC 5th         
#>  4   40254972 collagenase… Drug      ATC           ATC 5th         
#>  5   21605533 indium (111… Drug      ATC           ATC 5th         
#>  6   21600738 insulin (po… Drug      ATC           ATC 5th         
#>  7   21604468 melevodopa … Drug      ATC           ATC 5th         
#>  8   21603219 dequalinium… Drug      ATC           ATC 5th         
#>  9   35807504 Immune effe… Drug      HemOnc        Component Class 
#> 10   21601955 amorolfine;… Drug      ATC           ATC 5th         
#> # … with 5 more variables: standard_concept <chr>, concept_code <chr>,
#> #   valid_start_date <date>, valid_end_date <date>, invalid_reason <chr>

# Test data can also be taken from tables other than Concept
get_test_data(table = "concept_relationship")

#> Connecting using PostgreSQL driver
#> [2021-01-10 14:53:25]	Overriding cache... Querying Athena...
#> [2021-01-10 14:53:25] 

#> ✓ Open connection

#> [2021-01-10 14:53:25] 

#> ✓ JDBC connection

#> [2021-01-10 14:53:25]	SQL: SELECT * FROM omop_vocabulary.concept_relationship ORDER BY RANDOM() LIMIT 100
#> [2021-01-10 14:53:25]	Querying...
#> [2021-01-10 14:53:57]	Querying...complete
#> [2021-01-10 14:53:57] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 14:53:57]	Caching resultset...
#> [2021-01-10 14:53:57]	Postgres connection to 'athena' closed

#> # A tibble: 100 x 6
#>    concept_id_1 concept_id_2 relationship_id valid_start_date valid_end_date
#>           <dbl>        <dbl> <chr>           <date>           <date>        
#>  1     19082079     40235262 RxNorm dose fo… 2011-03-07       2099-12-31    
#>  2     46274818       970835 RxNorm - SPL    1970-01-01       2099-12-31    
#>  3     35909811     35941408 Schema to Value 1970-01-01       2099-12-31    
#>  4     36624278      1319136 Maps to         1970-01-01       2099-12-31    
#>  5     19082104     44045778 RxNorm dose fo… 2017-06-25       2099-12-31    
#>  6      1029597     36304993 Property of     2019-12-13       2099-12-31    
#>  7      3489634      3224948 Dir proc site … 2019-08-16       2099-12-31    
#>  8     36781322     36781322 Mapped from     2018-10-25       2099-12-31    
#>  9     41327211     41325881 Tradename of    2017-07-18       2099-12-31    
#> 10      1506426     42863879 RxNorm - VAPro… 2014-12-01       2099-12-31    
#> # … with 90 more rows, and 1 more variable: invalid_reason <chr>

get_test_data(table = "concept_ancestor", limit = 10)

#> Connecting using PostgreSQL driver
#> [2021-01-10 14:53:57]	Overriding cache... Querying Athena...
#> [2021-01-10 14:53:57] 

#> ✓ Open connection

#> [2021-01-10 14:53:57] 

#> ✓ JDBC connection

#> [2021-01-10 14:53:57]	SQL: SELECT * FROM omop_vocabulary.concept_ancestor ORDER BY RANDOM() LIMIT 10
#> [2021-01-10 14:53:57]	Querying...
#> [2021-01-10 14:54:30]	Querying...complete
#> [2021-01-10 14:54:30] 

#> ✓ Returned data has more than 0 rows

#> [2021-01-10 14:54:30]	Caching resultset...
#> [2021-01-10 14:54:30]	Postgres connection to 'athena' closed

#> # A tibble: 10 x 4
#>    ancestor_concept_… descendant_concep… min_levels_of_sepa… max_levels_of_sepa…
#>                 <dbl>              <dbl>               <dbl>               <dbl>
#>  1            4090447           37115161                   4                   4
#>  2           36217216           43282419                   3                   4
#>  3           40046606           41332732                   2                   3
#>  4           43767923           40999804                   2                   2
#>  5           45695840           21141896                   2                   2
#>  6           42651656           44076584                   2                   2
#>  7           35161843           35149615                   0                   0
#>  8           44374885           44374885                   0                   0
#>  9           45771096           36714528                   3                   3
#> 10           45796386           41092078                   2                   2