Query the Common Data Model with Annotations

Query a Common Data Model table with automatic joins with any Concept Id to the Concept table with the appropriate prefix.

queryCDM(
  sql_statement,
  write_schema = "patelm9",
  vocab_schema = "omop_vocabulary",
  conn,
  conn_fun = "connectAthena()",
  skip_cache = FALSE,
  override_cache = FALSE,
  cache_only = FALSE,
  cache_resultset = TRUE,
  render_sql = TRUE,
  render_only = FALSE,
  verbose = TRUE,
  sleepTime = 1
)

Arguments

sql_statement
SQL query

write_schema
Schema to write the staged data to for a join to Concept table.

vocab_schema
Schema where to Concept table can be found.

conn
Connection object. If provided, diverts queries to the connection instead of the local Athena instance without caching features.

skip_cache
Skip the caching altogether and directly query the database.

override_cache
If TRUE, the cache will not be loaded and will be overwritten by a new query. For override_cache to take effect, skip_cache should be FALSE.

cache_only
Loads from the cache and does not query the database. A NULL object is returned if a resultset was not cached.

cache_resultset
(deprecated) If TRUE, the resultset from the query will first be loaded from the cache. The query will be executed if a cached resultset is not retrieved for this particular query, after which the resultset will be cached. If FALSE, Athena or conn will be directly queried without any caching operations.

render_sql
If TRUE, the SQL will be printed back in the console prior to execution. Default: FALSE

verbose
If TRUE, prints loading and querying operations messages to the console. Default: FALSE

sleepTime
Argument for `Sys.sleep()` in between queries to allow for halting function execution, especially in cases where other chariot functions are executing multiple queries in succession and require cancellation.

Value

a [tibble][tibble::tibble-package]

See also

Examples

library(pg13)

#> Loading required package: DatabaseConnector

#> 
#> Attaching package: ‘pg13’

#> The following object is masked from ‘package:DatabaseConnector’:
#> 
#>     connect

#> The following object is masked from ‘package:readr’:
#> 
#>     read_table

#> The following objects are masked from ‘package:chariot’:
#> 
#>     cache_query, load_cached_query

library(tidyverse)

conn <- local_connect("polyester")

#> Connecting using PostgreSQL driver

queryCDM(sql_statement = "SELECT * FROM omop_cdm.drug_exposure ORDER BY RANDOM() LIMIT 5;",
         write_schema = "omop_cdm",
         vocab_schema = "omop_cdm",
         conn = conn)

#> [2021-01-09 15:12:17]	Loading Cache...
#> [2021-01-09 15:12:17]	Cached SQL: SELECT * FROM omop_cdm.drug_exposure ORDER BY RANDOM() LIMIT 5;
#> [2021-01-09 15:12:17]	Cached resultset found...
#> [2021-01-09 15:12:17]	Target column: drug_concept_id
#> [2021-01-09 15:12:17]	Dropping omop_cdm.V20210109151217...
#> [2021-01-09 15:12:17] 

#> ✓ Open connection

#> [2021-01-09 15:12:17] 

#> ✓ JDBC connection

#> [2021-01-09 15:12:17]	SQL: DROP TABLE IF EXISTS omop_cdm.V20210109151217;
#> [2021-01-09 15:12:17]	Sending...
#> [2021-01-09 15:12:17]	Sending...complete
#> [2021-01-09 15:12:17]	Dropping omop_cdm.V20210109151217...complete
#> [2021-01-09 15:12:17] 

#> ✓ Data 'data' has more than 0 rows

#> [2021-01-09 15:12:17] 

#> ✓ Table name 'V20210109151217' is not a reserved word

#> [2021-01-09 15:12:17] 

#> ✓ Field name 'drug_exposure_id' is not a reserved word

#> [2021-01-09 15:12:17] 

#> ✓ Field name 'person_id' is not a reserved word

#> [2021-01-09 15:12:17] 

#> ✓ Field name 'drug_concept_id' is not a reserved word

#> [2021-01-09 15:12:17] 

#> ✓ Field name 'drug_exposure_start_date' is not a reserved word

#> [2021-01-09 15:12:17] 

#> ✓ Field name 'drug_exposure_start_datetime' is not a reserved word

#> [2021-01-09 15:12:17] 

#> ✓ Field name 'drug_exposure_end_date' is not a reserved word

#> [2021-01-09 15:12:17] 

#> ✓ Field name 'drug_exposure_end_datetime' is not a reserved word

#> [2021-01-09 15:12:17] 

#> ✓ Field name 'verbatim_end_date' is not a reserved word

#> [2021-01-09 15:12:17] 

#> ✓ Field name 'drug_type_concept_id' is not a reserved word

#> [2021-01-09 15:12:17] 

#> ✓ Field name 'stop_reason' is not a reserved word

#> [2021-01-09 15:12:17] 

#> ✓ Field name 'refills' is not a reserved word

#> [2021-01-09 15:12:17] 

#> ✓ Field name 'quantity' is not a reserved word

#> [2021-01-09 15:12:17] 

#> ✓ Field name 'days_supply' is not a reserved word

#> [2021-01-09 15:12:17] 

#> ✓ Field name 'sig' is not a reserved word

#> [2021-01-09 15:12:17] 

#> ✓ Field name 'route_concept_id' is not a reserved word

#> [2021-01-09 15:12:17] 

#> ✓ Field name 'lot_number' is not a reserved word

#> [2021-01-09 15:12:17] 

#> ✓ Field name 'provider_id' is not a reserved word

#> [2021-01-09 15:12:17] 

#> ✓ Field name 'visit_occurrence_id' is not a reserved word

#> [2021-01-09 15:12:17] 

#> ✓ Field name 'visit_detail_id' is not a reserved word

#> [2021-01-09 15:12:17] 

#> ✓ Field name 'drug_source_value' is not a reserved word

#> [2021-01-09 15:12:17] 

#> ✓ Field name 'drug_source_concept_id' is not a reserved word

#> [2021-01-09 15:12:17] 

#> ✓ Field name 'route_source_value' is not a reserved word

#> [2021-01-09 15:12:17] 

#> ✓ Field name 'dose_unit_source_value' is not a reserved word

#> [2021-01-09 15:12:17]	Writing omop_cdm.V20210109151217...

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> [2021-01-09 15:12:17]	Writing omop_cdm.V20210109151217...complete
#> [2021-01-09 15:12:17] 

#> ✓ Open connection

#> [2021-01-09 15:12:17] 

#> ✓ JDBC connection

#> [2021-01-09 15:12:17]	SQL: SELECT a.*, b.* FROM omop_cdm.V20210109151217 a LEFT JOIN omop_cdm.concept b ON a.drug_concept_id = b.concept_id WHERE b.invalid_reason IS NULL
#> [2021-01-09 15:12:17]	Querying...
#> [2021-01-09 15:12:18]	Querying...complete
#> [2021-01-09 15:12:18] 

#> x Returned data has 0 rows

#> [2021-01-09 15:12:18]	Dropping omop_cdm.V20210109151217...
#> [2021-01-09 15:12:18] 

#> ✓ Open connection

#> [2021-01-09 15:12:18] 

#> ✓ JDBC connection

#> [2021-01-09 15:12:18]	SQL: DROP TABLE IF EXISTS omop_cdm.V20210109151217;
#> [2021-01-09 15:12:18]	Sending...
#> [2021-01-09 15:12:18]	Sending...complete
#> [2021-01-09 15:12:18]	Dropping omop_cdm.V20210109151217...complete

#> Joining, by = "drug_concept_id"

#> [2021-01-09 15:12:18]	Target column: drug_type_concept_id
#> [2021-01-09 15:12:18]	Dropping omop_cdm.V20210109151218...
#> [2021-01-09 15:12:18] 

#> ✓ Open connection

#> [2021-01-09 15:12:18] 

#> ✓ JDBC connection

#> [2021-01-09 15:12:18]	SQL: DROP TABLE IF EXISTS omop_cdm.V20210109151218;
#> [2021-01-09 15:12:18]	Sending...
#> [2021-01-09 15:12:18]	Sending...complete
#> [2021-01-09 15:12:18]	Dropping omop_cdm.V20210109151218...complete
#> [2021-01-09 15:12:18] 

#> ✓ Data 'data' has more than 0 rows

#> [2021-01-09 15:12:18] 

#> ✓ Table name 'V20210109151218' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'drug_exposure_id' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'person_id' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'drug_concept_id' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'drug_exposure_start_date' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'drug_exposure_start_datetime' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'drug_exposure_end_date' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'drug_exposure_end_datetime' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'verbatim_end_date' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'drug_type_concept_id' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'stop_reason' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'refills' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'quantity' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'days_supply' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'sig' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'route_concept_id' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'lot_number' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'provider_id' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'visit_occurrence_id' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'visit_detail_id' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'drug_source_value' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'drug_source_concept_id' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'route_source_value' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'dose_unit_source_value' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'drug_concept_name' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'drug_domain_id' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'drug_vocabulary_id' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'drug_concept_class_id' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'drug_standard_concept' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'drug_concept_code' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'drug_valid_start_date' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'drug_valid_end_date' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'drug_invalid_reason' is not a reserved word

#> [2021-01-09 15:12:18]	Writing omop_cdm.V20210109151218...

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> [2021-01-09 15:12:18]	Writing omop_cdm.V20210109151218...complete
#> [2021-01-09 15:12:18] 

#> ✓ Open connection

#> [2021-01-09 15:12:18] 

#> ✓ JDBC connection

#> [2021-01-09 15:12:18]	SQL: SELECT a.*, b.* FROM omop_cdm.V20210109151218 a LEFT JOIN omop_cdm.concept b ON a.drug_type_concept_id = b.concept_id WHERE b.invalid_reason IS NULL
#> [2021-01-09 15:12:18]	Querying...
#> [2021-01-09 15:12:18]	Querying...complete
#> [2021-01-09 15:12:18] 

#> x Returned data has 0 rows

#> [2021-01-09 15:12:18]	Dropping omop_cdm.V20210109151218...
#> [2021-01-09 15:12:18] 

#> ✓ Open connection

#> [2021-01-09 15:12:18] 

#> ✓ JDBC connection

#> [2021-01-09 15:12:18]	SQL: DROP TABLE IF EXISTS omop_cdm.V20210109151218;
#> [2021-01-09 15:12:18]	Sending...
#> [2021-01-09 15:12:18]	Sending...complete
#> [2021-01-09 15:12:18]	Dropping omop_cdm.V20210109151218...complete

#> Joining, by = "drug_type_concept_id"

#> [2021-01-09 15:12:18]	Target column: route_concept_id
#> [2021-01-09 15:12:18]	Dropping omop_cdm.V20210109151218...
#> [2021-01-09 15:12:18] 

#> ✓ Open connection

#> [2021-01-09 15:12:18] 

#> ✓ JDBC connection

#> [2021-01-09 15:12:18]	SQL: DROP TABLE IF EXISTS omop_cdm.V20210109151218;
#> [2021-01-09 15:12:18]	Sending...
#> [2021-01-09 15:12:18]	Sending...complete
#> [2021-01-09 15:12:18]	Dropping omop_cdm.V20210109151218...complete
#> [2021-01-09 15:12:18] 

#> ✓ Data 'data' has more than 0 rows

#> [2021-01-09 15:12:18] 

#> ✓ Table name 'V20210109151218' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'drug_exposure_id' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'person_id' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'drug_concept_id' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'drug_exposure_start_date' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'drug_exposure_start_datetime' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'drug_exposure_end_date' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'drug_exposure_end_datetime' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'verbatim_end_date' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'drug_type_concept_id' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'stop_reason' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'refills' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'quantity' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'days_supply' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'sig' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'route_concept_id' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'lot_number' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'provider_id' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'visit_occurrence_id' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'visit_detail_id' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'drug_source_value' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'drug_source_concept_id' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'route_source_value' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'dose_unit_source_value' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'drug_concept_name' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'drug_domain_id' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'drug_vocabulary_id' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'drug_concept_class_id' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'drug_standard_concept' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'drug_concept_code' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'drug_valid_start_date' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'drug_valid_end_date' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'drug_invalid_reason' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'drug_type_concept_name' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'drug_type_domain_id' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'drug_type_vocabulary_id' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'drug_type_concept_class_id' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'drug_type_standard_concept' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'drug_type_concept_code' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'drug_type_valid_start_date' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'drug_type_valid_end_date' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'drug_type_invalid_reason' is not a reserved word

#> [2021-01-09 15:12:18]	Writing omop_cdm.V20210109151218...

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> [2021-01-09 15:12:18]	Writing omop_cdm.V20210109151218...complete
#> [2021-01-09 15:12:18] 

#> ✓ Open connection

#> [2021-01-09 15:12:18] 

#> ✓ JDBC connection

#> [2021-01-09 15:12:18]	SQL: SELECT a.*, b.* FROM omop_cdm.V20210109151218 a LEFT JOIN omop_cdm.concept b ON a.route_concept_id = b.concept_id WHERE b.invalid_reason IS NULL
#> [2021-01-09 15:12:18]	Querying...
#> [2021-01-09 15:12:18]	Querying...complete
#> [2021-01-09 15:12:18] 

#> x Returned data has 0 rows

#> [2021-01-09 15:12:18]	Dropping omop_cdm.V20210109151218...
#> [2021-01-09 15:12:18] 

#> ✓ Open connection

#> [2021-01-09 15:12:18] 

#> ✓ JDBC connection

#> [2021-01-09 15:12:18]	SQL: DROP TABLE IF EXISTS omop_cdm.V20210109151218;
#> [2021-01-09 15:12:18]	Sending...
#> [2021-01-09 15:12:18]	Sending...complete
#> [2021-01-09 15:12:18]	Dropping omop_cdm.V20210109151218...complete

#> Joining, by = "route_concept_id"

#> [2021-01-09 15:12:18]	Target column: drug_source_concept_id
#> [2021-01-09 15:12:18]	Dropping omop_cdm.V20210109151218...
#> [2021-01-09 15:12:18] 

#> ✓ Open connection

#> [2021-01-09 15:12:18] 

#> ✓ JDBC connection

#> [2021-01-09 15:12:18]	SQL: DROP TABLE IF EXISTS omop_cdm.V20210109151218;
#> [2021-01-09 15:12:18]	Sending...
#> [2021-01-09 15:12:18]	Sending...complete
#> [2021-01-09 15:12:18]	Dropping omop_cdm.V20210109151218...complete
#> [2021-01-09 15:12:18] 

#> ✓ Data 'data' has more than 0 rows

#> [2021-01-09 15:12:18] 

#> ✓ Table name 'V20210109151218' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'drug_exposure_id' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'person_id' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'drug_concept_id' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'drug_exposure_start_date' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'drug_exposure_start_datetime' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'drug_exposure_end_date' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'drug_exposure_end_datetime' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'verbatim_end_date' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'drug_type_concept_id' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'stop_reason' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'refills' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'quantity' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'days_supply' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'sig' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'route_concept_id' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'lot_number' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'provider_id' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'visit_occurrence_id' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'visit_detail_id' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'drug_source_value' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'drug_source_concept_id' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'route_source_value' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'dose_unit_source_value' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'drug_concept_name' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'drug_domain_id' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'drug_vocabulary_id' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'drug_concept_class_id' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'drug_standard_concept' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'drug_concept_code' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'drug_valid_start_date' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'drug_valid_end_date' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'drug_invalid_reason' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'drug_type_concept_name' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'drug_type_domain_id' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'drug_type_vocabulary_id' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'drug_type_concept_class_id' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'drug_type_standard_concept' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'drug_type_concept_code' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'drug_type_valid_start_date' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'drug_type_valid_end_date' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'drug_type_invalid_reason' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'route_concept_name' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'route_domain_id' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'route_vocabulary_id' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'route_concept_class_id' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'route_standard_concept' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'route_concept_code' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'route_valid_start_date' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'route_valid_end_date' is not a reserved word

#> [2021-01-09 15:12:18] 

#> ✓ Field name 'route_invalid_reason' is not a reserved word

#> [2021-01-09 15:12:18]	Writing omop_cdm.V20210109151218...

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> [2021-01-09 15:12:18]	Writing omop_cdm.V20210109151218...complete
#> [2021-01-09 15:12:18] 

#> ✓ Open connection

#> [2021-01-09 15:12:18] 

#> ✓ JDBC connection

#> [2021-01-09 15:12:18]	SQL: SELECT a.*, b.* FROM omop_cdm.V20210109151218 a LEFT JOIN omop_cdm.concept b ON a.drug_source_concept_id = b.concept_id WHERE b.invalid_reason IS NULL
#> [2021-01-09 15:12:18]	Querying...
#> [2021-01-09 15:12:18]	Querying...complete
#> [2021-01-09 15:12:18] 

#> x Returned data has 0 rows

#> [2021-01-09 15:12:18]	Dropping omop_cdm.V20210109151218...
#> [2021-01-09 15:12:18] 

#> ✓ Open connection

#> [2021-01-09 15:12:18] 

#> ✓ JDBC connection

#> [2021-01-09 15:12:18]	SQL: DROP TABLE IF EXISTS omop_cdm.V20210109151218;
#> [2021-01-09 15:12:18]	Sending...
#> [2021-01-09 15:12:18]	Sending...complete
#> [2021-01-09 15:12:18]	Dropping omop_cdm.V20210109151218...complete

#> Joining, by = "drug_source_concept_id"

#> # A tibble: 5 x 59
#>   drug_exposure_id person_id drug_concept_id drug_exposure_s…
#>              <dbl>     <dbl>           <dbl> <date>          
#> 1           201638      3274        19080128 2020-07-16      
#> 2           190383      3077        40163924 1995-05-04      
#> 3           273683      4400        19073094 2002-12-04      
#> 4           335087      5277        19078106 2016-04-12      
#> 5           164759      2684        40163554 1997-01-31      
#> # … with 55 more variables: drug_exposure_start_datetime <dttm>,
#> #   drug_exposure_end_date <date>, drug_exposure_end_datetime <dttm>,
#> #   verbatim_end_date <date>, drug_type_concept_id <dbl>, stop_reason <chr>,
#> #   refills <dbl>, quantity <dbl>, days_supply <dbl>, sig <chr>,
#> #   route_concept_id <dbl>, lot_number <chr>, provider_id <dbl>,
#> #   visit_occurrence_id <dbl>, visit_detail_id <dbl>, drug_source_value <chr>,
#> #   drug_source_concept_id <dbl>, route_source_value <chr>,
#> #   dose_unit_source_value <chr>, drug_concept_name <chr>,
#> #   drug_domain_id <chr>, drug_vocabulary_id <chr>,
#> #   drug_concept_class_id <chr>, drug_standard_concept <chr>,
#> #   drug_concept_code <chr>, drug_valid_start_date <date>,
#> #   drug_valid_end_date <date>, drug_invalid_reason <chr>,
#> #   drug_type_concept_name <chr>, drug_type_domain_id <chr>,
#> #   drug_type_vocabulary_id <chr>, drug_type_concept_class_id <chr>,
#> #   drug_type_standard_concept <chr>, drug_type_concept_code <chr>,
#> #   drug_type_valid_start_date <date>, drug_type_valid_end_date <date>,
#> #   drug_type_invalid_reason <chr>, route_concept_name <chr>,
#> #   route_domain_id <chr>, route_vocabulary_id <chr>,
#> #   route_concept_class_id <chr>, route_standard_concept <chr>,
#> #   route_concept_code <chr>, route_valid_start_date <date>,
#> #   route_valid_end_date <date>, route_invalid_reason <chr>,
#> #   drug_source_concept_name <chr>, drug_source_domain_id <chr>,
#> #   drug_source_vocabulary_id <chr>, drug_source_concept_class_id <chr>,
#> #   drug_source_standard_concept <chr>, drug_source_concept_code <chr>,
#> #   drug_source_valid_start_date <date>, drug_source_valid_end_date <date>,
#> #   drug_source_invalid_reason <chr>

dc(conn = conn)

#> [2021-01-09 15:12:19]	Postgres connection to 'polyester' closed